+++ corporate_services/public/js/employee_timesheet_entry.js
@@ -1,4 +1,4 @@
-(function(){"use strict";function Oc(e){return e&&e.__esModule&&Object.prototype.hasOwnProperty.call(e,"default")?e.default:e}var Zi={exports:{}},mr={},Ji={exports:{}},D={};/**
+(function(){"use strict";function Mc(e){return e&&e.__esModule&&Object.prototype.hasOwnProperty.call(e,"default")?e.default:e}var Zi={exports:{}},mr={},Ji={exports:{}},D={};/**
  * @license React
  * react.production.min.js
  *
@@ -6,7 +6,7 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
- */var _n=Symbol.for("react.element"),Mc=Symbol.for("react.portal"),Fc=Symbol.for("react.fragment"),Dc=Symbol.for("react.strict_mode"),Ic=Symbol.for("react.profiler"),Ac=Symbol.for("react.provider"),Uc=Symbol.for("react.context"),$c=Symbol.for("react.forward_ref"),Vc=Symbol.for("react.suspense"),Bc=Symbol.for("react.memo"),Wc=Symbol.for("react.lazy"),qi=Symbol.iterator;function Hc(e){return e===null||typeof e!="object"?null:(e=qi&&e[qi]||e["@@iterator"],typeof e=="function"?e:null)}var bi={isMounted:function(){return!1},enqueueForceUpdate:function(){},enqueueReplaceState:function(){},enqueueSetState:function(){}},eu=Object.assign,tu={};function Yt(e,t,n){this.props=e,this.context=t,this.refs=tu,this.updater=n||bi}Yt.prototype.isReactComponent={},Yt.prototype.setState=function(e,t){if(typeof e!="object"&&typeof e!="function"&&e!=null)throw Error("setState(...): takes an object of state variables to update or a function which returns an object of state variables.");this.updater.enqueueSetState(this,e,t,"setState")},Yt.prototype.forceUpdate=function(e){this.updater.enqueueForceUpdate(this,e,"forceUpdate")};function nu(){}nu.prototype=Yt.prototype;function Rl(e,t,n){this.props=e,this.context=t,this.refs=tu,this.updater=n||bi}var Ol=Rl.prototype=new nu;Ol.constructor=Rl,eu(Ol,Yt.prototype),Ol.isPureReactComponent=!0;var ru=Array.isArray,lu=Object.prototype.hasOwnProperty,Ml={current:null},ou={key:!0,ref:!0,__self:!0,__source:!0};function iu(e,t,n){var r,l={},o=null,i=null;if(t!=null)for(r in t.ref!==void 0&&(i=t.ref),t.key!==void 0&&(o=""+t.key),t)lu.call(t,r)&&!ou.hasOwnProperty(r)&&(l[r]=t[r]);var u=arguments.length-2;if(u===1)l.children=n;else if(1<u){for(var s=Array(u),c=0;c<u;c++)s[c]=arguments[c+2];l.children=s}if(e&&e.defaultProps)for(r in u=e.defaultProps,u)l[r]===void 0&&(l[r]=u[r]);return{$$typeof:_n,type:e,key:o,ref:i,props:l,_owner:Ml.current}}function Qc(e,t){return{$$typeof:_n,type:e.type,key:t,ref:e.ref,props:e.props,_owner:e._owner}}function Fl(e){return typeof e=="object"&&e!==null&&e.$$typeof===_n}function Kc(e){var t={"=":"=0",":":"=2"};return"$"+e.replace(/[=:]/g,function(n){return t[n]})}var uu=/\/+/g;function Dl(e,t){return typeof e=="object"&&e!==null&&e.key!=null?Kc(""+e.key):t.toString(36)}function hr(e,t,n,r,l){var o=typeof e;(o==="undefined"||o==="boolean")&&(e=null);var i=!1;if(e===null)i=!0;else switch(o){case"string":case"number":i=!0;break;case"object":switch(e.$$typeof){case _n:case Mc:i=!0}}if(i)return i=e,l=l(i),e=r===""?"."+Dl(i,0):r,ru(l)?(n="",e!=null&&(n=e.replace(uu,"$&/")+"/"),hr(l,t,n,"",function(c){return c})):l!=null&&(Fl(l)&&(l=Qc(l,n+(!l.key||i&&i.key===l.key?"":(""+l.key).replace(uu,"$&/")+"/")+e)),t.push(l)),1;if(i=0,r=r===""?".":r+":",ru(e))for(var u=0;u<e.length;u++){o=e[u];var s=r+Dl(o,u);i+=hr(o,t,n,s,l)}else if(s=Hc(e),typeof s=="function")for(e=s.call(e),u=0;!(o=e.next()).done;)o=o.value,s=r+Dl(o,u++),i+=hr(o,t,n,s,l);else if(o==="object")throw t=String(e),Error("Objects are not valid as a React child (found: "+(t==="[object Object]"?"object with keys {"+Object.keys(e).join(", ")+"}":t)+"). If you meant to render a collection of children, use an array instead.");return i}function vr(e,t,n){if(e==null)return e;var r=[],l=0;return hr(e,r,"","",function(o){return t.call(n,o,l++)}),r}function Yc(e){if(e._status===-1){var t=e._result;t=t(),t.then(function(n){(e._status===0||e._status===-1)&&(e._status=1,e._result=n)},function(n){(e._status===0||e._status===-1)&&(e._status=2,e._result=n)}),e._status===-1&&(e._status=0,e._result=t)}if(e._status===1)return e._result.default;throw e._result}var pe={current:null},yr={transition:null},Xc={ReactCurrentDispatcher:pe,ReactCurrentBatchConfig:yr,ReactCurrentOwner:Ml};function su(){throw Error("act(...) is not supported in production builds of React.")}D.Children={map:vr,forEach:function(e,t,n){vr(e,function(){t.apply(this,arguments)},n)},count:function(e){var t=0;return vr(e,function(){t++}),t},toArray:function(e){return vr(e,function(t){return t})||[]},only:function(e){if(!Fl(e))throw Error("React.Children.only expected to receive a single React element child.");return e}},D.Component=Yt,D.Fragment=Fc,D.Profiler=Ic,D.PureComponent=Rl,D.StrictMode=Dc,D.Suspense=Vc,D.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=Xc,D.act=su,D.cloneElement=function(e,t,n){if(e==null)throw Error("React.cloneElement(...): The argument must be a React element, but you passed "+e+".");var r=eu({},e.props),l=e.key,o=e.ref,i=e._owner;if(t!=null){if(t.ref!==void 0&&(o=t.ref,i=Ml.current),t.key!==void 0&&(l=""+t.key),e.type&&e.type.defaultProps)var u=e.type.defaultProps;for(s in t)lu.call(t,s)&&!ou.hasOwnProperty(s)&&(r[s]=t[s]===void 0&&u!==void 0?u[s]:t[s])}var s=arguments.length-2;if(s===1)r.children=n;else if(1<s){u=Array(s);for(var c=0;c<s;c++)u[c]=arguments[c+2];r.children=u}return{$$typeof:_n,type:e.type,key:l,ref:o,props:r,_owner:i}},D.createContext=function(e){return e={$$typeof:Uc,_currentValue:e,_currentValue2:e,_threadCount:0,Provider:null,Consumer:null,_defaultValue:null,_globalName:null},e.Provider={$$typeof:Ac,_context:e},e.Consumer=e},D.createElement=iu,D.createFactory=function(e){var t=iu.bind(null,e);return t.type=e,t},D.createRef=function(){return{current:null}},D.forwardRef=function(e){return{$$typeof:$c,render:e}},D.isValidElement=Fl,D.lazy=function(e){return{$$typeof:Wc,_payload:{_status:-1,_result:e},_init:Yc}},D.memo=function(e,t){return{$$typeof:Bc,type:e,compare:t===void 0?null:t}},D.startTransition=function(e){var t=yr.transition;yr.transition={};try{e()}finally{yr.transition=t}},D.unstable_act=su,D.useCallback=function(e,t){return pe.current.useCallback(e,t)},D.useContext=function(e){return pe.current.useContext(e)},D.useDebugValue=function(){},D.useDeferredValue=function(e){return pe.current.useDeferredValue(e)},D.useEffect=function(e,t){return pe.current.useEffect(e,t)},D.useId=function(){return pe.current.useId()},D.useImperativeHandle=function(e,t,n){return pe.current.useImperativeHandle(e,t,n)},D.useInsertionEffect=function(e,t){return pe.current.useInsertionEffect(e,t)},D.useLayoutEffect=function(e,t){return pe.current.useLayoutEffect(e,t)},D.useMemo=function(e,t){return pe.current.useMemo(e,t)},D.useReducer=function(e,t,n){return pe.current.useReducer(e,t,n)},D.useRef=function(e){return pe.current.useRef(e)},D.useState=function(e){return pe.current.useState(e)},D.useSyncExternalStore=function(e,t,n){return pe.current.useSyncExternalStore(e,t,n)},D.useTransition=function(){return pe.current.useTransition()},D.version="18.3.1",Ji.exports=D;var R=Ji.exports;const Gc=Oc(R);/**
+ */var _n=Symbol.for("react.element"),Oc=Symbol.for("react.portal"),Fc=Symbol.for("react.fragment"),Dc=Symbol.for("react.strict_mode"),Ic=Symbol.for("react.profiler"),Ac=Symbol.for("react.provider"),Uc=Symbol.for("react.context"),$c=Symbol.for("react.forward_ref"),Vc=Symbol.for("react.suspense"),Bc=Symbol.for("react.memo"),Wc=Symbol.for("react.lazy"),qi=Symbol.iterator;function Hc(e){return e===null||typeof e!="object"?null:(e=qi&&e[qi]||e["@@iterator"],typeof e=="function"?e:null)}var bi={isMounted:function(){return!1},enqueueForceUpdate:function(){},enqueueReplaceState:function(){},enqueueSetState:function(){}},eu=Object.assign,tu={};function Yt(e,t,n){this.props=e,this.context=t,this.refs=tu,this.updater=n||bi}Yt.prototype.isReactComponent={},Yt.prototype.setState=function(e,t){if(typeof e!="object"&&typeof e!="function"&&e!=null)throw Error("setState(...): takes an object of state variables to update or a function which returns an object of state variables.");this.updater.enqueueSetState(this,e,t,"setState")},Yt.prototype.forceUpdate=function(e){this.updater.enqueueForceUpdate(this,e,"forceUpdate")};function nu(){}nu.prototype=Yt.prototype;function Rl(e,t,n){this.props=e,this.context=t,this.refs=tu,this.updater=n||bi}var Ml=Rl.prototype=new nu;Ml.constructor=Rl,eu(Ml,Yt.prototype),Ml.isPureReactComponent=!0;var ru=Array.isArray,lu=Object.prototype.hasOwnProperty,Ol={current:null},ou={key:!0,ref:!0,__self:!0,__source:!0};function iu(e,t,n){var r,l={},o=null,i=null;if(t!=null)for(r in t.ref!==void 0&&(i=t.ref),t.key!==void 0&&(o=""+t.key),t)lu.call(t,r)&&!ou.hasOwnProperty(r)&&(l[r]=t[r]);var u=arguments.length-2;if(u===1)l.children=n;else if(1<u){for(var s=Array(u),c=0;c<u;c++)s[c]=arguments[c+2];l.children=s}if(e&&e.defaultProps)for(r in u=e.defaultProps,u)l[r]===void 0&&(l[r]=u[r]);return{$$typeof:_n,type:e,key:o,ref:i,props:l,_owner:Ol.current}}function Qc(e,t){return{$$typeof:_n,type:e.type,key:t,ref:e.ref,props:e.props,_owner:e._owner}}function Fl(e){return typeof e=="object"&&e!==null&&e.$$typeof===_n}function Kc(e){var t={"=":"=0",":":"=2"};return"$"+e.replace(/[=:]/g,function(n){return t[n]})}var uu=/\/+/g;function Dl(e,t){return typeof e=="object"&&e!==null&&e.key!=null?Kc(""+e.key):t.toString(36)}function hr(e,t,n,r,l){var o=typeof e;(o==="undefined"||o==="boolean")&&(e=null);var i=!1;if(e===null)i=!0;else switch(o){case"string":case"number":i=!0;break;case"object":switch(e.$$typeof){case _n:case Oc:i=!0}}if(i)return i=e,l=l(i),e=r===""?"."+Dl(i,0):r,ru(l)?(n="",e!=null&&(n=e.replace(uu,"$&/")+"/"),hr(l,t,n,"",function(c){return c})):l!=null&&(Fl(l)&&(l=Qc(l,n+(!l.key||i&&i.key===l.key?"":(""+l.key).replace(uu,"$&/")+"/")+e)),t.push(l)),1;if(i=0,r=r===""?".":r+":",ru(e))for(var u=0;u<e.length;u++){o=e[u];var s=r+Dl(o,u);i+=hr(o,t,n,s,l)}else if(s=Hc(e),typeof s=="function")for(e=s.call(e),u=0;!(o=e.next()).done;)o=o.value,s=r+Dl(o,u++),i+=hr(o,t,n,s,l);else if(o==="object")throw t=String(e),Error("Objects are not valid as a React child (found: "+(t==="[object Object]"?"object with keys {"+Object.keys(e).join(", ")+"}":t)+"). If you meant to render a collection of children, use an array instead.");return i}function vr(e,t,n){if(e==null)return e;var r=[],l=0;return hr(e,r,"","",function(o){return t.call(n,o,l++)}),r}function Yc(e){if(e._status===-1){var t=e._result;t=t(),t.then(function(n){(e._status===0||e._status===-1)&&(e._status=1,e._result=n)},function(n){(e._status===0||e._status===-1)&&(e._status=2,e._result=n)}),e._status===-1&&(e._status=0,e._result=t)}if(e._status===1)return e._result.default;throw e._result}var pe={current:null},yr={transition:null},Xc={ReactCurrentDispatcher:pe,ReactCurrentBatchConfig:yr,ReactCurrentOwner:Ol};function su(){throw Error("act(...) is not supported in production builds of React.")}D.Children={map:vr,forEach:function(e,t,n){vr(e,function(){t.apply(this,arguments)},n)},count:function(e){var t=0;return vr(e,function(){t++}),t},toArray:function(e){return vr(e,function(t){return t})||[]},only:function(e){if(!Fl(e))throw Error("React.Children.only expected to receive a single React element child.");return e}},D.Component=Yt,D.Fragment=Fc,D.Profiler=Ic,D.PureComponent=Rl,D.StrictMode=Dc,D.Suspense=Vc,D.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=Xc,D.act=su,D.cloneElement=function(e,t,n){if(e==null)throw Error("React.cloneElement(...): The argument must be a React element, but you passed "+e+".");var r=eu({},e.props),l=e.key,o=e.ref,i=e._owner;if(t!=null){if(t.ref!==void 0&&(o=t.ref,i=Ol.current),t.key!==void 0&&(l=""+t.key),e.type&&e.type.defaultProps)var u=e.type.defaultProps;for(s in t)lu.call(t,s)&&!ou.hasOwnProperty(s)&&(r[s]=t[s]===void 0&&u!==void 0?u[s]:t[s])}var s=arguments.length-2;if(s===1)r.children=n;else if(1<s){u=Array(s);for(var c=0;c<s;c++)u[c]=arguments[c+2];r.children=u}return{$$typeof:_n,type:e.type,key:l,ref:o,props:r,_owner:i}},D.createContext=function(e){return e={$$typeof:Uc,_currentValue:e,_currentValue2:e,_threadCount:0,Provider:null,Consumer:null,_defaultValue:null,_globalName:null},e.Provider={$$typeof:Ac,_context:e},e.Consumer=e},D.createElement=iu,D.createFactory=function(e){var t=iu.bind(null,e);return t.type=e,t},D.createRef=function(){return{current:null}},D.forwardRef=function(e){return{$$typeof:$c,render:e}},D.isValidElement=Fl,D.lazy=function(e){return{$$typeof:Wc,_payload:{_status:-1,_result:e},_init:Yc}},D.memo=function(e,t){return{$$typeof:Bc,type:e,compare:t===void 0?null:t}},D.startTransition=function(e){var t=yr.transition;yr.transition={};try{e()}finally{yr.transition=t}},D.unstable_act=su,D.useCallback=function(e,t){return pe.current.useCallback(e,t)},D.useContext=function(e){return pe.current.useContext(e)},D.useDebugValue=function(){},D.useDeferredValue=function(e){return pe.current.useDeferredValue(e)},D.useEffect=function(e,t){return pe.current.useEffect(e,t)},D.useId=function(){return pe.current.useId()},D.useImperativeHandle=function(e,t,n){return pe.current.useImperativeHandle(e,t,n)},D.useInsertionEffect=function(e,t){return pe.current.useInsertionEffect(e,t)},D.useLayoutEffect=function(e,t){return pe.current.useLayoutEffect(e,t)},D.useMemo=function(e,t){return pe.current.useMemo(e,t)},D.useReducer=function(e,t,n){return pe.current.useReducer(e,t,n)},D.useRef=function(e){return pe.current.useRef(e)},D.useState=function(e){return pe.current.useState(e)},D.useSyncExternalStore=function(e,t,n){return pe.current.useSyncExternalStore(e,t,n)},D.useTransition=function(){return pe.current.useTransition()},D.version="18.3.1",Ji.exports=D;var R=Ji.exports;const Gc=Mc(R);/**
  * @license React
  * react-jsx-runtime.production.min.js
  *
@@ -14,7 +14,7 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
- */var Zc=R,Jc=Symbol.for("react.element"),qc=Symbol.for("react.fragment"),bc=Object.prototype.hasOwnProperty,ef=Zc.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.ReactCurrentOwner,tf={key:!0,ref:!0,__self:!0,__source:!0};function au(e,t,n){var r,l={},o=null,i=null;n!==void 0&&(o=""+n),t.key!==void 0&&(o=""+t.key),t.ref!==void 0&&(i=t.ref);for(r in t)bc.call(t,r)&&!tf.hasOwnProperty(r)&&(l[r]=t[r]);if(e&&e.defaultProps)for(r in t=e.defaultProps,t)l[r]===void 0&&(l[r]=t[r]);return{$$typeof:Jc,type:e,key:o,ref:i,props:l,_owner:ef.current}}mr.Fragment=qc,mr.jsx=au,mr.jsxs=au,Zi.exports=mr;var h=Zi.exports,cu={exports:{}},_e={},fu={exports:{}},du={};/**
+ */var Zc=R,Jc=Symbol.for("react.element"),qc=Symbol.for("react.fragment"),bc=Object.prototype.hasOwnProperty,ef=Zc.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.ReactCurrentOwner,tf={key:!0,ref:!0,__self:!0,__source:!0};function au(e,t,n){var r,l={},o=null,i=null;n!==void 0&&(o=""+n),t.key!==void 0&&(o=""+t.key),t.ref!==void 0&&(i=t.ref);for(r in t)bc.call(t,r)&&!tf.hasOwnProperty(r)&&(l[r]=t[r]);if(e&&e.defaultProps)for(r in t=e.defaultProps,t)l[r]===void 0&&(l[r]=t[r]);return{$$typeof:Jc,type:e,key:o,ref:i,props:l,_owner:ef.current}}mr.Fragment=qc,mr.jsx=au,mr.jsxs=au,Zi.exports=mr;var m=Zi.exports,cu={exports:{}},xe={},fu={exports:{}},du={};/**
  * @license React
  * scheduler.production.min.js
  *
@@ -22,7 +22,7 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
- */(function(e){function t(_,O){var M=_.length;_.push(O);e:for(;0<M;){var W=M-1>>>1,ee=_[W];if(0<l(ee,O))_[W]=O,_[M]=ee,M=W;else break e}}function n(_){return _.length===0?null:_[0]}function r(_){if(_.length===0)return null;var O=_[0],M=_.pop();if(M!==O){_[0]=M;e:for(var W=0,ee=_.length,q=ee>>>1;W<q;){var Te=2*(W+1)-1,Lt=_[Te],De=Te+1,at=_[De];if(0>l(Lt,M))De<ee&&0>l(at,Lt)?(_[W]=at,_[De]=M,W=De):(_[W]=Lt,_[Te]=M,W=Te);else if(De<ee&&0>l(at,M))_[W]=at,_[De]=M,W=De;else break e}}return O}function l(_,O){var M=_.sortIndex-O.sortIndex;return M!==0?M:_.id-O.id}if(typeof performance=="object"&&typeof performance.now=="function"){var o=performance;e.unstable_now=function(){return o.now()}}else{var i=Date,u=i.now();e.unstable_now=function(){return i.now()-u}}var s=[],c=[],y=1,v=null,m=3,k=!1,S=!1,C=!1,A=typeof setTimeout=="function"?setTimeout:null,f=typeof clearTimeout=="function"?clearTimeout:null,a=typeof setImmediate<"u"?setImmediate:null;typeof navigator<"u"&&navigator.scheduling!==void 0&&navigator.scheduling.isInputPending!==void 0&&navigator.scheduling.isInputPending.bind(navigator.scheduling);function d(_){for(var O=n(c);O!==null;){if(O.callback===null)r(c);else if(O.startTime<=_)r(c),O.sortIndex=O.expirationTime,t(s,O);else break;O=n(c)}}function g(_){if(C=!1,d(_),!S)if(n(s)!==null)S=!0,ut(N);else{var O=n(c);O!==null&&st(g,O.startTime-_)}}function N(_,O){S=!1,C&&(C=!1,f(L),L=-1),k=!0;var M=m;try{for(d(O),v=n(s);v!==null&&(!(v.expirationTime>O)||_&&!xe());){var W=v.callback;if(typeof W=="function"){v.callback=null,m=v.priorityLevel;var ee=W(v.expirationTime<=O);O=e.unstable_now(),typeof ee=="function"?v.callback=ee:v===n(s)&&r(s),d(O)}else r(s);v=n(s)}if(v!==null)var q=!0;else{var Te=n(c);Te!==null&&st(g,Te.startTime-O),q=!1}return q}finally{v=null,m=M,k=!1}}var P=!1,T=null,L=-1,Q=5,F=-1;function xe(){return!(e.unstable_now()-F<Q)}function He(){if(T!==null){var _=e.unstable_now();F=_;var O=!0;try{O=T(!0,_)}finally{O?Ze():(P=!1,T=null)}}else P=!1}var Ze;if(typeof a=="function")Ze=function(){a(He)};else if(typeof MessageChannel<"u"){var zt=new MessageChannel,dr=zt.port2;zt.port1.onmessage=He,Ze=function(){dr.postMessage(null)}}else Ze=function(){A(He,0)};function ut(_){T=_,P||(P=!0,Ze())}function st(_,O){L=A(function(){_(e.unstable_now())},O)}e.unstable_IdlePriority=5,e.unstable_ImmediatePriority=1,e.unstable_LowPriority=4,e.unstable_NormalPriority=3,e.unstable_Profiling=null,e.unstable_UserBlockingPriority=2,e.unstable_cancelCallback=function(_){_.callback=null},e.unstable_continueExecution=function(){S||k||(S=!0,ut(N))},e.unstable_forceFrameRate=function(_){0>_||125<_?console.error("forceFrameRate takes a positive int between 0 and 125, forcing frame rates higher than 125 fps is not supported"):Q=0<_?Math.floor(1e3/_):5},e.unstable_getCurrentPriorityLevel=function(){return m},e.unstable_getFirstCallbackNode=function(){return n(s)},e.unstable_next=function(_){switch(m){case 1:case 2:case 3:var O=3;break;default:O=m}var M=m;m=O;try{return _()}finally{m=M}},e.unstable_pauseExecution=function(){},e.unstable_requestPaint=function(){},e.unstable_runWithPriority=function(_,O){switch(_){case 1:case 2:case 3:case 4:case 5:break;default:_=3}var M=m;m=_;try{return O()}finally{m=M}},e.unstable_scheduleCallback=function(_,O,M){var W=e.unstable_now();switch(typeof M=="object"&&M!==null?(M=M.delay,M=typeof M=="number"&&0<M?W+M:W):M=W,_){case 1:var ee=-1;break;case 2:ee=250;break;case 5:ee=1073741823;break;case 4:ee=1e4;break;default:ee=5e3}return ee=M+ee,_={id:y++,callback:O,priorityLevel:_,startTime:M,expirationTime:ee,sortIndex:-1},M>W?(_.sortIndex=M,t(c,_),n(s)===null&&_===n(c)&&(C?(f(L),L=-1):C=!0,st(g,M-W))):(_.sortIndex=ee,t(s,_),S||k||(S=!0,ut(N))),_},e.unstable_shouldYield=xe,e.unstable_wrapCallback=function(_){var O=m;return function(){var M=m;m=O;try{return _.apply(this,arguments)}finally{m=M}}}})(du),fu.exports=du;var nf=fu.exports;/**
+ */(function(e){function t(x,M){var O=x.length;x.push(M);e:for(;0<O;){var W=O-1>>>1,ee=x[W];if(0<l(ee,M))x[W]=M,x[O]=ee,O=W;else break e}}function n(x){return x.length===0?null:x[0]}function r(x){if(x.length===0)return null;var M=x[0],O=x.pop();if(O!==M){x[0]=O;e:for(var W=0,ee=x.length,q=ee>>>1;W<q;){var Te=2*(W+1)-1,Lt=x[Te],De=Te+1,at=x[De];if(0>l(Lt,O))De<ee&&0>l(at,Lt)?(x[W]=at,x[De]=O,W=De):(x[W]=Lt,x[Te]=O,W=Te);else if(De<ee&&0>l(at,O))x[W]=at,x[De]=O,W=De;else break e}}return M}function l(x,M){var O=x.sortIndex-M.sortIndex;return O!==0?O:x.id-M.id}if(typeof performance=="object"&&typeof performance.now=="function"){var o=performance;e.unstable_now=function(){return o.now()}}else{var i=Date,u=i.now();e.unstable_now=function(){return i.now()-u}}var s=[],c=[],y=1,h=null,p=3,w=!1,S=!1,E=!1,U=typeof setTimeout=="function"?setTimeout:null,f=typeof clearTimeout=="function"?clearTimeout:null,a=typeof setImmediate<"u"?setImmediate:null;typeof navigator<"u"&&navigator.scheduling!==void 0&&navigator.scheduling.isInputPending!==void 0&&navigator.scheduling.isInputPending.bind(navigator.scheduling);function d(x){for(var M=n(c);M!==null;){if(M.callback===null)r(c);else if(M.startTime<=x)r(c),M.sortIndex=M.expirationTime,t(s,M);else break;M=n(c)}}function g(x){if(E=!1,d(x),!S)if(n(s)!==null)S=!0,ut(N);else{var M=n(c);M!==null&&st(g,M.startTime-x)}}function N(x,M){S=!1,E&&(E=!1,f(L),L=-1),w=!0;var O=p;try{for(d(M),h=n(s);h!==null&&(!(h.expirationTime>M)||x&&!Pe());){var W=h.callback;if(typeof W=="function"){h.callback=null,p=h.priorityLevel;var ee=W(h.expirationTime<=M);M=e.unstable_now(),typeof ee=="function"?h.callback=ee:h===n(s)&&r(s),d(M)}else r(s);h=n(s)}if(h!==null)var q=!0;else{var Te=n(c);Te!==null&&st(g,Te.startTime-M),q=!1}return q}finally{h=null,p=O,w=!1}}var P=!1,T=null,L=-1,Q=5,F=-1;function Pe(){return!(e.unstable_now()-F<Q)}function He(){if(T!==null){var x=e.unstable_now();F=x;var M=!0;try{M=T(!0,x)}finally{M?it():(P=!1,T=null)}}else P=!1}var it;if(typeof a=="function")it=function(){a(He)};else if(typeof MessageChannel<"u"){var zt=new MessageChannel,dr=zt.port2;zt.port1.onmessage=He,it=function(){dr.postMessage(null)}}else it=function(){U(He,0)};function ut(x){T=x,P||(P=!0,it())}function st(x,M){L=U(function(){x(e.unstable_now())},M)}e.unstable_IdlePriority=5,e.unstable_ImmediatePriority=1,e.unstable_LowPriority=4,e.unstable_NormalPriority=3,e.unstable_Profiling=null,e.unstable_UserBlockingPriority=2,e.unstable_cancelCallback=function(x){x.callback=null},e.unstable_continueExecution=function(){S||w||(S=!0,ut(N))},e.unstable_forceFrameRate=function(x){0>x||125<x?console.error("forceFrameRate takes a positive int between 0 and 125, forcing frame rates higher than 125 fps is not supported"):Q=0<x?Math.floor(1e3/x):5},e.unstable_getCurrentPriorityLevel=function(){return p},e.unstable_getFirstCallbackNode=function(){return n(s)},e.unstable_next=function(x){switch(p){case 1:case 2:case 3:var M=3;break;default:M=p}var O=p;p=M;try{return x()}finally{p=O}},e.unstable_pauseExecution=function(){},e.unstable_requestPaint=function(){},e.unstable_runWithPriority=function(x,M){switch(x){case 1:case 2:case 3:case 4:case 5:break;default:x=3}var O=p;p=x;try{return M()}finally{p=O}},e.unstable_scheduleCallback=function(x,M,O){var W=e.unstable_now();switch(typeof O=="object"&&O!==null?(O=O.delay,O=typeof O=="number"&&0<O?W+O:W):O=W,x){case 1:var ee=-1;break;case 2:ee=250;break;case 5:ee=1073741823;break;case 4:ee=1e4;break;default:ee=5e3}return ee=O+ee,x={id:y++,callback:M,priorityLevel:x,startTime:O,expirationTime:ee,sortIndex:-1},O>W?(x.sortIndex=O,t(c,x),n(s)===null&&x===n(c)&&(E?(f(L),L=-1):E=!0,st(g,O-W))):(x.sortIndex=ee,t(s,x),S||w||(S=!0,ut(N))),x},e.unstable_shouldYield=Pe,e.unstable_wrapCallback=function(x){var M=p;return function(){var O=p;p=M;try{return x.apply(this,arguments)}finally{p=O}}}})(du),fu.exports=du;var nf=fu.exports;/**
  * @license React
  * react-dom.production.min.js
  *
@@ -30,11 +30,11 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
- */var rf=R,Ee=nf;function w(e){for(var t="https://reactjs.org/docs/error-decoder.html?invariant="+e,n=1;n<arguments.length;n++)t+="&args[]="+encodeURIComponent(arguments[n]);return"Minified React error #"+e+"; visit "+t+" for the full message or use the non-minified dev environment for full errors and additional helpful warnings."}var pu=new Set,En={};function Rt(e,t){Xt(e,t),Xt(e+"Capture",t)}function Xt(e,t){for(En[e]=t,e=0;e<t.length;e++)pu.add(t[e])}var Je=!(typeof window>"u"||typeof window.document>"u"||typeof window.document.createElement>"u"),Il=Object.prototype.hasOwnProperty,lf=/^[:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD][:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD\-.0-9\u00B7\u0300-\u036F\u203F-\u2040]*$/,mu={},hu={};function of(e){return Il.call(hu,e)?!0:Il.call(mu,e)?!1:lf.test(e)?hu[e]=!0:(mu[e]=!0,!1)}function uf(e,t,n,r){if(n!==null&&n.type===0)return!1;switch(typeof t){case"function":case"symbol":return!0;case"boolean":return r?!1:n!==null?!n.acceptsBooleans:(e=e.toLowerCase().slice(0,5),e!=="data-"&&e!=="aria-");default:return!1}}function sf(e,t,n,r){if(t===null||typeof t>"u"||uf(e,t,n,r))return!0;if(r)return!1;if(n!==null)switch(n.type){case 3:return!t;case 4:return t===!1;case 5:return isNaN(t);case 6:return isNaN(t)||1>t}return!1}function me(e,t,n,r,l,o,i){this.acceptsBooleans=t===2||t===3||t===4,this.attributeName=r,this.attributeNamespace=l,this.mustUseProperty=n,this.propertyName=e,this.type=t,this.sanitizeURL=o,this.removeEmptyString=i}var oe={};"children dangerouslySetInnerHTML defaultValue defaultChecked innerHTML suppressContentEditableWarning suppressHydrationWarning style".split(" ").forEach(function(e){oe[e]=new me(e,0,!1,e,null,!1,!1)}),[["acceptCharset","accept-charset"],["className","class"],["htmlFor","for"],["httpEquiv","http-equiv"]].forEach(function(e){var t=e[0];oe[t]=new me(t,1,!1,e[1],null,!1,!1)}),["contentEditable","draggable","spellCheck","value"].forEach(function(e){oe[e]=new me(e,2,!1,e.toLowerCase(),null,!1,!1)}),["autoReverse","externalResourcesRequired","focusable","preserveAlpha"].forEach(function(e){oe[e]=new me(e,2,!1,e,null,!1,!1)}),"allowFullScreen async autoFocus autoPlay controls default defer disabled disablePictureInPicture disableRemotePlayback formNoValidate hidden loop noModule noValidate open playsInline readOnly required reversed scoped seamless itemScope".split(" ").forEach(function(e){oe[e]=new me(e,3,!1,e.toLowerCase(),null,!1,!1)}),["checked","multiple","muted","selected"].forEach(function(e){oe[e]=new me(e,3,!0,e,null,!1,!1)}),["capture","download"].forEach(function(e){oe[e]=new me(e,4,!1,e,null,!1,!1)}),["cols","rows","size","span"].forEach(function(e){oe[e]=new me(e,6,!1,e,null,!1,!1)}),["rowSpan","start"].forEach(function(e){oe[e]=new me(e,5,!1,e.toLowerCase(),null,!1,!1)});var Al=/[\-:]([a-z])/g;function Ul(e){return e[1].toUpperCase()}"accent-height alignment-baseline arabic-form baseline-shift cap-height clip-path clip-rule color-interpolation color-interpolation-filters color-profile color-rendering dominant-baseline enable-background fill-opacity fill-rule flood-color flood-opacity font-family font-size font-size-adjust font-stretch font-style font-variant font-weight glyph-name glyph-orientation-horizontal glyph-orientation-vertical horiz-adv-x horiz-origin-x image-rendering letter-spacing lighting-color marker-end marker-mid marker-start overline-position overline-thickness paint-order panose-1 pointer-events rendering-intent shape-rendering stop-color stop-opacity strikethrough-position strikethrough-thickness stroke-dasharray stroke-dashoffset stroke-linecap stroke-linejoin stroke-miterlimit stroke-opacity stroke-width text-anchor text-decoration text-rendering underline-position underline-thickness unicode-bidi unicode-range units-per-em v-alphabetic v-hanging v-ideographic v-mathematical vector-effect vert-adv-y vert-origin-x vert-origin-y word-spacing writing-mode xmlns:xlink x-height".split(" ").forEach(function(e){var t=e.replace(Al,Ul);oe[t]=new me(t,1,!1,e,null,!1,!1)}),"xlink:actuate xlink:arcrole xlink:role xlink:show xlink:title xlink:type".split(" ").forEach(function(e){var t=e.replace(Al,Ul);oe[t]=new me(t,1,!1,e,"http://www.w3.org/1999/xlink",!1,!1)}),["xml:base","xml:lang","xml:space"].forEach(function(e){var t=e.replace(Al,Ul);oe[t]=new me(t,1,!1,e,"http://www.w3.org/XML/1998/namespace",!1,!1)}),["tabIndex","crossOrigin"].forEach(function(e){oe[e]=new me(e,1,!1,e.toLowerCase(),null,!1,!1)}),oe.xlinkHref=new me("xlinkHref",1,!1,"xlink:href","http://www.w3.org/1999/xlink",!0,!1),["src","href","action","formAction"].forEach(function(e){oe[e]=new me(e,1,!1,e.toLowerCase(),null,!0,!0)});function $l(e,t,n,r){var l=oe.hasOwnProperty(t)?oe[t]:null;(l!==null?l.type!==0:r||!(2<t.length)||t[0]!=="o"&&t[0]!=="O"||t[1]!=="n"&&t[1]!=="N")&&(sf(t,n,l,r)&&(n=null),r||l===null?of(t)&&(n===null?e.removeAttribute(t):e.setAttribute(t,""+n)):l.mustUseProperty?e[l.propertyName]=n===null?l.type===3?!1:"":n:(t=l.attributeName,r=l.attributeNamespace,n===null?e.removeAttribute(t):(l=l.type,n=l===3||l===4&&n===!0?"":""+n,r?e.setAttributeNS(r,t,n):e.setAttribute(t,n))))}var qe=rf.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,gr=Symbol.for("react.element"),Gt=Symbol.for("react.portal"),Zt=Symbol.for("react.fragment"),Vl=Symbol.for("react.strict_mode"),Bl=Symbol.for("react.profiler"),vu=Symbol.for("react.provider"),yu=Symbol.for("react.context"),Wl=Symbol.for("react.forward_ref"),Hl=Symbol.for("react.suspense"),Ql=Symbol.for("react.suspense_list"),Kl=Symbol.for("react.memo"),ct=Symbol.for("react.lazy"),gu=Symbol.for("react.offscreen"),wu=Symbol.iterator;function Cn(e){return e===null||typeof e!="object"?null:(e=wu&&e[wu]||e["@@iterator"],typeof e=="function"?e:null)}var K=Object.assign,Yl;function Nn(e){if(Yl===void 0)try{throw Error()}catch(n){var t=n.stack.trim().match(/\n( *(at )?)/);Yl=t&&t[1]||""}return`
+ */var rf=R,_e=nf;function k(e){for(var t="https://reactjs.org/docs/error-decoder.html?invariant="+e,n=1;n<arguments.length;n++)t+="&args[]="+encodeURIComponent(arguments[n]);return"Minified React error #"+e+"; visit "+t+" for the full message or use the non-minified dev environment for full errors and additional helpful warnings."}var pu=new Set,En={};function Rt(e,t){Xt(e,t),Xt(e+"Capture",t)}function Xt(e,t){for(En[e]=t,e=0;e<t.length;e++)pu.add(t[e])}var Ze=!(typeof window>"u"||typeof window.document>"u"||typeof window.document.createElement>"u"),Il=Object.prototype.hasOwnProperty,lf=/^[:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD][:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD\-.0-9\u00B7\u0300-\u036F\u203F-\u2040]*$/,mu={},hu={};function of(e){return Il.call(hu,e)?!0:Il.call(mu,e)?!1:lf.test(e)?hu[e]=!0:(mu[e]=!0,!1)}function uf(e,t,n,r){if(n!==null&&n.type===0)return!1;switch(typeof t){case"function":case"symbol":return!0;case"boolean":return r?!1:n!==null?!n.acceptsBooleans:(e=e.toLowerCase().slice(0,5),e!=="data-"&&e!=="aria-");default:return!1}}function sf(e,t,n,r){if(t===null||typeof t>"u"||uf(e,t,n,r))return!0;if(r)return!1;if(n!==null)switch(n.type){case 3:return!t;case 4:return t===!1;case 5:return isNaN(t);case 6:return isNaN(t)||1>t}return!1}function me(e,t,n,r,l,o,i){this.acceptsBooleans=t===2||t===3||t===4,this.attributeName=r,this.attributeNamespace=l,this.mustUseProperty=n,this.propertyName=e,this.type=t,this.sanitizeURL=o,this.removeEmptyString=i}var oe={};"children dangerouslySetInnerHTML defaultValue defaultChecked innerHTML suppressContentEditableWarning suppressHydrationWarning style".split(" ").forEach(function(e){oe[e]=new me(e,0,!1,e,null,!1,!1)}),[["acceptCharset","accept-charset"],["className","class"],["htmlFor","for"],["httpEquiv","http-equiv"]].forEach(function(e){var t=e[0];oe[t]=new me(t,1,!1,e[1],null,!1,!1)}),["contentEditable","draggable","spellCheck","value"].forEach(function(e){oe[e]=new me(e,2,!1,e.toLowerCase(),null,!1,!1)}),["autoReverse","externalResourcesRequired","focusable","preserveAlpha"].forEach(function(e){oe[e]=new me(e,2,!1,e,null,!1,!1)}),"allowFullScreen async autoFocus autoPlay controls default defer disabled disablePictureInPicture disableRemotePlayback formNoValidate hidden loop noModule noValidate open playsInline readOnly required reversed scoped seamless itemScope".split(" ").forEach(function(e){oe[e]=new me(e,3,!1,e.toLowerCase(),null,!1,!1)}),["checked","multiple","muted","selected"].forEach(function(e){oe[e]=new me(e,3,!0,e,null,!1,!1)}),["capture","download"].forEach(function(e){oe[e]=new me(e,4,!1,e,null,!1,!1)}),["cols","rows","size","span"].forEach(function(e){oe[e]=new me(e,6,!1,e,null,!1,!1)}),["rowSpan","start"].forEach(function(e){oe[e]=new me(e,5,!1,e.toLowerCase(),null,!1,!1)});var Al=/[\-:]([a-z])/g;function Ul(e){return e[1].toUpperCase()}"accent-height alignment-baseline arabic-form baseline-shift cap-height clip-path clip-rule color-interpolation color-interpolation-filters color-profile color-rendering dominant-baseline enable-background fill-opacity fill-rule flood-color flood-opacity font-family font-size font-size-adjust font-stretch font-style font-variant font-weight glyph-name glyph-orientation-horizontal glyph-orientation-vertical horiz-adv-x horiz-origin-x image-rendering letter-spacing lighting-color marker-end marker-mid marker-start overline-position overline-thickness paint-order panose-1 pointer-events rendering-intent shape-rendering stop-color stop-opacity strikethrough-position strikethrough-thickness stroke-dasharray stroke-dashoffset stroke-linecap stroke-linejoin stroke-miterlimit stroke-opacity stroke-width text-anchor text-decoration text-rendering underline-position underline-thickness unicode-bidi unicode-range units-per-em v-alphabetic v-hanging v-ideographic v-mathematical vector-effect vert-adv-y vert-origin-x vert-origin-y word-spacing writing-mode xmlns:xlink x-height".split(" ").forEach(function(e){var t=e.replace(Al,Ul);oe[t]=new me(t,1,!1,e,null,!1,!1)}),"xlink:actuate xlink:arcrole xlink:role xlink:show xlink:title xlink:type".split(" ").forEach(function(e){var t=e.replace(Al,Ul);oe[t]=new me(t,1,!1,e,"http://www.w3.org/1999/xlink",!1,!1)}),["xml:base","xml:lang","xml:space"].forEach(function(e){var t=e.replace(Al,Ul);oe[t]=new me(t,1,!1,e,"http://www.w3.org/XML/1998/namespace",!1,!1)}),["tabIndex","crossOrigin"].forEach(function(e){oe[e]=new me(e,1,!1,e.toLowerCase(),null,!1,!1)}),oe.xlinkHref=new me("xlinkHref",1,!1,"xlink:href","http://www.w3.org/1999/xlink",!0,!1),["src","href","action","formAction"].forEach(function(e){oe[e]=new me(e,1,!1,e.toLowerCase(),null,!0,!0)});function $l(e,t,n,r){var l=oe.hasOwnProperty(t)?oe[t]:null;(l!==null?l.type!==0:r||!(2<t.length)||t[0]!=="o"&&t[0]!=="O"||t[1]!=="n"&&t[1]!=="N")&&(sf(t,n,l,r)&&(n=null),r||l===null?of(t)&&(n===null?e.removeAttribute(t):e.setAttribute(t,""+n)):l.mustUseProperty?e[l.propertyName]=n===null?l.type===3?!1:"":n:(t=l.attributeName,r=l.attributeNamespace,n===null?e.removeAttribute(t):(l=l.type,n=l===3||l===4&&n===!0?"":""+n,r?e.setAttributeNS(r,t,n):e.setAttribute(t,n))))}var Je=rf.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,gr=Symbol.for("react.element"),Gt=Symbol.for("react.portal"),Zt=Symbol.for("react.fragment"),Vl=Symbol.for("react.strict_mode"),Bl=Symbol.for("react.profiler"),vu=Symbol.for("react.provider"),yu=Symbol.for("react.context"),Wl=Symbol.for("react.forward_ref"),Hl=Symbol.for("react.suspense"),Ql=Symbol.for("react.suspense_list"),Kl=Symbol.for("react.memo"),ct=Symbol.for("react.lazy"),gu=Symbol.for("react.offscreen"),ku=Symbol.iterator;function Cn(e){return e===null||typeof e!="object"?null:(e=ku&&e[ku]||e["@@iterator"],typeof e=="function"?e:null)}var K=Object.assign,Yl;function Nn(e){if(Yl===void 0)try{throw Error()}catch(n){var t=n.stack.trim().match(/\n( *(at )?)/);Yl=t&&t[1]||""}return`
 `+Yl+e}var Xl=!1;function Gl(e,t){if(!e||Xl)return"";Xl=!0;var n=Error.prepareStackTrace;Error.prepareStackTrace=void 0;try{if(t)if(t=function(){throw Error()},Object.defineProperty(t.prototype,"props",{set:function(){throw Error()}}),typeof Reflect=="object"&&Reflect.construct){try{Reflect.construct(t,[])}catch(c){var r=c}Reflect.construct(e,[],t)}else{try{t.call()}catch(c){r=c}e.call(t.prototype)}else{try{throw Error()}catch(c){r=c}e()}}catch(c){if(c&&r&&typeof c.stack=="string"){for(var l=c.stack.split(`
 `),o=r.stack.split(`
 `),i=l.length-1,u=o.length-1;1<=i&&0<=u&&l[i]!==o[u];)u--;for(;1<=i&&0<=u;i--,u--)if(l[i]!==o[u]){if(i!==1||u!==1)do if(i--,u--,0>u||l[i]!==o[u]){var s=`
-`+l[i].replace(" at new "," at ");return e.displayName&&s.includes("<anonymous>")&&(s=s.replace("<anonymous>",e.displayName)),s}while(1<=i&&0<=u);break}}}finally{Xl=!1,Error.prepareStackTrace=n}return(e=e?e.displayName||e.name:"")?Nn(e):""}function af(e){switch(e.tag){case 5:return Nn(e.type);case 16:return Nn("Lazy");case 13:return Nn("Suspense");case 19:return Nn("SuspenseList");case 0:case 2:case 15:return e=Gl(e.type,!1),e;case 11:return e=Gl(e.type.render,!1),e;case 1:return e=Gl(e.type,!0),e;default:return""}}function Zl(e){if(e==null)return null;if(typeof e=="function")return e.displayName||e.name||null;if(typeof e=="string")return e;switch(e){case Zt:return"Fragment";case Gt:return"Portal";case Bl:return"Profiler";case Vl:return"StrictMode";case Hl:return"Suspense";case Ql:return"SuspenseList"}if(typeof e=="object")switch(e.$$typeof){case yu:return(e.displayName||"Context")+".Consumer";case vu:return(e._context.displayName||"Context")+".Provider";case Wl:var t=e.render;return e=e.displayName,e||(e=t.displayName||t.name||"",e=e!==""?"ForwardRef("+e+")":"ForwardRef"),e;case Kl:return t=e.displayName||null,t!==null?t:Zl(e.type)||"Memo";case ct:t=e._payload,e=e._init;try{return Zl(e(t))}catch{}}return null}function cf(e){var t=e.type;switch(e.tag){case 24:return"Cache";case 9:return(t.displayName||"Context")+".Consumer";case 10:return(t._context.displayName||"Context")+".Provider";case 18:return"DehydratedFragment";case 11:return e=t.render,e=e.displayName||e.name||"",t.displayName||(e!==""?"ForwardRef("+e+")":"ForwardRef");case 7:return"Fragment";case 5:return t;case 4:return"Portal";case 3:return"Root";case 6:return"Text";case 16:return Zl(t);case 8:return t===Vl?"StrictMode":"Mode";case 22:return"Offscreen";case 12:return"Profiler";case 21:return"Scope";case 13:return"Suspense";case 19:return"SuspenseList";case 25:return"TracingMarker";case 1:case 0:case 17:case 2:case 14:case 15:if(typeof t=="function")return t.displayName||t.name||null;if(typeof t=="string")return t}return null}function ft(e){switch(typeof e){case"boolean":case"number":case"string":case"undefined":return e;case"object":return e;default:return""}}function ku(e){var t=e.type;return(e=e.nodeName)&&e.toLowerCase()==="input"&&(t==="checkbox"||t==="radio")}function ff(e){var t=ku(e)?"checked":"value",n=Object.getOwnPropertyDescriptor(e.constructor.prototype,t),r=""+e[t];if(!e.hasOwnProperty(t)&&typeof n<"u"&&typeof n.get=="function"&&typeof n.set=="function"){var l=n.get,o=n.set;return Object.defineProperty(e,t,{configurable:!0,get:function(){return l.call(this)},set:function(i){r=""+i,o.call(this,i)}}),Object.defineProperty(e,t,{enumerable:n.enumerable}),{getValue:function(){return r},setValue:function(i){r=""+i},stopTracking:function(){e._valueTracker=null,delete e[t]}}}}function wr(e){e._valueTracker||(e._valueTracker=ff(e))}function Su(e){if(!e)return!1;var t=e._valueTracker;if(!t)return!0;var n=t.getValue(),r="";return e&&(r=ku(e)?e.checked?"true":"false":e.value),e=r,e!==n?(t.setValue(e),!0):!1}function kr(e){if(e=e||(typeof document<"u"?document:void 0),typeof e>"u")return null;try{return e.activeElement||e.body}catch{return e.body}}function Jl(e,t){var n=t.checked;return K({},t,{defaultChecked:void 0,defaultValue:void 0,value:void 0,checked:n??e._wrapperState.initialChecked})}function xu(e,t){var n=t.defaultValue==null?"":t.defaultValue,r=t.checked!=null?t.checked:t.defaultChecked;n=ft(t.value!=null?t.value:n),e._wrapperState={initialChecked:r,initialValue:n,controlled:t.type==="checkbox"||t.type==="radio"?t.checked!=null:t.value!=null}}function _u(e,t){t=t.checked,t!=null&&$l(e,"checked",t,!1)}function ql(e,t){_u(e,t);var n=ft(t.value),r=t.type;if(n!=null)r==="number"?(n===0&&e.value===""||e.value!=n)&&(e.value=""+n):e.value!==""+n&&(e.value=""+n);else if(r==="submit"||r==="reset"){e.removeAttribute("value");return}t.hasOwnProperty("value")?bl(e,t.type,n):t.hasOwnProperty("defaultValue")&&bl(e,t.type,ft(t.defaultValue)),t.checked==null&&t.defaultChecked!=null&&(e.defaultChecked=!!t.defaultChecked)}function Eu(e,t,n){if(t.hasOwnProperty("value")||t.hasOwnProperty("defaultValue")){var r=t.type;if(!(r!=="submit"&&r!=="reset"||t.value!==void 0&&t.value!==null))return;t=""+e._wrapperState.initialValue,n||t===e.value||(e.value=t),e.defaultValue=t}n=e.name,n!==""&&(e.name=""),e.defaultChecked=!!e._wrapperState.initialChecked,n!==""&&(e.name=n)}function bl(e,t,n){(t!=="number"||kr(e.ownerDocument)!==e)&&(n==null?e.defaultValue=""+e._wrapperState.initialValue:e.defaultValue!==""+n&&(e.defaultValue=""+n))}var jn=Array.isArray;function Jt(e,t,n,r){if(e=e.options,t){t={};for(var l=0;l<n.length;l++)t["$"+n[l]]=!0;for(n=0;n<e.length;n++)l=t.hasOwnProperty("$"+e[n].value),e[n].selected!==l&&(e[n].selected=l),l&&r&&(e[n].defaultSelected=!0)}else{for(n=""+ft(n),t=null,l=0;l<e.length;l++){if(e[l].value===n){e[l].selected=!0,r&&(e[l].defaultSelected=!0);return}t!==null||e[l].disabled||(t=e[l])}t!==null&&(t.selected=!0)}}function eo(e,t){if(t.dangerouslySetInnerHTML!=null)throw Error(w(91));return K({},t,{value:void 0,defaultValue:void 0,children:""+e._wrapperState.initialValue})}function Cu(e,t){var n=t.value;if(n==null){if(n=t.children,t=t.defaultValue,n!=null){if(t!=null)throw Error(w(92));if(jn(n)){if(1<n.length)throw Error(w(93));n=n[0]}t=n}t==null&&(t=""),n=t}e._wrapperState={initialValue:ft(n)}}function Nu(e,t){var n=ft(t.value),r=ft(t.defaultValue);n!=null&&(n=""+n,n!==e.value&&(e.value=n),t.defaultValue==null&&e.defaultValue!==n&&(e.defaultValue=n)),r!=null&&(e.defaultValue=""+r)}function ju(e){var t=e.textContent;t===e._wrapperState.initialValue&&t!==""&&t!==null&&(e.value=t)}function Pu(e){switch(e){case"svg":return"http://www.w3.org/2000/svg";case"math":return"http://www.w3.org/1998/Math/MathML";default:return"http://www.w3.org/1999/xhtml"}}function to(e,t){return e==null||e==="http://www.w3.org/1999/xhtml"?Pu(t):e==="http://www.w3.org/2000/svg"&&t==="foreignObject"?"http://www.w3.org/1999/xhtml":e}var Sr,Tu=function(e){return typeof MSApp<"u"&&MSApp.execUnsafeLocalFunction?function(t,n,r,l){MSApp.execUnsafeLocalFunction(function(){return e(t,n,r,l)})}:e}(function(e,t){if(e.namespaceURI!=="http://www.w3.org/2000/svg"||"innerHTML"in e)e.innerHTML=t;else{for(Sr=Sr||document.createElement("div"),Sr.innerHTML="<svg>"+t.valueOf().toString()+"</svg>",t=Sr.firstChild;e.firstChild;)e.removeChild(e.firstChild);for(;t.firstChild;)e.appendChild(t.firstChild)}});function Pn(e,t){if(t){var n=e.firstChild;if(n&&n===e.lastChild&&n.nodeType===3){n.nodeValue=t;return}}e.textContent=t}var Tn={animationIterationCount:!0,aspectRatio:!0,borderImageOutset:!0,borderImageSlice:!0,borderImageWidth:!0,boxFlex:!0,boxFlexGroup:!0,boxOrdinalGroup:!0,columnCount:!0,columns:!0,flex:!0,flexGrow:!0,flexPositive:!0,flexShrink:!0,flexNegative:!0,flexOrder:!0,gridArea:!0,gridRow:!0,gridRowEnd:!0,gridRowSpan:!0,gridRowStart:!0,gridColumn:!0,gridColumnEnd:!0,gridColumnSpan:!0,gridColumnStart:!0,fontWeight:!0,lineClamp:!0,lineHeight:!0,opacity:!0,order:!0,orphans:!0,tabSize:!0,widows:!0,zIndex:!0,zoom:!0,fillOpacity:!0,floodOpacity:!0,stopOpacity:!0,strokeDasharray:!0,strokeDashoffset:!0,strokeMiterlimit:!0,strokeOpacity:!0,strokeWidth:!0},df=["Webkit","ms","Moz","O"];Object.keys(Tn).forEach(function(e){df.forEach(function(t){t=t+e.charAt(0).toUpperCase()+e.substring(1),Tn[t]=Tn[e]})});function zu(e,t,n){return t==null||typeof t=="boolean"||t===""?"":n||typeof t!="number"||t===0||Tn.hasOwnProperty(e)&&Tn[e]?(""+t).trim():t+"px"}function Lu(e,t){e=e.style;for(var n in t)if(t.hasOwnProperty(n)){var r=n.indexOf("--")===0,l=zu(n,t[n],r);n==="float"&&(n="cssFloat"),r?e.setProperty(n,l):e[n]=l}}var pf=K({menuitem:!0},{area:!0,base:!0,br:!0,col:!0,embed:!0,hr:!0,img:!0,input:!0,keygen:!0,link:!0,meta:!0,param:!0,source:!0,track:!0,wbr:!0});function no(e,t){if(t){if(pf[e]&&(t.children!=null||t.dangerouslySetInnerHTML!=null))throw Error(w(137,e));if(t.dangerouslySetInnerHTML!=null){if(t.children!=null)throw Error(w(60));if(typeof t.dangerouslySetInnerHTML!="object"||!("__html"in t.dangerouslySetInnerHTML))throw Error(w(61))}if(t.style!=null&&typeof t.style!="object")throw Error(w(62))}}function ro(e,t){if(e.indexOf("-")===-1)return typeof t.is=="string";switch(e){case"annotation-xml":case"color-profile":case"font-face":case"font-face-src":case"font-face-uri":case"font-face-format":case"font-face-name":case"missing-glyph":return!1;default:return!0}}var lo=null;function oo(e){return e=e.target||e.srcElement||window,e.correspondingUseElement&&(e=e.correspondingUseElement),e.nodeType===3?e.parentNode:e}var io=null,qt=null,bt=null;function Ru(e){if(e=Jn(e)){if(typeof io!="function")throw Error(w(280));var t=e.stateNode;t&&(t=Hr(t),io(e.stateNode,e.type,t))}}function Ou(e){qt?bt?bt.push(e):bt=[e]:qt=e}function Mu(){if(qt){var e=qt,t=bt;if(bt=qt=null,Ru(e),t)for(e=0;e<t.length;e++)Ru(t[e])}}function Fu(e,t){return e(t)}function Du(){}var uo=!1;function Iu(e,t,n){if(uo)return e(t,n);uo=!0;try{return Fu(e,t,n)}finally{uo=!1,(qt!==null||bt!==null)&&(Du(),Mu())}}function zn(e,t){var n=e.stateNode;if(n===null)return null;var r=Hr(n);if(r===null)return null;n=r[t];e:switch(t){case"onClick":case"onClickCapture":case"onDoubleClick":case"onDoubleClickCapture":case"onMouseDown":case"onMouseDownCapture":case"onMouseMove":case"onMouseMoveCapture":case"onMouseUp":case"onMouseUpCapture":case"onMouseEnter":(r=!r.disabled)||(e=e.type,r=!(e==="button"||e==="input"||e==="select"||e==="textarea")),e=!r;break e;default:e=!1}if(e)return null;if(n&&typeof n!="function")throw Error(w(231,t,typeof n));return n}var so=!1;if(Je)try{var Ln={};Object.defineProperty(Ln,"passive",{get:function(){so=!0}}),window.addEventListener("test",Ln,Ln),window.removeEventListener("test",Ln,Ln)}catch{so=!1}function mf(e,t,n,r,l,o,i,u,s){var c=Array.prototype.slice.call(arguments,3);try{t.apply(n,c)}catch(y){this.onError(y)}}var Rn=!1,xr=null,_r=!1,ao=null,hf={onError:function(e){Rn=!0,xr=e}};function vf(e,t,n,r,l,o,i,u,s){Rn=!1,xr=null,mf.apply(hf,arguments)}function yf(e,t,n,r,l,o,i,u,s){if(vf.apply(this,arguments),Rn){if(Rn){var c=xr;Rn=!1,xr=null}else throw Error(w(198));_r||(_r=!0,ao=c)}}function Ot(e){var t=e,n=e;if(e.alternate)for(;t.return;)t=t.return;else{e=t;do t=e,t.flags&4098&&(n=t.return),e=t.return;while(e)}return t.tag===3?n:null}function Au(e){if(e.tag===13){var t=e.memoizedState;if(t===null&&(e=e.alternate,e!==null&&(t=e.memoizedState)),t!==null)return t.dehydrated}return null}function Uu(e){if(Ot(e)!==e)throw Error(w(188))}function gf(e){var t=e.alternate;if(!t){if(t=Ot(e),t===null)throw Error(w(188));return t!==e?null:e}for(var n=e,r=t;;){var l=n.return;if(l===null)break;var o=l.alternate;if(o===null){if(r=l.return,r!==null){n=r;continue}break}if(l.child===o.child){for(o=l.child;o;){if(o===n)return Uu(l),e;if(o===r)return Uu(l),t;o=o.sibling}throw Error(w(188))}if(n.return!==r.return)n=l,r=o;else{for(var i=!1,u=l.child;u;){if(u===n){i=!0,n=l,r=o;break}if(u===r){i=!0,r=l,n=o;break}u=u.sibling}if(!i){for(u=o.child;u;){if(u===n){i=!0,n=o,r=l;break}if(u===r){i=!0,r=o,n=l;break}u=u.sibling}if(!i)throw Error(w(189))}}if(n.alternate!==r)throw Error(w(190))}if(n.tag!==3)throw Error(w(188));return n.stateNode.current===n?e:t}function $u(e){return e=gf(e),e!==null?Vu(e):null}function Vu(e){if(e.tag===5||e.tag===6)return e;for(e=e.child;e!==null;){var t=Vu(e);if(t!==null)return t;e=e.sibling}return null}var Bu=Ee.unstable_scheduleCallback,Wu=Ee.unstable_cancelCallback,wf=Ee.unstable_shouldYield,kf=Ee.unstable_requestPaint,J=Ee.unstable_now,Sf=Ee.unstable_getCurrentPriorityLevel,co=Ee.unstable_ImmediatePriority,Hu=Ee.unstable_UserBlockingPriority,Er=Ee.unstable_NormalPriority,xf=Ee.unstable_LowPriority,Qu=Ee.unstable_IdlePriority,Cr=null,Qe=null;function _f(e){if(Qe&&typeof Qe.onCommitFiberRoot=="function")try{Qe.onCommitFiberRoot(Cr,e,void 0,(e.current.flags&128)===128)}catch{}}var Ie=Math.clz32?Math.clz32:Nf,Ef=Math.log,Cf=Math.LN2;function Nf(e){return e>>>=0,e===0?32:31-(Ef(e)/Cf|0)|0}var Nr=64,jr=4194304;function On(e){switch(e&-e){case 1:return 1;case 2:return 2;case 4:return 4;case 8:return 8;case 16:return 16;case 32:return 32;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return e&4194240;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return e&130023424;case 134217728:return 134217728;case 268435456:return 268435456;case 536870912:return 536870912;case 1073741824:return 1073741824;default:return e}}function Pr(e,t){var n=e.pendingLanes;if(n===0)return 0;var r=0,l=e.suspendedLanes,o=e.pingedLanes,i=n&268435455;if(i!==0){var u=i&~l;u!==0?r=On(u):(o&=i,o!==0&&(r=On(o)))}else i=n&~l,i!==0?r=On(i):o!==0&&(r=On(o));if(r===0)return 0;if(t!==0&&t!==r&&!(t&l)&&(l=r&-r,o=t&-t,l>=o||l===16&&(o&4194240)!==0))return t;if(r&4&&(r|=n&16),t=e.entangledLanes,t!==0)for(e=e.entanglements,t&=r;0<t;)n=31-Ie(t),l=1<<n,r|=e[n],t&=~l;return r}function jf(e,t){switch(e){case 1:case 2:case 4:return t+250;case 8:case 16:case 32:case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return t+5e3;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return-1;case 134217728:case 268435456:case 536870912:case 1073741824:return-1;default:return-1}}function Pf(e,t){for(var n=e.suspendedLanes,r=e.pingedLanes,l=e.expirationTimes,o=e.pendingLanes;0<o;){var i=31-Ie(o),u=1<<i,s=l[i];s===-1?(!(u&n)||u&r)&&(l[i]=jf(u,t)):s<=t&&(e.expiredLanes|=u),o&=~u}}function fo(e){return e=e.pendingLanes&-1073741825,e!==0?e:e&1073741824?1073741824:0}function Ku(){var e=Nr;return Nr<<=1,!(Nr&4194240)&&(Nr=64),e}function po(e){for(var t=[],n=0;31>n;n++)t.push(e);return t}function Mn(e,t,n){e.pendingLanes|=t,t!==536870912&&(e.suspendedLanes=0,e.pingedLanes=0),e=e.eventTimes,t=31-Ie(t),e[t]=n}function Tf(e,t){var n=e.pendingLanes&~t;e.pendingLanes=t,e.suspendedLanes=0,e.pingedLanes=0,e.expiredLanes&=t,e.mutableReadLanes&=t,e.entangledLanes&=t,t=e.entanglements;var r=e.eventTimes;for(e=e.expirationTimes;0<n;){var l=31-Ie(n),o=1<<l;t[l]=0,r[l]=-1,e[l]=-1,n&=~o}}function mo(e,t){var n=e.entangledLanes|=t;for(e=e.entanglements;n;){var r=31-Ie(n),l=1<<r;l&t|e[r]&t&&(e[r]|=t),n&=~l}}var U=0;function Yu(e){return e&=-e,1<e?4<e?e&268435455?16:536870912:4:1}var Xu,ho,Gu,Zu,Ju,vo=!1,Tr=[],dt=null,pt=null,mt=null,Fn=new Map,Dn=new Map,ht=[],zf="mousedown mouseup touchcancel touchend touchstart auxclick dblclick pointercancel pointerdown pointerup dragend dragstart drop compositionend compositionstart keydown keypress keyup input textInput copy cut paste click change contextmenu reset submit".split(" ");function qu(e,t){switch(e){case"focusin":case"focusout":dt=null;break;case"dragenter":case"dragleave":pt=null;break;case"mouseover":case"mouseout":mt=null;break;case"pointerover":case"pointerout":Fn.delete(t.pointerId);break;case"gotpointercapture":case"lostpointercapture":Dn.delete(t.pointerId)}}function In(e,t,n,r,l,o){return e===null||e.nativeEvent!==o?(e={blockedOn:t,domEventName:n,eventSystemFlags:r,nativeEvent:o,targetContainers:[l]},t!==null&&(t=Jn(t),t!==null&&ho(t)),e):(e.eventSystemFlags|=r,t=e.targetContainers,l!==null&&t.indexOf(l)===-1&&t.push(l),e)}function Lf(e,t,n,r,l){switch(t){case"focusin":return dt=In(dt,e,t,n,r,l),!0;case"dragenter":return pt=In(pt,e,t,n,r,l),!0;case"mouseover":return mt=In(mt,e,t,n,r,l),!0;case"pointerover":var o=l.pointerId;return Fn.set(o,In(Fn.get(o)||null,e,t,n,r,l)),!0;case"gotpointercapture":return o=l.pointerId,Dn.set(o,In(Dn.get(o)||null,e,t,n,r,l)),!0}return!1}function bu(e){var t=Mt(e.target);if(t!==null){var n=Ot(t);if(n!==null){if(t=n.tag,t===13){if(t=Au(n),t!==null){e.blockedOn=t,Ju(e.priority,function(){Gu(n)});return}}else if(t===3&&n.stateNode.current.memoizedState.isDehydrated){e.blockedOn=n.tag===3?n.stateNode.containerInfo:null;return}}}e.blockedOn=null}function zr(e){if(e.blockedOn!==null)return!1;for(var t=e.targetContainers;0<t.length;){var n=go(e.domEventName,e.eventSystemFlags,t[0],e.nativeEvent);if(n===null){n=e.nativeEvent;var r=new n.constructor(n.type,n);lo=r,n.target.dispatchEvent(r),lo=null}else return t=Jn(n),t!==null&&ho(t),e.blockedOn=n,!1;t.shift()}return!0}function es(e,t,n){zr(e)&&n.delete(t)}function Rf(){vo=!1,dt!==null&&zr(dt)&&(dt=null),pt!==null&&zr(pt)&&(pt=null),mt!==null&&zr(mt)&&(mt=null),Fn.forEach(es),Dn.forEach(es)}function An(e,t){e.blockedOn===t&&(e.blockedOn=null,vo||(vo=!0,Ee.unstable_scheduleCallback(Ee.unstable_NormalPriority,Rf)))}function Un(e){function t(l){return An(l,e)}if(0<Tr.length){An(Tr[0],e);for(var n=1;n<Tr.length;n++){var r=Tr[n];r.blockedOn===e&&(r.blockedOn=null)}}for(dt!==null&&An(dt,e),pt!==null&&An(pt,e),mt!==null&&An(mt,e),Fn.forEach(t),Dn.forEach(t),n=0;n<ht.length;n++)r=ht[n],r.blockedOn===e&&(r.blockedOn=null);for(;0<ht.length&&(n=ht[0],n.blockedOn===null);)bu(n),n.blockedOn===null&&ht.shift()}var en=qe.ReactCurrentBatchConfig,Lr=!0;function Of(e,t,n,r){var l=U,o=en.transition;en.transition=null;try{U=1,yo(e,t,n,r)}finally{U=l,en.transition=o}}function Mf(e,t,n,r){var l=U,o=en.transition;en.transition=null;try{U=4,yo(e,t,n,r)}finally{U=l,en.transition=o}}function yo(e,t,n,r){if(Lr){var l=go(e,t,n,r);if(l===null)Fo(e,t,r,Rr,n),qu(e,r);else if(Lf(l,e,t,n,r))r.stopPropagation();else if(qu(e,r),t&4&&-1<zf.indexOf(e)){for(;l!==null;){var o=Jn(l);if(o!==null&&Xu(o),o=go(e,t,n,r),o===null&&Fo(e,t,r,Rr,n),o===l)break;l=o}l!==null&&r.stopPropagation()}else Fo(e,t,r,null,n)}}var Rr=null;function go(e,t,n,r){if(Rr=null,e=oo(r),e=Mt(e),e!==null)if(t=Ot(e),t===null)e=null;else if(n=t.tag,n===13){if(e=Au(t),e!==null)return e;e=null}else if(n===3){if(t.stateNode.current.memoizedState.isDehydrated)return t.tag===3?t.stateNode.containerInfo:null;e=null}else t!==e&&(e=null);return Rr=e,null}function ts(e){switch(e){case"cancel":case"click":case"close":case"contextmenu":case"copy":case"cut":case"auxclick":case"dblclick":case"dragend":case"dragstart":case"drop":case"focusin":case"focusout":case"input":case"invalid":case"keydown":case"keypress":case"keyup":case"mousedown":case"mouseup":case"paste":case"pause":case"play":case"pointercancel":case"pointerdown":case"pointerup":case"ratechange":case"reset":case"resize":case"seeked":case"submit":case"touchcancel":case"touchend":case"touchstart":case"volumechange":case"change":case"selectionchange":case"textInput":case"compositionstart":case"compositionend":case"compositionupdate":case"beforeblur":case"afterblur":case"beforeinput":case"blur":case"fullscreenchange":case"focus":case"hashchange":case"popstate":case"select":case"selectstart":return 1;case"drag":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"mousemove":case"mouseout":case"mouseover":case"pointermove":case"pointerout":case"pointerover":case"scroll":case"toggle":case"touchmove":case"wheel":case"mouseenter":case"mouseleave":case"pointerenter":case"pointerleave":return 4;case"message":switch(Sf()){case co:return 1;case Hu:return 4;case Er:case xf:return 16;case Qu:return 536870912;default:return 16}default:return 16}}var vt=null,wo=null,Or=null;function ns(){if(Or)return Or;var e,t=wo,n=t.length,r,l="value"in vt?vt.value:vt.textContent,o=l.length;for(e=0;e<n&&t[e]===l[e];e++);var i=n-e;for(r=1;r<=i&&t[n-r]===l[o-r];r++);return Or=l.slice(e,1<r?1-r:void 0)}function Mr(e){var t=e.keyCode;return"charCode"in e?(e=e.charCode,e===0&&t===13&&(e=13)):e=t,e===10&&(e=13),32<=e||e===13?e:0}function Fr(){return!0}function rs(){return!1}function Ce(e){function t(n,r,l,o,i){this._reactName=n,this._targetInst=l,this.type=r,this.nativeEvent=o,this.target=i,this.currentTarget=null;for(var u in e)e.hasOwnProperty(u)&&(n=e[u],this[u]=n?n(o):o[u]);return this.isDefaultPrevented=(o.defaultPrevented!=null?o.defaultPrevented:o.returnValue===!1)?Fr:rs,this.isPropagationStopped=rs,this}return K(t.prototype,{preventDefault:function(){this.defaultPrevented=!0;var n=this.nativeEvent;n&&(n.preventDefault?n.preventDefault():typeof n.returnValue!="unknown"&&(n.returnValue=!1),this.isDefaultPrevented=Fr)},stopPropagation:function(){var n=this.nativeEvent;n&&(n.stopPropagation?n.stopPropagation():typeof n.cancelBubble!="unknown"&&(n.cancelBubble=!0),this.isPropagationStopped=Fr)},persist:function(){},isPersistent:Fr}),t}var tn={eventPhase:0,bubbles:0,cancelable:0,timeStamp:function(e){return e.timeStamp||Date.now()},defaultPrevented:0,isTrusted:0},ko=Ce(tn),$n=K({},tn,{view:0,detail:0}),Ff=Ce($n),So,xo,Vn,Dr=K({},$n,{screenX:0,screenY:0,clientX:0,clientY:0,pageX:0,pageY:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,getModifierState:Eo,button:0,buttons:0,relatedTarget:function(e){return e.relatedTarget===void 0?e.fromElement===e.srcElement?e.toElement:e.fromElement:e.relatedTarget},movementX:function(e){return"movementX"in e?e.movementX:(e!==Vn&&(Vn&&e.type==="mousemove"?(So=e.screenX-Vn.screenX,xo=e.screenY-Vn.screenY):xo=So=0,Vn=e),So)},movementY:function(e){return"movementY"in e?e.movementY:xo}}),ls=Ce(Dr),Df=K({},Dr,{dataTransfer:0}),If=Ce(Df),Af=K({},$n,{relatedTarget:0}),_o=Ce(Af),Uf=K({},tn,{animationName:0,elapsedTime:0,pseudoElement:0}),$f=Ce(Uf),Vf=K({},tn,{clipboardData:function(e){return"clipboardData"in e?e.clipboardData:window.clipboardData}}),Bf=Ce(Vf),Wf=K({},tn,{data:0}),os=Ce(Wf),Hf={Esc:"Escape",Spacebar:" ",Left:"ArrowLeft",Up:"ArrowUp",Right:"ArrowRight",Down:"ArrowDown",Del:"Delete",Win:"OS",Menu:"ContextMenu",Apps:"ContextMenu",Scroll:"ScrollLock",MozPrintableKey:"Unidentified"},Qf={8:"Backspace",9:"Tab",12:"Clear",13:"Enter",16:"Shift",17:"Control",18:"Alt",19:"Pause",20:"CapsLock",27:"Escape",32:" ",33:"PageUp",34:"PageDown",35:"End",36:"Home",37:"ArrowLeft",38:"ArrowUp",39:"ArrowRight",40:"ArrowDown",45:"Insert",46:"Delete",112:"F1",113:"F2",114:"F3",115:"F4",116:"F5",117:"F6",118:"F7",119:"F8",120:"F9",121:"F10",122:"F11",123:"F12",144:"NumLock",145:"ScrollLock",224:"Meta"},Kf={Alt:"altKey",Control:"ctrlKey",Meta:"metaKey",Shift:"shiftKey"};function Yf(e){var t=this.nativeEvent;return t.getModifierState?t.getModifierState(e):(e=Kf[e])?!!t[e]:!1}function Eo(){return Yf}var Xf=K({},$n,{key:function(e){if(e.key){var t=Hf[e.key]||e.key;if(t!=="Unidentified")return t}return e.type==="keypress"?(e=Mr(e),e===13?"Enter":String.fromCharCode(e)):e.type==="keydown"||e.type==="keyup"?Qf[e.keyCode]||"Unidentified":""},code:0,location:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,repeat:0,locale:0,getModifierState:Eo,charCode:function(e){return e.type==="keypress"?Mr(e):0},keyCode:function(e){return e.type==="keydown"||e.type==="keyup"?e.keyCode:0},which:function(e){return e.type==="keypress"?Mr(e):e.type==="keydown"||e.type==="keyup"?e.keyCode:0}}),Gf=Ce(Xf),Zf=K({},Dr,{pointerId:0,width:0,height:0,pressure:0,tangentialPressure:0,tiltX:0,tiltY:0,twist:0,pointerType:0,isPrimary:0}),is=Ce(Zf),Jf=K({},$n,{touches:0,targetTouches:0,changedTouches:0,altKey:0,metaKey:0,ctrlKey:0,shiftKey:0,getModifierState:Eo}),qf=Ce(Jf),bf=K({},tn,{propertyName:0,elapsedTime:0,pseudoElement:0}),ed=Ce(bf),td=K({},Dr,{deltaX:function(e){return"deltaX"in e?e.deltaX:"wheelDeltaX"in e?-e.wheelDeltaX:0},deltaY:function(e){return"deltaY"in e?e.deltaY:"wheelDeltaY"in e?-e.wheelDeltaY:"wheelDelta"in e?-e.wheelDelta:0},deltaZ:0,deltaMode:0}),nd=Ce(td),rd=[9,13,27,32],Co=Je&&"CompositionEvent"in window,Bn=null;Je&&"documentMode"in document&&(Bn=document.documentMode);var ld=Je&&"TextEvent"in window&&!Bn,us=Je&&(!Co||Bn&&8<Bn&&11>=Bn),ss=" ",as=!1;function cs(e,t){switch(e){case"keyup":return rd.indexOf(t.keyCode)!==-1;case"keydown":return t.keyCode!==229;case"keypress":case"mousedown":case"focusout":return!0;default:return!1}}function fs(e){return e=e.detail,typeof e=="object"&&"data"in e?e.data:null}var nn=!1;function od(e,t){switch(e){case"compositionend":return fs(t);case"keypress":return t.which!==32?null:(as=!0,ss);case"textInput":return e=t.data,e===ss&&as?null:e;default:return null}}function id(e,t){if(nn)return e==="compositionend"||!Co&&cs(e,t)?(e=ns(),Or=wo=vt=null,nn=!1,e):null;switch(e){case"paste":return null;case"keypress":if(!(t.ctrlKey||t.altKey||t.metaKey)||t.ctrlKey&&t.altKey){if(t.char&&1<t.char.length)return t.char;if(t.which)return String.fromCharCode(t.which)}return null;case"compositionend":return us&&t.locale!=="ko"?null:t.data;default:return null}}var ud={color:!0,date:!0,datetime:!0,"datetime-local":!0,email:!0,month:!0,number:!0,password:!0,range:!0,search:!0,tel:!0,text:!0,time:!0,url:!0,week:!0};function ds(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t==="input"?!!ud[e.type]:t==="textarea"}function ps(e,t,n,r){Ou(r),t=Vr(t,"onChange"),0<t.length&&(n=new ko("onChange","change",null,n,r),e.push({event:n,listeners:t}))}var Wn=null,Hn=null;function sd(e){Ls(e,0)}function Ir(e){var t=sn(e);if(Su(t))return e}function ad(e,t){if(e==="change")return t}var ms=!1;if(Je){var No;if(Je){var jo="oninput"in document;if(!jo){var hs=document.createElement("div");hs.setAttribute("oninput","return;"),jo=typeof hs.oninput=="function"}No=jo}else No=!1;ms=No&&(!document.documentMode||9<document.documentMode)}function vs(){Wn&&(Wn.detachEvent("onpropertychange",ys),Hn=Wn=null)}function ys(e){if(e.propertyName==="value"&&Ir(Hn)){var t=[];ps(t,Hn,e,oo(e)),Iu(sd,t)}}function cd(e,t,n){e==="focusin"?(vs(),Wn=t,Hn=n,Wn.attachEvent("onpropertychange",ys)):e==="focusout"&&vs()}function fd(e){if(e==="selectionchange"||e==="keyup"||e==="keydown")return Ir(Hn)}function dd(e,t){if(e==="click")return Ir(t)}function pd(e,t){if(e==="input"||e==="change")return Ir(t)}function md(e,t){return e===t&&(e!==0||1/e===1/t)||e!==e&&t!==t}var Ae=typeof Object.is=="function"?Object.is:md;function Qn(e,t){if(Ae(e,t))return!0;if(typeof e!="object"||e===null||typeof t!="object"||t===null)return!1;var n=Object.keys(e),r=Object.keys(t);if(n.length!==r.length)return!1;for(r=0;r<n.length;r++){var l=n[r];if(!Il.call(t,l)||!Ae(e[l],t[l]))return!1}return!0}function gs(e){for(;e&&e.firstChild;)e=e.firstChild;return e}function ws(e,t){var n=gs(e);e=0;for(var r;n;){if(n.nodeType===3){if(r=e+n.textContent.length,e<=t&&r>=t)return{node:n,offset:t-e};e=r}e:{for(;n;){if(n.nextSibling){n=n.nextSibling;break e}n=n.parentNode}n=void 0}n=gs(n)}}function ks(e,t){return e&&t?e===t?!0:e&&e.nodeType===3?!1:t&&t.nodeType===3?ks(e,t.parentNode):"contains"in e?e.contains(t):e.compareDocumentPosition?!!(e.compareDocumentPosition(t)&16):!1:!1}function Ss(){for(var e=window,t=kr();t instanceof e.HTMLIFrameElement;){try{var n=typeof t.contentWindow.location.href=="string"}catch{n=!1}if(n)e=t.contentWindow;else break;t=kr(e.document)}return t}function Po(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t&&(t==="input"&&(e.type==="text"||e.type==="search"||e.type==="tel"||e.type==="url"||e.type==="password")||t==="textarea"||e.contentEditable==="true")}function hd(e){var t=Ss(),n=e.focusedElem,r=e.selectionRange;if(t!==n&&n&&n.ownerDocument&&ks(n.ownerDocument.documentElement,n)){if(r!==null&&Po(n)){if(t=r.start,e=r.end,e===void 0&&(e=t),"selectionStart"in n)n.selectionStart=t,n.selectionEnd=Math.min(e,n.value.length);else if(e=(t=n.ownerDocument||document)&&t.defaultView||window,e.getSelection){e=e.getSelection();var l=n.textContent.length,o=Math.min(r.start,l);r=r.end===void 0?o:Math.min(r.end,l),!e.extend&&o>r&&(l=r,r=o,o=l),l=ws(n,o);var i=ws(n,r);l&&i&&(e.rangeCount!==1||e.anchorNode!==l.node||e.anchorOffset!==l.offset||e.focusNode!==i.node||e.focusOffset!==i.offset)&&(t=t.createRange(),t.setStart(l.node,l.offset),e.removeAllRanges(),o>r?(e.addRange(t),e.extend(i.node,i.offset)):(t.setEnd(i.node,i.offset),e.addRange(t)))}}for(t=[],e=n;e=e.parentNode;)e.nodeType===1&&t.push({element:e,left:e.scrollLeft,top:e.scrollTop});for(typeof n.focus=="function"&&n.focus(),n=0;n<t.length;n++)e=t[n],e.element.scrollLeft=e.left,e.element.scrollTop=e.top}}var vd=Je&&"documentMode"in document&&11>=document.documentMode,rn=null,To=null,Kn=null,zo=!1;function xs(e,t,n){var r=n.window===n?n.document:n.nodeType===9?n:n.ownerDocument;zo||rn==null||rn!==kr(r)||(r=rn,"selectionStart"in r&&Po(r)?r={start:r.selectionStart,end:r.selectionEnd}:(r=(r.ownerDocument&&r.ownerDocument.defaultView||window).getSelection(),r={anchorNode:r.anchorNode,anchorOffset:r.anchorOffset,focusNode:r.focusNode,focusOffset:r.focusOffset}),Kn&&Qn(Kn,r)||(Kn=r,r=Vr(To,"onSelect"),0<r.length&&(t=new ko("onSelect","select",null,t,n),e.push({event:t,listeners:r}),t.target=rn)))}function Ar(e,t){var n={};return n[e.toLowerCase()]=t.toLowerCase(),n["Webkit"+e]="webkit"+t,n["Moz"+e]="moz"+t,n}var ln={animationend:Ar("Animation","AnimationEnd"),animationiteration:Ar("Animation","AnimationIteration"),animationstart:Ar("Animation","AnimationStart"),transitionend:Ar("Transition","TransitionEnd")},Lo={},_s={};Je&&(_s=document.createElement("div").style,"AnimationEvent"in window||(delete ln.animationend.animation,delete ln.animationiteration.animation,delete ln.animationstart.animation),"TransitionEvent"in window||delete ln.transitionend.transition);function Ur(e){if(Lo[e])return Lo[e];if(!ln[e])return e;var t=ln[e],n;for(n in t)if(t.hasOwnProperty(n)&&n in _s)return Lo[e]=t[n];return e}var Es=Ur("animationend"),Cs=Ur("animationiteration"),Ns=Ur("animationstart"),js=Ur("transitionend"),Ps=new Map,Ts="abort auxClick cancel canPlay canPlayThrough click close contextMenu copy cut drag dragEnd dragEnter dragExit dragLeave dragOver dragStart drop durationChange emptied encrypted ended error gotPointerCapture input invalid keyDown keyPress keyUp load loadedData loadedMetadata loadStart lostPointerCapture mouseDown mouseMove mouseOut mouseOver mouseUp paste pause play playing pointerCancel pointerDown pointerMove pointerOut pointerOver pointerUp progress rateChange reset resize seeked seeking stalled submit suspend timeUpdate touchCancel touchEnd touchStart volumeChange scroll toggle touchMove waiting wheel".split(" ");function yt(e,t){Ps.set(e,t),Rt(t,[e])}for(var Ro=0;Ro<Ts.length;Ro++){var Oo=Ts[Ro],yd=Oo.toLowerCase(),gd=Oo[0].toUpperCase()+Oo.slice(1);yt(yd,"on"+gd)}yt(Es,"onAnimationEnd"),yt(Cs,"onAnimationIteration"),yt(Ns,"onAnimationStart"),yt("dblclick","onDoubleClick"),yt("focusin","onFocus"),yt("focusout","onBlur"),yt(js,"onTransitionEnd"),Xt("onMouseEnter",["mouseout","mouseover"]),Xt("onMouseLeave",["mouseout","mouseover"]),Xt("onPointerEnter",["pointerout","pointerover"]),Xt("onPointerLeave",["pointerout","pointerover"]),Rt("onChange","change click focusin focusout input keydown keyup selectionchange".split(" ")),Rt("onSelect","focusout contextmenu dragend focusin keydown keyup mousedown mouseup selectionchange".split(" ")),Rt("onBeforeInput",["compositionend","keypress","textInput","paste"]),Rt("onCompositionEnd","compositionend focusout keydown keypress keyup mousedown".split(" ")),Rt("onCompositionStart","compositionstart focusout keydown keypress keyup mousedown".split(" ")),Rt("onCompositionUpdate","compositionupdate focusout keydown keypress keyup mousedown".split(" "));var Yn="abort canplay canplaythrough durationchange emptied encrypted ended error loadeddata loadedmetadata loadstart pause play playing progress ratechange resize seeked seeking stalled suspend timeupdate volumechange waiting".split(" "),wd=new Set("cancel close invalid load scroll toggle".split(" ").concat(Yn));function zs(e,t,n){var r=e.type||"unknown-event";e.currentTarget=n,yf(r,t,void 0,e),e.currentTarget=null}function Ls(e,t){t=(t&4)!==0;for(var n=0;n<e.length;n++){var r=e[n],l=r.event;r=r.listeners;e:{var o=void 0;if(t)for(var i=r.length-1;0<=i;i--){var u=r[i],s=u.instance,c=u.currentTarget;if(u=u.listener,s!==o&&l.isPropagationStopped())break e;zs(l,u,c),o=s}else for(i=0;i<r.length;i++){if(u=r[i],s=u.instance,c=u.currentTarget,u=u.listener,s!==o&&l.isPropagationStopped())break e;zs(l,u,c),o=s}}}if(_r)throw e=ao,_r=!1,ao=null,e}function V(e,t){var n=t[Vo];n===void 0&&(n=t[Vo]=new Set);var r=e+"__bubble";n.has(r)||(Rs(t,e,2,!1),n.add(r))}function Mo(e,t,n){var r=0;t&&(r|=4),Rs(n,e,r,t)}var $r="_reactListening"+Math.random().toString(36).slice(2);function Xn(e){if(!e[$r]){e[$r]=!0,pu.forEach(function(n){n!=="selectionchange"&&(wd.has(n)||Mo(n,!1,e),Mo(n,!0,e))});var t=e.nodeType===9?e:e.ownerDocument;t===null||t[$r]||(t[$r]=!0,Mo("selectionchange",!1,t))}}function Rs(e,t,n,r){switch(ts(t)){case 1:var l=Of;break;case 4:l=Mf;break;default:l=yo}n=l.bind(null,t,n,e),l=void 0,!so||t!=="touchstart"&&t!=="touchmove"&&t!=="wheel"||(l=!0),r?l!==void 0?e.addEventListener(t,n,{capture:!0,passive:l}):e.addEventListener(t,n,!0):l!==void 0?e.addEventListener(t,n,{passive:l}):e.addEventListener(t,n,!1)}function Fo(e,t,n,r,l){var o=r;if(!(t&1)&&!(t&2)&&r!==null)e:for(;;){if(r===null)return;var i=r.tag;if(i===3||i===4){var u=r.stateNode.containerInfo;if(u===l||u.nodeType===8&&u.parentNode===l)break;if(i===4)for(i=r.return;i!==null;){var s=i.tag;if((s===3||s===4)&&(s=i.stateNode.containerInfo,s===l||s.nodeType===8&&s.parentNode===l))return;i=i.return}for(;u!==null;){if(i=Mt(u),i===null)return;if(s=i.tag,s===5||s===6){r=o=i;continue e}u=u.parentNode}}r=r.return}Iu(function(){var c=o,y=oo(n),v=[];e:{var m=Ps.get(e);if(m!==void 0){var k=ko,S=e;switch(e){case"keypress":if(Mr(n)===0)break e;case"keydown":case"keyup":k=Gf;break;case"focusin":S="focus",k=_o;break;case"focusout":S="blur",k=_o;break;case"beforeblur":case"afterblur":k=_o;break;case"click":if(n.button===2)break e;case"auxclick":case"dblclick":case"mousedown":case"mousemove":case"mouseup":case"mouseout":case"mouseover":case"contextmenu":k=ls;break;case"drag":case"dragend":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"dragstart":case"drop":k=If;break;case"touchcancel":case"touchend":case"touchmove":case"touchstart":k=qf;break;case Es:case Cs:case Ns:k=$f;break;case js:k=ed;break;case"scroll":k=Ff;break;case"wheel":k=nd;break;case"copy":case"cut":case"paste":k=Bf;break;case"gotpointercapture":case"lostpointercapture":case"pointercancel":case"pointerdown":case"pointermove":case"pointerout":case"pointerover":case"pointerup":k=is}var C=(t&4)!==0,A=!C&&e==="scroll",f=C?m!==null?m+"Capture":null:m;C=[];for(var a=c,d;a!==null;){d=a;var g=d.stateNode;if(d.tag===5&&g!==null&&(d=g,f!==null&&(g=zn(a,f),g!=null&&C.push(Gn(a,g,d)))),A)break;a=a.return}0<C.length&&(m=new k(m,S,null,n,y),v.push({event:m,listeners:C}))}}if(!(t&7)){e:{if(m=e==="mouseover"||e==="pointerover",k=e==="mouseout"||e==="pointerout",m&&n!==lo&&(S=n.relatedTarget||n.fromElement)&&(Mt(S)||S[be]))break e;if((k||m)&&(m=y.window===y?y:(m=y.ownerDocument)?m.defaultView||m.parentWindow:window,k?(S=n.relatedTarget||n.toElement,k=c,S=S?Mt(S):null,S!==null&&(A=Ot(S),S!==A||S.tag!==5&&S.tag!==6)&&(S=null)):(k=null,S=c),k!==S)){if(C=ls,g="onMouseLeave",f="onMouseEnter",a="mouse",(e==="pointerout"||e==="pointerover")&&(C=is,g="onPointerLeave",f="onPointerEnter",a="pointer"),A=k==null?m:sn(k),d=S==null?m:sn(S),m=new C(g,a+"leave",k,n,y),m.target=A,m.relatedTarget=d,g=null,Mt(y)===c&&(C=new C(f,a+"enter",S,n,y),C.target=d,C.relatedTarget=A,g=C),A=g,k&&S)t:{for(C=k,f=S,a=0,d=C;d;d=on(d))a++;for(d=0,g=f;g;g=on(g))d++;for(;0<a-d;)C=on(C),a--;for(;0<d-a;)f=on(f),d--;for(;a--;){if(C===f||f!==null&&C===f.alternate)break t;C=on(C),f=on(f)}C=null}else C=null;k!==null&&Os(v,m,k,C,!1),S!==null&&A!==null&&Os(v,A,S,C,!0)}}e:{if(m=c?sn(c):window,k=m.nodeName&&m.nodeName.toLowerCase(),k==="select"||k==="input"&&m.type==="file")var N=ad;else if(ds(m))if(ms)N=pd;else{N=fd;var P=cd}else(k=m.nodeName)&&k.toLowerCase()==="input"&&(m.type==="checkbox"||m.type==="radio")&&(N=dd);if(N&&(N=N(e,c))){ps(v,N,n,y);break e}P&&P(e,m,c),e==="focusout"&&(P=m._wrapperState)&&P.controlled&&m.type==="number"&&bl(m,"number",m.value)}switch(P=c?sn(c):window,e){case"focusin":(ds(P)||P.contentEditable==="true")&&(rn=P,To=c,Kn=null);break;case"focusout":Kn=To=rn=null;break;case"mousedown":zo=!0;break;case"contextmenu":case"mouseup":case"dragend":zo=!1,xs(v,n,y);break;case"selectionchange":if(vd)break;case"keydown":case"keyup":xs(v,n,y)}var T;if(Co)e:{switch(e){case"compositionstart":var L="onCompositionStart";break e;case"compositionend":L="onCompositionEnd";break e;case"compositionupdate":L="onCompositionUpdate";break e}L=void 0}else nn?cs(e,n)&&(L="onCompositionEnd"):e==="keydown"&&n.keyCode===229&&(L="onCompositionStart");L&&(us&&n.locale!=="ko"&&(nn||L!=="onCompositionStart"?L==="onCompositionEnd"&&nn&&(T=ns()):(vt=y,wo="value"in vt?vt.value:vt.textContent,nn=!0)),P=Vr(c,L),0<P.length&&(L=new os(L,e,null,n,y),v.push({event:L,listeners:P}),T?L.data=T:(T=fs(n),T!==null&&(L.data=T)))),(T=ld?od(e,n):id(e,n))&&(c=Vr(c,"onBeforeInput"),0<c.length&&(y=new os("onBeforeInput","beforeinput",null,n,y),v.push({event:y,listeners:c}),y.data=T))}Ls(v,t)})}function Gn(e,t,n){return{instance:e,listener:t,currentTarget:n}}function Vr(e,t){for(var n=t+"Capture",r=[];e!==null;){var l=e,o=l.stateNode;l.tag===5&&o!==null&&(l=o,o=zn(e,n),o!=null&&r.unshift(Gn(e,o,l)),o=zn(e,t),o!=null&&r.push(Gn(e,o,l))),e=e.return}return r}function on(e){if(e===null)return null;do e=e.return;while(e&&e.tag!==5);return e||null}function Os(e,t,n,r,l){for(var o=t._reactName,i=[];n!==null&&n!==r;){var u=n,s=u.alternate,c=u.stateNode;if(s!==null&&s===r)break;u.tag===5&&c!==null&&(u=c,l?(s=zn(n,o),s!=null&&i.unshift(Gn(n,s,u))):l||(s=zn(n,o),s!=null&&i.push(Gn(n,s,u)))),n=n.return}i.length!==0&&e.push({event:t,listeners:i})}var kd=/\r\n?/g,Sd=/\u0000|\uFFFD/g;function Ms(e){return(typeof e=="string"?e:""+e).replace(kd,`
-`).replace(Sd,"")}function Br(e,t,n){if(t=Ms(t),Ms(e)!==t&&n)throw Error(w(425))}function Wr(){}var Do=null,Io=null;function Ao(e,t){return e==="textarea"||e==="noscript"||typeof t.children=="string"||typeof t.children=="number"||typeof t.dangerouslySetInnerHTML=="object"&&t.dangerouslySetInnerHTML!==null&&t.dangerouslySetInnerHTML.__html!=null}var Uo=typeof setTimeout=="function"?setTimeout:void 0,xd=typeof clearTimeout=="function"?clearTimeout:void 0,Fs=typeof Promise=="function"?Promise:void 0,_d=typeof queueMicrotask=="function"?queueMicrotask:typeof Fs<"u"?function(e){return Fs.resolve(null).then(e).catch(Ed)}:Uo;function Ed(e){setTimeout(function(){throw e})}function $o(e,t){var n=t,r=0;do{var l=n.nextSibling;if(e.removeChild(n),l&&l.nodeType===8)if(n=l.data,n==="/$"){if(r===0){e.removeChild(l),Un(t);return}r--}else n!=="$"&&n!=="$?"&&n!=="$!"||r++;n=l}while(n);Un(t)}function gt(e){for(;e!=null;e=e.nextSibling){var t=e.nodeType;if(t===1||t===3)break;if(t===8){if(t=e.data,t==="$"||t==="$!"||t==="$?")break;if(t==="/$")return null}}return e}function Ds(e){e=e.previousSibling;for(var t=0;e;){if(e.nodeType===8){var n=e.data;if(n==="$"||n==="$!"||n==="$?"){if(t===0)return e;t--}else n==="/$"&&t++}e=e.previousSibling}return null}var un=Math.random().toString(36).slice(2),Ke="__reactFiber$"+un,Zn="__reactProps$"+un,be="__reactContainer$"+un,Vo="__reactEvents$"+un,Cd="__reactListeners$"+un,Nd="__reactHandles$"+un;function Mt(e){var t=e[Ke];if(t)return t;for(var n=e.parentNode;n;){if(t=n[be]||n[Ke]){if(n=t.alternate,t.child!==null||n!==null&&n.child!==null)for(e=Ds(e);e!==null;){if(n=e[Ke])return n;e=Ds(e)}return t}e=n,n=e.parentNode}return null}function Jn(e){return e=e[Ke]||e[be],!e||e.tag!==5&&e.tag!==6&&e.tag!==13&&e.tag!==3?null:e}function sn(e){if(e.tag===5||e.tag===6)return e.stateNode;throw Error(w(33))}function Hr(e){return e[Zn]||null}var Bo=[],an=-1;function wt(e){return{current:e}}function B(e){0>an||(e.current=Bo[an],Bo[an]=null,an--)}function $(e,t){an++,Bo[an]=e.current,e.current=t}var kt={},se=wt(kt),ye=wt(!1),Ft=kt;function cn(e,t){var n=e.type.contextTypes;if(!n)return kt;var r=e.stateNode;if(r&&r.__reactInternalMemoizedUnmaskedChildContext===t)return r.__reactInternalMemoizedMaskedChildContext;var l={},o;for(o in n)l[o]=t[o];return r&&(e=e.stateNode,e.__reactInternalMemoizedUnmaskedChildContext=t,e.__reactInternalMemoizedMaskedChildContext=l),l}function ge(e){return e=e.childContextTypes,e!=null}function Qr(){B(ye),B(se)}function Is(e,t,n){if(se.current!==kt)throw Error(w(168));$(se,t),$(ye,n)}function As(e,t,n){var r=e.stateNode;if(t=t.childContextTypes,typeof r.getChildContext!="function")return n;r=r.getChildContext();for(var l in r)if(!(l in t))throw Error(w(108,cf(e)||"Unknown",l));return K({},n,r)}function Kr(e){return e=(e=e.stateNode)&&e.__reactInternalMemoizedMergedChildContext||kt,Ft=se.current,$(se,e),$(ye,ye.current),!0}function Us(e,t,n){var r=e.stateNode;if(!r)throw Error(w(169));n?(e=As(e,t,Ft),r.__reactInternalMemoizedMergedChildContext=e,B(ye),B(se),$(se,e)):B(ye),$(ye,n)}var et=null,Yr=!1,Wo=!1;function $s(e){et===null?et=[e]:et.push(e)}function jd(e){Yr=!0,$s(e)}function St(){if(!Wo&&et!==null){Wo=!0;var e=0,t=U;try{var n=et;for(U=1;e<n.length;e++){var r=n[e];do r=r(!0);while(r!==null)}et=null,Yr=!1}catch(l){throw et!==null&&(et=et.slice(e+1)),Bu(co,St),l}finally{U=t,Wo=!1}}return null}var fn=[],dn=0,Xr=null,Gr=0,ze=[],Le=0,Dt=null,tt=1,nt="";function It(e,t){fn[dn++]=Gr,fn[dn++]=Xr,Xr=e,Gr=t}function Vs(e,t,n){ze[Le++]=tt,ze[Le++]=nt,ze[Le++]=Dt,Dt=e;var r=tt;e=nt;var l=32-Ie(r)-1;r&=~(1<<l),n+=1;var o=32-Ie(t)+l;if(30<o){var i=l-l%5;o=(r&(1<<i)-1).toString(32),r>>=i,l-=i,tt=1<<32-Ie(t)+l|n<<l|r,nt=o+e}else tt=1<<o|n<<l|r,nt=e}function Ho(e){e.return!==null&&(It(e,1),Vs(e,1,0))}function Qo(e){for(;e===Xr;)Xr=fn[--dn],fn[dn]=null,Gr=fn[--dn],fn[dn]=null;for(;e===Dt;)Dt=ze[--Le],ze[Le]=null,nt=ze[--Le],ze[Le]=null,tt=ze[--Le],ze[Le]=null}var Ne=null,je=null,H=!1,Ue=null;function Bs(e,t){var n=Fe(5,null,null,0);n.elementType="DELETED",n.stateNode=t,n.return=e,t=e.deletions,t===null?(e.deletions=[n],e.flags|=16):t.push(n)}function Ws(e,t){switch(e.tag){case 5:var n=e.type;return t=t.nodeType!==1||n.toLowerCase()!==t.nodeName.toLowerCase()?null:t,t!==null?(e.stateNode=t,Ne=e,je=gt(t.firstChild),!0):!1;case 6:return t=e.pendingProps===""||t.nodeType!==3?null:t,t!==null?(e.stateNode=t,Ne=e,je=null,!0):!1;case 13:return t=t.nodeType!==8?null:t,t!==null?(n=Dt!==null?{id:tt,overflow:nt}:null,e.memoizedState={dehydrated:t,treeContext:n,retryLane:1073741824},n=Fe(18,null,null,0),n.stateNode=t,n.return=e,e.child=n,Ne=e,je=null,!0):!1;default:return!1}}function Ko(e){return(e.mode&1)!==0&&(e.flags&128)===0}function Yo(e){if(H){var t=je;if(t){var n=t;if(!Ws(e,t)){if(Ko(e))throw Error(w(418));t=gt(n.nextSibling);var r=Ne;t&&Ws(e,t)?Bs(r,n):(e.flags=e.flags&-4097|2,H=!1,Ne=e)}}else{if(Ko(e))throw Error(w(418));e.flags=e.flags&-4097|2,H=!1,Ne=e}}}function Hs(e){for(e=e.return;e!==null&&e.tag!==5&&e.tag!==3&&e.tag!==13;)e=e.return;Ne=e}function Zr(e){if(e!==Ne)return!1;if(!H)return Hs(e),H=!0,!1;var t;if((t=e.tag!==3)&&!(t=e.tag!==5)&&(t=e.type,t=t!=="head"&&t!=="body"&&!Ao(e.type,e.memoizedProps)),t&&(t=je)){if(Ko(e))throw Qs(),Error(w(418));for(;t;)Bs(e,t),t=gt(t.nextSibling)}if(Hs(e),e.tag===13){if(e=e.memoizedState,e=e!==null?e.dehydrated:null,!e)throw Error(w(317));e:{for(e=e.nextSibling,t=0;e;){if(e.nodeType===8){var n=e.data;if(n==="/$"){if(t===0){je=gt(e.nextSibling);break e}t--}else n!=="$"&&n!=="$!"&&n!=="$?"||t++}e=e.nextSibling}je=null}}else je=Ne?gt(e.stateNode.nextSibling):null;return!0}function Qs(){for(var e=je;e;)e=gt(e.nextSibling)}function pn(){je=Ne=null,H=!1}function Xo(e){Ue===null?Ue=[e]:Ue.push(e)}var Pd=qe.ReactCurrentBatchConfig;function qn(e,t,n){if(e=n.ref,e!==null&&typeof e!="function"&&typeof e!="object"){if(n._owner){if(n=n._owner,n){if(n.tag!==1)throw Error(w(309));var r=n.stateNode}if(!r)throw Error(w(147,e));var l=r,o=""+e;return t!==null&&t.ref!==null&&typeof t.ref=="function"&&t.ref._stringRef===o?t.ref:(t=function(i){var u=l.refs;i===null?delete u[o]:u[o]=i},t._stringRef=o,t)}if(typeof e!="string")throw Error(w(284));if(!n._owner)throw Error(w(290,e))}return e}function Jr(e,t){throw e=Object.prototype.toString.call(t),Error(w(31,e==="[object Object]"?"object with keys {"+Object.keys(t).join(", ")+"}":e))}function Ks(e){var t=e._init;return t(e._payload)}function Ys(e){function t(f,a){if(e){var d=f.deletions;d===null?(f.deletions=[a],f.flags|=16):d.push(a)}}function n(f,a){if(!e)return null;for(;a!==null;)t(f,a),a=a.sibling;return null}function r(f,a){for(f=new Map;a!==null;)a.key!==null?f.set(a.key,a):f.set(a.index,a),a=a.sibling;return f}function l(f,a){return f=Tt(f,a),f.index=0,f.sibling=null,f}function o(f,a,d){return f.index=d,e?(d=f.alternate,d!==null?(d=d.index,d<a?(f.flags|=2,a):d):(f.flags|=2,a)):(f.flags|=1048576,a)}function i(f){return e&&f.alternate===null&&(f.flags|=2),f}function u(f,a,d,g){return a===null||a.tag!==6?(a=Ui(d,f.mode,g),a.return=f,a):(a=l(a,d),a.return=f,a)}function s(f,a,d,g){var N=d.type;return N===Zt?y(f,a,d.props.children,g,d.key):a!==null&&(a.elementType===N||typeof N=="object"&&N!==null&&N.$$typeof===ct&&Ks(N)===a.type)?(g=l(a,d.props),g.ref=qn(f,a,d),g.return=f,g):(g=Sl(d.type,d.key,d.props,null,f.mode,g),g.ref=qn(f,a,d),g.return=f,g)}function c(f,a,d,g){return a===null||a.tag!==4||a.stateNode.containerInfo!==d.containerInfo||a.stateNode.implementation!==d.implementation?(a=$i(d,f.mode,g),a.return=f,a):(a=l(a,d.children||[]),a.return=f,a)}function y(f,a,d,g,N){return a===null||a.tag!==7?(a=Qt(d,f.mode,g,N),a.return=f,a):(a=l(a,d),a.return=f,a)}function v(f,a,d){if(typeof a=="string"&&a!==""||typeof a=="number")return a=Ui(""+a,f.mode,d),a.return=f,a;if(typeof a=="object"&&a!==null){switch(a.$$typeof){case gr:return d=Sl(a.type,a.key,a.props,null,f.mode,d),d.ref=qn(f,null,a),d.return=f,d;case Gt:return a=$i(a,f.mode,d),a.return=f,a;case ct:var g=a._init;return v(f,g(a._payload),d)}if(jn(a)||Cn(a))return a=Qt(a,f.mode,d,null),a.return=f,a;Jr(f,a)}return null}function m(f,a,d,g){var N=a!==null?a.key:null;if(typeof d=="string"&&d!==""||typeof d=="number")return N!==null?null:u(f,a,""+d,g);if(typeof d=="object"&&d!==null){switch(d.$$typeof){case gr:return d.key===N?s(f,a,d,g):null;case Gt:return d.key===N?c(f,a,d,g):null;case ct:return N=d._init,m(f,a,N(d._payload),g)}if(jn(d)||Cn(d))return N!==null?null:y(f,a,d,g,null);Jr(f,d)}return null}function k(f,a,d,g,N){if(typeof g=="string"&&g!==""||typeof g=="number")return f=f.get(d)||null,u(a,f,""+g,N);if(typeof g=="object"&&g!==null){switch(g.$$typeof){case gr:return f=f.get(g.key===null?d:g.key)||null,s(a,f,g,N);case Gt:return f=f.get(g.key===null?d:g.key)||null,c(a,f,g,N);case ct:var P=g._init;return k(f,a,d,P(g._payload),N)}if(jn(g)||Cn(g))return f=f.get(d)||null,y(a,f,g,N,null);Jr(a,g)}return null}function S(f,a,d,g){for(var N=null,P=null,T=a,L=a=0,Q=null;T!==null&&L<d.length;L++){T.index>L?(Q=T,T=null):Q=T.sibling;var F=m(f,T,d[L],g);if(F===null){T===null&&(T=Q);break}e&&T&&F.alternate===null&&t(f,T),a=o(F,a,L),P===null?N=F:P.sibling=F,P=F,T=Q}if(L===d.length)return n(f,T),H&&It(f,L),N;if(T===null){for(;L<d.length;L++)T=v(f,d[L],g),T!==null&&(a=o(T,a,L),P===null?N=T:P.sibling=T,P=T);return H&&It(f,L),N}for(T=r(f,T);L<d.length;L++)Q=k(T,f,L,d[L],g),Q!==null&&(e&&Q.alternate!==null&&T.delete(Q.key===null?L:Q.key),a=o(Q,a,L),P===null?N=Q:P.sibling=Q,P=Q);return e&&T.forEach(function(xe){return t(f,xe)}),H&&It(f,L),N}function C(f,a,d,g){var N=Cn(d);if(typeof N!="function")throw Error(w(150));if(d=N.call(d),d==null)throw Error(w(151));for(var P=N=null,T=a,L=a=0,Q=null,F=d.next();T!==null&&!F.done;L++,F=d.next()){T.index>L?(Q=T,T=null):Q=T.sibling;var xe=m(f,T,F.value,g);if(xe===null){T===null&&(T=Q);break}e&&T&&xe.alternate===null&&t(f,T),a=o(xe,a,L),P===null?N=xe:P.sibling=xe,P=xe,T=Q}if(F.done)return n(f,T),H&&It(f,L),N;if(T===null){for(;!F.done;L++,F=d.next())F=v(f,F.value,g),F!==null&&(a=o(F,a,L),P===null?N=F:P.sibling=F,P=F);return H&&It(f,L),N}for(T=r(f,T);!F.done;L++,F=d.next())F=k(T,f,L,F.value,g),F!==null&&(e&&F.alternate!==null&&T.delete(F.key===null?L:F.key),a=o(F,a,L),P===null?N=F:P.sibling=F,P=F);return e&&T.forEach(function(He){return t(f,He)}),H&&It(f,L),N}function A(f,a,d,g){if(typeof d=="object"&&d!==null&&d.type===Zt&&d.key===null&&(d=d.props.children),typeof d=="object"&&d!==null){switch(d.$$typeof){case gr:e:{for(var N=d.key,P=a;P!==null;){if(P.key===N){if(N=d.type,N===Zt){if(P.tag===7){n(f,P.sibling),a=l(P,d.props.children),a.return=f,f=a;break e}}else if(P.elementType===N||typeof N=="object"&&N!==null&&N.$$typeof===ct&&Ks(N)===P.type){n(f,P.sibling),a=l(P,d.props),a.ref=qn(f,P,d),a.return=f,f=a;break e}n(f,P);break}else t(f,P);P=P.sibling}d.type===Zt?(a=Qt(d.props.children,f.mode,g,d.key),a.return=f,f=a):(g=Sl(d.type,d.key,d.props,null,f.mode,g),g.ref=qn(f,a,d),g.return=f,f=g)}return i(f);case Gt:e:{for(P=d.key;a!==null;){if(a.key===P)if(a.tag===4&&a.stateNode.containerInfo===d.containerInfo&&a.stateNode.implementation===d.implementation){n(f,a.sibling),a=l(a,d.children||[]),a.return=f,f=a;break e}else{n(f,a);break}else t(f,a);a=a.sibling}a=$i(d,f.mode,g),a.return=f,f=a}return i(f);case ct:return P=d._init,A(f,a,P(d._payload),g)}if(jn(d))return S(f,a,d,g);if(Cn(d))return C(f,a,d,g);Jr(f,d)}return typeof d=="string"&&d!==""||typeof d=="number"?(d=""+d,a!==null&&a.tag===6?(n(f,a.sibling),a=l(a,d),a.return=f,f=a):(n(f,a),a=Ui(d,f.mode,g),a.return=f,f=a),i(f)):n(f,a)}return A}var mn=Ys(!0),Xs=Ys(!1),qr=wt(null),br=null,hn=null,Go=null;function Zo(){Go=hn=br=null}function Jo(e){var t=qr.current;B(qr),e._currentValue=t}function qo(e,t,n){for(;e!==null;){var r=e.alternate;if((e.childLanes&t)!==t?(e.childLanes|=t,r!==null&&(r.childLanes|=t)):r!==null&&(r.childLanes&t)!==t&&(r.childLanes|=t),e===n)break;e=e.return}}function vn(e,t){br=e,Go=hn=null,e=e.dependencies,e!==null&&e.firstContext!==null&&(e.lanes&t&&(we=!0),e.firstContext=null)}function Re(e){var t=e._currentValue;if(Go!==e)if(e={context:e,memoizedValue:t,next:null},hn===null){if(br===null)throw Error(w(308));hn=e,br.dependencies={lanes:0,firstContext:e}}else hn=hn.next=e;return t}var At=null;function bo(e){At===null?At=[e]:At.push(e)}function Gs(e,t,n,r){var l=t.interleaved;return l===null?(n.next=n,bo(t)):(n.next=l.next,l.next=n),t.interleaved=n,rt(e,r)}function rt(e,t){e.lanes|=t;var n=e.alternate;for(n!==null&&(n.lanes|=t),n=e,e=e.return;e!==null;)e.childLanes|=t,n=e.alternate,n!==null&&(n.childLanes|=t),n=e,e=e.return;return n.tag===3?n.stateNode:null}var xt=!1;function ei(e){e.updateQueue={baseState:e.memoizedState,firstBaseUpdate:null,lastBaseUpdate:null,shared:{pending:null,interleaved:null,lanes:0},effects:null}}function Zs(e,t){e=e.updateQueue,t.updateQueue===e&&(t.updateQueue={baseState:e.baseState,firstBaseUpdate:e.firstBaseUpdate,lastBaseUpdate:e.lastBaseUpdate,shared:e.shared,effects:e.effects})}function lt(e,t){return{eventTime:e,lane:t,tag:0,payload:null,callback:null,next:null}}function _t(e,t,n){var r=e.updateQueue;if(r===null)return null;if(r=r.shared,I&2){var l=r.pending;return l===null?t.next=t:(t.next=l.next,l.next=t),r.pending=t,rt(e,n)}return l=r.interleaved,l===null?(t.next=t,bo(r)):(t.next=l.next,l.next=t),r.interleaved=t,rt(e,n)}function el(e,t,n){if(t=t.updateQueue,t!==null&&(t=t.shared,(n&4194240)!==0)){var r=t.lanes;r&=e.pendingLanes,n|=r,t.lanes=n,mo(e,n)}}function Js(e,t){var n=e.updateQueue,r=e.alternate;if(r!==null&&(r=r.updateQueue,n===r)){var l=null,o=null;if(n=n.firstBaseUpdate,n!==null){do{var i={eventTime:n.eventTime,lane:n.lane,tag:n.tag,payload:n.payload,callback:n.callback,next:null};o===null?l=o=i:o=o.next=i,n=n.next}while(n!==null);o===null?l=o=t:o=o.next=t}else l=o=t;n={baseState:r.baseState,firstBaseUpdate:l,lastBaseUpdate:o,shared:r.shared,effects:r.effects},e.updateQueue=n;return}e=n.lastBaseUpdate,e===null?n.firstBaseUpdate=t:e.next=t,n.lastBaseUpdate=t}function tl(e,t,n,r){var l=e.updateQueue;xt=!1;var o=l.firstBaseUpdate,i=l.lastBaseUpdate,u=l.shared.pending;if(u!==null){l.shared.pending=null;var s=u,c=s.next;s.next=null,i===null?o=c:i.next=c,i=s;var y=e.alternate;y!==null&&(y=y.updateQueue,u=y.lastBaseUpdate,u!==i&&(u===null?y.firstBaseUpdate=c:u.next=c,y.lastBaseUpdate=s))}if(o!==null){var v=l.baseState;i=0,y=c=s=null,u=o;do{var m=u.lane,k=u.eventTime;if((r&m)===m){y!==null&&(y=y.next={eventTime:k,lane:0,tag:u.tag,payload:u.payload,callback:u.callback,next:null});e:{var S=e,C=u;switch(m=t,k=n,C.tag){case 1:if(S=C.payload,typeof S=="function"){v=S.call(k,v,m);break e}v=S;break e;case 3:S.flags=S.flags&-65537|128;case 0:if(S=C.payload,m=typeof S=="function"?S.call(k,v,m):S,m==null)break e;v=K({},v,m);break e;case 2:xt=!0}}u.callback!==null&&u.lane!==0&&(e.flags|=64,m=l.effects,m===null?l.effects=[u]:m.push(u))}else k={eventTime:k,lane:m,tag:u.tag,payload:u.payload,callback:u.callback,next:null},y===null?(c=y=k,s=v):y=y.next=k,i|=m;if(u=u.next,u===null){if(u=l.shared.pending,u===null)break;m=u,u=m.next,m.next=null,l.lastBaseUpdate=m,l.shared.pending=null}}while(!0);if(y===null&&(s=v),l.baseState=s,l.firstBaseUpdate=c,l.lastBaseUpdate=y,t=l.shared.interleaved,t!==null){l=t;do i|=l.lane,l=l.next;while(l!==t)}else o===null&&(l.shared.lanes=0);Vt|=i,e.lanes=i,e.memoizedState=v}}function qs(e,t,n){if(e=t.effects,t.effects=null,e!==null)for(t=0;t<e.length;t++){var r=e[t],l=r.callback;if(l!==null){if(r.callback=null,r=n,typeof l!="function")throw Error(w(191,l));l.call(r)}}}var bn={},Ye=wt(bn),er=wt(bn),tr=wt(bn);function Ut(e){if(e===bn)throw Error(w(174));return e}function ti(e,t){switch($(tr,t),$(er,e),$(Ye,bn),e=t.nodeType,e){case 9:case 11:t=(t=t.documentElement)?t.namespaceURI:to(null,"");break;default:e=e===8?t.parentNode:t,t=e.namespaceURI||null,e=e.tagName,t=to(t,e)}B(Ye),$(Ye,t)}function yn(){B(Ye),B(er),B(tr)}function bs(e){Ut(tr.current);var t=Ut(Ye.current),n=to(t,e.type);t!==n&&($(er,e),$(Ye,n))}function ni(e){er.current===e&&(B(Ye),B(er))}var Y=wt(0);function nl(e){for(var t=e;t!==null;){if(t.tag===13){var n=t.memoizedState;if(n!==null&&(n=n.dehydrated,n===null||n.data==="$?"||n.data==="$!"))return t}else if(t.tag===19&&t.memoizedProps.revealOrder!==void 0){if(t.flags&128)return t}else if(t.child!==null){t.child.return=t,t=t.child;continue}if(t===e)break;for(;t.sibling===null;){if(t.return===null||t.return===e)return null;t=t.return}t.sibling.return=t.return,t=t.sibling}return null}var ri=[];function li(){for(var e=0;e<ri.length;e++)ri[e]._workInProgressVersionPrimary=null;ri.length=0}var rl=qe.ReactCurrentDispatcher,oi=qe.ReactCurrentBatchConfig,$t=0,X=null,te=null,re=null,ll=!1,nr=!1,rr=0,Td=0;function ae(){throw Error(w(321))}function ii(e,t){if(t===null)return!1;for(var n=0;n<t.length&&n<e.length;n++)if(!Ae(e[n],t[n]))return!1;return!0}function ui(e,t,n,r,l,o){if($t=o,X=t,t.memoizedState=null,t.updateQueue=null,t.lanes=0,rl.current=e===null||e.memoizedState===null?Od:Md,e=n(r,l),nr){o=0;do{if(nr=!1,rr=0,25<=o)throw Error(w(301));o+=1,re=te=null,t.updateQueue=null,rl.current=Fd,e=n(r,l)}while(nr)}if(rl.current=ul,t=te!==null&&te.next!==null,$t=0,re=te=X=null,ll=!1,t)throw Error(w(300));return e}function si(){var e=rr!==0;return rr=0,e}function Xe(){var e={memoizedState:null,baseState:null,baseQueue:null,queue:null,next:null};return re===null?X.memoizedState=re=e:re=re.next=e,re}function Oe(){if(te===null){var e=X.alternate;e=e!==null?e.memoizedState:null}else e=te.next;var t=re===null?X.memoizedState:re.next;if(t!==null)re=t,te=e;else{if(e===null)throw Error(w(310));te=e,e={memoizedState:te.memoizedState,baseState:te.baseState,baseQueue:te.baseQueue,queue:te.queue,next:null},re===null?X.memoizedState=re=e:re=re.next=e}return re}function lr(e,t){return typeof t=="function"?t(e):t}function ai(e){var t=Oe(),n=t.queue;if(n===null)throw Error(w(311));n.lastRenderedReducer=e;var r=te,l=r.baseQueue,o=n.pending;if(o!==null){if(l!==null){var i=l.next;l.next=o.next,o.next=i}r.baseQueue=l=o,n.pending=null}if(l!==null){o=l.next,r=r.baseState;var u=i=null,s=null,c=o;do{var y=c.lane;if(($t&y)===y)s!==null&&(s=s.next={lane:0,action:c.action,hasEagerState:c.hasEagerState,eagerState:c.eagerState,next:null}),r=c.hasEagerState?c.eagerState:e(r,c.action);else{var v={lane:y,action:c.action,hasEagerState:c.hasEagerState,eagerState:c.eagerState,next:null};s===null?(u=s=v,i=r):s=s.next=v,X.lanes|=y,Vt|=y}c=c.next}while(c!==null&&c!==o);s===null?i=r:s.next=u,Ae(r,t.memoizedState)||(we=!0),t.memoizedState=r,t.baseState=i,t.baseQueue=s,n.lastRenderedState=r}if(e=n.interleaved,e!==null){l=e;do o=l.lane,X.lanes|=o,Vt|=o,l=l.next;while(l!==e)}else l===null&&(n.lanes=0);return[t.memoizedState,n.dispatch]}function ci(e){var t=Oe(),n=t.queue;if(n===null)throw Error(w(311));n.lastRenderedReducer=e;var r=n.dispatch,l=n.pending,o=t.memoizedState;if(l!==null){n.pending=null;var i=l=l.next;do o=e(o,i.action),i=i.next;while(i!==l);Ae(o,t.memoizedState)||(we=!0),t.memoizedState=o,t.baseQueue===null&&(t.baseState=o),n.lastRenderedState=o}return[o,r]}function ea(){}function ta(e,t){var n=X,r=Oe(),l=t(),o=!Ae(r.memoizedState,l);if(o&&(r.memoizedState=l,we=!0),r=r.queue,fi(la.bind(null,n,r,e),[e]),r.getSnapshot!==t||o||re!==null&&re.memoizedState.tag&1){if(n.flags|=2048,or(9,ra.bind(null,n,r,l,t),void 0,null),le===null)throw Error(w(349));$t&30||na(n,t,l)}return l}function na(e,t,n){e.flags|=16384,e={getSnapshot:t,value:n},t=X.updateQueue,t===null?(t={lastEffect:null,stores:null},X.updateQueue=t,t.stores=[e]):(n=t.stores,n===null?t.stores=[e]:n.push(e))}function ra(e,t,n,r){t.value=n,t.getSnapshot=r,oa(t)&&ia(e)}function la(e,t,n){return n(function(){oa(t)&&ia(e)})}function oa(e){var t=e.getSnapshot;e=e.value;try{var n=t();return!Ae(e,n)}catch{return!0}}function ia(e){var t=rt(e,1);t!==null&&We(t,e,1,-1)}function ua(e){var t=Xe();return typeof e=="function"&&(e=e()),t.memoizedState=t.baseState=e,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:lr,lastRenderedState:e},t.queue=e,e=e.dispatch=Rd.bind(null,X,e),[t.memoizedState,e]}function or(e,t,n,r){return e={tag:e,create:t,destroy:n,deps:r,next:null},t=X.updateQueue,t===null?(t={lastEffect:null,stores:null},X.updateQueue=t,t.lastEffect=e.next=e):(n=t.lastEffect,n===null?t.lastEffect=e.next=e:(r=n.next,n.next=e,e.next=r,t.lastEffect=e)),e}function sa(){return Oe().memoizedState}function ol(e,t,n,r){var l=Xe();X.flags|=e,l.memoizedState=or(1|t,n,void 0,r===void 0?null:r)}function il(e,t,n,r){var l=Oe();r=r===void 0?null:r;var o=void 0;if(te!==null){var i=te.memoizedState;if(o=i.destroy,r!==null&&ii(r,i.deps)){l.memoizedState=or(t,n,o,r);return}}X.flags|=e,l.memoizedState=or(1|t,n,o,r)}function aa(e,t){return ol(8390656,8,e,t)}function fi(e,t){return il(2048,8,e,t)}function ca(e,t){return il(4,2,e,t)}function fa(e,t){return il(4,4,e,t)}function da(e,t){if(typeof t=="function")return e=e(),t(e),function(){t(null)};if(t!=null)return e=e(),t.current=e,function(){t.current=null}}function pa(e,t,n){return n=n!=null?n.concat([e]):null,il(4,4,da.bind(null,t,e),n)}function di(){}function ma(e,t){var n=Oe();t=t===void 0?null:t;var r=n.memoizedState;return r!==null&&t!==null&&ii(t,r[1])?r[0]:(n.memoizedState=[e,t],e)}function ha(e,t){var n=Oe();t=t===void 0?null:t;var r=n.memoizedState;return r!==null&&t!==null&&ii(t,r[1])?r[0]:(e=e(),n.memoizedState=[e,t],e)}function va(e,t,n){return $t&21?(Ae(n,t)||(n=Ku(),X.lanes|=n,Vt|=n,e.baseState=!0),t):(e.baseState&&(e.baseState=!1,we=!0),e.memoizedState=n)}function zd(e,t){var n=U;U=n!==0&&4>n?n:4,e(!0);var r=oi.transition;oi.transition={};try{e(!1),t()}finally{U=n,oi.transition=r}}function ya(){return Oe().memoizedState}function Ld(e,t,n){var r=jt(e);if(n={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null},ga(e))wa(t,n);else if(n=Gs(e,t,n,r),n!==null){var l=ve();We(n,e,r,l),ka(n,t,r)}}function Rd(e,t,n){var r=jt(e),l={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null};if(ga(e))wa(t,l);else{var o=e.alternate;if(e.lanes===0&&(o===null||o.lanes===0)&&(o=t.lastRenderedReducer,o!==null))try{var i=t.lastRenderedState,u=o(i,n);if(l.hasEagerState=!0,l.eagerState=u,Ae(u,i)){var s=t.interleaved;s===null?(l.next=l,bo(t)):(l.next=s.next,s.next=l),t.interleaved=l;return}}catch{}finally{}n=Gs(e,t,l,r),n!==null&&(l=ve(),We(n,e,r,l),ka(n,t,r))}}function ga(e){var t=e.alternate;return e===X||t!==null&&t===X}function wa(e,t){nr=ll=!0;var n=e.pending;n===null?t.next=t:(t.next=n.next,n.next=t),e.pending=t}function ka(e,t,n){if(n&4194240){var r=t.lanes;r&=e.pendingLanes,n|=r,t.lanes=n,mo(e,n)}}var ul={readContext:Re,useCallback:ae,useContext:ae,useEffect:ae,useImperativeHandle:ae,useInsertionEffect:ae,useLayoutEffect:ae,useMemo:ae,useReducer:ae,useRef:ae,useState:ae,useDebugValue:ae,useDeferredValue:ae,useTransition:ae,useMutableSource:ae,useSyncExternalStore:ae,useId:ae,unstable_isNewReconciler:!1},Od={readContext:Re,useCallback:function(e,t){return Xe().memoizedState=[e,t===void 0?null:t],e},useContext:Re,useEffect:aa,useImperativeHandle:function(e,t,n){return n=n!=null?n.concat([e]):null,ol(4194308,4,da.bind(null,t,e),n)},useLayoutEffect:function(e,t){return ol(4194308,4,e,t)},useInsertionEffect:function(e,t){return ol(4,2,e,t)},useMemo:function(e,t){var n=Xe();return t=t===void 0?null:t,e=e(),n.memoizedState=[e,t],e},useReducer:function(e,t,n){var r=Xe();return t=n!==void 0?n(t):t,r.memoizedState=r.baseState=t,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:e,lastRenderedState:t},r.queue=e,e=e.dispatch=Ld.bind(null,X,e),[r.memoizedState,e]},useRef:function(e){var t=Xe();return e={current:e},t.memoizedState=e},useState:ua,useDebugValue:di,useDeferredValue:function(e){return Xe().memoizedState=e},useTransition:function(){var e=ua(!1),t=e[0];return e=zd.bind(null,e[1]),Xe().memoizedState=e,[t,e]},useMutableSource:function(){},useSyncExternalStore:function(e,t,n){var r=X,l=Xe();if(H){if(n===void 0)throw Error(w(407));n=n()}else{if(n=t(),le===null)throw Error(w(349));$t&30||na(r,t,n)}l.memoizedState=n;var o={value:n,getSnapshot:t};return l.queue=o,aa(la.bind(null,r,o,e),[e]),r.flags|=2048,or(9,ra.bind(null,r,o,n,t),void 0,null),n},useId:function(){var e=Xe(),t=le.identifierPrefix;if(H){var n=nt,r=tt;n=(r&~(1<<32-Ie(r)-1)).toString(32)+n,t=":"+t+"R"+n,n=rr++,0<n&&(t+="H"+n.toString(32)),t+=":"}else n=Td++,t=":"+t+"r"+n.toString(32)+":";return e.memoizedState=t},unstable_isNewReconciler:!1},Md={readContext:Re,useCallback:ma,useContext:Re,useEffect:fi,useImperativeHandle:pa,useInsertionEffect:ca,useLayoutEffect:fa,useMemo:ha,useReducer:ai,useRef:sa,useState:function(){return ai(lr)},useDebugValue:di,useDeferredValue:function(e){var t=Oe();return va(t,te.memoizedState,e)},useTransition:function(){var e=ai(lr)[0],t=Oe().memoizedState;return[e,t]},useMutableSource:ea,useSyncExternalStore:ta,useId:ya,unstable_isNewReconciler:!1},Fd={readContext:Re,useCallback:ma,useContext:Re,useEffect:fi,useImperativeHandle:pa,useInsertionEffect:ca,useLayoutEffect:fa,useMemo:ha,useReducer:ci,useRef:sa,useState:function(){return ci(lr)},useDebugValue:di,useDeferredValue:function(e){var t=Oe();return te===null?t.memoizedState=e:va(t,te.memoizedState,e)},useTransition:function(){var e=ci(lr)[0],t=Oe().memoizedState;return[e,t]},useMutableSource:ea,useSyncExternalStore:ta,useId:ya,unstable_isNewReconciler:!1};function $e(e,t){if(e&&e.defaultProps){t=K({},t),e=e.defaultProps;for(var n in e)t[n]===void 0&&(t[n]=e[n]);return t}return t}function pi(e,t,n,r){t=e.memoizedState,n=n(r,t),n=n==null?t:K({},t,n),e.memoizedState=n,e.lanes===0&&(e.updateQueue.baseState=n)}var sl={isMounted:function(e){return(e=e._reactInternals)?Ot(e)===e:!1},enqueueSetState:function(e,t,n){e=e._reactInternals;var r=ve(),l=jt(e),o=lt(r,l);o.payload=t,n!=null&&(o.callback=n),t=_t(e,o,l),t!==null&&(We(t,e,l,r),el(t,e,l))},enqueueReplaceState:function(e,t,n){e=e._reactInternals;var r=ve(),l=jt(e),o=lt(r,l);o.tag=1,o.payload=t,n!=null&&(o.callback=n),t=_t(e,o,l),t!==null&&(We(t,e,l,r),el(t,e,l))},enqueueForceUpdate:function(e,t){e=e._reactInternals;var n=ve(),r=jt(e),l=lt(n,r);l.tag=2,t!=null&&(l.callback=t),t=_t(e,l,r),t!==null&&(We(t,e,r,n),el(t,e,r))}};function Sa(e,t,n,r,l,o,i){return e=e.stateNode,typeof e.shouldComponentUpdate=="function"?e.shouldComponentUpdate(r,o,i):t.prototype&&t.prototype.isPureReactComponent?!Qn(n,r)||!Qn(l,o):!0}function xa(e,t,n){var r=!1,l=kt,o=t.contextType;return typeof o=="object"&&o!==null?o=Re(o):(l=ge(t)?Ft:se.current,r=t.contextTypes,o=(r=r!=null)?cn(e,l):kt),t=new t(n,o),e.memoizedState=t.state!==null&&t.state!==void 0?t.state:null,t.updater=sl,e.stateNode=t,t._reactInternals=e,r&&(e=e.stateNode,e.__reactInternalMemoizedUnmaskedChildContext=l,e.__reactInternalMemoizedMaskedChildContext=o),t}function _a(e,t,n,r){e=t.state,typeof t.componentWillReceiveProps=="function"&&t.componentWillReceiveProps(n,r),typeof t.UNSAFE_componentWillReceiveProps=="function"&&t.UNSAFE_componentWillReceiveProps(n,r),t.state!==e&&sl.enqueueReplaceState(t,t.state,null)}function mi(e,t,n,r){var l=e.stateNode;l.props=n,l.state=e.memoizedState,l.refs={},ei(e);var o=t.contextType;typeof o=="object"&&o!==null?l.context=Re(o):(o=ge(t)?Ft:se.current,l.context=cn(e,o)),l.state=e.memoizedState,o=t.getDerivedStateFromProps,typeof o=="function"&&(pi(e,t,o,n),l.state=e.memoizedState),typeof t.getDerivedStateFromProps=="function"||typeof l.getSnapshotBeforeUpdate=="function"||typeof l.UNSAFE_componentWillMount!="function"&&typeof l.componentWillMount!="function"||(t=l.state,typeof l.componentWillMount=="function"&&l.componentWillMount(),typeof l.UNSAFE_componentWillMount=="function"&&l.UNSAFE_componentWillMount(),t!==l.state&&sl.enqueueReplaceState(l,l.state,null),tl(e,n,l,r),l.state=e.memoizedState),typeof l.componentDidMount=="function"&&(e.flags|=4194308)}function gn(e,t){try{var n="",r=t;do n+=af(r),r=r.return;while(r);var l=n}catch(o){l=`
+`+l[i].replace(" at new "," at ");return e.displayName&&s.includes("<anonymous>")&&(s=s.replace("<anonymous>",e.displayName)),s}while(1<=i&&0<=u);break}}}finally{Xl=!1,Error.prepareStackTrace=n}return(e=e?e.displayName||e.name:"")?Nn(e):""}function af(e){switch(e.tag){case 5:return Nn(e.type);case 16:return Nn("Lazy");case 13:return Nn("Suspense");case 19:return Nn("SuspenseList");case 0:case 2:case 15:return e=Gl(e.type,!1),e;case 11:return e=Gl(e.type.render,!1),e;case 1:return e=Gl(e.type,!0),e;default:return""}}function Zl(e){if(e==null)return null;if(typeof e=="function")return e.displayName||e.name||null;if(typeof e=="string")return e;switch(e){case Zt:return"Fragment";case Gt:return"Portal";case Bl:return"Profiler";case Vl:return"StrictMode";case Hl:return"Suspense";case Ql:return"SuspenseList"}if(typeof e=="object")switch(e.$$typeof){case yu:return(e.displayName||"Context")+".Consumer";case vu:return(e._context.displayName||"Context")+".Provider";case Wl:var t=e.render;return e=e.displayName,e||(e=t.displayName||t.name||"",e=e!==""?"ForwardRef("+e+")":"ForwardRef"),e;case Kl:return t=e.displayName||null,t!==null?t:Zl(e.type)||"Memo";case ct:t=e._payload,e=e._init;try{return Zl(e(t))}catch{}}return null}function cf(e){var t=e.type;switch(e.tag){case 24:return"Cache";case 9:return(t.displayName||"Context")+".Consumer";case 10:return(t._context.displayName||"Context")+".Provider";case 18:return"DehydratedFragment";case 11:return e=t.render,e=e.displayName||e.name||"",t.displayName||(e!==""?"ForwardRef("+e+")":"ForwardRef");case 7:return"Fragment";case 5:return t;case 4:return"Portal";case 3:return"Root";case 6:return"Text";case 16:return Zl(t);case 8:return t===Vl?"StrictMode":"Mode";case 22:return"Offscreen";case 12:return"Profiler";case 21:return"Scope";case 13:return"Suspense";case 19:return"SuspenseList";case 25:return"TracingMarker";case 1:case 0:case 17:case 2:case 14:case 15:if(typeof t=="function")return t.displayName||t.name||null;if(typeof t=="string")return t}return null}function ft(e){switch(typeof e){case"boolean":case"number":case"string":case"undefined":return e;case"object":return e;default:return""}}function wu(e){var t=e.type;return(e=e.nodeName)&&e.toLowerCase()==="input"&&(t==="checkbox"||t==="radio")}function ff(e){var t=wu(e)?"checked":"value",n=Object.getOwnPropertyDescriptor(e.constructor.prototype,t),r=""+e[t];if(!e.hasOwnProperty(t)&&typeof n<"u"&&typeof n.get=="function"&&typeof n.set=="function"){var l=n.get,o=n.set;return Object.defineProperty(e,t,{configurable:!0,get:function(){return l.call(this)},set:function(i){r=""+i,o.call(this,i)}}),Object.defineProperty(e,t,{enumerable:n.enumerable}),{getValue:function(){return r},setValue:function(i){r=""+i},stopTracking:function(){e._valueTracker=null,delete e[t]}}}}function kr(e){e._valueTracker||(e._valueTracker=ff(e))}function Su(e){if(!e)return!1;var t=e._valueTracker;if(!t)return!0;var n=t.getValue(),r="";return e&&(r=wu(e)?e.checked?"true":"false":e.value),e=r,e!==n?(t.setValue(e),!0):!1}function wr(e){if(e=e||(typeof document<"u"?document:void 0),typeof e>"u")return null;try{return e.activeElement||e.body}catch{return e.body}}function Jl(e,t){var n=t.checked;return K({},t,{defaultChecked:void 0,defaultValue:void 0,value:void 0,checked:n??e._wrapperState.initialChecked})}function xu(e,t){var n=t.defaultValue==null?"":t.defaultValue,r=t.checked!=null?t.checked:t.defaultChecked;n=ft(t.value!=null?t.value:n),e._wrapperState={initialChecked:r,initialValue:n,controlled:t.type==="checkbox"||t.type==="radio"?t.checked!=null:t.value!=null}}function _u(e,t){t=t.checked,t!=null&&$l(e,"checked",t,!1)}function ql(e,t){_u(e,t);var n=ft(t.value),r=t.type;if(n!=null)r==="number"?(n===0&&e.value===""||e.value!=n)&&(e.value=""+n):e.value!==""+n&&(e.value=""+n);else if(r==="submit"||r==="reset"){e.removeAttribute("value");return}t.hasOwnProperty("value")?bl(e,t.type,n):t.hasOwnProperty("defaultValue")&&bl(e,t.type,ft(t.defaultValue)),t.checked==null&&t.defaultChecked!=null&&(e.defaultChecked=!!t.defaultChecked)}function Eu(e,t,n){if(t.hasOwnProperty("value")||t.hasOwnProperty("defaultValue")){var r=t.type;if(!(r!=="submit"&&r!=="reset"||t.value!==void 0&&t.value!==null))return;t=""+e._wrapperState.initialValue,n||t===e.value||(e.value=t),e.defaultValue=t}n=e.name,n!==""&&(e.name=""),e.defaultChecked=!!e._wrapperState.initialChecked,n!==""&&(e.name=n)}function bl(e,t,n){(t!=="number"||wr(e.ownerDocument)!==e)&&(n==null?e.defaultValue=""+e._wrapperState.initialValue:e.defaultValue!==""+n&&(e.defaultValue=""+n))}var jn=Array.isArray;function Jt(e,t,n,r){if(e=e.options,t){t={};for(var l=0;l<n.length;l++)t["$"+n[l]]=!0;for(n=0;n<e.length;n++)l=t.hasOwnProperty("$"+e[n].value),e[n].selected!==l&&(e[n].selected=l),l&&r&&(e[n].defaultSelected=!0)}else{for(n=""+ft(n),t=null,l=0;l<e.length;l++){if(e[l].value===n){e[l].selected=!0,r&&(e[l].defaultSelected=!0);return}t!==null||e[l].disabled||(t=e[l])}t!==null&&(t.selected=!0)}}function eo(e,t){if(t.dangerouslySetInnerHTML!=null)throw Error(k(91));return K({},t,{value:void 0,defaultValue:void 0,children:""+e._wrapperState.initialValue})}function Cu(e,t){var n=t.value;if(n==null){if(n=t.children,t=t.defaultValue,n!=null){if(t!=null)throw Error(k(92));if(jn(n)){if(1<n.length)throw Error(k(93));n=n[0]}t=n}t==null&&(t=""),n=t}e._wrapperState={initialValue:ft(n)}}function Nu(e,t){var n=ft(t.value),r=ft(t.defaultValue);n!=null&&(n=""+n,n!==e.value&&(e.value=n),t.defaultValue==null&&e.defaultValue!==n&&(e.defaultValue=n)),r!=null&&(e.defaultValue=""+r)}function ju(e){var t=e.textContent;t===e._wrapperState.initialValue&&t!==""&&t!==null&&(e.value=t)}function Pu(e){switch(e){case"svg":return"http://www.w3.org/2000/svg";case"math":return"http://www.w3.org/1998/Math/MathML";default:return"http://www.w3.org/1999/xhtml"}}function to(e,t){return e==null||e==="http://www.w3.org/1999/xhtml"?Pu(t):e==="http://www.w3.org/2000/svg"&&t==="foreignObject"?"http://www.w3.org/1999/xhtml":e}var Sr,Tu=function(e){return typeof MSApp<"u"&&MSApp.execUnsafeLocalFunction?function(t,n,r,l){MSApp.execUnsafeLocalFunction(function(){return e(t,n,r,l)})}:e}(function(e,t){if(e.namespaceURI!=="http://www.w3.org/2000/svg"||"innerHTML"in e)e.innerHTML=t;else{for(Sr=Sr||document.createElement("div"),Sr.innerHTML="<svg>"+t.valueOf().toString()+"</svg>",t=Sr.firstChild;e.firstChild;)e.removeChild(e.firstChild);for(;t.firstChild;)e.appendChild(t.firstChild)}});function Pn(e,t){if(t){var n=e.firstChild;if(n&&n===e.lastChild&&n.nodeType===3){n.nodeValue=t;return}}e.textContent=t}var Tn={animationIterationCount:!0,aspectRatio:!0,borderImageOutset:!0,borderImageSlice:!0,borderImageWidth:!0,boxFlex:!0,boxFlexGroup:!0,boxOrdinalGroup:!0,columnCount:!0,columns:!0,flex:!0,flexGrow:!0,flexPositive:!0,flexShrink:!0,flexNegative:!0,flexOrder:!0,gridArea:!0,gridRow:!0,gridRowEnd:!0,gridRowSpan:!0,gridRowStart:!0,gridColumn:!0,gridColumnEnd:!0,gridColumnSpan:!0,gridColumnStart:!0,fontWeight:!0,lineClamp:!0,lineHeight:!0,opacity:!0,order:!0,orphans:!0,tabSize:!0,widows:!0,zIndex:!0,zoom:!0,fillOpacity:!0,floodOpacity:!0,stopOpacity:!0,strokeDasharray:!0,strokeDashoffset:!0,strokeMiterlimit:!0,strokeOpacity:!0,strokeWidth:!0},df=["Webkit","ms","Moz","O"];Object.keys(Tn).forEach(function(e){df.forEach(function(t){t=t+e.charAt(0).toUpperCase()+e.substring(1),Tn[t]=Tn[e]})});function zu(e,t,n){return t==null||typeof t=="boolean"||t===""?"":n||typeof t!="number"||t===0||Tn.hasOwnProperty(e)&&Tn[e]?(""+t).trim():t+"px"}function Lu(e,t){e=e.style;for(var n in t)if(t.hasOwnProperty(n)){var r=n.indexOf("--")===0,l=zu(n,t[n],r);n==="float"&&(n="cssFloat"),r?e.setProperty(n,l):e[n]=l}}var pf=K({menuitem:!0},{area:!0,base:!0,br:!0,col:!0,embed:!0,hr:!0,img:!0,input:!0,keygen:!0,link:!0,meta:!0,param:!0,source:!0,track:!0,wbr:!0});function no(e,t){if(t){if(pf[e]&&(t.children!=null||t.dangerouslySetInnerHTML!=null))throw Error(k(137,e));if(t.dangerouslySetInnerHTML!=null){if(t.children!=null)throw Error(k(60));if(typeof t.dangerouslySetInnerHTML!="object"||!("__html"in t.dangerouslySetInnerHTML))throw Error(k(61))}if(t.style!=null&&typeof t.style!="object")throw Error(k(62))}}function ro(e,t){if(e.indexOf("-")===-1)return typeof t.is=="string";switch(e){case"annotation-xml":case"color-profile":case"font-face":case"font-face-src":case"font-face-uri":case"font-face-format":case"font-face-name":case"missing-glyph":return!1;default:return!0}}var lo=null;function oo(e){return e=e.target||e.srcElement||window,e.correspondingUseElement&&(e=e.correspondingUseElement),e.nodeType===3?e.parentNode:e}var io=null,qt=null,bt=null;function Ru(e){if(e=Jn(e)){if(typeof io!="function")throw Error(k(280));var t=e.stateNode;t&&(t=Hr(t),io(e.stateNode,e.type,t))}}function Mu(e){qt?bt?bt.push(e):bt=[e]:qt=e}function Ou(){if(qt){var e=qt,t=bt;if(bt=qt=null,Ru(e),t)for(e=0;e<t.length;e++)Ru(t[e])}}function Fu(e,t){return e(t)}function Du(){}var uo=!1;function Iu(e,t,n){if(uo)return e(t,n);uo=!0;try{return Fu(e,t,n)}finally{uo=!1,(qt!==null||bt!==null)&&(Du(),Ou())}}function zn(e,t){var n=e.stateNode;if(n===null)return null;var r=Hr(n);if(r===null)return null;n=r[t];e:switch(t){case"onClick":case"onClickCapture":case"onDoubleClick":case"onDoubleClickCapture":case"onMouseDown":case"onMouseDownCapture":case"onMouseMove":case"onMouseMoveCapture":case"onMouseUp":case"onMouseUpCapture":case"onMouseEnter":(r=!r.disabled)||(e=e.type,r=!(e==="button"||e==="input"||e==="select"||e==="textarea")),e=!r;break e;default:e=!1}if(e)return null;if(n&&typeof n!="function")throw Error(k(231,t,typeof n));return n}var so=!1;if(Ze)try{var Ln={};Object.defineProperty(Ln,"passive",{get:function(){so=!0}}),window.addEventListener("test",Ln,Ln),window.removeEventListener("test",Ln,Ln)}catch{so=!1}function mf(e,t,n,r,l,o,i,u,s){var c=Array.prototype.slice.call(arguments,3);try{t.apply(n,c)}catch(y){this.onError(y)}}var Rn=!1,xr=null,_r=!1,ao=null,hf={onError:function(e){Rn=!0,xr=e}};function vf(e,t,n,r,l,o,i,u,s){Rn=!1,xr=null,mf.apply(hf,arguments)}function yf(e,t,n,r,l,o,i,u,s){if(vf.apply(this,arguments),Rn){if(Rn){var c=xr;Rn=!1,xr=null}else throw Error(k(198));_r||(_r=!0,ao=c)}}function Mt(e){var t=e,n=e;if(e.alternate)for(;t.return;)t=t.return;else{e=t;do t=e,t.flags&4098&&(n=t.return),e=t.return;while(e)}return t.tag===3?n:null}function Au(e){if(e.tag===13){var t=e.memoizedState;if(t===null&&(e=e.alternate,e!==null&&(t=e.memoizedState)),t!==null)return t.dehydrated}return null}function Uu(e){if(Mt(e)!==e)throw Error(k(188))}function gf(e){var t=e.alternate;if(!t){if(t=Mt(e),t===null)throw Error(k(188));return t!==e?null:e}for(var n=e,r=t;;){var l=n.return;if(l===null)break;var o=l.alternate;if(o===null){if(r=l.return,r!==null){n=r;continue}break}if(l.child===o.child){for(o=l.child;o;){if(o===n)return Uu(l),e;if(o===r)return Uu(l),t;o=o.sibling}throw Error(k(188))}if(n.return!==r.return)n=l,r=o;else{for(var i=!1,u=l.child;u;){if(u===n){i=!0,n=l,r=o;break}if(u===r){i=!0,r=l,n=o;break}u=u.sibling}if(!i){for(u=o.child;u;){if(u===n){i=!0,n=o,r=l;break}if(u===r){i=!0,r=o,n=l;break}u=u.sibling}if(!i)throw Error(k(189))}}if(n.alternate!==r)throw Error(k(190))}if(n.tag!==3)throw Error(k(188));return n.stateNode.current===n?e:t}function $u(e){return e=gf(e),e!==null?Vu(e):null}function Vu(e){if(e.tag===5||e.tag===6)return e;for(e=e.child;e!==null;){var t=Vu(e);if(t!==null)return t;e=e.sibling}return null}var Bu=_e.unstable_scheduleCallback,Wu=_e.unstable_cancelCallback,kf=_e.unstable_shouldYield,wf=_e.unstable_requestPaint,J=_e.unstable_now,Sf=_e.unstable_getCurrentPriorityLevel,co=_e.unstable_ImmediatePriority,Hu=_e.unstable_UserBlockingPriority,Er=_e.unstable_NormalPriority,xf=_e.unstable_LowPriority,Qu=_e.unstable_IdlePriority,Cr=null,Qe=null;function _f(e){if(Qe&&typeof Qe.onCommitFiberRoot=="function")try{Qe.onCommitFiberRoot(Cr,e,void 0,(e.current.flags&128)===128)}catch{}}var Ie=Math.clz32?Math.clz32:Nf,Ef=Math.log,Cf=Math.LN2;function Nf(e){return e>>>=0,e===0?32:31-(Ef(e)/Cf|0)|0}var Nr=64,jr=4194304;function Mn(e){switch(e&-e){case 1:return 1;case 2:return 2;case 4:return 4;case 8:return 8;case 16:return 16;case 32:return 32;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return e&4194240;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return e&130023424;case 134217728:return 134217728;case 268435456:return 268435456;case 536870912:return 536870912;case 1073741824:return 1073741824;default:return e}}function Pr(e,t){var n=e.pendingLanes;if(n===0)return 0;var r=0,l=e.suspendedLanes,o=e.pingedLanes,i=n&268435455;if(i!==0){var u=i&~l;u!==0?r=Mn(u):(o&=i,o!==0&&(r=Mn(o)))}else i=n&~l,i!==0?r=Mn(i):o!==0&&(r=Mn(o));if(r===0)return 0;if(t!==0&&t!==r&&!(t&l)&&(l=r&-r,o=t&-t,l>=o||l===16&&(o&4194240)!==0))return t;if(r&4&&(r|=n&16),t=e.entangledLanes,t!==0)for(e=e.entanglements,t&=r;0<t;)n=31-Ie(t),l=1<<n,r|=e[n],t&=~l;return r}function jf(e,t){switch(e){case 1:case 2:case 4:return t+250;case 8:case 16:case 32:case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return t+5e3;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return-1;case 134217728:case 268435456:case 536870912:case 1073741824:return-1;default:return-1}}function Pf(e,t){for(var n=e.suspendedLanes,r=e.pingedLanes,l=e.expirationTimes,o=e.pendingLanes;0<o;){var i=31-Ie(o),u=1<<i,s=l[i];s===-1?(!(u&n)||u&r)&&(l[i]=jf(u,t)):s<=t&&(e.expiredLanes|=u),o&=~u}}function fo(e){return e=e.pendingLanes&-1073741825,e!==0?e:e&1073741824?1073741824:0}function Ku(){var e=Nr;return Nr<<=1,!(Nr&4194240)&&(Nr=64),e}function po(e){for(var t=[],n=0;31>n;n++)t.push(e);return t}function On(e,t,n){e.pendingLanes|=t,t!==536870912&&(e.suspendedLanes=0,e.pingedLanes=0),e=e.eventTimes,t=31-Ie(t),e[t]=n}function Tf(e,t){var n=e.pendingLanes&~t;e.pendingLanes=t,e.suspendedLanes=0,e.pingedLanes=0,e.expiredLanes&=t,e.mutableReadLanes&=t,e.entangledLanes&=t,t=e.entanglements;var r=e.eventTimes;for(e=e.expirationTimes;0<n;){var l=31-Ie(n),o=1<<l;t[l]=0,r[l]=-1,e[l]=-1,n&=~o}}function mo(e,t){var n=e.entangledLanes|=t;for(e=e.entanglements;n;){var r=31-Ie(n),l=1<<r;l&t|e[r]&t&&(e[r]|=t),n&=~l}}var A=0;function Yu(e){return e&=-e,1<e?4<e?e&268435455?16:536870912:4:1}var Xu,ho,Gu,Zu,Ju,vo=!1,Tr=[],dt=null,pt=null,mt=null,Fn=new Map,Dn=new Map,ht=[],zf="mousedown mouseup touchcancel touchend touchstart auxclick dblclick pointercancel pointerdown pointerup dragend dragstart drop compositionend compositionstart keydown keypress keyup input textInput copy cut paste click change contextmenu reset submit".split(" ");function qu(e,t){switch(e){case"focusin":case"focusout":dt=null;break;case"dragenter":case"dragleave":pt=null;break;case"mouseover":case"mouseout":mt=null;break;case"pointerover":case"pointerout":Fn.delete(t.pointerId);break;case"gotpointercapture":case"lostpointercapture":Dn.delete(t.pointerId)}}function In(e,t,n,r,l,o){return e===null||e.nativeEvent!==o?(e={blockedOn:t,domEventName:n,eventSystemFlags:r,nativeEvent:o,targetContainers:[l]},t!==null&&(t=Jn(t),t!==null&&ho(t)),e):(e.eventSystemFlags|=r,t=e.targetContainers,l!==null&&t.indexOf(l)===-1&&t.push(l),e)}function Lf(e,t,n,r,l){switch(t){case"focusin":return dt=In(dt,e,t,n,r,l),!0;case"dragenter":return pt=In(pt,e,t,n,r,l),!0;case"mouseover":return mt=In(mt,e,t,n,r,l),!0;case"pointerover":var o=l.pointerId;return Fn.set(o,In(Fn.get(o)||null,e,t,n,r,l)),!0;case"gotpointercapture":return o=l.pointerId,Dn.set(o,In(Dn.get(o)||null,e,t,n,r,l)),!0}return!1}function bu(e){var t=Ot(e.target);if(t!==null){var n=Mt(t);if(n!==null){if(t=n.tag,t===13){if(t=Au(n),t!==null){e.blockedOn=t,Ju(e.priority,function(){Gu(n)});return}}else if(t===3&&n.stateNode.current.memoizedState.isDehydrated){e.blockedOn=n.tag===3?n.stateNode.containerInfo:null;return}}}e.blockedOn=null}function zr(e){if(e.blockedOn!==null)return!1;for(var t=e.targetContainers;0<t.length;){var n=go(e.domEventName,e.eventSystemFlags,t[0],e.nativeEvent);if(n===null){n=e.nativeEvent;var r=new n.constructor(n.type,n);lo=r,n.target.dispatchEvent(r),lo=null}else return t=Jn(n),t!==null&&ho(t),e.blockedOn=n,!1;t.shift()}return!0}function es(e,t,n){zr(e)&&n.delete(t)}function Rf(){vo=!1,dt!==null&&zr(dt)&&(dt=null),pt!==null&&zr(pt)&&(pt=null),mt!==null&&zr(mt)&&(mt=null),Fn.forEach(es),Dn.forEach(es)}function An(e,t){e.blockedOn===t&&(e.blockedOn=null,vo||(vo=!0,_e.unstable_scheduleCallback(_e.unstable_NormalPriority,Rf)))}function Un(e){function t(l){return An(l,e)}if(0<Tr.length){An(Tr[0],e);for(var n=1;n<Tr.length;n++){var r=Tr[n];r.blockedOn===e&&(r.blockedOn=null)}}for(dt!==null&&An(dt,e),pt!==null&&An(pt,e),mt!==null&&An(mt,e),Fn.forEach(t),Dn.forEach(t),n=0;n<ht.length;n++)r=ht[n],r.blockedOn===e&&(r.blockedOn=null);for(;0<ht.length&&(n=ht[0],n.blockedOn===null);)bu(n),n.blockedOn===null&&ht.shift()}var en=Je.ReactCurrentBatchConfig,Lr=!0;function Mf(e,t,n,r){var l=A,o=en.transition;en.transition=null;try{A=1,yo(e,t,n,r)}finally{A=l,en.transition=o}}function Of(e,t,n,r){var l=A,o=en.transition;en.transition=null;try{A=4,yo(e,t,n,r)}finally{A=l,en.transition=o}}function yo(e,t,n,r){if(Lr){var l=go(e,t,n,r);if(l===null)Fo(e,t,r,Rr,n),qu(e,r);else if(Lf(l,e,t,n,r))r.stopPropagation();else if(qu(e,r),t&4&&-1<zf.indexOf(e)){for(;l!==null;){var o=Jn(l);if(o!==null&&Xu(o),o=go(e,t,n,r),o===null&&Fo(e,t,r,Rr,n),o===l)break;l=o}l!==null&&r.stopPropagation()}else Fo(e,t,r,null,n)}}var Rr=null;function go(e,t,n,r){if(Rr=null,e=oo(r),e=Ot(e),e!==null)if(t=Mt(e),t===null)e=null;else if(n=t.tag,n===13){if(e=Au(t),e!==null)return e;e=null}else if(n===3){if(t.stateNode.current.memoizedState.isDehydrated)return t.tag===3?t.stateNode.containerInfo:null;e=null}else t!==e&&(e=null);return Rr=e,null}function ts(e){switch(e){case"cancel":case"click":case"close":case"contextmenu":case"copy":case"cut":case"auxclick":case"dblclick":case"dragend":case"dragstart":case"drop":case"focusin":case"focusout":case"input":case"invalid":case"keydown":case"keypress":case"keyup":case"mousedown":case"mouseup":case"paste":case"pause":case"play":case"pointercancel":case"pointerdown":case"pointerup":case"ratechange":case"reset":case"resize":case"seeked":case"submit":case"touchcancel":case"touchend":case"touchstart":case"volumechange":case"change":case"selectionchange":case"textInput":case"compositionstart":case"compositionend":case"compositionupdate":case"beforeblur":case"afterblur":case"beforeinput":case"blur":case"fullscreenchange":case"focus":case"hashchange":case"popstate":case"select":case"selectstart":return 1;case"drag":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"mousemove":case"mouseout":case"mouseover":case"pointermove":case"pointerout":case"pointerover":case"scroll":case"toggle":case"touchmove":case"wheel":case"mouseenter":case"mouseleave":case"pointerenter":case"pointerleave":return 4;case"message":switch(Sf()){case co:return 1;case Hu:return 4;case Er:case xf:return 16;case Qu:return 536870912;default:return 16}default:return 16}}var vt=null,ko=null,Mr=null;function ns(){if(Mr)return Mr;var e,t=ko,n=t.length,r,l="value"in vt?vt.value:vt.textContent,o=l.length;for(e=0;e<n&&t[e]===l[e];e++);var i=n-e;for(r=1;r<=i&&t[n-r]===l[o-r];r++);return Mr=l.slice(e,1<r?1-r:void 0)}function Or(e){var t=e.keyCode;return"charCode"in e?(e=e.charCode,e===0&&t===13&&(e=13)):e=t,e===10&&(e=13),32<=e||e===13?e:0}function Fr(){return!0}function rs(){return!1}function Ee(e){function t(n,r,l,o,i){this._reactName=n,this._targetInst=l,this.type=r,this.nativeEvent=o,this.target=i,this.currentTarget=null;for(var u in e)e.hasOwnProperty(u)&&(n=e[u],this[u]=n?n(o):o[u]);return this.isDefaultPrevented=(o.defaultPrevented!=null?o.defaultPrevented:o.returnValue===!1)?Fr:rs,this.isPropagationStopped=rs,this}return K(t.prototype,{preventDefault:function(){this.defaultPrevented=!0;var n=this.nativeEvent;n&&(n.preventDefault?n.preventDefault():typeof n.returnValue!="unknown"&&(n.returnValue=!1),this.isDefaultPrevented=Fr)},stopPropagation:function(){var n=this.nativeEvent;n&&(n.stopPropagation?n.stopPropagation():typeof n.cancelBubble!="unknown"&&(n.cancelBubble=!0),this.isPropagationStopped=Fr)},persist:function(){},isPersistent:Fr}),t}var tn={eventPhase:0,bubbles:0,cancelable:0,timeStamp:function(e){return e.timeStamp||Date.now()},defaultPrevented:0,isTrusted:0},wo=Ee(tn),$n=K({},tn,{view:0,detail:0}),Ff=Ee($n),So,xo,Vn,Dr=K({},$n,{screenX:0,screenY:0,clientX:0,clientY:0,pageX:0,pageY:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,getModifierState:Eo,button:0,buttons:0,relatedTarget:function(e){return e.relatedTarget===void 0?e.fromElement===e.srcElement?e.toElement:e.fromElement:e.relatedTarget},movementX:function(e){return"movementX"in e?e.movementX:(e!==Vn&&(Vn&&e.type==="mousemove"?(So=e.screenX-Vn.screenX,xo=e.screenY-Vn.screenY):xo=So=0,Vn=e),So)},movementY:function(e){return"movementY"in e?e.movementY:xo}}),ls=Ee(Dr),Df=K({},Dr,{dataTransfer:0}),If=Ee(Df),Af=K({},$n,{relatedTarget:0}),_o=Ee(Af),Uf=K({},tn,{animationName:0,elapsedTime:0,pseudoElement:0}),$f=Ee(Uf),Vf=K({},tn,{clipboardData:function(e){return"clipboardData"in e?e.clipboardData:window.clipboardData}}),Bf=Ee(Vf),Wf=K({},tn,{data:0}),os=Ee(Wf),Hf={Esc:"Escape",Spacebar:" ",Left:"ArrowLeft",Up:"ArrowUp",Right:"ArrowRight",Down:"ArrowDown",Del:"Delete",Win:"OS",Menu:"ContextMenu",Apps:"ContextMenu",Scroll:"ScrollLock",MozPrintableKey:"Unidentified"},Qf={8:"Backspace",9:"Tab",12:"Clear",13:"Enter",16:"Shift",17:"Control",18:"Alt",19:"Pause",20:"CapsLock",27:"Escape",32:" ",33:"PageUp",34:"PageDown",35:"End",36:"Home",37:"ArrowLeft",38:"ArrowUp",39:"ArrowRight",40:"ArrowDown",45:"Insert",46:"Delete",112:"F1",113:"F2",114:"F3",115:"F4",116:"F5",117:"F6",118:"F7",119:"F8",120:"F9",121:"F10",122:"F11",123:"F12",144:"NumLock",145:"ScrollLock",224:"Meta"},Kf={Alt:"altKey",Control:"ctrlKey",Meta:"metaKey",Shift:"shiftKey"};function Yf(e){var t=this.nativeEvent;return t.getModifierState?t.getModifierState(e):(e=Kf[e])?!!t[e]:!1}function Eo(){return Yf}var Xf=K({},$n,{key:function(e){if(e.key){var t=Hf[e.key]||e.key;if(t!=="Unidentified")return t}return e.type==="keypress"?(e=Or(e),e===13?"Enter":String.fromCharCode(e)):e.type==="keydown"||e.type==="keyup"?Qf[e.keyCode]||"Unidentified":""},code:0,location:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,repeat:0,locale:0,getModifierState:Eo,charCode:function(e){return e.type==="keypress"?Or(e):0},keyCode:function(e){return e.type==="keydown"||e.type==="keyup"?e.keyCode:0},which:function(e){return e.type==="keypress"?Or(e):e.type==="keydown"||e.type==="keyup"?e.keyCode:0}}),Gf=Ee(Xf),Zf=K({},Dr,{pointerId:0,width:0,height:0,pressure:0,tangentialPressure:0,tiltX:0,tiltY:0,twist:0,pointerType:0,isPrimary:0}),is=Ee(Zf),Jf=K({},$n,{touches:0,targetTouches:0,changedTouches:0,altKey:0,metaKey:0,ctrlKey:0,shiftKey:0,getModifierState:Eo}),qf=Ee(Jf),bf=K({},tn,{propertyName:0,elapsedTime:0,pseudoElement:0}),ed=Ee(bf),td=K({},Dr,{deltaX:function(e){return"deltaX"in e?e.deltaX:"wheelDeltaX"in e?-e.wheelDeltaX:0},deltaY:function(e){return"deltaY"in e?e.deltaY:"wheelDeltaY"in e?-e.wheelDeltaY:"wheelDelta"in e?-e.wheelDelta:0},deltaZ:0,deltaMode:0}),nd=Ee(td),rd=[9,13,27,32],Co=Ze&&"CompositionEvent"in window,Bn=null;Ze&&"documentMode"in document&&(Bn=document.documentMode);var ld=Ze&&"TextEvent"in window&&!Bn,us=Ze&&(!Co||Bn&&8<Bn&&11>=Bn),ss=" ",as=!1;function cs(e,t){switch(e){case"keyup":return rd.indexOf(t.keyCode)!==-1;case"keydown":return t.keyCode!==229;case"keypress":case"mousedown":case"focusout":return!0;default:return!1}}function fs(e){return e=e.detail,typeof e=="object"&&"data"in e?e.data:null}var nn=!1;function od(e,t){switch(e){case"compositionend":return fs(t);case"keypress":return t.which!==32?null:(as=!0,ss);case"textInput":return e=t.data,e===ss&&as?null:e;default:return null}}function id(e,t){if(nn)return e==="compositionend"||!Co&&cs(e,t)?(e=ns(),Mr=ko=vt=null,nn=!1,e):null;switch(e){case"paste":return null;case"keypress":if(!(t.ctrlKey||t.altKey||t.metaKey)||t.ctrlKey&&t.altKey){if(t.char&&1<t.char.length)return t.char;if(t.which)return String.fromCharCode(t.which)}return null;case"compositionend":return us&&t.locale!=="ko"?null:t.data;default:return null}}var ud={color:!0,date:!0,datetime:!0,"datetime-local":!0,email:!0,month:!0,number:!0,password:!0,range:!0,search:!0,tel:!0,text:!0,time:!0,url:!0,week:!0};function ds(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t==="input"?!!ud[e.type]:t==="textarea"}function ps(e,t,n,r){Mu(r),t=Vr(t,"onChange"),0<t.length&&(n=new wo("onChange","change",null,n,r),e.push({event:n,listeners:t}))}var Wn=null,Hn=null;function sd(e){Ls(e,0)}function Ir(e){var t=sn(e);if(Su(t))return e}function ad(e,t){if(e==="change")return t}var ms=!1;if(Ze){var No;if(Ze){var jo="oninput"in document;if(!jo){var hs=document.createElement("div");hs.setAttribute("oninput","return;"),jo=typeof hs.oninput=="function"}No=jo}else No=!1;ms=No&&(!document.documentMode||9<document.documentMode)}function vs(){Wn&&(Wn.detachEvent("onpropertychange",ys),Hn=Wn=null)}function ys(e){if(e.propertyName==="value"&&Ir(Hn)){var t=[];ps(t,Hn,e,oo(e)),Iu(sd,t)}}function cd(e,t,n){e==="focusin"?(vs(),Wn=t,Hn=n,Wn.attachEvent("onpropertychange",ys)):e==="focusout"&&vs()}function fd(e){if(e==="selectionchange"||e==="keyup"||e==="keydown")return Ir(Hn)}function dd(e,t){if(e==="click")return Ir(t)}function pd(e,t){if(e==="input"||e==="change")return Ir(t)}function md(e,t){return e===t&&(e!==0||1/e===1/t)||e!==e&&t!==t}var Ae=typeof Object.is=="function"?Object.is:md;function Qn(e,t){if(Ae(e,t))return!0;if(typeof e!="object"||e===null||typeof t!="object"||t===null)return!1;var n=Object.keys(e),r=Object.keys(t);if(n.length!==r.length)return!1;for(r=0;r<n.length;r++){var l=n[r];if(!Il.call(t,l)||!Ae(e[l],t[l]))return!1}return!0}function gs(e){for(;e&&e.firstChild;)e=e.firstChild;return e}function ks(e,t){var n=gs(e);e=0;for(var r;n;){if(n.nodeType===3){if(r=e+n.textContent.length,e<=t&&r>=t)return{node:n,offset:t-e};e=r}e:{for(;n;){if(n.nextSibling){n=n.nextSibling;break e}n=n.parentNode}n=void 0}n=gs(n)}}function ws(e,t){return e&&t?e===t?!0:e&&e.nodeType===3?!1:t&&t.nodeType===3?ws(e,t.parentNode):"contains"in e?e.contains(t):e.compareDocumentPosition?!!(e.compareDocumentPosition(t)&16):!1:!1}function Ss(){for(var e=window,t=wr();t instanceof e.HTMLIFrameElement;){try{var n=typeof t.contentWindow.location.href=="string"}catch{n=!1}if(n)e=t.contentWindow;else break;t=wr(e.document)}return t}function Po(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t&&(t==="input"&&(e.type==="text"||e.type==="search"||e.type==="tel"||e.type==="url"||e.type==="password")||t==="textarea"||e.contentEditable==="true")}function hd(e){var t=Ss(),n=e.focusedElem,r=e.selectionRange;if(t!==n&&n&&n.ownerDocument&&ws(n.ownerDocument.documentElement,n)){if(r!==null&&Po(n)){if(t=r.start,e=r.end,e===void 0&&(e=t),"selectionStart"in n)n.selectionStart=t,n.selectionEnd=Math.min(e,n.value.length);else if(e=(t=n.ownerDocument||document)&&t.defaultView||window,e.getSelection){e=e.getSelection();var l=n.textContent.length,o=Math.min(r.start,l);r=r.end===void 0?o:Math.min(r.end,l),!e.extend&&o>r&&(l=r,r=o,o=l),l=ks(n,o);var i=ks(n,r);l&&i&&(e.rangeCount!==1||e.anchorNode!==l.node||e.anchorOffset!==l.offset||e.focusNode!==i.node||e.focusOffset!==i.offset)&&(t=t.createRange(),t.setStart(l.node,l.offset),e.removeAllRanges(),o>r?(e.addRange(t),e.extend(i.node,i.offset)):(t.setEnd(i.node,i.offset),e.addRange(t)))}}for(t=[],e=n;e=e.parentNode;)e.nodeType===1&&t.push({element:e,left:e.scrollLeft,top:e.scrollTop});for(typeof n.focus=="function"&&n.focus(),n=0;n<t.length;n++)e=t[n],e.element.scrollLeft=e.left,e.element.scrollTop=e.top}}var vd=Ze&&"documentMode"in document&&11>=document.documentMode,rn=null,To=null,Kn=null,zo=!1;function xs(e,t,n){var r=n.window===n?n.document:n.nodeType===9?n:n.ownerDocument;zo||rn==null||rn!==wr(r)||(r=rn,"selectionStart"in r&&Po(r)?r={start:r.selectionStart,end:r.selectionEnd}:(r=(r.ownerDocument&&r.ownerDocument.defaultView||window).getSelection(),r={anchorNode:r.anchorNode,anchorOffset:r.anchorOffset,focusNode:r.focusNode,focusOffset:r.focusOffset}),Kn&&Qn(Kn,r)||(Kn=r,r=Vr(To,"onSelect"),0<r.length&&(t=new wo("onSelect","select",null,t,n),e.push({event:t,listeners:r}),t.target=rn)))}function Ar(e,t){var n={};return n[e.toLowerCase()]=t.toLowerCase(),n["Webkit"+e]="webkit"+t,n["Moz"+e]="moz"+t,n}var ln={animationend:Ar("Animation","AnimationEnd"),animationiteration:Ar("Animation","AnimationIteration"),animationstart:Ar("Animation","AnimationStart"),transitionend:Ar("Transition","TransitionEnd")},Lo={},_s={};Ze&&(_s=document.createElement("div").style,"AnimationEvent"in window||(delete ln.animationend.animation,delete ln.animationiteration.animation,delete ln.animationstart.animation),"TransitionEvent"in window||delete ln.transitionend.transition);function Ur(e){if(Lo[e])return Lo[e];if(!ln[e])return e;var t=ln[e],n;for(n in t)if(t.hasOwnProperty(n)&&n in _s)return Lo[e]=t[n];return e}var Es=Ur("animationend"),Cs=Ur("animationiteration"),Ns=Ur("animationstart"),js=Ur("transitionend"),Ps=new Map,Ts="abort auxClick cancel canPlay canPlayThrough click close contextMenu copy cut drag dragEnd dragEnter dragExit dragLeave dragOver dragStart drop durationChange emptied encrypted ended error gotPointerCapture input invalid keyDown keyPress keyUp load loadedData loadedMetadata loadStart lostPointerCapture mouseDown mouseMove mouseOut mouseOver mouseUp paste pause play playing pointerCancel pointerDown pointerMove pointerOut pointerOver pointerUp progress rateChange reset resize seeked seeking stalled submit suspend timeUpdate touchCancel touchEnd touchStart volumeChange scroll toggle touchMove waiting wheel".split(" ");function yt(e,t){Ps.set(e,t),Rt(t,[e])}for(var Ro=0;Ro<Ts.length;Ro++){var Mo=Ts[Ro],yd=Mo.toLowerCase(),gd=Mo[0].toUpperCase()+Mo.slice(1);yt(yd,"on"+gd)}yt(Es,"onAnimationEnd"),yt(Cs,"onAnimationIteration"),yt(Ns,"onAnimationStart"),yt("dblclick","onDoubleClick"),yt("focusin","onFocus"),yt("focusout","onBlur"),yt(js,"onTransitionEnd"),Xt("onMouseEnter",["mouseout","mouseover"]),Xt("onMouseLeave",["mouseout","mouseover"]),Xt("onPointerEnter",["pointerout","pointerover"]),Xt("onPointerLeave",["pointerout","pointerover"]),Rt("onChange","change click focusin focusout input keydown keyup selectionchange".split(" ")),Rt("onSelect","focusout contextmenu dragend focusin keydown keyup mousedown mouseup selectionchange".split(" ")),Rt("onBeforeInput",["compositionend","keypress","textInput","paste"]),Rt("onCompositionEnd","compositionend focusout keydown keypress keyup mousedown".split(" ")),Rt("onCompositionStart","compositionstart focusout keydown keypress keyup mousedown".split(" ")),Rt("onCompositionUpdate","compositionupdate focusout keydown keypress keyup mousedown".split(" "));var Yn="abort canplay canplaythrough durationchange emptied encrypted ended error loadeddata loadedmetadata loadstart pause play playing progress ratechange resize seeked seeking stalled suspend timeupdate volumechange waiting".split(" "),kd=new Set("cancel close invalid load scroll toggle".split(" ").concat(Yn));function zs(e,t,n){var r=e.type||"unknown-event";e.currentTarget=n,yf(r,t,void 0,e),e.currentTarget=null}function Ls(e,t){t=(t&4)!==0;for(var n=0;n<e.length;n++){var r=e[n],l=r.event;r=r.listeners;e:{var o=void 0;if(t)for(var i=r.length-1;0<=i;i--){var u=r[i],s=u.instance,c=u.currentTarget;if(u=u.listener,s!==o&&l.isPropagationStopped())break e;zs(l,u,c),o=s}else for(i=0;i<r.length;i++){if(u=r[i],s=u.instance,c=u.currentTarget,u=u.listener,s!==o&&l.isPropagationStopped())break e;zs(l,u,c),o=s}}}if(_r)throw e=ao,_r=!1,ao=null,e}function V(e,t){var n=t[Vo];n===void 0&&(n=t[Vo]=new Set);var r=e+"__bubble";n.has(r)||(Rs(t,e,2,!1),n.add(r))}function Oo(e,t,n){var r=0;t&&(r|=4),Rs(n,e,r,t)}var $r="_reactListening"+Math.random().toString(36).slice(2);function Xn(e){if(!e[$r]){e[$r]=!0,pu.forEach(function(n){n!=="selectionchange"&&(kd.has(n)||Oo(n,!1,e),Oo(n,!0,e))});var t=e.nodeType===9?e:e.ownerDocument;t===null||t[$r]||(t[$r]=!0,Oo("selectionchange",!1,t))}}function Rs(e,t,n,r){switch(ts(t)){case 1:var l=Mf;break;case 4:l=Of;break;default:l=yo}n=l.bind(null,t,n,e),l=void 0,!so||t!=="touchstart"&&t!=="touchmove"&&t!=="wheel"||(l=!0),r?l!==void 0?e.addEventListener(t,n,{capture:!0,passive:l}):e.addEventListener(t,n,!0):l!==void 0?e.addEventListener(t,n,{passive:l}):e.addEventListener(t,n,!1)}function Fo(e,t,n,r,l){var o=r;if(!(t&1)&&!(t&2)&&r!==null)e:for(;;){if(r===null)return;var i=r.tag;if(i===3||i===4){var u=r.stateNode.containerInfo;if(u===l||u.nodeType===8&&u.parentNode===l)break;if(i===4)for(i=r.return;i!==null;){var s=i.tag;if((s===3||s===4)&&(s=i.stateNode.containerInfo,s===l||s.nodeType===8&&s.parentNode===l))return;i=i.return}for(;u!==null;){if(i=Ot(u),i===null)return;if(s=i.tag,s===5||s===6){r=o=i;continue e}u=u.parentNode}}r=r.return}Iu(function(){var c=o,y=oo(n),h=[];e:{var p=Ps.get(e);if(p!==void 0){var w=wo,S=e;switch(e){case"keypress":if(Or(n)===0)break e;case"keydown":case"keyup":w=Gf;break;case"focusin":S="focus",w=_o;break;case"focusout":S="blur",w=_o;break;case"beforeblur":case"afterblur":w=_o;break;case"click":if(n.button===2)break e;case"auxclick":case"dblclick":case"mousedown":case"mousemove":case"mouseup":case"mouseout":case"mouseover":case"contextmenu":w=ls;break;case"drag":case"dragend":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"dragstart":case"drop":w=If;break;case"touchcancel":case"touchend":case"touchmove":case"touchstart":w=qf;break;case Es:case Cs:case Ns:w=$f;break;case js:w=ed;break;case"scroll":w=Ff;break;case"wheel":w=nd;break;case"copy":case"cut":case"paste":w=Bf;break;case"gotpointercapture":case"lostpointercapture":case"pointercancel":case"pointerdown":case"pointermove":case"pointerout":case"pointerover":case"pointerup":w=is}var E=(t&4)!==0,U=!E&&e==="scroll",f=E?p!==null?p+"Capture":null:p;E=[];for(var a=c,d;a!==null;){d=a;var g=d.stateNode;if(d.tag===5&&g!==null&&(d=g,f!==null&&(g=zn(a,f),g!=null&&E.push(Gn(a,g,d)))),U)break;a=a.return}0<E.length&&(p=new w(p,S,null,n,y),h.push({event:p,listeners:E}))}}if(!(t&7)){e:{if(p=e==="mouseover"||e==="pointerover",w=e==="mouseout"||e==="pointerout",p&&n!==lo&&(S=n.relatedTarget||n.fromElement)&&(Ot(S)||S[qe]))break e;if((w||p)&&(p=y.window===y?y:(p=y.ownerDocument)?p.defaultView||p.parentWindow:window,w?(S=n.relatedTarget||n.toElement,w=c,S=S?Ot(S):null,S!==null&&(U=Mt(S),S!==U||S.tag!==5&&S.tag!==6)&&(S=null)):(w=null,S=c),w!==S)){if(E=ls,g="onMouseLeave",f="onMouseEnter",a="mouse",(e==="pointerout"||e==="pointerover")&&(E=is,g="onPointerLeave",f="onPointerEnter",a="pointer"),U=w==null?p:sn(w),d=S==null?p:sn(S),p=new E(g,a+"leave",w,n,y),p.target=U,p.relatedTarget=d,g=null,Ot(y)===c&&(E=new E(f,a+"enter",S,n,y),E.target=d,E.relatedTarget=U,g=E),U=g,w&&S)t:{for(E=w,f=S,a=0,d=E;d;d=on(d))a++;for(d=0,g=f;g;g=on(g))d++;for(;0<a-d;)E=on(E),a--;for(;0<d-a;)f=on(f),d--;for(;a--;){if(E===f||f!==null&&E===f.alternate)break t;E=on(E),f=on(f)}E=null}else E=null;w!==null&&Ms(h,p,w,E,!1),S!==null&&U!==null&&Ms(h,U,S,E,!0)}}e:{if(p=c?sn(c):window,w=p.nodeName&&p.nodeName.toLowerCase(),w==="select"||w==="input"&&p.type==="file")var N=ad;else if(ds(p))if(ms)N=pd;else{N=fd;var P=cd}else(w=p.nodeName)&&w.toLowerCase()==="input"&&(p.type==="checkbox"||p.type==="radio")&&(N=dd);if(N&&(N=N(e,c))){ps(h,N,n,y);break e}P&&P(e,p,c),e==="focusout"&&(P=p._wrapperState)&&P.controlled&&p.type==="number"&&bl(p,"number",p.value)}switch(P=c?sn(c):window,e){case"focusin":(ds(P)||P.contentEditable==="true")&&(rn=P,To=c,Kn=null);break;case"focusout":Kn=To=rn=null;break;case"mousedown":zo=!0;break;case"contextmenu":case"mouseup":case"dragend":zo=!1,xs(h,n,y);break;case"selectionchange":if(vd)break;case"keydown":case"keyup":xs(h,n,y)}var T;if(Co)e:{switch(e){case"compositionstart":var L="onCompositionStart";break e;case"compositionend":L="onCompositionEnd";break e;case"compositionupdate":L="onCompositionUpdate";break e}L=void 0}else nn?cs(e,n)&&(L="onCompositionEnd"):e==="keydown"&&n.keyCode===229&&(L="onCompositionStart");L&&(us&&n.locale!=="ko"&&(nn||L!=="onCompositionStart"?L==="onCompositionEnd"&&nn&&(T=ns()):(vt=y,ko="value"in vt?vt.value:vt.textContent,nn=!0)),P=Vr(c,L),0<P.length&&(L=new os(L,e,null,n,y),h.push({event:L,listeners:P}),T?L.data=T:(T=fs(n),T!==null&&(L.data=T)))),(T=ld?od(e,n):id(e,n))&&(c=Vr(c,"onBeforeInput"),0<c.length&&(y=new os("onBeforeInput","beforeinput",null,n,y),h.push({event:y,listeners:c}),y.data=T))}Ls(h,t)})}function Gn(e,t,n){return{instance:e,listener:t,currentTarget:n}}function Vr(e,t){for(var n=t+"Capture",r=[];e!==null;){var l=e,o=l.stateNode;l.tag===5&&o!==null&&(l=o,o=zn(e,n),o!=null&&r.unshift(Gn(e,o,l)),o=zn(e,t),o!=null&&r.push(Gn(e,o,l))),e=e.return}return r}function on(e){if(e===null)return null;do e=e.return;while(e&&e.tag!==5);return e||null}function Ms(e,t,n,r,l){for(var o=t._reactName,i=[];n!==null&&n!==r;){var u=n,s=u.alternate,c=u.stateNode;if(s!==null&&s===r)break;u.tag===5&&c!==null&&(u=c,l?(s=zn(n,o),s!=null&&i.unshift(Gn(n,s,u))):l||(s=zn(n,o),s!=null&&i.push(Gn(n,s,u)))),n=n.return}i.length!==0&&e.push({event:t,listeners:i})}var wd=/\r\n?/g,Sd=/\u0000|\uFFFD/g;function Os(e){return(typeof e=="string"?e:""+e).replace(wd,`
+`).replace(Sd,"")}function Br(e,t,n){if(t=Os(t),Os(e)!==t&&n)throw Error(k(425))}function Wr(){}var Do=null,Io=null;function Ao(e,t){return e==="textarea"||e==="noscript"||typeof t.children=="string"||typeof t.children=="number"||typeof t.dangerouslySetInnerHTML=="object"&&t.dangerouslySetInnerHTML!==null&&t.dangerouslySetInnerHTML.__html!=null}var Uo=typeof setTimeout=="function"?setTimeout:void 0,xd=typeof clearTimeout=="function"?clearTimeout:void 0,Fs=typeof Promise=="function"?Promise:void 0,_d=typeof queueMicrotask=="function"?queueMicrotask:typeof Fs<"u"?function(e){return Fs.resolve(null).then(e).catch(Ed)}:Uo;function Ed(e){setTimeout(function(){throw e})}function $o(e,t){var n=t,r=0;do{var l=n.nextSibling;if(e.removeChild(n),l&&l.nodeType===8)if(n=l.data,n==="/$"){if(r===0){e.removeChild(l),Un(t);return}r--}else n!=="$"&&n!=="$?"&&n!=="$!"||r++;n=l}while(n);Un(t)}function gt(e){for(;e!=null;e=e.nextSibling){var t=e.nodeType;if(t===1||t===3)break;if(t===8){if(t=e.data,t==="$"||t==="$!"||t==="$?")break;if(t==="/$")return null}}return e}function Ds(e){e=e.previousSibling;for(var t=0;e;){if(e.nodeType===8){var n=e.data;if(n==="$"||n==="$!"||n==="$?"){if(t===0)return e;t--}else n==="/$"&&t++}e=e.previousSibling}return null}var un=Math.random().toString(36).slice(2),Ke="__reactFiber$"+un,Zn="__reactProps$"+un,qe="__reactContainer$"+un,Vo="__reactEvents$"+un,Cd="__reactListeners$"+un,Nd="__reactHandles$"+un;function Ot(e){var t=e[Ke];if(t)return t;for(var n=e.parentNode;n;){if(t=n[qe]||n[Ke]){if(n=t.alternate,t.child!==null||n!==null&&n.child!==null)for(e=Ds(e);e!==null;){if(n=e[Ke])return n;e=Ds(e)}return t}e=n,n=e.parentNode}return null}function Jn(e){return e=e[Ke]||e[qe],!e||e.tag!==5&&e.tag!==6&&e.tag!==13&&e.tag!==3?null:e}function sn(e){if(e.tag===5||e.tag===6)return e.stateNode;throw Error(k(33))}function Hr(e){return e[Zn]||null}var Bo=[],an=-1;function kt(e){return{current:e}}function B(e){0>an||(e.current=Bo[an],Bo[an]=null,an--)}function $(e,t){an++,Bo[an]=e.current,e.current=t}var wt={},se=kt(wt),ye=kt(!1),Ft=wt;function cn(e,t){var n=e.type.contextTypes;if(!n)return wt;var r=e.stateNode;if(r&&r.__reactInternalMemoizedUnmaskedChildContext===t)return r.__reactInternalMemoizedMaskedChildContext;var l={},o;for(o in n)l[o]=t[o];return r&&(e=e.stateNode,e.__reactInternalMemoizedUnmaskedChildContext=t,e.__reactInternalMemoizedMaskedChildContext=l),l}function ge(e){return e=e.childContextTypes,e!=null}function Qr(){B(ye),B(se)}function Is(e,t,n){if(se.current!==wt)throw Error(k(168));$(se,t),$(ye,n)}function As(e,t,n){var r=e.stateNode;if(t=t.childContextTypes,typeof r.getChildContext!="function")return n;r=r.getChildContext();for(var l in r)if(!(l in t))throw Error(k(108,cf(e)||"Unknown",l));return K({},n,r)}function Kr(e){return e=(e=e.stateNode)&&e.__reactInternalMemoizedMergedChildContext||wt,Ft=se.current,$(se,e),$(ye,ye.current),!0}function Us(e,t,n){var r=e.stateNode;if(!r)throw Error(k(169));n?(e=As(e,t,Ft),r.__reactInternalMemoizedMergedChildContext=e,B(ye),B(se),$(se,e)):B(ye),$(ye,n)}var be=null,Yr=!1,Wo=!1;function $s(e){be===null?be=[e]:be.push(e)}function jd(e){Yr=!0,$s(e)}function St(){if(!Wo&&be!==null){Wo=!0;var e=0,t=A;try{var n=be;for(A=1;e<n.length;e++){var r=n[e];do r=r(!0);while(r!==null)}be=null,Yr=!1}catch(l){throw be!==null&&(be=be.slice(e+1)),Bu(co,St),l}finally{A=t,Wo=!1}}return null}var fn=[],dn=0,Xr=null,Gr=0,ze=[],Le=0,Dt=null,et=1,tt="";function It(e,t){fn[dn++]=Gr,fn[dn++]=Xr,Xr=e,Gr=t}function Vs(e,t,n){ze[Le++]=et,ze[Le++]=tt,ze[Le++]=Dt,Dt=e;var r=et;e=tt;var l=32-Ie(r)-1;r&=~(1<<l),n+=1;var o=32-Ie(t)+l;if(30<o){var i=l-l%5;o=(r&(1<<i)-1).toString(32),r>>=i,l-=i,et=1<<32-Ie(t)+l|n<<l|r,tt=o+e}else et=1<<o|n<<l|r,tt=e}function Ho(e){e.return!==null&&(It(e,1),Vs(e,1,0))}function Qo(e){for(;e===Xr;)Xr=fn[--dn],fn[dn]=null,Gr=fn[--dn],fn[dn]=null;for(;e===Dt;)Dt=ze[--Le],ze[Le]=null,tt=ze[--Le],ze[Le]=null,et=ze[--Le],ze[Le]=null}var Ce=null,Ne=null,H=!1,Ue=null;function Bs(e,t){var n=Fe(5,null,null,0);n.elementType="DELETED",n.stateNode=t,n.return=e,t=e.deletions,t===null?(e.deletions=[n],e.flags|=16):t.push(n)}function Ws(e,t){switch(e.tag){case 5:var n=e.type;return t=t.nodeType!==1||n.toLowerCase()!==t.nodeName.toLowerCase()?null:t,t!==null?(e.stateNode=t,Ce=e,Ne=gt(t.firstChild),!0):!1;case 6:return t=e.pendingProps===""||t.nodeType!==3?null:t,t!==null?(e.stateNode=t,Ce=e,Ne=null,!0):!1;case 13:return t=t.nodeType!==8?null:t,t!==null?(n=Dt!==null?{id:et,overflow:tt}:null,e.memoizedState={dehydrated:t,treeContext:n,retryLane:1073741824},n=Fe(18,null,null,0),n.stateNode=t,n.return=e,e.child=n,Ce=e,Ne=null,!0):!1;default:return!1}}function Ko(e){return(e.mode&1)!==0&&(e.flags&128)===0}function Yo(e){if(H){var t=Ne;if(t){var n=t;if(!Ws(e,t)){if(Ko(e))throw Error(k(418));t=gt(n.nextSibling);var r=Ce;t&&Ws(e,t)?Bs(r,n):(e.flags=e.flags&-4097|2,H=!1,Ce=e)}}else{if(Ko(e))throw Error(k(418));e.flags=e.flags&-4097|2,H=!1,Ce=e}}}function Hs(e){for(e=e.return;e!==null&&e.tag!==5&&e.tag!==3&&e.tag!==13;)e=e.return;Ce=e}function Zr(e){if(e!==Ce)return!1;if(!H)return Hs(e),H=!0,!1;var t;if((t=e.tag!==3)&&!(t=e.tag!==5)&&(t=e.type,t=t!=="head"&&t!=="body"&&!Ao(e.type,e.memoizedProps)),t&&(t=Ne)){if(Ko(e))throw Qs(),Error(k(418));for(;t;)Bs(e,t),t=gt(t.nextSibling)}if(Hs(e),e.tag===13){if(e=e.memoizedState,e=e!==null?e.dehydrated:null,!e)throw Error(k(317));e:{for(e=e.nextSibling,t=0;e;){if(e.nodeType===8){var n=e.data;if(n==="/$"){if(t===0){Ne=gt(e.nextSibling);break e}t--}else n!=="$"&&n!=="$!"&&n!=="$?"||t++}e=e.nextSibling}Ne=null}}else Ne=Ce?gt(e.stateNode.nextSibling):null;return!0}function Qs(){for(var e=Ne;e;)e=gt(e.nextSibling)}function pn(){Ne=Ce=null,H=!1}function Xo(e){Ue===null?Ue=[e]:Ue.push(e)}var Pd=Je.ReactCurrentBatchConfig;function qn(e,t,n){if(e=n.ref,e!==null&&typeof e!="function"&&typeof e!="object"){if(n._owner){if(n=n._owner,n){if(n.tag!==1)throw Error(k(309));var r=n.stateNode}if(!r)throw Error(k(147,e));var l=r,o=""+e;return t!==null&&t.ref!==null&&typeof t.ref=="function"&&t.ref._stringRef===o?t.ref:(t=function(i){var u=l.refs;i===null?delete u[o]:u[o]=i},t._stringRef=o,t)}if(typeof e!="string")throw Error(k(284));if(!n._owner)throw Error(k(290,e))}return e}function Jr(e,t){throw e=Object.prototype.toString.call(t),Error(k(31,e==="[object Object]"?"object with keys {"+Object.keys(t).join(", ")+"}":e))}function Ks(e){var t=e._init;return t(e._payload)}function Ys(e){function t(f,a){if(e){var d=f.deletions;d===null?(f.deletions=[a],f.flags|=16):d.push(a)}}function n(f,a){if(!e)return null;for(;a!==null;)t(f,a),a=a.sibling;return null}function r(f,a){for(f=new Map;a!==null;)a.key!==null?f.set(a.key,a):f.set(a.index,a),a=a.sibling;return f}function l(f,a){return f=Tt(f,a),f.index=0,f.sibling=null,f}function o(f,a,d){return f.index=d,e?(d=f.alternate,d!==null?(d=d.index,d<a?(f.flags|=2,a):d):(f.flags|=2,a)):(f.flags|=1048576,a)}function i(f){return e&&f.alternate===null&&(f.flags|=2),f}function u(f,a,d,g){return a===null||a.tag!==6?(a=Ui(d,f.mode,g),a.return=f,a):(a=l(a,d),a.return=f,a)}function s(f,a,d,g){var N=d.type;return N===Zt?y(f,a,d.props.children,g,d.key):a!==null&&(a.elementType===N||typeof N=="object"&&N!==null&&N.$$typeof===ct&&Ks(N)===a.type)?(g=l(a,d.props),g.ref=qn(f,a,d),g.return=f,g):(g=Sl(d.type,d.key,d.props,null,f.mode,g),g.ref=qn(f,a,d),g.return=f,g)}function c(f,a,d,g){return a===null||a.tag!==4||a.stateNode.containerInfo!==d.containerInfo||a.stateNode.implementation!==d.implementation?(a=$i(d,f.mode,g),a.return=f,a):(a=l(a,d.children||[]),a.return=f,a)}function y(f,a,d,g,N){return a===null||a.tag!==7?(a=Qt(d,f.mode,g,N),a.return=f,a):(a=l(a,d),a.return=f,a)}function h(f,a,d){if(typeof a=="string"&&a!==""||typeof a=="number")return a=Ui(""+a,f.mode,d),a.return=f,a;if(typeof a=="object"&&a!==null){switch(a.$$typeof){case gr:return d=Sl(a.type,a.key,a.props,null,f.mode,d),d.ref=qn(f,null,a),d.return=f,d;case Gt:return a=$i(a,f.mode,d),a.return=f,a;case ct:var g=a._init;return h(f,g(a._payload),d)}if(jn(a)||Cn(a))return a=Qt(a,f.mode,d,null),a.return=f,a;Jr(f,a)}return null}function p(f,a,d,g){var N=a!==null?a.key:null;if(typeof d=="string"&&d!==""||typeof d=="number")return N!==null?null:u(f,a,""+d,g);if(typeof d=="object"&&d!==null){switch(d.$$typeof){case gr:return d.key===N?s(f,a,d,g):null;case Gt:return d.key===N?c(f,a,d,g):null;case ct:return N=d._init,p(f,a,N(d._payload),g)}if(jn(d)||Cn(d))return N!==null?null:y(f,a,d,g,null);Jr(f,d)}return null}function w(f,a,d,g,N){if(typeof g=="string"&&g!==""||typeof g=="number")return f=f.get(d)||null,u(a,f,""+g,N);if(typeof g=="object"&&g!==null){switch(g.$$typeof){case gr:return f=f.get(g.key===null?d:g.key)||null,s(a,f,g,N);case Gt:return f=f.get(g.key===null?d:g.key)||null,c(a,f,g,N);case ct:var P=g._init;return w(f,a,d,P(g._payload),N)}if(jn(g)||Cn(g))return f=f.get(d)||null,y(a,f,g,N,null);Jr(a,g)}return null}function S(f,a,d,g){for(var N=null,P=null,T=a,L=a=0,Q=null;T!==null&&L<d.length;L++){T.index>L?(Q=T,T=null):Q=T.sibling;var F=p(f,T,d[L],g);if(F===null){T===null&&(T=Q);break}e&&T&&F.alternate===null&&t(f,T),a=o(F,a,L),P===null?N=F:P.sibling=F,P=F,T=Q}if(L===d.length)return n(f,T),H&&It(f,L),N;if(T===null){for(;L<d.length;L++)T=h(f,d[L],g),T!==null&&(a=o(T,a,L),P===null?N=T:P.sibling=T,P=T);return H&&It(f,L),N}for(T=r(f,T);L<d.length;L++)Q=w(T,f,L,d[L],g),Q!==null&&(e&&Q.alternate!==null&&T.delete(Q.key===null?L:Q.key),a=o(Q,a,L),P===null?N=Q:P.sibling=Q,P=Q);return e&&T.forEach(function(Pe){return t(f,Pe)}),H&&It(f,L),N}function E(f,a,d,g){var N=Cn(d);if(typeof N!="function")throw Error(k(150));if(d=N.call(d),d==null)throw Error(k(151));for(var P=N=null,T=a,L=a=0,Q=null,F=d.next();T!==null&&!F.done;L++,F=d.next()){T.index>L?(Q=T,T=null):Q=T.sibling;var Pe=p(f,T,F.value,g);if(Pe===null){T===null&&(T=Q);break}e&&T&&Pe.alternate===null&&t(f,T),a=o(Pe,a,L),P===null?N=Pe:P.sibling=Pe,P=Pe,T=Q}if(F.done)return n(f,T),H&&It(f,L),N;if(T===null){for(;!F.done;L++,F=d.next())F=h(f,F.value,g),F!==null&&(a=o(F,a,L),P===null?N=F:P.sibling=F,P=F);return H&&It(f,L),N}for(T=r(f,T);!F.done;L++,F=d.next())F=w(T,f,L,F.value,g),F!==null&&(e&&F.alternate!==null&&T.delete(F.key===null?L:F.key),a=o(F,a,L),P===null?N=F:P.sibling=F,P=F);return e&&T.forEach(function(He){return t(f,He)}),H&&It(f,L),N}function U(f,a,d,g){if(typeof d=="object"&&d!==null&&d.type===Zt&&d.key===null&&(d=d.props.children),typeof d=="object"&&d!==null){switch(d.$$typeof){case gr:e:{for(var N=d.key,P=a;P!==null;){if(P.key===N){if(N=d.type,N===Zt){if(P.tag===7){n(f,P.sibling),a=l(P,d.props.children),a.return=f,f=a;break e}}else if(P.elementType===N||typeof N=="object"&&N!==null&&N.$$typeof===ct&&Ks(N)===P.type){n(f,P.sibling),a=l(P,d.props),a.ref=qn(f,P,d),a.return=f,f=a;break e}n(f,P);break}else t(f,P);P=P.sibling}d.type===Zt?(a=Qt(d.props.children,f.mode,g,d.key),a.return=f,f=a):(g=Sl(d.type,d.key,d.props,null,f.mode,g),g.ref=qn(f,a,d),g.return=f,f=g)}return i(f);case Gt:e:{for(P=d.key;a!==null;){if(a.key===P)if(a.tag===4&&a.stateNode.containerInfo===d.containerInfo&&a.stateNode.implementation===d.implementation){n(f,a.sibling),a=l(a,d.children||[]),a.return=f,f=a;break e}else{n(f,a);break}else t(f,a);a=a.sibling}a=$i(d,f.mode,g),a.return=f,f=a}return i(f);case ct:return P=d._init,U(f,a,P(d._payload),g)}if(jn(d))return S(f,a,d,g);if(Cn(d))return E(f,a,d,g);Jr(f,d)}return typeof d=="string"&&d!==""||typeof d=="number"?(d=""+d,a!==null&&a.tag===6?(n(f,a.sibling),a=l(a,d),a.return=f,f=a):(n(f,a),a=Ui(d,f.mode,g),a.return=f,f=a),i(f)):n(f,a)}return U}var mn=Ys(!0),Xs=Ys(!1),qr=kt(null),br=null,hn=null,Go=null;function Zo(){Go=hn=br=null}function Jo(e){var t=qr.current;B(qr),e._currentValue=t}function qo(e,t,n){for(;e!==null;){var r=e.alternate;if((e.childLanes&t)!==t?(e.childLanes|=t,r!==null&&(r.childLanes|=t)):r!==null&&(r.childLanes&t)!==t&&(r.childLanes|=t),e===n)break;e=e.return}}function vn(e,t){br=e,Go=hn=null,e=e.dependencies,e!==null&&e.firstContext!==null&&(e.lanes&t&&(ke=!0),e.firstContext=null)}function Re(e){var t=e._currentValue;if(Go!==e)if(e={context:e,memoizedValue:t,next:null},hn===null){if(br===null)throw Error(k(308));hn=e,br.dependencies={lanes:0,firstContext:e}}else hn=hn.next=e;return t}var At=null;function bo(e){At===null?At=[e]:At.push(e)}function Gs(e,t,n,r){var l=t.interleaved;return l===null?(n.next=n,bo(t)):(n.next=l.next,l.next=n),t.interleaved=n,nt(e,r)}function nt(e,t){e.lanes|=t;var n=e.alternate;for(n!==null&&(n.lanes|=t),n=e,e=e.return;e!==null;)e.childLanes|=t,n=e.alternate,n!==null&&(n.childLanes|=t),n=e,e=e.return;return n.tag===3?n.stateNode:null}var xt=!1;function ei(e){e.updateQueue={baseState:e.memoizedState,firstBaseUpdate:null,lastBaseUpdate:null,shared:{pending:null,interleaved:null,lanes:0},effects:null}}function Zs(e,t){e=e.updateQueue,t.updateQueue===e&&(t.updateQueue={baseState:e.baseState,firstBaseUpdate:e.firstBaseUpdate,lastBaseUpdate:e.lastBaseUpdate,shared:e.shared,effects:e.effects})}function rt(e,t){return{eventTime:e,lane:t,tag:0,payload:null,callback:null,next:null}}function _t(e,t,n){var r=e.updateQueue;if(r===null)return null;if(r=r.shared,I&2){var l=r.pending;return l===null?t.next=t:(t.next=l.next,l.next=t),r.pending=t,nt(e,n)}return l=r.interleaved,l===null?(t.next=t,bo(r)):(t.next=l.next,l.next=t),r.interleaved=t,nt(e,n)}function el(e,t,n){if(t=t.updateQueue,t!==null&&(t=t.shared,(n&4194240)!==0)){var r=t.lanes;r&=e.pendingLanes,n|=r,t.lanes=n,mo(e,n)}}function Js(e,t){var n=e.updateQueue,r=e.alternate;if(r!==null&&(r=r.updateQueue,n===r)){var l=null,o=null;if(n=n.firstBaseUpdate,n!==null){do{var i={eventTime:n.eventTime,lane:n.lane,tag:n.tag,payload:n.payload,callback:n.callback,next:null};o===null?l=o=i:o=o.next=i,n=n.next}while(n!==null);o===null?l=o=t:o=o.next=t}else l=o=t;n={baseState:r.baseState,firstBaseUpdate:l,lastBaseUpdate:o,shared:r.shared,effects:r.effects},e.updateQueue=n;return}e=n.lastBaseUpdate,e===null?n.firstBaseUpdate=t:e.next=t,n.lastBaseUpdate=t}function tl(e,t,n,r){var l=e.updateQueue;xt=!1;var o=l.firstBaseUpdate,i=l.lastBaseUpdate,u=l.shared.pending;if(u!==null){l.shared.pending=null;var s=u,c=s.next;s.next=null,i===null?o=c:i.next=c,i=s;var y=e.alternate;y!==null&&(y=y.updateQueue,u=y.lastBaseUpdate,u!==i&&(u===null?y.firstBaseUpdate=c:u.next=c,y.lastBaseUpdate=s))}if(o!==null){var h=l.baseState;i=0,y=c=s=null,u=o;do{var p=u.lane,w=u.eventTime;if((r&p)===p){y!==null&&(y=y.next={eventTime:w,lane:0,tag:u.tag,payload:u.payload,callback:u.callback,next:null});e:{var S=e,E=u;switch(p=t,w=n,E.tag){case 1:if(S=E.payload,typeof S=="function"){h=S.call(w,h,p);break e}h=S;break e;case 3:S.flags=S.flags&-65537|128;case 0:if(S=E.payload,p=typeof S=="function"?S.call(w,h,p):S,p==null)break e;h=K({},h,p);break e;case 2:xt=!0}}u.callback!==null&&u.lane!==0&&(e.flags|=64,p=l.effects,p===null?l.effects=[u]:p.push(u))}else w={eventTime:w,lane:p,tag:u.tag,payload:u.payload,callback:u.callback,next:null},y===null?(c=y=w,s=h):y=y.next=w,i|=p;if(u=u.next,u===null){if(u=l.shared.pending,u===null)break;p=u,u=p.next,p.next=null,l.lastBaseUpdate=p,l.shared.pending=null}}while(!0);if(y===null&&(s=h),l.baseState=s,l.firstBaseUpdate=c,l.lastBaseUpdate=y,t=l.shared.interleaved,t!==null){l=t;do i|=l.lane,l=l.next;while(l!==t)}else o===null&&(l.shared.lanes=0);Vt|=i,e.lanes=i,e.memoizedState=h}}function qs(e,t,n){if(e=t.effects,t.effects=null,e!==null)for(t=0;t<e.length;t++){var r=e[t],l=r.callback;if(l!==null){if(r.callback=null,r=n,typeof l!="function")throw Error(k(191,l));l.call(r)}}}var bn={},Ye=kt(bn),er=kt(bn),tr=kt(bn);function Ut(e){if(e===bn)throw Error(k(174));return e}function ti(e,t){switch($(tr,t),$(er,e),$(Ye,bn),e=t.nodeType,e){case 9:case 11:t=(t=t.documentElement)?t.namespaceURI:to(null,"");break;default:e=e===8?t.parentNode:t,t=e.namespaceURI||null,e=e.tagName,t=to(t,e)}B(Ye),$(Ye,t)}function yn(){B(Ye),B(er),B(tr)}function bs(e){Ut(tr.current);var t=Ut(Ye.current),n=to(t,e.type);t!==n&&($(er,e),$(Ye,n))}function ni(e){er.current===e&&(B(Ye),B(er))}var Y=kt(0);function nl(e){for(var t=e;t!==null;){if(t.tag===13){var n=t.memoizedState;if(n!==null&&(n=n.dehydrated,n===null||n.data==="$?"||n.data==="$!"))return t}else if(t.tag===19&&t.memoizedProps.revealOrder!==void 0){if(t.flags&128)return t}else if(t.child!==null){t.child.return=t,t=t.child;continue}if(t===e)break;for(;t.sibling===null;){if(t.return===null||t.return===e)return null;t=t.return}t.sibling.return=t.return,t=t.sibling}return null}var ri=[];function li(){for(var e=0;e<ri.length;e++)ri[e]._workInProgressVersionPrimary=null;ri.length=0}var rl=Je.ReactCurrentDispatcher,oi=Je.ReactCurrentBatchConfig,$t=0,X=null,te=null,re=null,ll=!1,nr=!1,rr=0,Td=0;function ae(){throw Error(k(321))}function ii(e,t){if(t===null)return!1;for(var n=0;n<t.length&&n<e.length;n++)if(!Ae(e[n],t[n]))return!1;return!0}function ui(e,t,n,r,l,o){if($t=o,X=t,t.memoizedState=null,t.updateQueue=null,t.lanes=0,rl.current=e===null||e.memoizedState===null?Md:Od,e=n(r,l),nr){o=0;do{if(nr=!1,rr=0,25<=o)throw Error(k(301));o+=1,re=te=null,t.updateQueue=null,rl.current=Fd,e=n(r,l)}while(nr)}if(rl.current=ul,t=te!==null&&te.next!==null,$t=0,re=te=X=null,ll=!1,t)throw Error(k(300));return e}function si(){var e=rr!==0;return rr=0,e}function Xe(){var e={memoizedState:null,baseState:null,baseQueue:null,queue:null,next:null};return re===null?X.memoizedState=re=e:re=re.next=e,re}function Me(){if(te===null){var e=X.alternate;e=e!==null?e.memoizedState:null}else e=te.next;var t=re===null?X.memoizedState:re.next;if(t!==null)re=t,te=e;else{if(e===null)throw Error(k(310));te=e,e={memoizedState:te.memoizedState,baseState:te.baseState,baseQueue:te.baseQueue,queue:te.queue,next:null},re===null?X.memoizedState=re=e:re=re.next=e}return re}function lr(e,t){return typeof t=="function"?t(e):t}function ai(e){var t=Me(),n=t.queue;if(n===null)throw Error(k(311));n.lastRenderedReducer=e;var r=te,l=r.baseQueue,o=n.pending;if(o!==null){if(l!==null){var i=l.next;l.next=o.next,o.next=i}r.baseQueue=l=o,n.pending=null}if(l!==null){o=l.next,r=r.baseState;var u=i=null,s=null,c=o;do{var y=c.lane;if(($t&y)===y)s!==null&&(s=s.next={lane:0,action:c.action,hasEagerState:c.hasEagerState,eagerState:c.eagerState,next:null}),r=c.hasEagerState?c.eagerState:e(r,c.action);else{var h={lane:y,action:c.action,hasEagerState:c.hasEagerState,eagerState:c.eagerState,next:null};s===null?(u=s=h,i=r):s=s.next=h,X.lanes|=y,Vt|=y}c=c.next}while(c!==null&&c!==o);s===null?i=r:s.next=u,Ae(r,t.memoizedState)||(ke=!0),t.memoizedState=r,t.baseState=i,t.baseQueue=s,n.lastRenderedState=r}if(e=n.interleaved,e!==null){l=e;do o=l.lane,X.lanes|=o,Vt|=o,l=l.next;while(l!==e)}else l===null&&(n.lanes=0);return[t.memoizedState,n.dispatch]}function ci(e){var t=Me(),n=t.queue;if(n===null)throw Error(k(311));n.lastRenderedReducer=e;var r=n.dispatch,l=n.pending,o=t.memoizedState;if(l!==null){n.pending=null;var i=l=l.next;do o=e(o,i.action),i=i.next;while(i!==l);Ae(o,t.memoizedState)||(ke=!0),t.memoizedState=o,t.baseQueue===null&&(t.baseState=o),n.lastRenderedState=o}return[o,r]}function ea(){}function ta(e,t){var n=X,r=Me(),l=t(),o=!Ae(r.memoizedState,l);if(o&&(r.memoizedState=l,ke=!0),r=r.queue,fi(la.bind(null,n,r,e),[e]),r.getSnapshot!==t||o||re!==null&&re.memoizedState.tag&1){if(n.flags|=2048,or(9,ra.bind(null,n,r,l,t),void 0,null),le===null)throw Error(k(349));$t&30||na(n,t,l)}return l}function na(e,t,n){e.flags|=16384,e={getSnapshot:t,value:n},t=X.updateQueue,t===null?(t={lastEffect:null,stores:null},X.updateQueue=t,t.stores=[e]):(n=t.stores,n===null?t.stores=[e]:n.push(e))}function ra(e,t,n,r){t.value=n,t.getSnapshot=r,oa(t)&&ia(e)}function la(e,t,n){return n(function(){oa(t)&&ia(e)})}function oa(e){var t=e.getSnapshot;e=e.value;try{var n=t();return!Ae(e,n)}catch{return!0}}function ia(e){var t=nt(e,1);t!==null&&We(t,e,1,-1)}function ua(e){var t=Xe();return typeof e=="function"&&(e=e()),t.memoizedState=t.baseState=e,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:lr,lastRenderedState:e},t.queue=e,e=e.dispatch=Rd.bind(null,X,e),[t.memoizedState,e]}function or(e,t,n,r){return e={tag:e,create:t,destroy:n,deps:r,next:null},t=X.updateQueue,t===null?(t={lastEffect:null,stores:null},X.updateQueue=t,t.lastEffect=e.next=e):(n=t.lastEffect,n===null?t.lastEffect=e.next=e:(r=n.next,n.next=e,e.next=r,t.lastEffect=e)),e}function sa(){return Me().memoizedState}function ol(e,t,n,r){var l=Xe();X.flags|=e,l.memoizedState=or(1|t,n,void 0,r===void 0?null:r)}function il(e,t,n,r){var l=Me();r=r===void 0?null:r;var o=void 0;if(te!==null){var i=te.memoizedState;if(o=i.destroy,r!==null&&ii(r,i.deps)){l.memoizedState=or(t,n,o,r);return}}X.flags|=e,l.memoizedState=or(1|t,n,o,r)}function aa(e,t){return ol(8390656,8,e,t)}function fi(e,t){return il(2048,8,e,t)}function ca(e,t){return il(4,2,e,t)}function fa(e,t){return il(4,4,e,t)}function da(e,t){if(typeof t=="function")return e=e(),t(e),function(){t(null)};if(t!=null)return e=e(),t.current=e,function(){t.current=null}}function pa(e,t,n){return n=n!=null?n.concat([e]):null,il(4,4,da.bind(null,t,e),n)}function di(){}function ma(e,t){var n=Me();t=t===void 0?null:t;var r=n.memoizedState;return r!==null&&t!==null&&ii(t,r[1])?r[0]:(n.memoizedState=[e,t],e)}function ha(e,t){var n=Me();t=t===void 0?null:t;var r=n.memoizedState;return r!==null&&t!==null&&ii(t,r[1])?r[0]:(e=e(),n.memoizedState=[e,t],e)}function va(e,t,n){return $t&21?(Ae(n,t)||(n=Ku(),X.lanes|=n,Vt|=n,e.baseState=!0),t):(e.baseState&&(e.baseState=!1,ke=!0),e.memoizedState=n)}function zd(e,t){var n=A;A=n!==0&&4>n?n:4,e(!0);var r=oi.transition;oi.transition={};try{e(!1),t()}finally{A=n,oi.transition=r}}function ya(){return Me().memoizedState}function Ld(e,t,n){var r=jt(e);if(n={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null},ga(e))ka(t,n);else if(n=Gs(e,t,n,r),n!==null){var l=ve();We(n,e,r,l),wa(n,t,r)}}function Rd(e,t,n){var r=jt(e),l={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null};if(ga(e))ka(t,l);else{var o=e.alternate;if(e.lanes===0&&(o===null||o.lanes===0)&&(o=t.lastRenderedReducer,o!==null))try{var i=t.lastRenderedState,u=o(i,n);if(l.hasEagerState=!0,l.eagerState=u,Ae(u,i)){var s=t.interleaved;s===null?(l.next=l,bo(t)):(l.next=s.next,s.next=l),t.interleaved=l;return}}catch{}finally{}n=Gs(e,t,l,r),n!==null&&(l=ve(),We(n,e,r,l),wa(n,t,r))}}function ga(e){var t=e.alternate;return e===X||t!==null&&t===X}function ka(e,t){nr=ll=!0;var n=e.pending;n===null?t.next=t:(t.next=n.next,n.next=t),e.pending=t}function wa(e,t,n){if(n&4194240){var r=t.lanes;r&=e.pendingLanes,n|=r,t.lanes=n,mo(e,n)}}var ul={readContext:Re,useCallback:ae,useContext:ae,useEffect:ae,useImperativeHandle:ae,useInsertionEffect:ae,useLayoutEffect:ae,useMemo:ae,useReducer:ae,useRef:ae,useState:ae,useDebugValue:ae,useDeferredValue:ae,useTransition:ae,useMutableSource:ae,useSyncExternalStore:ae,useId:ae,unstable_isNewReconciler:!1},Md={readContext:Re,useCallback:function(e,t){return Xe().memoizedState=[e,t===void 0?null:t],e},useContext:Re,useEffect:aa,useImperativeHandle:function(e,t,n){return n=n!=null?n.concat([e]):null,ol(4194308,4,da.bind(null,t,e),n)},useLayoutEffect:function(e,t){return ol(4194308,4,e,t)},useInsertionEffect:function(e,t){return ol(4,2,e,t)},useMemo:function(e,t){var n=Xe();return t=t===void 0?null:t,e=e(),n.memoizedState=[e,t],e},useReducer:function(e,t,n){var r=Xe();return t=n!==void 0?n(t):t,r.memoizedState=r.baseState=t,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:e,lastRenderedState:t},r.queue=e,e=e.dispatch=Ld.bind(null,X,e),[r.memoizedState,e]},useRef:function(e){var t=Xe();return e={current:e},t.memoizedState=e},useState:ua,useDebugValue:di,useDeferredValue:function(e){return Xe().memoizedState=e},useTransition:function(){var e=ua(!1),t=e[0];return e=zd.bind(null,e[1]),Xe().memoizedState=e,[t,e]},useMutableSource:function(){},useSyncExternalStore:function(e,t,n){var r=X,l=Xe();if(H){if(n===void 0)throw Error(k(407));n=n()}else{if(n=t(),le===null)throw Error(k(349));$t&30||na(r,t,n)}l.memoizedState=n;var o={value:n,getSnapshot:t};return l.queue=o,aa(la.bind(null,r,o,e),[e]),r.flags|=2048,or(9,ra.bind(null,r,o,n,t),void 0,null),n},useId:function(){var e=Xe(),t=le.identifierPrefix;if(H){var n=tt,r=et;n=(r&~(1<<32-Ie(r)-1)).toString(32)+n,t=":"+t+"R"+n,n=rr++,0<n&&(t+="H"+n.toString(32)),t+=":"}else n=Td++,t=":"+t+"r"+n.toString(32)+":";return e.memoizedState=t},unstable_isNewReconciler:!1},Od={readContext:Re,useCallback:ma,useContext:Re,useEffect:fi,useImperativeHandle:pa,useInsertionEffect:ca,useLayoutEffect:fa,useMemo:ha,useReducer:ai,useRef:sa,useState:function(){return ai(lr)},useDebugValue:di,useDeferredValue:function(e){var t=Me();return va(t,te.memoizedState,e)},useTransition:function(){var e=ai(lr)[0],t=Me().memoizedState;return[e,t]},useMutableSource:ea,useSyncExternalStore:ta,useId:ya,unstable_isNewReconciler:!1},Fd={readContext:Re,useCallback:ma,useContext:Re,useEffect:fi,useImperativeHandle:pa,useInsertionEffect:ca,useLayoutEffect:fa,useMemo:ha,useReducer:ci,useRef:sa,useState:function(){return ci(lr)},useDebugValue:di,useDeferredValue:function(e){var t=Me();return te===null?t.memoizedState=e:va(t,te.memoizedState,e)},useTransition:function(){var e=ci(lr)[0],t=Me().memoizedState;return[e,t]},useMutableSource:ea,useSyncExternalStore:ta,useId:ya,unstable_isNewReconciler:!1};function $e(e,t){if(e&&e.defaultProps){t=K({},t),e=e.defaultProps;for(var n in e)t[n]===void 0&&(t[n]=e[n]);return t}return t}function pi(e,t,n,r){t=e.memoizedState,n=n(r,t),n=n==null?t:K({},t,n),e.memoizedState=n,e.lanes===0&&(e.updateQueue.baseState=n)}var sl={isMounted:function(e){return(e=e._reactInternals)?Mt(e)===e:!1},enqueueSetState:function(e,t,n){e=e._reactInternals;var r=ve(),l=jt(e),o=rt(r,l);o.payload=t,n!=null&&(o.callback=n),t=_t(e,o,l),t!==null&&(We(t,e,l,r),el(t,e,l))},enqueueReplaceState:function(e,t,n){e=e._reactInternals;var r=ve(),l=jt(e),o=rt(r,l);o.tag=1,o.payload=t,n!=null&&(o.callback=n),t=_t(e,o,l),t!==null&&(We(t,e,l,r),el(t,e,l))},enqueueForceUpdate:function(e,t){e=e._reactInternals;var n=ve(),r=jt(e),l=rt(n,r);l.tag=2,t!=null&&(l.callback=t),t=_t(e,l,r),t!==null&&(We(t,e,r,n),el(t,e,r))}};function Sa(e,t,n,r,l,o,i){return e=e.stateNode,typeof e.shouldComponentUpdate=="function"?e.shouldComponentUpdate(r,o,i):t.prototype&&t.prototype.isPureReactComponent?!Qn(n,r)||!Qn(l,o):!0}function xa(e,t,n){var r=!1,l=wt,o=t.contextType;return typeof o=="object"&&o!==null?o=Re(o):(l=ge(t)?Ft:se.current,r=t.contextTypes,o=(r=r!=null)?cn(e,l):wt),t=new t(n,o),e.memoizedState=t.state!==null&&t.state!==void 0?t.state:null,t.updater=sl,e.stateNode=t,t._reactInternals=e,r&&(e=e.stateNode,e.__reactInternalMemoizedUnmaskedChildContext=l,e.__reactInternalMemoizedMaskedChildContext=o),t}function _a(e,t,n,r){e=t.state,typeof t.componentWillReceiveProps=="function"&&t.componentWillReceiveProps(n,r),typeof t.UNSAFE_componentWillReceiveProps=="function"&&t.UNSAFE_componentWillReceiveProps(n,r),t.state!==e&&sl.enqueueReplaceState(t,t.state,null)}function mi(e,t,n,r){var l=e.stateNode;l.props=n,l.state=e.memoizedState,l.refs={},ei(e);var o=t.contextType;typeof o=="object"&&o!==null?l.context=Re(o):(o=ge(t)?Ft:se.current,l.context=cn(e,o)),l.state=e.memoizedState,o=t.getDerivedStateFromProps,typeof o=="function"&&(pi(e,t,o,n),l.state=e.memoizedState),typeof t.getDerivedStateFromProps=="function"||typeof l.getSnapshotBeforeUpdate=="function"||typeof l.UNSAFE_componentWillMount!="function"&&typeof l.componentWillMount!="function"||(t=l.state,typeof l.componentWillMount=="function"&&l.componentWillMount(),typeof l.UNSAFE_componentWillMount=="function"&&l.UNSAFE_componentWillMount(),t!==l.state&&sl.enqueueReplaceState(l,l.state,null),tl(e,n,l,r),l.state=e.memoizedState),typeof l.componentDidMount=="function"&&(e.flags|=4194308)}function gn(e,t){try{var n="",r=t;do n+=af(r),r=r.return;while(r);var l=n}catch(o){l=`
 Error generating stack: `+o.message+`
-`+o.stack}return{value:e,source:t,stack:l,digest:null}}function hi(e,t,n){return{value:e,source:null,stack:n??null,digest:t??null}}function vi(e,t){try{console.error(t.value)}catch(n){setTimeout(function(){throw n})}}var Dd=typeof WeakMap=="function"?WeakMap:Map;function Ea(e,t,n){n=lt(-1,n),n.tag=3,n.payload={element:null};var r=t.value;return n.callback=function(){hl||(hl=!0,Li=r),vi(e,t)},n}function Ca(e,t,n){n=lt(-1,n),n.tag=3;var r=e.type.getDerivedStateFromError;if(typeof r=="function"){var l=t.value;n.payload=function(){return r(l)},n.callback=function(){vi(e,t)}}var o=e.stateNode;return o!==null&&typeof o.componentDidCatch=="function"&&(n.callback=function(){vi(e,t),typeof r!="function"&&(Ct===null?Ct=new Set([this]):Ct.add(this));var i=t.stack;this.componentDidCatch(t.value,{componentStack:i!==null?i:""})}),n}function Na(e,t,n){var r=e.pingCache;if(r===null){r=e.pingCache=new Dd;var l=new Set;r.set(t,l)}else l=r.get(t),l===void 0&&(l=new Set,r.set(t,l));l.has(n)||(l.add(n),e=Zd.bind(null,e,t,n),t.then(e,e))}function ja(e){do{var t;if((t=e.tag===13)&&(t=e.memoizedState,t=t!==null?t.dehydrated!==null:!0),t)return e;e=e.return}while(e!==null);return null}function Pa(e,t,n,r,l){return e.mode&1?(e.flags|=65536,e.lanes=l,e):(e===t?e.flags|=65536:(e.flags|=128,n.flags|=131072,n.flags&=-52805,n.tag===1&&(n.alternate===null?n.tag=17:(t=lt(-1,1),t.tag=2,_t(n,t,1))),n.lanes|=1),e)}var Id=qe.ReactCurrentOwner,we=!1;function he(e,t,n,r){t.child=e===null?Xs(t,null,n,r):mn(t,e.child,n,r)}function Ta(e,t,n,r,l){n=n.render;var o=t.ref;return vn(t,l),r=ui(e,t,n,r,o,l),n=si(),e!==null&&!we?(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~l,ot(e,t,l)):(H&&n&&Ho(t),t.flags|=1,he(e,t,r,l),t.child)}function za(e,t,n,r,l){if(e===null){var o=n.type;return typeof o=="function"&&!Ai(o)&&o.defaultProps===void 0&&n.compare===null&&n.defaultProps===void 0?(t.tag=15,t.type=o,La(e,t,o,r,l)):(e=Sl(n.type,null,r,t,t.mode,l),e.ref=t.ref,e.return=t,t.child=e)}if(o=e.child,!(e.lanes&l)){var i=o.memoizedProps;if(n=n.compare,n=n!==null?n:Qn,n(i,r)&&e.ref===t.ref)return ot(e,t,l)}return t.flags|=1,e=Tt(o,r),e.ref=t.ref,e.return=t,t.child=e}function La(e,t,n,r,l){if(e!==null){var o=e.memoizedProps;if(Qn(o,r)&&e.ref===t.ref)if(we=!1,t.pendingProps=r=o,(e.lanes&l)!==0)e.flags&131072&&(we=!0);else return t.lanes=e.lanes,ot(e,t,l)}return yi(e,t,n,r,l)}function Ra(e,t,n){var r=t.pendingProps,l=r.children,o=e!==null?e.memoizedState:null;if(r.mode==="hidden")if(!(t.mode&1))t.memoizedState={baseLanes:0,cachePool:null,transitions:null},$(kn,Pe),Pe|=n;else{if(!(n&1073741824))return e=o!==null?o.baseLanes|n:n,t.lanes=t.childLanes=1073741824,t.memoizedState={baseLanes:e,cachePool:null,transitions:null},t.updateQueue=null,$(kn,Pe),Pe|=e,null;t.memoizedState={baseLanes:0,cachePool:null,transitions:null},r=o!==null?o.baseLanes:n,$(kn,Pe),Pe|=r}else o!==null?(r=o.baseLanes|n,t.memoizedState=null):r=n,$(kn,Pe),Pe|=r;return he(e,t,l,n),t.child}function Oa(e,t){var n=t.ref;(e===null&&n!==null||e!==null&&e.ref!==n)&&(t.flags|=512,t.flags|=2097152)}function yi(e,t,n,r,l){var o=ge(n)?Ft:se.current;return o=cn(t,o),vn(t,l),n=ui(e,t,n,r,o,l),r=si(),e!==null&&!we?(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~l,ot(e,t,l)):(H&&r&&Ho(t),t.flags|=1,he(e,t,n,l),t.child)}function Ma(e,t,n,r,l){if(ge(n)){var o=!0;Kr(t)}else o=!1;if(vn(t,l),t.stateNode===null)cl(e,t),xa(t,n,r),mi(t,n,r,l),r=!0;else if(e===null){var i=t.stateNode,u=t.memoizedProps;i.props=u;var s=i.context,c=n.contextType;typeof c=="object"&&c!==null?c=Re(c):(c=ge(n)?Ft:se.current,c=cn(t,c));var y=n.getDerivedStateFromProps,v=typeof y=="function"||typeof i.getSnapshotBeforeUpdate=="function";v||typeof i.UNSAFE_componentWillReceiveProps!="function"&&typeof i.componentWillReceiveProps!="function"||(u!==r||s!==c)&&_a(t,i,r,c),xt=!1;var m=t.memoizedState;i.state=m,tl(t,r,i,l),s=t.memoizedState,u!==r||m!==s||ye.current||xt?(typeof y=="function"&&(pi(t,n,y,r),s=t.memoizedState),(u=xt||Sa(t,n,u,r,m,s,c))?(v||typeof i.UNSAFE_componentWillMount!="function"&&typeof i.componentWillMount!="function"||(typeof i.componentWillMount=="function"&&i.componentWillMount(),typeof i.UNSAFE_componentWillMount=="function"&&i.UNSAFE_componentWillMount()),typeof i.componentDidMount=="function"&&(t.flags|=4194308)):(typeof i.componentDidMount=="function"&&(t.flags|=4194308),t.memoizedProps=r,t.memoizedState=s),i.props=r,i.state=s,i.context=c,r=u):(typeof i.componentDidMount=="function"&&(t.flags|=4194308),r=!1)}else{i=t.stateNode,Zs(e,t),u=t.memoizedProps,c=t.type===t.elementType?u:$e(t.type,u),i.props=c,v=t.pendingProps,m=i.context,s=n.contextType,typeof s=="object"&&s!==null?s=Re(s):(s=ge(n)?Ft:se.current,s=cn(t,s));var k=n.getDerivedStateFromProps;(y=typeof k=="function"||typeof i.getSnapshotBeforeUpdate=="function")||typeof i.UNSAFE_componentWillReceiveProps!="function"&&typeof i.componentWillReceiveProps!="function"||(u!==v||m!==s)&&_a(t,i,r,s),xt=!1,m=t.memoizedState,i.state=m,tl(t,r,i,l);var S=t.memoizedState;u!==v||m!==S||ye.current||xt?(typeof k=="function"&&(pi(t,n,k,r),S=t.memoizedState),(c=xt||Sa(t,n,c,r,m,S,s)||!1)?(y||typeof i.UNSAFE_componentWillUpdate!="function"&&typeof i.componentWillUpdate!="function"||(typeof i.componentWillUpdate=="function"&&i.componentWillUpdate(r,S,s),typeof i.UNSAFE_componentWillUpdate=="function"&&i.UNSAFE_componentWillUpdate(r,S,s)),typeof i.componentDidUpdate=="function"&&(t.flags|=4),typeof i.getSnapshotBeforeUpdate=="function"&&(t.flags|=1024)):(typeof i.componentDidUpdate!="function"||u===e.memoizedProps&&m===e.memoizedState||(t.flags|=4),typeof i.getSnapshotBeforeUpdate!="function"||u===e.memoizedProps&&m===e.memoizedState||(t.flags|=1024),t.memoizedProps=r,t.memoizedState=S),i.props=r,i.state=S,i.context=s,r=c):(typeof i.componentDidUpdate!="function"||u===e.memoizedProps&&m===e.memoizedState||(t.flags|=4),typeof i.getSnapshotBeforeUpdate!="function"||u===e.memoizedProps&&m===e.memoizedState||(t.flags|=1024),r=!1)}return gi(e,t,n,r,o,l)}function gi(e,t,n,r,l,o){Oa(e,t);var i=(t.flags&128)!==0;if(!r&&!i)return l&&Us(t,n,!1),ot(e,t,o);r=t.stateNode,Id.current=t;var u=i&&typeof n.getDerivedStateFromError!="function"?null:r.render();return t.flags|=1,e!==null&&i?(t.child=mn(t,e.child,null,o),t.child=mn(t,null,u,o)):he(e,t,u,o),t.memoizedState=r.state,l&&Us(t,n,!0),t.child}function Fa(e){var t=e.stateNode;t.pendingContext?Is(e,t.pendingContext,t.pendingContext!==t.context):t.context&&Is(e,t.context,!1),ti(e,t.containerInfo)}function Da(e,t,n,r,l){return pn(),Xo(l),t.flags|=256,he(e,t,n,r),t.child}var wi={dehydrated:null,treeContext:null,retryLane:0};function ki(e){return{baseLanes:e,cachePool:null,transitions:null}}function Ia(e,t,n){var r=t.pendingProps,l=Y.current,o=!1,i=(t.flags&128)!==0,u;if((u=i)||(u=e!==null&&e.memoizedState===null?!1:(l&2)!==0),u?(o=!0,t.flags&=-129):(e===null||e.memoizedState!==null)&&(l|=1),$(Y,l&1),e===null)return Yo(t),e=t.memoizedState,e!==null&&(e=e.dehydrated,e!==null)?(t.mode&1?e.data==="$!"?t.lanes=8:t.lanes=1073741824:t.lanes=1,null):(i=r.children,e=r.fallback,o?(r=t.mode,o=t.child,i={mode:"hidden",children:i},!(r&1)&&o!==null?(o.childLanes=0,o.pendingProps=i):o=xl(i,r,0,null),e=Qt(e,r,n,null),o.return=t,e.return=t,o.sibling=e,t.child=o,t.child.memoizedState=ki(n),t.memoizedState=wi,e):Si(t,i));if(l=e.memoizedState,l!==null&&(u=l.dehydrated,u!==null))return Ad(e,t,i,r,u,l,n);if(o){o=r.fallback,i=t.mode,l=e.child,u=l.sibling;var s={mode:"hidden",children:r.children};return!(i&1)&&t.child!==l?(r=t.child,r.childLanes=0,r.pendingProps=s,t.deletions=null):(r=Tt(l,s),r.subtreeFlags=l.subtreeFlags&14680064),u!==null?o=Tt(u,o):(o=Qt(o,i,n,null),o.flags|=2),o.return=t,r.return=t,r.sibling=o,t.child=r,r=o,o=t.child,i=e.child.memoizedState,i=i===null?ki(n):{baseLanes:i.baseLanes|n,cachePool:null,transitions:i.transitions},o.memoizedState=i,o.childLanes=e.childLanes&~n,t.memoizedState=wi,r}return o=e.child,e=o.sibling,r=Tt(o,{mode:"visible",children:r.children}),!(t.mode&1)&&(r.lanes=n),r.return=t,r.sibling=null,e!==null&&(n=t.deletions,n===null?(t.deletions=[e],t.flags|=16):n.push(e)),t.child=r,t.memoizedState=null,r}function Si(e,t){return t=xl({mode:"visible",children:t},e.mode,0,null),t.return=e,e.child=t}function al(e,t,n,r){return r!==null&&Xo(r),mn(t,e.child,null,n),e=Si(t,t.pendingProps.children),e.flags|=2,t.memoizedState=null,e}function Ad(e,t,n,r,l,o,i){if(n)return t.flags&256?(t.flags&=-257,r=hi(Error(w(422))),al(e,t,i,r)):t.memoizedState!==null?(t.child=e.child,t.flags|=128,null):(o=r.fallback,l=t.mode,r=xl({mode:"visible",children:r.children},l,0,null),o=Qt(o,l,i,null),o.flags|=2,r.return=t,o.return=t,r.sibling=o,t.child=r,t.mode&1&&mn(t,e.child,null,i),t.child.memoizedState=ki(i),t.memoizedState=wi,o);if(!(t.mode&1))return al(e,t,i,null);if(l.data==="$!"){if(r=l.nextSibling&&l.nextSibling.dataset,r)var u=r.dgst;return r=u,o=Error(w(419)),r=hi(o,r,void 0),al(e,t,i,r)}if(u=(i&e.childLanes)!==0,we||u){if(r=le,r!==null){switch(i&-i){case 4:l=2;break;case 16:l=8;break;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:l=32;break;case 536870912:l=268435456;break;default:l=0}l=l&(r.suspendedLanes|i)?0:l,l!==0&&l!==o.retryLane&&(o.retryLane=l,rt(e,l),We(r,e,l,-1))}return Ii(),r=hi(Error(w(421))),al(e,t,i,r)}return l.data==="$?"?(t.flags|=128,t.child=e.child,t=Jd.bind(null,e),l._reactRetry=t,null):(e=o.treeContext,je=gt(l.nextSibling),Ne=t,H=!0,Ue=null,e!==null&&(ze[Le++]=tt,ze[Le++]=nt,ze[Le++]=Dt,tt=e.id,nt=e.overflow,Dt=t),t=Si(t,r.children),t.flags|=4096,t)}function Aa(e,t,n){e.lanes|=t;var r=e.alternate;r!==null&&(r.lanes|=t),qo(e.return,t,n)}function xi(e,t,n,r,l){var o=e.memoizedState;o===null?e.memoizedState={isBackwards:t,rendering:null,renderingStartTime:0,last:r,tail:n,tailMode:l}:(o.isBackwards=t,o.rendering=null,o.renderingStartTime=0,o.last=r,o.tail=n,o.tailMode=l)}function Ua(e,t,n){var r=t.pendingProps,l=r.revealOrder,o=r.tail;if(he(e,t,r.children,n),r=Y.current,r&2)r=r&1|2,t.flags|=128;else{if(e!==null&&e.flags&128)e:for(e=t.child;e!==null;){if(e.tag===13)e.memoizedState!==null&&Aa(e,n,t);else if(e.tag===19)Aa(e,n,t);else if(e.child!==null){e.child.return=e,e=e.child;continue}if(e===t)break e;for(;e.sibling===null;){if(e.return===null||e.return===t)break e;e=e.return}e.sibling.return=e.return,e=e.sibling}r&=1}if($(Y,r),!(t.mode&1))t.memoizedState=null;else switch(l){case"forwards":for(n=t.child,l=null;n!==null;)e=n.alternate,e!==null&&nl(e)===null&&(l=n),n=n.sibling;n=l,n===null?(l=t.child,t.child=null):(l=n.sibling,n.sibling=null),xi(t,!1,l,n,o);break;case"backwards":for(n=null,l=t.child,t.child=null;l!==null;){if(e=l.alternate,e!==null&&nl(e)===null){t.child=l;break}e=l.sibling,l.sibling=n,n=l,l=e}xi(t,!0,n,null,o);break;case"together":xi(t,!1,null,null,void 0);break;default:t.memoizedState=null}return t.child}function cl(e,t){!(t.mode&1)&&e!==null&&(e.alternate=null,t.alternate=null,t.flags|=2)}function ot(e,t,n){if(e!==null&&(t.dependencies=e.dependencies),Vt|=t.lanes,!(n&t.childLanes))return null;if(e!==null&&t.child!==e.child)throw Error(w(153));if(t.child!==null){for(e=t.child,n=Tt(e,e.pendingProps),t.child=n,n.return=t;e.sibling!==null;)e=e.sibling,n=n.sibling=Tt(e,e.pendingProps),n.return=t;n.sibling=null}return t.child}function Ud(e,t,n){switch(t.tag){case 3:Fa(t),pn();break;case 5:bs(t);break;case 1:ge(t.type)&&Kr(t);break;case 4:ti(t,t.stateNode.containerInfo);break;case 10:var r=t.type._context,l=t.memoizedProps.value;$(qr,r._currentValue),r._currentValue=l;break;case 13:if(r=t.memoizedState,r!==null)return r.dehydrated!==null?($(Y,Y.current&1),t.flags|=128,null):n&t.child.childLanes?Ia(e,t,n):($(Y,Y.current&1),e=ot(e,t,n),e!==null?e.sibling:null);$(Y,Y.current&1);break;case 19:if(r=(n&t.childLanes)!==0,e.flags&128){if(r)return Ua(e,t,n);t.flags|=128}if(l=t.memoizedState,l!==null&&(l.rendering=null,l.tail=null,l.lastEffect=null),$(Y,Y.current),r)break;return null;case 22:case 23:return t.lanes=0,Ra(e,t,n)}return ot(e,t,n)}var $a,_i,Va,Ba;$a=function(e,t){for(var n=t.child;n!==null;){if(n.tag===5||n.tag===6)e.appendChild(n.stateNode);else if(n.tag!==4&&n.child!==null){n.child.return=n,n=n.child;continue}if(n===t)break;for(;n.sibling===null;){if(n.return===null||n.return===t)return;n=n.return}n.sibling.return=n.return,n=n.sibling}},_i=function(){},Va=function(e,t,n,r){var l=e.memoizedProps;if(l!==r){e=t.stateNode,Ut(Ye.current);var o=null;switch(n){case"input":l=Jl(e,l),r=Jl(e,r),o=[];break;case"select":l=K({},l,{value:void 0}),r=K({},r,{value:void 0}),o=[];break;case"textarea":l=eo(e,l),r=eo(e,r),o=[];break;default:typeof l.onClick!="function"&&typeof r.onClick=="function"&&(e.onclick=Wr)}no(n,r);var i;n=null;for(c in l)if(!r.hasOwnProperty(c)&&l.hasOwnProperty(c)&&l[c]!=null)if(c==="style"){var u=l[c];for(i in u)u.hasOwnProperty(i)&&(n||(n={}),n[i]="")}else c!=="dangerouslySetInnerHTML"&&c!=="children"&&c!=="suppressContentEditableWarning"&&c!=="suppressHydrationWarning"&&c!=="autoFocus"&&(En.hasOwnProperty(c)?o||(o=[]):(o=o||[]).push(c,null));for(c in r){var s=r[c];if(u=l!=null?l[c]:void 0,r.hasOwnProperty(c)&&s!==u&&(s!=null||u!=null))if(c==="style")if(u){for(i in u)!u.hasOwnProperty(i)||s&&s.hasOwnProperty(i)||(n||(n={}),n[i]="");for(i in s)s.hasOwnProperty(i)&&u[i]!==s[i]&&(n||(n={}),n[i]=s[i])}else n||(o||(o=[]),o.push(c,n)),n=s;else c==="dangerouslySetInnerHTML"?(s=s?s.__html:void 0,u=u?u.__html:void 0,s!=null&&u!==s&&(o=o||[]).push(c,s)):c==="children"?typeof s!="string"&&typeof s!="number"||(o=o||[]).push(c,""+s):c!=="suppressContentEditableWarning"&&c!=="suppressHydrationWarning"&&(En.hasOwnProperty(c)?(s!=null&&c==="onScroll"&&V("scroll",e),o||u===s||(o=[])):(o=o||[]).push(c,s))}n&&(o=o||[]).push("style",n);var c=o;(t.updateQueue=c)&&(t.flags|=4)}},Ba=function(e,t,n,r){n!==r&&(t.flags|=4)};function ir(e,t){if(!H)switch(e.tailMode){case"hidden":t=e.tail;for(var n=null;t!==null;)t.alternate!==null&&(n=t),t=t.sibling;n===null?e.tail=null:n.sibling=null;break;case"collapsed":n=e.tail;for(var r=null;n!==null;)n.alternate!==null&&(r=n),n=n.sibling;r===null?t||e.tail===null?e.tail=null:e.tail.sibling=null:r.sibling=null}}function ce(e){var t=e.alternate!==null&&e.alternate.child===e.child,n=0,r=0;if(t)for(var l=e.child;l!==null;)n|=l.lanes|l.childLanes,r|=l.subtreeFlags&14680064,r|=l.flags&14680064,l.return=e,l=l.sibling;else for(l=e.child;l!==null;)n|=l.lanes|l.childLanes,r|=l.subtreeFlags,r|=l.flags,l.return=e,l=l.sibling;return e.subtreeFlags|=r,e.childLanes=n,t}function $d(e,t,n){var r=t.pendingProps;switch(Qo(t),t.tag){case 2:case 16:case 15:case 0:case 11:case 7:case 8:case 12:case 9:case 14:return ce(t),null;case 1:return ge(t.type)&&Qr(),ce(t),null;case 3:return r=t.stateNode,yn(),B(ye),B(se),li(),r.pendingContext&&(r.context=r.pendingContext,r.pendingContext=null),(e===null||e.child===null)&&(Zr(t)?t.flags|=4:e===null||e.memoizedState.isDehydrated&&!(t.flags&256)||(t.flags|=1024,Ue!==null&&(Mi(Ue),Ue=null))),_i(e,t),ce(t),null;case 5:ni(t);var l=Ut(tr.current);if(n=t.type,e!==null&&t.stateNode!=null)Va(e,t,n,r,l),e.ref!==t.ref&&(t.flags|=512,t.flags|=2097152);else{if(!r){if(t.stateNode===null)throw Error(w(166));return ce(t),null}if(e=Ut(Ye.current),Zr(t)){r=t.stateNode,n=t.type;var o=t.memoizedProps;switch(r[Ke]=t,r[Zn]=o,e=(t.mode&1)!==0,n){case"dialog":V("cancel",r),V("close",r);break;case"iframe":case"object":case"embed":V("load",r);break;case"video":case"audio":for(l=0;l<Yn.length;l++)V(Yn[l],r);break;case"source":V("error",r);break;case"img":case"image":case"link":V("error",r),V("load",r);break;case"details":V("toggle",r);break;case"input":xu(r,o),V("invalid",r);break;case"select":r._wrapperState={wasMultiple:!!o.multiple},V("invalid",r);break;case"textarea":Cu(r,o),V("invalid",r)}no(n,o),l=null;for(var i in o)if(o.hasOwnProperty(i)){var u=o[i];i==="children"?typeof u=="string"?r.textContent!==u&&(o.suppressHydrationWarning!==!0&&Br(r.textContent,u,e),l=["children",u]):typeof u=="number"&&r.textContent!==""+u&&(o.suppressHydrationWarning!==!0&&Br(r.textContent,u,e),l=["children",""+u]):En.hasOwnProperty(i)&&u!=null&&i==="onScroll"&&V("scroll",r)}switch(n){case"input":wr(r),Eu(r,o,!0);break;case"textarea":wr(r),ju(r);break;case"select":case"option":break;default:typeof o.onClick=="function"&&(r.onclick=Wr)}r=l,t.updateQueue=r,r!==null&&(t.flags|=4)}else{i=l.nodeType===9?l:l.ownerDocument,e==="http://www.w3.org/1999/xhtml"&&(e=Pu(n)),e==="http://www.w3.org/1999/xhtml"?n==="script"?(e=i.createElement("div"),e.innerHTML="<script><\/script>",e=e.removeChild(e.firstChild)):typeof r.is=="string"?e=i.createElement(n,{is:r.is}):(e=i.createElement(n),n==="select"&&(i=e,r.multiple?i.multiple=!0:r.size&&(i.size=r.size))):e=i.createElementNS(e,n),e[Ke]=t,e[Zn]=r,$a(e,t,!1,!1),t.stateNode=e;e:{switch(i=ro(n,r),n){case"dialog":V("cancel",e),V("close",e),l=r;break;case"iframe":case"object":case"embed":V("load",e),l=r;break;case"video":case"audio":for(l=0;l<Yn.length;l++)V(Yn[l],e);l=r;break;case"source":V("error",e),l=r;break;case"img":case"image":case"link":V("error",e),V("load",e),l=r;break;case"details":V("toggle",e),l=r;break;case"input":xu(e,r),l=Jl(e,r),V("invalid",e);break;case"option":l=r;break;case"select":e._wrapperState={wasMultiple:!!r.multiple},l=K({},r,{value:void 0}),V("invalid",e);break;case"textarea":Cu(e,r),l=eo(e,r),V("invalid",e);break;default:l=r}no(n,l),u=l;for(o in u)if(u.hasOwnProperty(o)){var s=u[o];o==="style"?Lu(e,s):o==="dangerouslySetInnerHTML"?(s=s?s.__html:void 0,s!=null&&Tu(e,s)):o==="children"?typeof s=="string"?(n!=="textarea"||s!=="")&&Pn(e,s):typeof s=="number"&&Pn(e,""+s):o!=="suppressContentEditableWarning"&&o!=="suppressHydrationWarning"&&o!=="autoFocus"&&(En.hasOwnProperty(o)?s!=null&&o==="onScroll"&&V("scroll",e):s!=null&&$l(e,o,s,i))}switch(n){case"input":wr(e),Eu(e,r,!1);break;case"textarea":wr(e),ju(e);break;case"option":r.value!=null&&e.setAttribute("value",""+ft(r.value));break;case"select":e.multiple=!!r.multiple,o=r.value,o!=null?Jt(e,!!r.multiple,o,!1):r.defaultValue!=null&&Jt(e,!!r.multiple,r.defaultValue,!0);break;default:typeof l.onClick=="function"&&(e.onclick=Wr)}switch(n){case"button":case"input":case"select":case"textarea":r=!!r.autoFocus;break e;case"img":r=!0;break e;default:r=!1}}r&&(t.flags|=4)}t.ref!==null&&(t.flags|=512,t.flags|=2097152)}return ce(t),null;case 6:if(e&&t.stateNode!=null)Ba(e,t,e.memoizedProps,r);else{if(typeof r!="string"&&t.stateNode===null)throw Error(w(166));if(n=Ut(tr.current),Ut(Ye.current),Zr(t)){if(r=t.stateNode,n=t.memoizedProps,r[Ke]=t,(o=r.nodeValue!==n)&&(e=Ne,e!==null))switch(e.tag){case 3:Br(r.nodeValue,n,(e.mode&1)!==0);break;case 5:e.memoizedProps.suppressHydrationWarning!==!0&&Br(r.nodeValue,n,(e.mode&1)!==0)}o&&(t.flags|=4)}else r=(n.nodeType===9?n:n.ownerDocument).createTextNode(r),r[Ke]=t,t.stateNode=r}return ce(t),null;case 13:if(B(Y),r=t.memoizedState,e===null||e.memoizedState!==null&&e.memoizedState.dehydrated!==null){if(H&&je!==null&&t.mode&1&&!(t.flags&128))Qs(),pn(),t.flags|=98560,o=!1;else if(o=Zr(t),r!==null&&r.dehydrated!==null){if(e===null){if(!o)throw Error(w(318));if(o=t.memoizedState,o=o!==null?o.dehydrated:null,!o)throw Error(w(317));o[Ke]=t}else pn(),!(t.flags&128)&&(t.memoizedState=null),t.flags|=4;ce(t),o=!1}else Ue!==null&&(Mi(Ue),Ue=null),o=!0;if(!o)return t.flags&65536?t:null}return t.flags&128?(t.lanes=n,t):(r=r!==null,r!==(e!==null&&e.memoizedState!==null)&&r&&(t.child.flags|=8192,t.mode&1&&(e===null||Y.current&1?ne===0&&(ne=3):Ii())),t.updateQueue!==null&&(t.flags|=4),ce(t),null);case 4:return yn(),_i(e,t),e===null&&Xn(t.stateNode.containerInfo),ce(t),null;case 10:return Jo(t.type._context),ce(t),null;case 17:return ge(t.type)&&Qr(),ce(t),null;case 19:if(B(Y),o=t.memoizedState,o===null)return ce(t),null;if(r=(t.flags&128)!==0,i=o.rendering,i===null)if(r)ir(o,!1);else{if(ne!==0||e!==null&&e.flags&128)for(e=t.child;e!==null;){if(i=nl(e),i!==null){for(t.flags|=128,ir(o,!1),r=i.updateQueue,r!==null&&(t.updateQueue=r,t.flags|=4),t.subtreeFlags=0,r=n,n=t.child;n!==null;)o=n,e=r,o.flags&=14680066,i=o.alternate,i===null?(o.childLanes=0,o.lanes=e,o.child=null,o.subtreeFlags=0,o.memoizedProps=null,o.memoizedState=null,o.updateQueue=null,o.dependencies=null,o.stateNode=null):(o.childLanes=i.childLanes,o.lanes=i.lanes,o.child=i.child,o.subtreeFlags=0,o.deletions=null,o.memoizedProps=i.memoizedProps,o.memoizedState=i.memoizedState,o.updateQueue=i.updateQueue,o.type=i.type,e=i.dependencies,o.dependencies=e===null?null:{lanes:e.lanes,firstContext:e.firstContext}),n=n.sibling;return $(Y,Y.current&1|2),t.child}e=e.sibling}o.tail!==null&&J()>Sn&&(t.flags|=128,r=!0,ir(o,!1),t.lanes=4194304)}else{if(!r)if(e=nl(i),e!==null){if(t.flags|=128,r=!0,n=e.updateQueue,n!==null&&(t.updateQueue=n,t.flags|=4),ir(o,!0),o.tail===null&&o.tailMode==="hidden"&&!i.alternate&&!H)return ce(t),null}else 2*J()-o.renderingStartTime>Sn&&n!==1073741824&&(t.flags|=128,r=!0,ir(o,!1),t.lanes=4194304);o.isBackwards?(i.sibling=t.child,t.child=i):(n=o.last,n!==null?n.sibling=i:t.child=i,o.last=i)}return o.tail!==null?(t=o.tail,o.rendering=t,o.tail=t.sibling,o.renderingStartTime=J(),t.sibling=null,n=Y.current,$(Y,r?n&1|2:n&1),t):(ce(t),null);case 22:case 23:return Di(),r=t.memoizedState!==null,e!==null&&e.memoizedState!==null!==r&&(t.flags|=8192),r&&t.mode&1?Pe&1073741824&&(ce(t),t.subtreeFlags&6&&(t.flags|=8192)):ce(t),null;case 24:return null;case 25:return null}throw Error(w(156,t.tag))}function Vd(e,t){switch(Qo(t),t.tag){case 1:return ge(t.type)&&Qr(),e=t.flags,e&65536?(t.flags=e&-65537|128,t):null;case 3:return yn(),B(ye),B(se),li(),e=t.flags,e&65536&&!(e&128)?(t.flags=e&-65537|128,t):null;case 5:return ni(t),null;case 13:if(B(Y),e=t.memoizedState,e!==null&&e.dehydrated!==null){if(t.alternate===null)throw Error(w(340));pn()}return e=t.flags,e&65536?(t.flags=e&-65537|128,t):null;case 19:return B(Y),null;case 4:return yn(),null;case 10:return Jo(t.type._context),null;case 22:case 23:return Di(),null;case 24:return null;default:return null}}var fl=!1,fe=!1,Bd=typeof WeakSet=="function"?WeakSet:Set,j=null;function wn(e,t){var n=e.ref;if(n!==null)if(typeof n=="function")try{n(null)}catch(r){Z(e,t,r)}else n.current=null}function Ei(e,t,n){try{n()}catch(r){Z(e,t,r)}}var Wa=!1;function Wd(e,t){if(Do=Lr,e=Ss(),Po(e)){if("selectionStart"in e)var n={start:e.selectionStart,end:e.selectionEnd};else e:{n=(n=e.ownerDocument)&&n.defaultView||window;var r=n.getSelection&&n.getSelection();if(r&&r.rangeCount!==0){n=r.anchorNode;var l=r.anchorOffset,o=r.focusNode;r=r.focusOffset;try{n.nodeType,o.nodeType}catch{n=null;break e}var i=0,u=-1,s=-1,c=0,y=0,v=e,m=null;t:for(;;){for(var k;v!==n||l!==0&&v.nodeType!==3||(u=i+l),v!==o||r!==0&&v.nodeType!==3||(s=i+r),v.nodeType===3&&(i+=v.nodeValue.length),(k=v.firstChild)!==null;)m=v,v=k;for(;;){if(v===e)break t;if(m===n&&++c===l&&(u=i),m===o&&++y===r&&(s=i),(k=v.nextSibling)!==null)break;v=m,m=v.parentNode}v=k}n=u===-1||s===-1?null:{start:u,end:s}}else n=null}n=n||{start:0,end:0}}else n=null;for(Io={focusedElem:e,selectionRange:n},Lr=!1,j=t;j!==null;)if(t=j,e=t.child,(t.subtreeFlags&1028)!==0&&e!==null)e.return=t,j=e;else for(;j!==null;){t=j;try{var S=t.alternate;if(t.flags&1024)switch(t.tag){case 0:case 11:case 15:break;case 1:if(S!==null){var C=S.memoizedProps,A=S.memoizedState,f=t.stateNode,a=f.getSnapshotBeforeUpdate(t.elementType===t.type?C:$e(t.type,C),A);f.__reactInternalSnapshotBeforeUpdate=a}break;case 3:var d=t.stateNode.containerInfo;d.nodeType===1?d.textContent="":d.nodeType===9&&d.documentElement&&d.removeChild(d.documentElement);break;case 5:case 6:case 4:case 17:break;default:throw Error(w(163))}}catch(g){Z(t,t.return,g)}if(e=t.sibling,e!==null){e.return=t.return,j=e;break}j=t.return}return S=Wa,Wa=!1,S}function ur(e,t,n){var r=t.updateQueue;if(r=r!==null?r.lastEffect:null,r!==null){var l=r=r.next;do{if((l.tag&e)===e){var o=l.destroy;l.destroy=void 0,o!==void 0&&Ei(t,n,o)}l=l.next}while(l!==r)}}function dl(e,t){if(t=t.updateQueue,t=t!==null?t.lastEffect:null,t!==null){var n=t=t.next;do{if((n.tag&e)===e){var r=n.create;n.destroy=r()}n=n.next}while(n!==t)}}function Ci(e){var t=e.ref;if(t!==null){var n=e.stateNode;switch(e.tag){case 5:e=n;break;default:e=n}typeof t=="function"?t(e):t.current=e}}function Ha(e){var t=e.alternate;t!==null&&(e.alternate=null,Ha(t)),e.child=null,e.deletions=null,e.sibling=null,e.tag===5&&(t=e.stateNode,t!==null&&(delete t[Ke],delete t[Zn],delete t[Vo],delete t[Cd],delete t[Nd])),e.stateNode=null,e.return=null,e.dependencies=null,e.memoizedProps=null,e.memoizedState=null,e.pendingProps=null,e.stateNode=null,e.updateQueue=null}function Qa(e){return e.tag===5||e.tag===3||e.tag===4}function Ka(e){e:for(;;){for(;e.sibling===null;){if(e.return===null||Qa(e.return))return null;e=e.return}for(e.sibling.return=e.return,e=e.sibling;e.tag!==5&&e.tag!==6&&e.tag!==18;){if(e.flags&2||e.child===null||e.tag===4)continue e;e.child.return=e,e=e.child}if(!(e.flags&2))return e.stateNode}}function Ni(e,t,n){var r=e.tag;if(r===5||r===6)e=e.stateNode,t?n.nodeType===8?n.parentNode.insertBefore(e,t):n.insertBefore(e,t):(n.nodeType===8?(t=n.parentNode,t.insertBefore(e,n)):(t=n,t.appendChild(e)),n=n._reactRootContainer,n!=null||t.onclick!==null||(t.onclick=Wr));else if(r!==4&&(e=e.child,e!==null))for(Ni(e,t,n),e=e.sibling;e!==null;)Ni(e,t,n),e=e.sibling}function ji(e,t,n){var r=e.tag;if(r===5||r===6)e=e.stateNode,t?n.insertBefore(e,t):n.appendChild(e);else if(r!==4&&(e=e.child,e!==null))for(ji(e,t,n),e=e.sibling;e!==null;)ji(e,t,n),e=e.sibling}var ie=null,Ve=!1;function Et(e,t,n){for(n=n.child;n!==null;)Ya(e,t,n),n=n.sibling}function Ya(e,t,n){if(Qe&&typeof Qe.onCommitFiberUnmount=="function")try{Qe.onCommitFiberUnmount(Cr,n)}catch{}switch(n.tag){case 5:fe||wn(n,t);case 6:var r=ie,l=Ve;ie=null,Et(e,t,n),ie=r,Ve=l,ie!==null&&(Ve?(e=ie,n=n.stateNode,e.nodeType===8?e.parentNode.removeChild(n):e.removeChild(n)):ie.removeChild(n.stateNode));break;case 18:ie!==null&&(Ve?(e=ie,n=n.stateNode,e.nodeType===8?$o(e.parentNode,n):e.nodeType===1&&$o(e,n),Un(e)):$o(ie,n.stateNode));break;case 4:r=ie,l=Ve,ie=n.stateNode.containerInfo,Ve=!0,Et(e,t,n),ie=r,Ve=l;break;case 0:case 11:case 14:case 15:if(!fe&&(r=n.updateQueue,r!==null&&(r=r.lastEffect,r!==null))){l=r=r.next;do{var o=l,i=o.destroy;o=o.tag,i!==void 0&&(o&2||o&4)&&Ei(n,t,i),l=l.next}while(l!==r)}Et(e,t,n);break;case 1:if(!fe&&(wn(n,t),r=n.stateNode,typeof r.componentWillUnmount=="function"))try{r.props=n.memoizedProps,r.state=n.memoizedState,r.componentWillUnmount()}catch(u){Z(n,t,u)}Et(e,t,n);break;case 21:Et(e,t,n);break;case 22:n.mode&1?(fe=(r=fe)||n.memoizedState!==null,Et(e,t,n),fe=r):Et(e,t,n);break;default:Et(e,t,n)}}function Xa(e){var t=e.updateQueue;if(t!==null){e.updateQueue=null;var n=e.stateNode;n===null&&(n=e.stateNode=new Bd),t.forEach(function(r){var l=qd.bind(null,e,r);n.has(r)||(n.add(r),r.then(l,l))})}}function Be(e,t){var n=t.deletions;if(n!==null)for(var r=0;r<n.length;r++){var l=n[r];try{var o=e,i=t,u=i;e:for(;u!==null;){switch(u.tag){case 5:ie=u.stateNode,Ve=!1;break e;case 3:ie=u.stateNode.containerInfo,Ve=!0;break e;case 4:ie=u.stateNode.containerInfo,Ve=!0;break e}u=u.return}if(ie===null)throw Error(w(160));Ya(o,i,l),ie=null,Ve=!1;var s=l.alternate;s!==null&&(s.return=null),l.return=null}catch(c){Z(l,t,c)}}if(t.subtreeFlags&12854)for(t=t.child;t!==null;)Ga(t,e),t=t.sibling}function Ga(e,t){var n=e.alternate,r=e.flags;switch(e.tag){case 0:case 11:case 14:case 15:if(Be(t,e),Ge(e),r&4){try{ur(3,e,e.return),dl(3,e)}catch(C){Z(e,e.return,C)}try{ur(5,e,e.return)}catch(C){Z(e,e.return,C)}}break;case 1:Be(t,e),Ge(e),r&512&&n!==null&&wn(n,n.return);break;case 5:if(Be(t,e),Ge(e),r&512&&n!==null&&wn(n,n.return),e.flags&32){var l=e.stateNode;try{Pn(l,"")}catch(C){Z(e,e.return,C)}}if(r&4&&(l=e.stateNode,l!=null)){var o=e.memoizedProps,i=n!==null?n.memoizedProps:o,u=e.type,s=e.updateQueue;if(e.updateQueue=null,s!==null)try{u==="input"&&o.type==="radio"&&o.name!=null&&_u(l,o),ro(u,i);var c=ro(u,o);for(i=0;i<s.length;i+=2){var y=s[i],v=s[i+1];y==="style"?Lu(l,v):y==="dangerouslySetInnerHTML"?Tu(l,v):y==="children"?Pn(l,v):$l(l,y,v,c)}switch(u){case"input":ql(l,o);break;case"textarea":Nu(l,o);break;case"select":var m=l._wrapperState.wasMultiple;l._wrapperState.wasMultiple=!!o.multiple;var k=o.value;k!=null?Jt(l,!!o.multiple,k,!1):m!==!!o.multiple&&(o.defaultValue!=null?Jt(l,!!o.multiple,o.defaultValue,!0):Jt(l,!!o.multiple,o.multiple?[]:"",!1))}l[Zn]=o}catch(C){Z(e,e.return,C)}}break;case 6:if(Be(t,e),Ge(e),r&4){if(e.stateNode===null)throw Error(w(162));l=e.stateNode,o=e.memoizedProps;try{l.nodeValue=o}catch(C){Z(e,e.return,C)}}break;case 3:if(Be(t,e),Ge(e),r&4&&n!==null&&n.memoizedState.isDehydrated)try{Un(t.containerInfo)}catch(C){Z(e,e.return,C)}break;case 4:Be(t,e),Ge(e);break;case 13:Be(t,e),Ge(e),l=e.child,l.flags&8192&&(o=l.memoizedState!==null,l.stateNode.isHidden=o,!o||l.alternate!==null&&l.alternate.memoizedState!==null||(zi=J())),r&4&&Xa(e);break;case 22:if(y=n!==null&&n.memoizedState!==null,e.mode&1?(fe=(c=fe)||y,Be(t,e),fe=c):Be(t,e),Ge(e),r&8192){if(c=e.memoizedState!==null,(e.stateNode.isHidden=c)&&!y&&e.mode&1)for(j=e,y=e.child;y!==null;){for(v=j=y;j!==null;){switch(m=j,k=m.child,m.tag){case 0:case 11:case 14:case 15:ur(4,m,m.return);break;case 1:wn(m,m.return);var S=m.stateNode;if(typeof S.componentWillUnmount=="function"){r=m,n=m.return;try{t=r,S.props=t.memoizedProps,S.state=t.memoizedState,S.componentWillUnmount()}catch(C){Z(r,n,C)}}break;case 5:wn(m,m.return);break;case 22:if(m.memoizedState!==null){qa(v);continue}}k!==null?(k.return=m,j=k):qa(v)}y=y.sibling}e:for(y=null,v=e;;){if(v.tag===5){if(y===null){y=v;try{l=v.stateNode,c?(o=l.style,typeof o.setProperty=="function"?o.setProperty("display","none","important"):o.display="none"):(u=v.stateNode,s=v.memoizedProps.style,i=s!=null&&s.hasOwnProperty("display")?s.display:null,u.style.display=zu("display",i))}catch(C){Z(e,e.return,C)}}}else if(v.tag===6){if(y===null)try{v.stateNode.nodeValue=c?"":v.memoizedProps}catch(C){Z(e,e.return,C)}}else if((v.tag!==22&&v.tag!==23||v.memoizedState===null||v===e)&&v.child!==null){v.child.return=v,v=v.child;continue}if(v===e)break e;for(;v.sibling===null;){if(v.return===null||v.return===e)break e;y===v&&(y=null),v=v.return}y===v&&(y=null),v.sibling.return=v.return,v=v.sibling}}break;case 19:Be(t,e),Ge(e),r&4&&Xa(e);break;case 21:break;default:Be(t,e),Ge(e)}}function Ge(e){var t=e.flags;if(t&2){try{e:{for(var n=e.return;n!==null;){if(Qa(n)){var r=n;break e}n=n.return}throw Error(w(160))}switch(r.tag){case 5:var l=r.stateNode;r.flags&32&&(Pn(l,""),r.flags&=-33);var o=Ka(e);ji(e,o,l);break;case 3:case 4:var i=r.stateNode.containerInfo,u=Ka(e);Ni(e,u,i);break;default:throw Error(w(161))}}catch(s){Z(e,e.return,s)}e.flags&=-3}t&4096&&(e.flags&=-4097)}function Hd(e,t,n){j=e,Za(e)}function Za(e,t,n){for(var r=(e.mode&1)!==0;j!==null;){var l=j,o=l.child;if(l.tag===22&&r){var i=l.memoizedState!==null||fl;if(!i){var u=l.alternate,s=u!==null&&u.memoizedState!==null||fe;u=fl;var c=fe;if(fl=i,(fe=s)&&!c)for(j=l;j!==null;)i=j,s=i.child,i.tag===22&&i.memoizedState!==null?ba(l):s!==null?(s.return=i,j=s):ba(l);for(;o!==null;)j=o,Za(o),o=o.sibling;j=l,fl=u,fe=c}Ja(e)}else l.subtreeFlags&8772&&o!==null?(o.return=l,j=o):Ja(e)}}function Ja(e){for(;j!==null;){var t=j;if(t.flags&8772){var n=t.alternate;try{if(t.flags&8772)switch(t.tag){case 0:case 11:case 15:fe||dl(5,t);break;case 1:var r=t.stateNode;if(t.flags&4&&!fe)if(n===null)r.componentDidMount();else{var l=t.elementType===t.type?n.memoizedProps:$e(t.type,n.memoizedProps);r.componentDidUpdate(l,n.memoizedState,r.__reactInternalSnapshotBeforeUpdate)}var o=t.updateQueue;o!==null&&qs(t,o,r);break;case 3:var i=t.updateQueue;if(i!==null){if(n=null,t.child!==null)switch(t.child.tag){case 5:n=t.child.stateNode;break;case 1:n=t.child.stateNode}qs(t,i,n)}break;case 5:var u=t.stateNode;if(n===null&&t.flags&4){n=u;var s=t.memoizedProps;switch(t.type){case"button":case"input":case"select":case"textarea":s.autoFocus&&n.focus();break;case"img":s.src&&(n.src=s.src)}}break;case 6:break;case 4:break;case 12:break;case 13:if(t.memoizedState===null){var c=t.alternate;if(c!==null){var y=c.memoizedState;if(y!==null){var v=y.dehydrated;v!==null&&Un(v)}}}break;case 19:case 17:case 21:case 22:case 23:case 25:break;default:throw Error(w(163))}fe||t.flags&512&&Ci(t)}catch(m){Z(t,t.return,m)}}if(t===e){j=null;break}if(n=t.sibling,n!==null){n.return=t.return,j=n;break}j=t.return}}function qa(e){for(;j!==null;){var t=j;if(t===e){j=null;break}var n=t.sibling;if(n!==null){n.return=t.return,j=n;break}j=t.return}}function ba(e){for(;j!==null;){var t=j;try{switch(t.tag){case 0:case 11:case 15:var n=t.return;try{dl(4,t)}catch(s){Z(t,n,s)}break;case 1:var r=t.stateNode;if(typeof r.componentDidMount=="function"){var l=t.return;try{r.componentDidMount()}catch(s){Z(t,l,s)}}var o=t.return;try{Ci(t)}catch(s){Z(t,o,s)}break;case 5:var i=t.return;try{Ci(t)}catch(s){Z(t,i,s)}}}catch(s){Z(t,t.return,s)}if(t===e){j=null;break}var u=t.sibling;if(u!==null){u.return=t.return,j=u;break}j=t.return}}var Qd=Math.ceil,pl=qe.ReactCurrentDispatcher,Pi=qe.ReactCurrentOwner,Me=qe.ReactCurrentBatchConfig,I=0,le=null,b=null,ue=0,Pe=0,kn=wt(0),ne=0,sr=null,Vt=0,ml=0,Ti=0,ar=null,ke=null,zi=0,Sn=1/0,it=null,hl=!1,Li=null,Ct=null,vl=!1,Nt=null,yl=0,cr=0,Ri=null,gl=-1,wl=0;function ve(){return I&6?J():gl!==-1?gl:gl=J()}function jt(e){return e.mode&1?I&2&&ue!==0?ue&-ue:Pd.transition!==null?(wl===0&&(wl=Ku()),wl):(e=U,e!==0||(e=window.event,e=e===void 0?16:ts(e.type)),e):1}function We(e,t,n,r){if(50<cr)throw cr=0,Ri=null,Error(w(185));Mn(e,n,r),(!(I&2)||e!==le)&&(e===le&&(!(I&2)&&(ml|=n),ne===4&&Pt(e,ue)),Se(e,r),n===1&&I===0&&!(t.mode&1)&&(Sn=J()+500,Yr&&St()))}function Se(e,t){var n=e.callbackNode;Pf(e,t);var r=Pr(e,e===le?ue:0);if(r===0)n!==null&&Wu(n),e.callbackNode=null,e.callbackPriority=0;else if(t=r&-r,e.callbackPriority!==t){if(n!=null&&Wu(n),t===1)e.tag===0?jd(tc.bind(null,e)):$s(tc.bind(null,e)),_d(function(){!(I&6)&&St()}),n=null;else{switch(Yu(r)){case 1:n=co;break;case 4:n=Hu;break;case 16:n=Er;break;case 536870912:n=Qu;break;default:n=Er}n=ac(n,ec.bind(null,e))}e.callbackPriority=t,e.callbackNode=n}}function ec(e,t){if(gl=-1,wl=0,I&6)throw Error(w(327));var n=e.callbackNode;if(xn()&&e.callbackNode!==n)return null;var r=Pr(e,e===le?ue:0);if(r===0)return null;if(r&30||r&e.expiredLanes||t)t=kl(e,r);else{t=r;var l=I;I|=2;var o=rc();(le!==e||ue!==t)&&(it=null,Sn=J()+500,Wt(e,t));do try{Xd();break}catch(u){nc(e,u)}while(!0);Zo(),pl.current=o,I=l,b!==null?t=0:(le=null,ue=0,t=ne)}if(t!==0){if(t===2&&(l=fo(e),l!==0&&(r=l,t=Oi(e,l))),t===1)throw n=sr,Wt(e,0),Pt(e,r),Se(e,J()),n;if(t===6)Pt(e,r);else{if(l=e.current.alternate,!(r&30)&&!Kd(l)&&(t=kl(e,r),t===2&&(o=fo(e),o!==0&&(r=o,t=Oi(e,o))),t===1))throw n=sr,Wt(e,0),Pt(e,r),Se(e,J()),n;switch(e.finishedWork=l,e.finishedLanes=r,t){case 0:case 1:throw Error(w(345));case 2:Ht(e,ke,it);break;case 3:if(Pt(e,r),(r&130023424)===r&&(t=zi+500-J(),10<t)){if(Pr(e,0)!==0)break;if(l=e.suspendedLanes,(l&r)!==r){ve(),e.pingedLanes|=e.suspendedLanes&l;break}e.timeoutHandle=Uo(Ht.bind(null,e,ke,it),t);break}Ht(e,ke,it);break;case 4:if(Pt(e,r),(r&4194240)===r)break;for(t=e.eventTimes,l=-1;0<r;){var i=31-Ie(r);o=1<<i,i=t[i],i>l&&(l=i),r&=~o}if(r=l,r=J()-r,r=(120>r?120:480>r?480:1080>r?1080:1920>r?1920:3e3>r?3e3:4320>r?4320:1960*Qd(r/1960))-r,10<r){e.timeoutHandle=Uo(Ht.bind(null,e,ke,it),r);break}Ht(e,ke,it);break;case 5:Ht(e,ke,it);break;default:throw Error(w(329))}}}return Se(e,J()),e.callbackNode===n?ec.bind(null,e):null}function Oi(e,t){var n=ar;return e.current.memoizedState.isDehydrated&&(Wt(e,t).flags|=256),e=kl(e,t),e!==2&&(t=ke,ke=n,t!==null&&Mi(t)),e}function Mi(e){ke===null?ke=e:ke.push.apply(ke,e)}function Kd(e){for(var t=e;;){if(t.flags&16384){var n=t.updateQueue;if(n!==null&&(n=n.stores,n!==null))for(var r=0;r<n.length;r++){var l=n[r],o=l.getSnapshot;l=l.value;try{if(!Ae(o(),l))return!1}catch{return!1}}}if(n=t.child,t.subtreeFlags&16384&&n!==null)n.return=t,t=n;else{if(t===e)break;for(;t.sibling===null;){if(t.return===null||t.return===e)return!0;t=t.return}t.sibling.return=t.return,t=t.sibling}}return!0}function Pt(e,t){for(t&=~Ti,t&=~ml,e.suspendedLanes|=t,e.pingedLanes&=~t,e=e.expirationTimes;0<t;){var n=31-Ie(t),r=1<<n;e[n]=-1,t&=~r}}function tc(e){if(I&6)throw Error(w(327));xn();var t=Pr(e,0);if(!(t&1))return Se(e,J()),null;var n=kl(e,t);if(e.tag!==0&&n===2){var r=fo(e);r!==0&&(t=r,n=Oi(e,r))}if(n===1)throw n=sr,Wt(e,0),Pt(e,t),Se(e,J()),n;if(n===6)throw Error(w(345));return e.finishedWork=e.current.alternate,e.finishedLanes=t,Ht(e,ke,it),Se(e,J()),null}function Fi(e,t){var n=I;I|=1;try{return e(t)}finally{I=n,I===0&&(Sn=J()+500,Yr&&St())}}function Bt(e){Nt!==null&&Nt.tag===0&&!(I&6)&&xn();var t=I;I|=1;var n=Me.transition,r=U;try{if(Me.transition=null,U=1,e)return e()}finally{U=r,Me.transition=n,I=t,!(I&6)&&St()}}function Di(){Pe=kn.current,B(kn)}function Wt(e,t){e.finishedWork=null,e.finishedLanes=0;var n=e.timeoutHandle;if(n!==-1&&(e.timeoutHandle=-1,xd(n)),b!==null)for(n=b.return;n!==null;){var r=n;switch(Qo(r),r.tag){case 1:r=r.type.childContextTypes,r!=null&&Qr();break;case 3:yn(),B(ye),B(se),li();break;case 5:ni(r);break;case 4:yn();break;case 13:B(Y);break;case 19:B(Y);break;case 10:Jo(r.type._context);break;case 22:case 23:Di()}n=n.return}if(le=e,b=e=Tt(e.current,null),ue=Pe=t,ne=0,sr=null,Ti=ml=Vt=0,ke=ar=null,At!==null){for(t=0;t<At.length;t++)if(n=At[t],r=n.interleaved,r!==null){n.interleaved=null;var l=r.next,o=n.pending;if(o!==null){var i=o.next;o.next=l,r.next=i}n.pending=r}At=null}return e}function nc(e,t){do{var n=b;try{if(Zo(),rl.current=ul,ll){for(var r=X.memoizedState;r!==null;){var l=r.queue;l!==null&&(l.pending=null),r=r.next}ll=!1}if($t=0,re=te=X=null,nr=!1,rr=0,Pi.current=null,n===null||n.return===null){ne=1,sr=t,b=null;break}e:{var o=e,i=n.return,u=n,s=t;if(t=ue,u.flags|=32768,s!==null&&typeof s=="object"&&typeof s.then=="function"){var c=s,y=u,v=y.tag;if(!(y.mode&1)&&(v===0||v===11||v===15)){var m=y.alternate;m?(y.updateQueue=m.updateQueue,y.memoizedState=m.memoizedState,y.lanes=m.lanes):(y.updateQueue=null,y.memoizedState=null)}var k=ja(i);if(k!==null){k.flags&=-257,Pa(k,i,u,o,t),k.mode&1&&Na(o,c,t),t=k,s=c;var S=t.updateQueue;if(S===null){var C=new Set;C.add(s),t.updateQueue=C}else S.add(s);break e}else{if(!(t&1)){Na(o,c,t),Ii();break e}s=Error(w(426))}}else if(H&&u.mode&1){var A=ja(i);if(A!==null){!(A.flags&65536)&&(A.flags|=256),Pa(A,i,u,o,t),Xo(gn(s,u));break e}}o=s=gn(s,u),ne!==4&&(ne=2),ar===null?ar=[o]:ar.push(o),o=i;do{switch(o.tag){case 3:o.flags|=65536,t&=-t,o.lanes|=t;var f=Ea(o,s,t);Js(o,f);break e;case 1:u=s;var a=o.type,d=o.stateNode;if(!(o.flags&128)&&(typeof a.getDerivedStateFromError=="function"||d!==null&&typeof d.componentDidCatch=="function"&&(Ct===null||!Ct.has(d)))){o.flags|=65536,t&=-t,o.lanes|=t;var g=Ca(o,u,t);Js(o,g);break e}}o=o.return}while(o!==null)}oc(n)}catch(N){t=N,b===n&&n!==null&&(b=n=n.return);continue}break}while(!0)}function rc(){var e=pl.current;return pl.current=ul,e===null?ul:e}function Ii(){(ne===0||ne===3||ne===2)&&(ne=4),le===null||!(Vt&268435455)&&!(ml&268435455)||Pt(le,ue)}function kl(e,t){var n=I;I|=2;var r=rc();(le!==e||ue!==t)&&(it=null,Wt(e,t));do try{Yd();break}catch(l){nc(e,l)}while(!0);if(Zo(),I=n,pl.current=r,b!==null)throw Error(w(261));return le=null,ue=0,ne}function Yd(){for(;b!==null;)lc(b)}function Xd(){for(;b!==null&&!wf();)lc(b)}function lc(e){var t=sc(e.alternate,e,Pe);e.memoizedProps=e.pendingProps,t===null?oc(e):b=t,Pi.current=null}function oc(e){var t=e;do{var n=t.alternate;if(e=t.return,t.flags&32768){if(n=Vd(n,t),n!==null){n.flags&=32767,b=n;return}if(e!==null)e.flags|=32768,e.subtreeFlags=0,e.deletions=null;else{ne=6,b=null;return}}else if(n=$d(n,t,Pe),n!==null){b=n;return}if(t=t.sibling,t!==null){b=t;return}b=t=e}while(t!==null);ne===0&&(ne=5)}function Ht(e,t,n){var r=U,l=Me.transition;try{Me.transition=null,U=1,Gd(e,t,n,r)}finally{Me.transition=l,U=r}return null}function Gd(e,t,n,r){do xn();while(Nt!==null);if(I&6)throw Error(w(327));n=e.finishedWork;var l=e.finishedLanes;if(n===null)return null;if(e.finishedWork=null,e.finishedLanes=0,n===e.current)throw Error(w(177));e.callbackNode=null,e.callbackPriority=0;var o=n.lanes|n.childLanes;if(Tf(e,o),e===le&&(b=le=null,ue=0),!(n.subtreeFlags&2064)&&!(n.flags&2064)||vl||(vl=!0,ac(Er,function(){return xn(),null})),o=(n.flags&15990)!==0,n.subtreeFlags&15990||o){o=Me.transition,Me.transition=null;var i=U;U=1;var u=I;I|=4,Pi.current=null,Wd(e,n),Ga(n,e),hd(Io),Lr=!!Do,Io=Do=null,e.current=n,Hd(n),kf(),I=u,U=i,Me.transition=o}else e.current=n;if(vl&&(vl=!1,Nt=e,yl=l),o=e.pendingLanes,o===0&&(Ct=null),_f(n.stateNode),Se(e,J()),t!==null)for(r=e.onRecoverableError,n=0;n<t.length;n++)l=t[n],r(l.value,{componentStack:l.stack,digest:l.digest});if(hl)throw hl=!1,e=Li,Li=null,e;return yl&1&&e.tag!==0&&xn(),o=e.pendingLanes,o&1?e===Ri?cr++:(cr=0,Ri=e):cr=0,St(),null}function xn(){if(Nt!==null){var e=Yu(yl),t=Me.transition,n=U;try{if(Me.transition=null,U=16>e?16:e,Nt===null)var r=!1;else{if(e=Nt,Nt=null,yl=0,I&6)throw Error(w(331));var l=I;for(I|=4,j=e.current;j!==null;){var o=j,i=o.child;if(j.flags&16){var u=o.deletions;if(u!==null){for(var s=0;s<u.length;s++){var c=u[s];for(j=c;j!==null;){var y=j;switch(y.tag){case 0:case 11:case 15:ur(8,y,o)}var v=y.child;if(v!==null)v.return=y,j=v;else for(;j!==null;){y=j;var m=y.sibling,k=y.return;if(Ha(y),y===c){j=null;break}if(m!==null){m.return=k,j=m;break}j=k}}}var S=o.alternate;if(S!==null){var C=S.child;if(C!==null){S.child=null;do{var A=C.sibling;C.sibling=null,C=A}while(C!==null)}}j=o}}if(o.subtreeFlags&2064&&i!==null)i.return=o,j=i;else e:for(;j!==null;){if(o=j,o.flags&2048)switch(o.tag){case 0:case 11:case 15:ur(9,o,o.return)}var f=o.sibling;if(f!==null){f.return=o.return,j=f;break e}j=o.return}}var a=e.current;for(j=a;j!==null;){i=j;var d=i.child;if(i.subtreeFlags&2064&&d!==null)d.return=i,j=d;else e:for(i=a;j!==null;){if(u=j,u.flags&2048)try{switch(u.tag){case 0:case 11:case 15:dl(9,u)}}catch(N){Z(u,u.return,N)}if(u===i){j=null;break e}var g=u.sibling;if(g!==null){g.return=u.return,j=g;break e}j=u.return}}if(I=l,St(),Qe&&typeof Qe.onPostCommitFiberRoot=="function")try{Qe.onPostCommitFiberRoot(Cr,e)}catch{}r=!0}return r}finally{U=n,Me.transition=t}}return!1}function ic(e,t,n){t=gn(n,t),t=Ea(e,t,1),e=_t(e,t,1),t=ve(),e!==null&&(Mn(e,1,t),Se(e,t))}function Z(e,t,n){if(e.tag===3)ic(e,e,n);else for(;t!==null;){if(t.tag===3){ic(t,e,n);break}else if(t.tag===1){var r=t.stateNode;if(typeof t.type.getDerivedStateFromError=="function"||typeof r.componentDidCatch=="function"&&(Ct===null||!Ct.has(r))){e=gn(n,e),e=Ca(t,e,1),t=_t(t,e,1),e=ve(),t!==null&&(Mn(t,1,e),Se(t,e));break}}t=t.return}}function Zd(e,t,n){var r=e.pingCache;r!==null&&r.delete(t),t=ve(),e.pingedLanes|=e.suspendedLanes&n,le===e&&(ue&n)===n&&(ne===4||ne===3&&(ue&130023424)===ue&&500>J()-zi?Wt(e,0):Ti|=n),Se(e,t)}function uc(e,t){t===0&&(e.mode&1?(t=jr,jr<<=1,!(jr&130023424)&&(jr=4194304)):t=1);var n=ve();e=rt(e,t),e!==null&&(Mn(e,t,n),Se(e,n))}function Jd(e){var t=e.memoizedState,n=0;t!==null&&(n=t.retryLane),uc(e,n)}function qd(e,t){var n=0;switch(e.tag){case 13:var r=e.stateNode,l=e.memoizedState;l!==null&&(n=l.retryLane);break;case 19:r=e.stateNode;break;default:throw Error(w(314))}r!==null&&r.delete(t),uc(e,n)}var sc;sc=function(e,t,n){if(e!==null)if(e.memoizedProps!==t.pendingProps||ye.current)we=!0;else{if(!(e.lanes&n)&&!(t.flags&128))return we=!1,Ud(e,t,n);we=!!(e.flags&131072)}else we=!1,H&&t.flags&1048576&&Vs(t,Gr,t.index);switch(t.lanes=0,t.tag){case 2:var r=t.type;cl(e,t),e=t.pendingProps;var l=cn(t,se.current);vn(t,n),l=ui(null,t,r,e,l,n);var o=si();return t.flags|=1,typeof l=="object"&&l!==null&&typeof l.render=="function"&&l.$$typeof===void 0?(t.tag=1,t.memoizedState=null,t.updateQueue=null,ge(r)?(o=!0,Kr(t)):o=!1,t.memoizedState=l.state!==null&&l.state!==void 0?l.state:null,ei(t),l.updater=sl,t.stateNode=l,l._reactInternals=t,mi(t,r,e,n),t=gi(null,t,r,!0,o,n)):(t.tag=0,H&&o&&Ho(t),he(null,t,l,n),t=t.child),t;case 16:r=t.elementType;e:{switch(cl(e,t),e=t.pendingProps,l=r._init,r=l(r._payload),t.type=r,l=t.tag=ep(r),e=$e(r,e),l){case 0:t=yi(null,t,r,e,n);break e;case 1:t=Ma(null,t,r,e,n);break e;case 11:t=Ta(null,t,r,e,n);break e;case 14:t=za(null,t,r,$e(r.type,e),n);break e}throw Error(w(306,r,""))}return t;case 0:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:$e(r,l),yi(e,t,r,l,n);case 1:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:$e(r,l),Ma(e,t,r,l,n);case 3:e:{if(Fa(t),e===null)throw Error(w(387));r=t.pendingProps,o=t.memoizedState,l=o.element,Zs(e,t),tl(t,r,null,n);var i=t.memoizedState;if(r=i.element,o.isDehydrated)if(o={element:r,isDehydrated:!1,cache:i.cache,pendingSuspenseBoundaries:i.pendingSuspenseBoundaries,transitions:i.transitions},t.updateQueue.baseState=o,t.memoizedState=o,t.flags&256){l=gn(Error(w(423)),t),t=Da(e,t,r,n,l);break e}else if(r!==l){l=gn(Error(w(424)),t),t=Da(e,t,r,n,l);break e}else for(je=gt(t.stateNode.containerInfo.firstChild),Ne=t,H=!0,Ue=null,n=Xs(t,null,r,n),t.child=n;n;)n.flags=n.flags&-3|4096,n=n.sibling;else{if(pn(),r===l){t=ot(e,t,n);break e}he(e,t,r,n)}t=t.child}return t;case 5:return bs(t),e===null&&Yo(t),r=t.type,l=t.pendingProps,o=e!==null?e.memoizedProps:null,i=l.children,Ao(r,l)?i=null:o!==null&&Ao(r,o)&&(t.flags|=32),Oa(e,t),he(e,t,i,n),t.child;case 6:return e===null&&Yo(t),null;case 13:return Ia(e,t,n);case 4:return ti(t,t.stateNode.containerInfo),r=t.pendingProps,e===null?t.child=mn(t,null,r,n):he(e,t,r,n),t.child;case 11:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:$e(r,l),Ta(e,t,r,l,n);case 7:return he(e,t,t.pendingProps,n),t.child;case 8:return he(e,t,t.pendingProps.children,n),t.child;case 12:return he(e,t,t.pendingProps.children,n),t.child;case 10:e:{if(r=t.type._context,l=t.pendingProps,o=t.memoizedProps,i=l.value,$(qr,r._currentValue),r._currentValue=i,o!==null)if(Ae(o.value,i)){if(o.children===l.children&&!ye.current){t=ot(e,t,n);break e}}else for(o=t.child,o!==null&&(o.return=t);o!==null;){var u=o.dependencies;if(u!==null){i=o.child;for(var s=u.firstContext;s!==null;){if(s.context===r){if(o.tag===1){s=lt(-1,n&-n),s.tag=2;var c=o.updateQueue;if(c!==null){c=c.shared;var y=c.pending;y===null?s.next=s:(s.next=y.next,y.next=s),c.pending=s}}o.lanes|=n,s=o.alternate,s!==null&&(s.lanes|=n),qo(o.return,n,t),u.lanes|=n;break}s=s.next}}else if(o.tag===10)i=o.type===t.type?null:o.child;else if(o.tag===18){if(i=o.return,i===null)throw Error(w(341));i.lanes|=n,u=i.alternate,u!==null&&(u.lanes|=n),qo(i,n,t),i=o.sibling}else i=o.child;if(i!==null)i.return=o;else for(i=o;i!==null;){if(i===t){i=null;break}if(o=i.sibling,o!==null){o.return=i.return,i=o;break}i=i.return}o=i}he(e,t,l.children,n),t=t.child}return t;case 9:return l=t.type,r=t.pendingProps.children,vn(t,n),l=Re(l),r=r(l),t.flags|=1,he(e,t,r,n),t.child;case 14:return r=t.type,l=$e(r,t.pendingProps),l=$e(r.type,l),za(e,t,r,l,n);case 15:return La(e,t,t.type,t.pendingProps,n);case 17:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:$e(r,l),cl(e,t),t.tag=1,ge(r)?(e=!0,Kr(t)):e=!1,vn(t,n),xa(t,r,l),mi(t,r,l,n),gi(null,t,r,!0,e,n);case 19:return Ua(e,t,n);case 22:return Ra(e,t,n)}throw Error(w(156,t.tag))};function ac(e,t){return Bu(e,t)}function bd(e,t,n,r){this.tag=e,this.key=n,this.sibling=this.child=this.return=this.stateNode=this.type=this.elementType=null,this.index=0,this.ref=null,this.pendingProps=t,this.dependencies=this.memoizedState=this.updateQueue=this.memoizedProps=null,this.mode=r,this.subtreeFlags=this.flags=0,this.deletions=null,this.childLanes=this.lanes=0,this.alternate=null}function Fe(e,t,n,r){return new bd(e,t,n,r)}function Ai(e){return e=e.prototype,!(!e||!e.isReactComponent)}function ep(e){if(typeof e=="function")return Ai(e)?1:0;if(e!=null){if(e=e.$$typeof,e===Wl)return 11;if(e===Kl)return 14}return 2}function Tt(e,t){var n=e.alternate;return n===null?(n=Fe(e.tag,t,e.key,e.mode),n.elementType=e.elementType,n.type=e.type,n.stateNode=e.stateNode,n.alternate=e,e.alternate=n):(n.pendingProps=t,n.type=e.type,n.flags=0,n.subtreeFlags=0,n.deletions=null),n.flags=e.flags&14680064,n.childLanes=e.childLanes,n.lanes=e.lanes,n.child=e.child,n.memoizedProps=e.memoizedProps,n.memoizedState=e.memoizedState,n.updateQueue=e.updateQueue,t=e.dependencies,n.dependencies=t===null?null:{lanes:t.lanes,firstContext:t.firstContext},n.sibling=e.sibling,n.index=e.index,n.ref=e.ref,n}function Sl(e,t,n,r,l,o){var i=2;if(r=e,typeof e=="function")Ai(e)&&(i=1);else if(typeof e=="string")i=5;else e:switch(e){case Zt:return Qt(n.children,l,o,t);case Vl:i=8,l|=8;break;case Bl:return e=Fe(12,n,t,l|2),e.elementType=Bl,e.lanes=o,e;case Hl:return e=Fe(13,n,t,l),e.elementType=Hl,e.lanes=o,e;case Ql:return e=Fe(19,n,t,l),e.elementType=Ql,e.lanes=o,e;case gu:return xl(n,l,o,t);default:if(typeof e=="object"&&e!==null)switch(e.$$typeof){case vu:i=10;break e;case yu:i=9;break e;case Wl:i=11;break e;case Kl:i=14;break e;case ct:i=16,r=null;break e}throw Error(w(130,e==null?e:typeof e,""))}return t=Fe(i,n,t,l),t.elementType=e,t.type=r,t.lanes=o,t}function Qt(e,t,n,r){return e=Fe(7,e,r,t),e.lanes=n,e}function xl(e,t,n,r){return e=Fe(22,e,r,t),e.elementType=gu,e.lanes=n,e.stateNode={isHidden:!1},e}function Ui(e,t,n){return e=Fe(6,e,null,t),e.lanes=n,e}function $i(e,t,n){return t=Fe(4,e.children!==null?e.children:[],e.key,t),t.lanes=n,t.stateNode={containerInfo:e.containerInfo,pendingChildren:null,implementation:e.implementation},t}function tp(e,t,n,r,l){this.tag=t,this.containerInfo=e,this.finishedWork=this.pingCache=this.current=this.pendingChildren=null,this.timeoutHandle=-1,this.callbackNode=this.pendingContext=this.context=null,this.callbackPriority=0,this.eventTimes=po(0),this.expirationTimes=po(-1),this.entangledLanes=this.finishedLanes=this.mutableReadLanes=this.expiredLanes=this.pingedLanes=this.suspendedLanes=this.pendingLanes=0,this.entanglements=po(0),this.identifierPrefix=r,this.onRecoverableError=l,this.mutableSourceEagerHydrationData=null}function Vi(e,t,n,r,l,o,i,u,s){return e=new tp(e,t,n,u,s),t===1?(t=1,o===!0&&(t|=8)):t=0,o=Fe(3,null,null,t),e.current=o,o.stateNode=e,o.memoizedState={element:r,isDehydrated:n,cache:null,transitions:null,pendingSuspenseBoundaries:null},ei(o),e}function np(e,t,n){var r=3<arguments.length&&arguments[3]!==void 0?arguments[3]:null;return{$$typeof:Gt,key:r==null?null:""+r,children:e,containerInfo:t,implementation:n}}function cc(e){if(!e)return kt;e=e._reactInternals;e:{if(Ot(e)!==e||e.tag!==1)throw Error(w(170));var t=e;do{switch(t.tag){case 3:t=t.stateNode.context;break e;case 1:if(ge(t.type)){t=t.stateNode.__reactInternalMemoizedMergedChildContext;break e}}t=t.return}while(t!==null);throw Error(w(171))}if(e.tag===1){var n=e.type;if(ge(n))return As(e,n,t)}return t}function fc(e,t,n,r,l,o,i,u,s){return e=Vi(n,r,!0,e,l,o,i,u,s),e.context=cc(null),n=e.current,r=ve(),l=jt(n),o=lt(r,l),o.callback=t??null,_t(n,o,l),e.current.lanes=l,Mn(e,l,r),Se(e,r),e}function _l(e,t,n,r){var l=t.current,o=ve(),i=jt(l);return n=cc(n),t.context===null?t.context=n:t.pendingContext=n,t=lt(o,i),t.payload={element:e},r=r===void 0?null:r,r!==null&&(t.callback=r),e=_t(l,t,i),e!==null&&(We(e,l,i,o),el(e,l,i)),i}function El(e){if(e=e.current,!e.child)return null;switch(e.child.tag){case 5:return e.child.stateNode;default:return e.child.stateNode}}function dc(e,t){if(e=e.memoizedState,e!==null&&e.dehydrated!==null){var n=e.retryLane;e.retryLane=n!==0&&n<t?n:t}}function Bi(e,t){dc(e,t),(e=e.alternate)&&dc(e,t)}function rp(){return null}var pc=typeof reportError=="function"?reportError:function(e){console.error(e)};function Wi(e){this._internalRoot=e}Cl.prototype.render=Wi.prototype.render=function(e){var t=this._internalRoot;if(t===null)throw Error(w(409));_l(e,t,null,null)},Cl.prototype.unmount=Wi.prototype.unmount=function(){var e=this._internalRoot;if(e!==null){this._internalRoot=null;var t=e.containerInfo;Bt(function(){_l(null,e,null,null)}),t[be]=null}};function Cl(e){this._internalRoot=e}Cl.prototype.unstable_scheduleHydration=function(e){if(e){var t=Zu();e={blockedOn:null,target:e,priority:t};for(var n=0;n<ht.length&&t!==0&&t<ht[n].priority;n++);ht.splice(n,0,e),n===0&&bu(e)}};function Hi(e){return!(!e||e.nodeType!==1&&e.nodeType!==9&&e.nodeType!==11)}function Nl(e){return!(!e||e.nodeType!==1&&e.nodeType!==9&&e.nodeType!==11&&(e.nodeType!==8||e.nodeValue!==" react-mount-point-unstable "))}function mc(){}function lp(e,t,n,r,l){if(l){if(typeof r=="function"){var o=r;r=function(){var c=El(i);o.call(c)}}var i=fc(t,r,e,0,null,!1,!1,"",mc);return e._reactRootContainer=i,e[be]=i.current,Xn(e.nodeType===8?e.parentNode:e),Bt(),i}for(;l=e.lastChild;)e.removeChild(l);if(typeof r=="function"){var u=r;r=function(){var c=El(s);u.call(c)}}var s=Vi(e,0,!1,null,null,!1,!1,"",mc);return e._reactRootContainer=s,e[be]=s.current,Xn(e.nodeType===8?e.parentNode:e),Bt(function(){_l(t,s,n,r)}),s}function jl(e,t,n,r,l){var o=n._reactRootContainer;if(o){var i=o;if(typeof l=="function"){var u=l;l=function(){var s=El(i);u.call(s)}}_l(t,i,e,l)}else i=lp(n,t,e,l,r);return El(i)}Xu=function(e){switch(e.tag){case 3:var t=e.stateNode;if(t.current.memoizedState.isDehydrated){var n=On(t.pendingLanes);n!==0&&(mo(t,n|1),Se(t,J()),!(I&6)&&(Sn=J()+500,St()))}break;case 13:Bt(function(){var r=rt(e,1);if(r!==null){var l=ve();We(r,e,1,l)}}),Bi(e,1)}},ho=function(e){if(e.tag===13){var t=rt(e,134217728);if(t!==null){var n=ve();We(t,e,134217728,n)}Bi(e,134217728)}},Gu=function(e){if(e.tag===13){var t=jt(e),n=rt(e,t);if(n!==null){var r=ve();We(n,e,t,r)}Bi(e,t)}},Zu=function(){return U},Ju=function(e,t){var n=U;try{return U=e,t()}finally{U=n}},io=function(e,t,n){switch(t){case"input":if(ql(e,n),t=n.name,n.type==="radio"&&t!=null){for(n=e;n.parentNode;)n=n.parentNode;for(n=n.querySelectorAll("input[name="+JSON.stringify(""+t)+'][type="radio"]'),t=0;t<n.length;t++){var r=n[t];if(r!==e&&r.form===e.form){var l=Hr(r);if(!l)throw Error(w(90));Su(r),ql(r,l)}}}break;case"textarea":Nu(e,n);break;case"select":t=n.value,t!=null&&Jt(e,!!n.multiple,t,!1)}},Fu=Fi,Du=Bt;var op={usingClientEntryPoint:!1,Events:[Jn,sn,Hr,Ou,Mu,Fi]},fr={findFiberByHostInstance:Mt,bundleType:0,version:"18.3.1",rendererPackageName:"react-dom"},ip={bundleType:fr.bundleType,version:fr.version,rendererPackageName:fr.rendererPackageName,rendererConfig:fr.rendererConfig,overrideHookState:null,overrideHookStateDeletePath:null,overrideHookStateRenamePath:null,overrideProps:null,overridePropsDeletePath:null,overridePropsRenamePath:null,setErrorHandler:null,setSuspenseHandler:null,scheduleUpdate:null,currentDispatcherRef:qe.ReactCurrentDispatcher,findHostInstanceByFiber:function(e){return e=$u(e),e===null?null:e.stateNode},findFiberByHostInstance:fr.findFiberByHostInstance||rp,findHostInstancesForRefresh:null,scheduleRefresh:null,scheduleRoot:null,setRefreshHandler:null,getCurrentFiber:null,reconcilerVersion:"18.3.1-next-f1338f8080-20240426"};if(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__<"u"){var Pl=__REACT_DEVTOOLS_GLOBAL_HOOK__;if(!Pl.isDisabled&&Pl.supportsFiber)try{Cr=Pl.inject(ip),Qe=Pl}catch{}}_e.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=op,_e.createPortal=function(e,t){var n=2<arguments.length&&arguments[2]!==void 0?arguments[2]:null;if(!Hi(t))throw Error(w(200));return np(e,t,null,n)},_e.createRoot=function(e,t){if(!Hi(e))throw Error(w(299));var n=!1,r="",l=pc;return t!=null&&(t.unstable_strictMode===!0&&(n=!0),t.identifierPrefix!==void 0&&(r=t.identifierPrefix),t.onRecoverableError!==void 0&&(l=t.onRecoverableError)),t=Vi(e,1,!1,null,null,n,!1,r,l),e[be]=t.current,Xn(e.nodeType===8?e.parentNode:e),new Wi(t)},_e.findDOMNode=function(e){if(e==null)return null;if(e.nodeType===1)return e;var t=e._reactInternals;if(t===void 0)throw typeof e.render=="function"?Error(w(188)):(e=Object.keys(e).join(","),Error(w(268,e)));return e=$u(t),e=e===null?null:e.stateNode,e},_e.flushSync=function(e){return Bt(e)},_e.hydrate=function(e,t,n){if(!Nl(t))throw Error(w(200));return jl(null,e,t,!0,n)},_e.hydrateRoot=function(e,t,n){if(!Hi(e))throw Error(w(405));var r=n!=null&&n.hydratedSources||null,l=!1,o="",i=pc;if(n!=null&&(n.unstable_strictMode===!0&&(l=!0),n.identifierPrefix!==void 0&&(o=n.identifierPrefix),n.onRecoverableError!==void 0&&(i=n.onRecoverableError)),t=fc(t,null,e,1,n??null,l,!1,o,i),e[be]=t.current,Xn(e),r)for(e=0;e<r.length;e++)n=r[e],l=n._getVersion,l=l(n._source),t.mutableSourceEagerHydrationData==null?t.mutableSourceEagerHydrationData=[n,l]:t.mutableSourceEagerHydrationData.push(n,l);return new Cl(t)},_e.render=function(e,t,n){if(!Nl(t))throw Error(w(200));return jl(null,e,t,!1,n)},_e.unmountComponentAtNode=function(e){if(!Nl(e))throw Error(w(40));return e._reactRootContainer?(Bt(function(){jl(null,null,e,!1,function(){e._reactRootContainer=null,e[be]=null})}),!0):!1},_e.unstable_batchedUpdates=Fi,_e.unstable_renderSubtreeIntoContainer=function(e,t,n,r){if(!Nl(n))throw Error(w(200));if(e==null||e._reactInternals===void 0)throw Error(w(38));return jl(e,t,n,!1,r)},_e.version="18.3.1-next-f1338f8080-20240426";function hc(){if(!(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__>"u"||typeof __REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE!="function"))try{__REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE(hc)}catch(e){console.error(e)}}hc(),cu.exports=_e;var Tl=cu.exports,vc,yc=Tl;vc=yc.createRoot,yc.hydrateRoot;let up=1;function gc(){return"row_"+up++}function Qi(){return{id:gc(),task:"",hours:{}}}function Ki(e,t){return{type:e,name:t,tasks:[Qi()]}}function sp(e){return(e||[]).map(t=>({type:t.type,name:t.name,tasks:(t.tasks||[]).length?t.tasks.map(n=>({id:gc(),task:n.task||"",hours:n.hours||{}})):[Qi()]}))}function wc(e){if(!e||typeof e!="string")return"";const[t,n]=e.split("-"),r=parseInt(t,10),l=parseInt(n,10);return!r||r<1||r>12||!l?e:`${new Date(l,r-1,1).toLocaleString("en-US",{month:"long"})} ${l}`}function ap({submissionName:e,onContextChange:t}){const[r,l]=R.useState(null),[o,i]=R.useState(e||""),[u,s]=R.useState([]),[c,y]=R.useState(!0),[v,m]=R.useState(!1),[k,S]=R.useState(!1),[C,A]=R.useState(!1),[f,a]=R.useState(!1),[d,g]=R.useState(""),[N,P]=R.useState(""),[T,L]=R.useState(""),[Q,F]=R.useState(""),[xe,He]=R.useState([]),[Ze,zt]=R.useState(!1),dr=R.useRef(!0),ut=R.useRef(null),st=R.useRef(!1),_=R.useRef(!1),[O,M]=R.useState(()=>window.innerWidth<768),[W,ee]=R.useState(null);R.useEffect(()=>{const p=()=>M(window.innerWidth<768);return window.addEventListener("resize",p),()=>window.removeEventListener("resize",p)},[]);const q=R.useCallback(()=>{st.current=!0,_.current=!1},[]),Te=R.useCallback(p=>{if(!p){y(!1),F("No submission specified. Please open this page from a Timesheet Submission.");return}y(!0),F(""),frappe.call({method:"corporate_services.icl_corporate_services.page.employee_timesheet_entry.employee_timesheet_entry.get_timesheet_context",args:{submission_name:p},callback(x){if(x.message){const E=x.message;l(E),t&&t(E);const z=sp(E.existing_sections||[]);if(z.length)s(z);else{const G=(E.projects||[]).map(de=>Ki("project",de.project_name));s(G)}i(p)}y(!1),dr.current=!1},error(){F("Failed to load timesheet context."),y(!1),dr.current=!1}})},[t]);R.useEffect(()=>{Te(e)},[e,Te]);const Lt=R.useCallback(p=>{if(!p){He([]);return}frappe.call({method:"frappe.client.get",args:{doctype:"Timesheet Submission",name:p},callback(x){const E=x.message;if(!E){He([]);return}frappe.call({method:"frappe.model.workflow.get_transitions",args:{doc:E},callback(z){const G=(z.message||[]).map(de=>de.action).filter(Boolean);He(G)},error(){He([])}})},error(){He([])}})},[]);R.useEffect(()=>{Lt(o)},[o,Lt]);const De=R.useCallback((p=!1)=>{if(v||!o)return;const x=u.map(E=>({type:E.type,name:E.name,tasks:E.tasks.map(z=>({task:z.task,hours:z.hours}))}));m(!0),p&&S(!0),frappe.call({method:"corporate_services.icl_corporate_services.page.employee_timesheet_entry.employee_timesheet_entry.save_web_timesheet",args:{submission_name:o,sections:JSON.stringify(x)},callback(E){m(!1),S(!1),st.current=!1;const z=E.message||{};z.status==="success"?(_.current=!1,p&&frappe.show_alert({message:`Timesheet saved. Total hours: ${z.total_hours.toFixed(1)}`,indicator:"green"},5)):p?frappe.msgprint({title:z.status==="locked"?"Timesheet Locked":"Save Failed",message:z.message||"Could not save timesheet.",indicator:"red"}):(st.current=!1,_.current||(_.current=!0,frappe.msgprint({title:z.status==="locked"?"Timesheet Locked":"Autosave Failed",message:z.message||"Could not autosave timesheet. Edit the timesheet to try again.",indicator:"red"})))},error(){m(!1),S(!1),p?frappe.msgprint({title:"Save Failed",message:"Could not save timesheet.",indicator:"red"}):(st.current=!1,_.current||(_.current=!0,frappe.msgprint({title:"Autosave Failed",message:"Could not autosave timesheet. Edit the timesheet to try again.",indicator:"red"})))}})},[u,o,v]);R.useEffect(()=>{if(st.current&&!(dr.current||c))return ut.current&&clearTimeout(ut.current),ut.current=setTimeout(()=>De(!1),1200),()=>{ut.current&&clearTimeout(ut.current)}},[u,c,De,!0]);const at=R.useMemo(()=>{if(!r)return{};const p={};for(const x of r.dates)p[x.date]=0;for(const x of u)for(const E of x.tasks)for(const[z,G]of Object.entries(E.hours)){const de=parseFloat(G)||0;de>0&&(p[z]=(p[z]||0)+de)}return p},[r,u]),Kt=R.useMemo(()=>Object.values(at).reduce((p,x)=>p+x,0),[at]),zl=R.useCallback(p=>Object.values(p.hours).reduce((x,E)=>x+(parseFloat(E)||0),0),[]),kc=R.useCallback(p=>p.tasks.reduce((x,E)=>x+Object.values(E.hours).reduce((z,G)=>z+(parseFloat(G)||0),0),0),[]),Sc=R.useCallback((p,x,E,z)=>{q(),s(G=>G.map((de,pr)=>pr!==p?de:{...de,tasks:de.tasks.map(Ll=>Ll.id!==x?Ll:{...Ll,hours:{...Ll.hours,[E]:z===""?0:parseFloat(z)||0}})}))},[q]),xc=R.useCallback((p,x,E)=>{q(),s(z=>z.map((G,de)=>de!==p?G:{...G,tasks:G.tasks.map(pr=>pr.id!==x?pr:{...pr,task:E})}))},[q]),_c=R.useCallback(p=>{q(),s(x=>x.map((E,z)=>z!==p?E:{...E,tasks:[...E.tasks,Qi()]}))},[q]),Ec=R.useCallback((p,x)=>{q(),s(E=>E.map((z,G)=>G!==p||z.tasks.length<=1?z:{...z,tasks:z.tasks.filter(de=>de.id!==x)}))},[q]),Cc=R.useCallback((p,x)=>{q(),s(E=>E.map((z,G)=>G!==p?z:{...z,name:x}))},[q]),Nc=R.useCallback(p=>{q(),s(x=>x.filter((E,z)=>z!==p))},[q]),cp=R.useCallback(p=>{q(),s(x=>[...x,Ki("project",p)]),A(!1),g("")},[q]),Yi=R.useCallback(p=>{q(),s(x=>[...x,Ki("activity",p)]),a(!1),P("")},[q]),fp=R.useCallback(()=>{const p=(T||"").trim();p&&frappe.call({method:"corporate_services.icl_corporate_services.page.employee_timesheet_entry.employee_timesheet_entry.create_activity_type",args:{name:p},callback:x=>{const E=x.message&&x.message.name||p;l(z=>new Set(((z==null?void 0:z.activity_types)||[]).map(de=>de.toLowerCase())).has(E.toLowerCase())?z:{...z,activity_types:[...(z==null?void 0:z.activity_types)||[],E].sort()}),Yi(E),L("")}})},[T,Yi]),dp=R.useCallback(p=>{!o||!p||Ze||(zt(!0),frappe.call({method:"frappe.client.get",args:{doctype:"Timesheet Submission",name:o},callback(x){const E=x.message;if(!E){zt(!1),frappe.msgprint(__("Unable to load linked Timesheet Submission."));return}frappe.call({method:"frappe.model.workflow.apply_workflow",args:{doc:E,action:p},callback(){zt(!1),frappe.show_alert({message:__("Workflow action applied: {0}",[p]),indicator:"green"}),Lt(o),Te(o)},error(){zt(!1)}})},error(){zt(!1)}}))},[o,Ze,Te,Lt]),Xi=R.useMemo(()=>{if(!r)return[];const p=new Set(u.filter(x=>x.type==="project").map(x=>x.name));return(r.all_projects||[]).filter(x=>!p.has(x.project_name))},[r,u]),Gi=R.useMemo(()=>{if(!r)return[];const p=new Set(u.filter(x=>x.type==="activity").map(x=>x.name));return(r.activity_types||[]).filter(x=>!p.has(x))},[r,u]),jc=R.useMemo(()=>{const p=d.toLowerCase();return p?Xi.filter(x=>x.project_name.toLowerCase().includes(p)):Xi},[Xi,d]),Pc=R.useMemo(()=>{const p=N.toLowerCase();return p?Gi.filter(x=>x.toLowerCase().includes(p)):Gi},[Gi,N]),Tc=R.useMemo(()=>{const p=((r==null?void 0:r.workflow_style)||"").toLowerCase();return{success:"green",danger:"red",primary:"blue",warning:"orange",info:"light-blue",inverse:"darkgrey"}[p]||"gray"},[r]);R.useMemo(()=>wc(r==null?void 0:r.month_year),[r]);const pp=R.useMemo(()=>(Array.isArray(r==null?void 0:r.submissions)?r.submissions:[]).slice(0,5),[r]);if(c)return h.jsx("div",{className:"ts-loading",children:"Loading timesheet..."});if(Q)return h.jsx("div",{className:"alert alert-danger",style:{margin:20},children:Q});if(!r)return null;const zc=document.getElementById("timesheet-entry-sidebar-root"),Lc=document.getElementById("timesheet-entry-context-root"),Rc=document.getElementById("timesheet-entry-actions-root"),mp=h.jsxs("div",{className:"frappe-card",style:{padding:10,height:"fit-content"},children:[h.jsx("div",{style:{fontWeight:600,marginBottom:8},children:"Available Months"}),pp.map(p=>h.jsx("button",{className:`btn btn-xs ${p.name===o?"btn-primary":"btn-default"}`,style:{width:"100%",marginBottom:6,textAlign:"left"},onClick:()=>Te(p.name),children:wc(p.month_year)},p.name))]}),hp=h.jsxs("div",{style:{display:"flex",gap:8,alignItems:"center"},children:[h.jsx("button",{className:"btn btn-default btn-sm",onClick:()=>o&&frappe.set_route("Form","Timesheet Submission",o),disabled:!o,children:"Open Submission"}),h.jsx("button",{className:"btn btn-default btn-sm",onClick:()=>o&&(r==null?void 0:r.employee)&&frappe.set_route("timesheet_workflow","employee",r.employee,"submission",o),disabled:!o||!(r!=null&&r.employee),children:"Review"}),h.jsxs("div",{className:"btn-group",children:[h.jsx("button",{className:"btn btn-default btn-sm dropdown-toggle","data-toggle":"dropdown","aria-expanded":"false",disabled:Ze,children:Ze?__("Working..."):__("Actions")}),h.jsxs("ul",{className:"dropdown-menu dropdown-menu-right ts-actions-menu",children:[h.jsx("li",{children:h.jsxs("a",{className:"ts-actions-link",href:"#",onClick:p=>{p.preventDefault(),a(!1),A(x=>!x),g("")},children:[h.jsx("i",{className:"fa fa-folder-open ts-actions-icon","aria-hidden":"true"}),h.jsx("span",{children:__("Add Project")})]})}),h.jsx("li",{children:h.jsxs("a",{className:"ts-actions-link",href:"#",onClick:p=>{p.preventDefault(),A(!1),a(x=>!x),P("")},children:[h.jsx("i",{className:"fa fa-plus-circle ts-actions-icon","aria-hidden":"true"}),h.jsx("span",{children:__("Add Activity")})]})}),xe.length>0&&h.jsx("li",{className:"divider"}),xe.map(p=>h.jsx("li",{children:h.jsxs("a",{className:"ts-actions-link",href:"#",onClick:x=>{x.preventDefault(),dp(p)},children:[h.jsx("i",{className:"fa fa-random ts-actions-icon","aria-hidden":"true"}),h.jsx("span",{children:p})]})},p))]})]}),h.jsx("button",{className:"btn btn-primary btn-sm",onClick:()=>De(!0),disabled:k,children:k?"Saving...":"Save"})]}),vp=h.jsx("div",{style:{display:"inline-flex",alignItems:"center"},children:h.jsx("span",{className:`indicator-pill ${Tc}`,style:{marginLeft:12,whiteSpace:"nowrap"},children:r.workflow_state||"Draft"})});return h.jsxs(h.Fragment,{children:[zc?Tl.createPortal(mp,zc):null,Lc?Tl.createPortal(vp,Lc):null,Rc?Tl.createPortal(hp,Rc):null,h.jsxs("div",{className:"ts-entry-wrap",children:[h.jsx("div",{className:"ts-entry-header"}),C&&h.jsxs("div",{className:"ts-add-project-panel",children:[h.jsx("input",{className:"form-control form-control-sm",placeholder:"Search projects...",value:d,onChange:p=>g(p.target.value),autoFocus:!0}),h.jsxs("div",{style:{maxHeight:200,overflowY:"auto",marginTop:6},children:[jc.length===0&&h.jsx("div",{style:{padding:"6px 10px",color:"#888",fontSize:12},children:"No projects available"}),jc.map(p=>h.jsx("div",{className:"ts-project-item",onClick:()=>cp(p.project_name),children:p.project_name},p.name))]})]}),f&&h.jsxs("div",{className:"ts-add-project-panel",children:[h.jsx("input",{className:"form-control form-control-sm",placeholder:"Search activity types...",value:N,onChange:p=>P(p.target.value),autoFocus:!0}),h.jsxs("div",{style:{maxHeight:200,overflowY:"auto",marginTop:6},children:[Pc.length===0&&h.jsx("div",{style:{padding:"6px 10px",color:"#888",fontSize:12},children:"No activity types available"}),Pc.map(p=>h.jsx("div",{className:"ts-project-item",onClick:()=>Yi(p),children:p},p))]}),h.jsxs("div",{style:{marginTop:8,borderTop:"1px solid #eee",paddingTop:8},children:[h.jsx("input",{className:"form-control form-control-sm",placeholder:"Create new activity type (e.g. Meetings, Leave, Holiday)",value:T,onChange:p=>L(p.target.value)}),h.jsx("button",{className:"btn btn-default btn-sm",style:{marginTop:6},onClick:fp,children:"Create & Add"})]})]}),O&&h.jsxs("div",{className:"ts-mobile-wrap",children:[u.map((p,x)=>h.jsxs("div",{className:"ts-mobile-section",children:[h.jsxs("div",{className:`ts-mobile-section-header ${p.type==="activity"?"ts-mobile-header-activity":""}`,children:[h.jsxs("div",{className:"ts-mobile-section-title",children:[p.type==="activity"?h.jsx("input",{className:"form-control form-control-sm ts-mobile-activity-input",value:p.name,onChange:E=>Cc(x,E.target.value),placeholder:"Activity name"}):h.jsx("span",{children:p.name}),h.jsx("span",{className:"ts-section-type-badge",children:p.type==="project"?"Project":"Activity"})]}),h.jsxs("div",{className:"ts-mobile-section-meta",children:[h.jsxs("span",{className:"ts-mobile-section-total",children:[kc(p).toFixed(1),"h"]}),p.type==="activity"&&h.jsx("button",{className:"btn btn-default btn-xs ts-mobile-delete-btn",onClick:()=>Nc(x),children:"Delete"})]})]}),p.tasks.map(E=>h.jsxs("div",{className:"ts-mobile-task",children:[h.jsxs("div",{className:"ts-mobile-task-row",children:[h.jsx("input",{className:"ts-task-input ts-mobile-task-name",type:"text",maxLength:500,value:E.task,onChange:z=>xc(x,E.id,z.target.value),placeholder:"Task description"}),h.jsxs("button",{className:`ts-mobile-hours-toggle ${W===E.id?"active":""}`,onClick:()=>ee(W===E.id?null:E.id),children:[zl(E)>0?zl(E).toFixed(1):"0","h ▾"]}),p.tasks.length>1&&h.jsx("button",{className:"ts-remove-btn",onClick:()=>Ec(x,E.id),title:"Remove row",children:"×"})]}),W===E.id&&h.jsx("div",{className:"ts-mobile-dates",children:r.dates.map(z=>h.jsxs("div",{className:`ts-mobile-date-row${z.is_weekend?" ts-weekend":""}`,children:[h.jsxs("span",{className:"ts-mobile-date-label",children:[z.day_short," ",z.date_num]}),h.jsx("input",{className:"ts-mobile-hours-input",type:"number",min:"0",step:"0.5",inputMode:"decimal",value:E.hours[z.date]||"",onChange:G=>Sc(x,E.id,z.date,G.target.value),placeholder:"0"})]},z.date))})]},E.id)),h.jsx("button",{className:"ts-mobile-addrow-btn",onClick:()=>_c(x),children:"+ Add task row"})]},p.type+"_"+p.name+"_"+x)),h.jsxs("div",{className:"ts-mobile-footer-actions",children:[h.jsx("button",{className:"btn btn-default btn-sm ts-mobile-action-btn",onClick:()=>{a(!1),A(p=>!p),g("")},children:"+ Add Project"}),h.jsx("button",{className:"btn btn-default btn-sm ts-mobile-action-btn",onClick:()=>{A(!1),a(p=>!p),P("")},children:"+ Add Activity"})]}),h.jsxs("div",{className:"ts-mobile-save-bar",children:[h.jsxs("div",{style:{fontSize:12,color:"#666"},children:["Total: ",h.jsxs("strong",{children:[Kt>0?Kt.toFixed(1):"0","h"]}),r.workflow_state&&h.jsx("span",{className:`indicator-pill ${Tc}`,style:{marginLeft:10},children:r.workflow_state})]}),h.jsx("button",{className:"btn btn-primary",onClick:()=>De(!0),disabled:k,children:k?"Saving...":"Save Timesheet"})]})]}),!O&&h.jsx("div",{className:"ts-table-wrap",children:h.jsxs("table",{className:"ts-table",children:[h.jsxs("thead",{children:[h.jsxs("tr",{className:"ts-row-dayname",children:[h.jsx("th",{className:"ts-col-sticky",style:{left:0,minWidth:180},children:"Projects / Activities"}),h.jsx("th",{className:"ts-col-sticky",style:{left:180,minWidth:220},children:"Tasks"}),r.dates.map(p=>h.jsx("th",{className:"ts-col-date"+(p.is_weekend?" ts-weekend":""),children:p.day_short},p.date)),h.jsx("th",{className:"ts-col-total",children:"Total Hours"}),h.jsx("th",{className:"ts-col-total",children:"Ratio %"})]}),h.jsxs("tr",{className:"ts-row-datenum",children:[h.jsx("th",{className:"ts-col-sticky",style:{left:0,minWidth:180,top:29}}),h.jsx("th",{className:"ts-col-sticky",style:{left:180,minWidth:220,top:29},children:"Hours"}),r.dates.map(p=>h.jsx("th",{className:"ts-col-date"+(p.is_weekend?" ts-weekend":""),style:{top:29},children:p.date_num},p.date)),h.jsx("th",{className:"ts-col-total",style:{top:29}}),h.jsx("th",{className:"ts-col-total",style:{top:29}})]})]}),h.jsxs("tbody",{children:[u.map((p,x)=>h.jsxs(Gc.Fragment,{children:[h.jsxs("tr",{className:`ts-section-header ${p.type==="activity"?"ts-section-header-activity":""}`,children:[h.jsxs("td",{colSpan:2,className:"ts-col-sticky ts-section-name",style:{left:0},children:[p.type==="activity"?h.jsx("input",{className:"form-control form-control-sm",style:{display:"inline-block",width:220,marginRight:8,height:28},value:p.name,onChange:E=>Cc(x,E.target.value),placeholder:"Activity name"}):p.name,h.jsx("span",{className:"ts-section-type-badge",children:p.type==="project"?"Project":"Activity"}),p.type==="activity"&&h.jsx("button",{className:"btn btn-default btn-xs ts-activity-delete-btn",style:{marginLeft:8},onClick:()=>Nc(x),title:"Delete activity section",children:"Delete"})]}),r.dates.map(E=>h.jsx("td",{className:E.is_weekend?"ts-weekend":""},E.date)),h.jsx("td",{}),h.jsx("td",{className:"ts-col-total",children:Kt>0?(kc(p)/Kt*100).toFixed(1)+"%":""})]}),p.tasks.map(E=>h.jsxs("tr",{children:[h.jsx("td",{className:"ts-col-sticky",style:{left:0,background:"#fff"}}),h.jsx("td",{className:"ts-col-sticky",style:{left:180,background:"#fff"},children:h.jsxs("div",{className:"ts-task-cell",children:[h.jsx("input",{className:"ts-task-input",type:"text",maxLength:500,value:E.task,onChange:z=>xc(x,E.id,z.target.value),placeholder:"Task description"}),p.tasks.length>1&&h.jsx("button",{className:"ts-remove-btn",onClick:()=>Ec(x,E.id),title:"Remove row",children:"×"})]})}),r.dates.map(z=>h.jsx("td",{className:"ts-col-date"+(z.is_weekend?" ts-weekend":""),children:h.jsx("input",{className:"ts-hours-input",type:"number",min:"0",step:"0.5",value:E.hours[z.date]||"",onChange:G=>Sc(x,E.id,z.date,G.target.value)})},z.date)),h.jsx("td",{className:"ts-col-total",children:zl(E)>0?zl(E).toFixed(1):""}),h.jsx("td",{className:"ts-col-total"})]},E.id)),h.jsxs("tr",{className:"ts-row-addtask",children:[h.jsx("td",{className:"ts-col-sticky",style:{left:0,background:"#fff"}}),h.jsx("td",{className:"ts-col-sticky",style:{left:180,background:"#fff"},children:h.jsx("button",{className:"ts-addtask-btn",onClick:()=>_c(x),children:"+ Add row"})}),r.dates.map(E=>h.jsx("td",{className:E.is_weekend?"ts-weekend":""},E.date)),h.jsx("td",{}),h.jsx("td",{})]})]},p.type+"_"+p.name+"_"+x)),h.jsxs("tr",{className:"ts-row-total",children:[h.jsx("td",{colSpan:2,className:"ts-col-sticky",style:{left:0},children:"TOTAL"}),r.dates.map(p=>h.jsx("td",{className:"ts-col-date"+(p.is_weekend?" ts-weekend":""),children:at[p.date]>0?at[p.date].toFixed(1):""},p.date)),h.jsx("td",{className:"ts-col-total",children:Kt>0?Kt.toFixed(1):""}),h.jsx("td",{className:"ts-col-total",children:Kt>0?"100%":""})]})]})]})})]})]})}window.initTimesheetEntry=function(e,t,n){const r=document.createElement("div");r.id="ts-entry-root",e.main[0].appendChild(r),vc(r).render(h.jsx(ap,{submissionName:t,onContextChange:n}))}})();
+`+o.stack}return{value:e,source:t,stack:l,digest:null}}function hi(e,t,n){return{value:e,source:null,stack:n??null,digest:t??null}}function vi(e,t){try{console.error(t.value)}catch(n){setTimeout(function(){throw n})}}var Dd=typeof WeakMap=="function"?WeakMap:Map;function Ea(e,t,n){n=rt(-1,n),n.tag=3,n.payload={element:null};var r=t.value;return n.callback=function(){hl||(hl=!0,Li=r),vi(e,t)},n}function Ca(e,t,n){n=rt(-1,n),n.tag=3;var r=e.type.getDerivedStateFromError;if(typeof r=="function"){var l=t.value;n.payload=function(){return r(l)},n.callback=function(){vi(e,t)}}var o=e.stateNode;return o!==null&&typeof o.componentDidCatch=="function"&&(n.callback=function(){vi(e,t),typeof r!="function"&&(Ct===null?Ct=new Set([this]):Ct.add(this));var i=t.stack;this.componentDidCatch(t.value,{componentStack:i!==null?i:""})}),n}function Na(e,t,n){var r=e.pingCache;if(r===null){r=e.pingCache=new Dd;var l=new Set;r.set(t,l)}else l=r.get(t),l===void 0&&(l=new Set,r.set(t,l));l.has(n)||(l.add(n),e=Zd.bind(null,e,t,n),t.then(e,e))}function ja(e){do{var t;if((t=e.tag===13)&&(t=e.memoizedState,t=t!==null?t.dehydrated!==null:!0),t)return e;e=e.return}while(e!==null);return null}function Pa(e,t,n,r,l){return e.mode&1?(e.flags|=65536,e.lanes=l,e):(e===t?e.flags|=65536:(e.flags|=128,n.flags|=131072,n.flags&=-52805,n.tag===1&&(n.alternate===null?n.tag=17:(t=rt(-1,1),t.tag=2,_t(n,t,1))),n.lanes|=1),e)}var Id=Je.ReactCurrentOwner,ke=!1;function he(e,t,n,r){t.child=e===null?Xs(t,null,n,r):mn(t,e.child,n,r)}function Ta(e,t,n,r,l){n=n.render;var o=t.ref;return vn(t,l),r=ui(e,t,n,r,o,l),n=si(),e!==null&&!ke?(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~l,lt(e,t,l)):(H&&n&&Ho(t),t.flags|=1,he(e,t,r,l),t.child)}function za(e,t,n,r,l){if(e===null){var o=n.type;return typeof o=="function"&&!Ai(o)&&o.defaultProps===void 0&&n.compare===null&&n.defaultProps===void 0?(t.tag=15,t.type=o,La(e,t,o,r,l)):(e=Sl(n.type,null,r,t,t.mode,l),e.ref=t.ref,e.return=t,t.child=e)}if(o=e.child,!(e.lanes&l)){var i=o.memoizedProps;if(n=n.compare,n=n!==null?n:Qn,n(i,r)&&e.ref===t.ref)return lt(e,t,l)}return t.flags|=1,e=Tt(o,r),e.ref=t.ref,e.return=t,t.child=e}function La(e,t,n,r,l){if(e!==null){var o=e.memoizedProps;if(Qn(o,r)&&e.ref===t.ref)if(ke=!1,t.pendingProps=r=o,(e.lanes&l)!==0)e.flags&131072&&(ke=!0);else return t.lanes=e.lanes,lt(e,t,l)}return yi(e,t,n,r,l)}function Ra(e,t,n){var r=t.pendingProps,l=r.children,o=e!==null?e.memoizedState:null;if(r.mode==="hidden")if(!(t.mode&1))t.memoizedState={baseLanes:0,cachePool:null,transitions:null},$(wn,je),je|=n;else{if(!(n&1073741824))return e=o!==null?o.baseLanes|n:n,t.lanes=t.childLanes=1073741824,t.memoizedState={baseLanes:e,cachePool:null,transitions:null},t.updateQueue=null,$(wn,je),je|=e,null;t.memoizedState={baseLanes:0,cachePool:null,transitions:null},r=o!==null?o.baseLanes:n,$(wn,je),je|=r}else o!==null?(r=o.baseLanes|n,t.memoizedState=null):r=n,$(wn,je),je|=r;return he(e,t,l,n),t.child}function Ma(e,t){var n=t.ref;(e===null&&n!==null||e!==null&&e.ref!==n)&&(t.flags|=512,t.flags|=2097152)}function yi(e,t,n,r,l){var o=ge(n)?Ft:se.current;return o=cn(t,o),vn(t,l),n=ui(e,t,n,r,o,l),r=si(),e!==null&&!ke?(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~l,lt(e,t,l)):(H&&r&&Ho(t),t.flags|=1,he(e,t,n,l),t.child)}function Oa(e,t,n,r,l){if(ge(n)){var o=!0;Kr(t)}else o=!1;if(vn(t,l),t.stateNode===null)cl(e,t),xa(t,n,r),mi(t,n,r,l),r=!0;else if(e===null){var i=t.stateNode,u=t.memoizedProps;i.props=u;var s=i.context,c=n.contextType;typeof c=="object"&&c!==null?c=Re(c):(c=ge(n)?Ft:se.current,c=cn(t,c));var y=n.getDerivedStateFromProps,h=typeof y=="function"||typeof i.getSnapshotBeforeUpdate=="function";h||typeof i.UNSAFE_componentWillReceiveProps!="function"&&typeof i.componentWillReceiveProps!="function"||(u!==r||s!==c)&&_a(t,i,r,c),xt=!1;var p=t.memoizedState;i.state=p,tl(t,r,i,l),s=t.memoizedState,u!==r||p!==s||ye.current||xt?(typeof y=="function"&&(pi(t,n,y,r),s=t.memoizedState),(u=xt||Sa(t,n,u,r,p,s,c))?(h||typeof i.UNSAFE_componentWillMount!="function"&&typeof i.componentWillMount!="function"||(typeof i.componentWillMount=="function"&&i.componentWillMount(),typeof i.UNSAFE_componentWillMount=="function"&&i.UNSAFE_componentWillMount()),typeof i.componentDidMount=="function"&&(t.flags|=4194308)):(typeof i.componentDidMount=="function"&&(t.flags|=4194308),t.memoizedProps=r,t.memoizedState=s),i.props=r,i.state=s,i.context=c,r=u):(typeof i.componentDidMount=="function"&&(t.flags|=4194308),r=!1)}else{i=t.stateNode,Zs(e,t),u=t.memoizedProps,c=t.type===t.elementType?u:$e(t.type,u),i.props=c,h=t.pendingProps,p=i.context,s=n.contextType,typeof s=="object"&&s!==null?s=Re(s):(s=ge(n)?Ft:se.current,s=cn(t,s));var w=n.getDerivedStateFromProps;(y=typeof w=="function"||typeof i.getSnapshotBeforeUpdate=="function")||typeof i.UNSAFE_componentWillReceiveProps!="function"&&typeof i.componentWillReceiveProps!="function"||(u!==h||p!==s)&&_a(t,i,r,s),xt=!1,p=t.memoizedState,i.state=p,tl(t,r,i,l);var S=t.memoizedState;u!==h||p!==S||ye.current||xt?(typeof w=="function"&&(pi(t,n,w,r),S=t.memoizedState),(c=xt||Sa(t,n,c,r,p,S,s)||!1)?(y||typeof i.UNSAFE_componentWillUpdate!="function"&&typeof i.componentWillUpdate!="function"||(typeof i.componentWillUpdate=="function"&&i.componentWillUpdate(r,S,s),typeof i.UNSAFE_componentWillUpdate=="function"&&i.UNSAFE_componentWillUpdate(r,S,s)),typeof i.componentDidUpdate=="function"&&(t.flags|=4),typeof i.getSnapshotBeforeUpdate=="function"&&(t.flags|=1024)):(typeof i.componentDidUpdate!="function"||u===e.memoizedProps&&p===e.memoizedState||(t.flags|=4),typeof i.getSnapshotBeforeUpdate!="function"||u===e.memoizedProps&&p===e.memoizedState||(t.flags|=1024),t.memoizedProps=r,t.memoizedState=S),i.props=r,i.state=S,i.context=s,r=c):(typeof i.componentDidUpdate!="function"||u===e.memoizedProps&&p===e.memoizedState||(t.flags|=4),typeof i.getSnapshotBeforeUpdate!="function"||u===e.memoizedProps&&p===e.memoizedState||(t.flags|=1024),r=!1)}return gi(e,t,n,r,o,l)}function gi(e,t,n,r,l,o){Ma(e,t);var i=(t.flags&128)!==0;if(!r&&!i)return l&&Us(t,n,!1),lt(e,t,o);r=t.stateNode,Id.current=t;var u=i&&typeof n.getDerivedStateFromError!="function"?null:r.render();return t.flags|=1,e!==null&&i?(t.child=mn(t,e.child,null,o),t.child=mn(t,null,u,o)):he(e,t,u,o),t.memoizedState=r.state,l&&Us(t,n,!0),t.child}function Fa(e){var t=e.stateNode;t.pendingContext?Is(e,t.pendingContext,t.pendingContext!==t.context):t.context&&Is(e,t.context,!1),ti(e,t.containerInfo)}function Da(e,t,n,r,l){return pn(),Xo(l),t.flags|=256,he(e,t,n,r),t.child}var ki={dehydrated:null,treeContext:null,retryLane:0};function wi(e){return{baseLanes:e,cachePool:null,transitions:null}}function Ia(e,t,n){var r=t.pendingProps,l=Y.current,o=!1,i=(t.flags&128)!==0,u;if((u=i)||(u=e!==null&&e.memoizedState===null?!1:(l&2)!==0),u?(o=!0,t.flags&=-129):(e===null||e.memoizedState!==null)&&(l|=1),$(Y,l&1),e===null)return Yo(t),e=t.memoizedState,e!==null&&(e=e.dehydrated,e!==null)?(t.mode&1?e.data==="$!"?t.lanes=8:t.lanes=1073741824:t.lanes=1,null):(i=r.children,e=r.fallback,o?(r=t.mode,o=t.child,i={mode:"hidden",children:i},!(r&1)&&o!==null?(o.childLanes=0,o.pendingProps=i):o=xl(i,r,0,null),e=Qt(e,r,n,null),o.return=t,e.return=t,o.sibling=e,t.child=o,t.child.memoizedState=wi(n),t.memoizedState=ki,e):Si(t,i));if(l=e.memoizedState,l!==null&&(u=l.dehydrated,u!==null))return Ad(e,t,i,r,u,l,n);if(o){o=r.fallback,i=t.mode,l=e.child,u=l.sibling;var s={mode:"hidden",children:r.children};return!(i&1)&&t.child!==l?(r=t.child,r.childLanes=0,r.pendingProps=s,t.deletions=null):(r=Tt(l,s),r.subtreeFlags=l.subtreeFlags&14680064),u!==null?o=Tt(u,o):(o=Qt(o,i,n,null),o.flags|=2),o.return=t,r.return=t,r.sibling=o,t.child=r,r=o,o=t.child,i=e.child.memoizedState,i=i===null?wi(n):{baseLanes:i.baseLanes|n,cachePool:null,transitions:i.transitions},o.memoizedState=i,o.childLanes=e.childLanes&~n,t.memoizedState=ki,r}return o=e.child,e=o.sibling,r=Tt(o,{mode:"visible",children:r.children}),!(t.mode&1)&&(r.lanes=n),r.return=t,r.sibling=null,e!==null&&(n=t.deletions,n===null?(t.deletions=[e],t.flags|=16):n.push(e)),t.child=r,t.memoizedState=null,r}function Si(e,t){return t=xl({mode:"visible",children:t},e.mode,0,null),t.return=e,e.child=t}function al(e,t,n,r){return r!==null&&Xo(r),mn(t,e.child,null,n),e=Si(t,t.pendingProps.children),e.flags|=2,t.memoizedState=null,e}function Ad(e,t,n,r,l,o,i){if(n)return t.flags&256?(t.flags&=-257,r=hi(Error(k(422))),al(e,t,i,r)):t.memoizedState!==null?(t.child=e.child,t.flags|=128,null):(o=r.fallback,l=t.mode,r=xl({mode:"visible",children:r.children},l,0,null),o=Qt(o,l,i,null),o.flags|=2,r.return=t,o.return=t,r.sibling=o,t.child=r,t.mode&1&&mn(t,e.child,null,i),t.child.memoizedState=wi(i),t.memoizedState=ki,o);if(!(t.mode&1))return al(e,t,i,null);if(l.data==="$!"){if(r=l.nextSibling&&l.nextSibling.dataset,r)var u=r.dgst;return r=u,o=Error(k(419)),r=hi(o,r,void 0),al(e,t,i,r)}if(u=(i&e.childLanes)!==0,ke||u){if(r=le,r!==null){switch(i&-i){case 4:l=2;break;case 16:l=8;break;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:l=32;break;case 536870912:l=268435456;break;default:l=0}l=l&(r.suspendedLanes|i)?0:l,l!==0&&l!==o.retryLane&&(o.retryLane=l,nt(e,l),We(r,e,l,-1))}return Ii(),r=hi(Error(k(421))),al(e,t,i,r)}return l.data==="$?"?(t.flags|=128,t.child=e.child,t=Jd.bind(null,e),l._reactRetry=t,null):(e=o.treeContext,Ne=gt(l.nextSibling),Ce=t,H=!0,Ue=null,e!==null&&(ze[Le++]=et,ze[Le++]=tt,ze[Le++]=Dt,et=e.id,tt=e.overflow,Dt=t),t=Si(t,r.children),t.flags|=4096,t)}function Aa(e,t,n){e.lanes|=t;var r=e.alternate;r!==null&&(r.lanes|=t),qo(e.return,t,n)}function xi(e,t,n,r,l){var o=e.memoizedState;o===null?e.memoizedState={isBackwards:t,rendering:null,renderingStartTime:0,last:r,tail:n,tailMode:l}:(o.isBackwards=t,o.rendering=null,o.renderingStartTime=0,o.last=r,o.tail=n,o.tailMode=l)}function Ua(e,t,n){var r=t.pendingProps,l=r.revealOrder,o=r.tail;if(he(e,t,r.children,n),r=Y.current,r&2)r=r&1|2,t.flags|=128;else{if(e!==null&&e.flags&128)e:for(e=t.child;e!==null;){if(e.tag===13)e.memoizedState!==null&&Aa(e,n,t);else if(e.tag===19)Aa(e,n,t);else if(e.child!==null){e.child.return=e,e=e.child;continue}if(e===t)break e;for(;e.sibling===null;){if(e.return===null||e.return===t)break e;e=e.return}e.sibling.return=e.return,e=e.sibling}r&=1}if($(Y,r),!(t.mode&1))t.memoizedState=null;else switch(l){case"forwards":for(n=t.child,l=null;n!==null;)e=n.alternate,e!==null&&nl(e)===null&&(l=n),n=n.sibling;n=l,n===null?(l=t.child,t.child=null):(l=n.sibling,n.sibling=null),xi(t,!1,l,n,o);break;case"backwards":for(n=null,l=t.child,t.child=null;l!==null;){if(e=l.alternate,e!==null&&nl(e)===null){t.child=l;break}e=l.sibling,l.sibling=n,n=l,l=e}xi(t,!0,n,null,o);break;case"together":xi(t,!1,null,null,void 0);break;default:t.memoizedState=null}return t.child}function cl(e,t){!(t.mode&1)&&e!==null&&(e.alternate=null,t.alternate=null,t.flags|=2)}function lt(e,t,n){if(e!==null&&(t.dependencies=e.dependencies),Vt|=t.lanes,!(n&t.childLanes))return null;if(e!==null&&t.child!==e.child)throw Error(k(153));if(t.child!==null){for(e=t.child,n=Tt(e,e.pendingProps),t.child=n,n.return=t;e.sibling!==null;)e=e.sibling,n=n.sibling=Tt(e,e.pendingProps),n.return=t;n.sibling=null}return t.child}function Ud(e,t,n){switch(t.tag){case 3:Fa(t),pn();break;case 5:bs(t);break;case 1:ge(t.type)&&Kr(t);break;case 4:ti(t,t.stateNode.containerInfo);break;case 10:var r=t.type._context,l=t.memoizedProps.value;$(qr,r._currentValue),r._currentValue=l;break;case 13:if(r=t.memoizedState,r!==null)return r.dehydrated!==null?($(Y,Y.current&1),t.flags|=128,null):n&t.child.childLanes?Ia(e,t,n):($(Y,Y.current&1),e=lt(e,t,n),e!==null?e.sibling:null);$(Y,Y.current&1);break;case 19:if(r=(n&t.childLanes)!==0,e.flags&128){if(r)return Ua(e,t,n);t.flags|=128}if(l=t.memoizedState,l!==null&&(l.rendering=null,l.tail=null,l.lastEffect=null),$(Y,Y.current),r)break;return null;case 22:case 23:return t.lanes=0,Ra(e,t,n)}return lt(e,t,n)}var $a,_i,Va,Ba;$a=function(e,t){for(var n=t.child;n!==null;){if(n.tag===5||n.tag===6)e.appendChild(n.stateNode);else if(n.tag!==4&&n.child!==null){n.child.return=n,n=n.child;continue}if(n===t)break;for(;n.sibling===null;){if(n.return===null||n.return===t)return;n=n.return}n.sibling.return=n.return,n=n.sibling}},_i=function(){},Va=function(e,t,n,r){var l=e.memoizedProps;if(l!==r){e=t.stateNode,Ut(Ye.current);var o=null;switch(n){case"input":l=Jl(e,l),r=Jl(e,r),o=[];break;case"select":l=K({},l,{value:void 0}),r=K({},r,{value:void 0}),o=[];break;case"textarea":l=eo(e,l),r=eo(e,r),o=[];break;default:typeof l.onClick!="function"&&typeof r.onClick=="function"&&(e.onclick=Wr)}no(n,r);var i;n=null;for(c in l)if(!r.hasOwnProperty(c)&&l.hasOwnProperty(c)&&l[c]!=null)if(c==="style"){var u=l[c];for(i in u)u.hasOwnProperty(i)&&(n||(n={}),n[i]="")}else c!=="dangerouslySetInnerHTML"&&c!=="children"&&c!=="suppressContentEditableWarning"&&c!=="suppressHydrationWarning"&&c!=="autoFocus"&&(En.hasOwnProperty(c)?o||(o=[]):(o=o||[]).push(c,null));for(c in r){var s=r[c];if(u=l!=null?l[c]:void 0,r.hasOwnProperty(c)&&s!==u&&(s!=null||u!=null))if(c==="style")if(u){for(i in u)!u.hasOwnProperty(i)||s&&s.hasOwnProperty(i)||(n||(n={}),n[i]="");for(i in s)s.hasOwnProperty(i)&&u[i]!==s[i]&&(n||(n={}),n[i]=s[i])}else n||(o||(o=[]),o.push(c,n)),n=s;else c==="dangerouslySetInnerHTML"?(s=s?s.__html:void 0,u=u?u.__html:void 0,s!=null&&u!==s&&(o=o||[]).push(c,s)):c==="children"?typeof s!="string"&&typeof s!="number"||(o=o||[]).push(c,""+s):c!=="suppressContentEditableWarning"&&c!=="suppressHydrationWarning"&&(En.hasOwnProperty(c)?(s!=null&&c==="onScroll"&&V("scroll",e),o||u===s||(o=[])):(o=o||[]).push(c,s))}n&&(o=o||[]).push("style",n);var c=o;(t.updateQueue=c)&&(t.flags|=4)}},Ba=function(e,t,n,r){n!==r&&(t.flags|=4)};function ir(e,t){if(!H)switch(e.tailMode){case"hidden":t=e.tail;for(var n=null;t!==null;)t.alternate!==null&&(n=t),t=t.sibling;n===null?e.tail=null:n.sibling=null;break;case"collapsed":n=e.tail;for(var r=null;n!==null;)n.alternate!==null&&(r=n),n=n.sibling;r===null?t||e.tail===null?e.tail=null:e.tail.sibling=null:r.sibling=null}}function ce(e){var t=e.alternate!==null&&e.alternate.child===e.child,n=0,r=0;if(t)for(var l=e.child;l!==null;)n|=l.lanes|l.childLanes,r|=l.subtreeFlags&14680064,r|=l.flags&14680064,l.return=e,l=l.sibling;else for(l=e.child;l!==null;)n|=l.lanes|l.childLanes,r|=l.subtreeFlags,r|=l.flags,l.return=e,l=l.sibling;return e.subtreeFlags|=r,e.childLanes=n,t}function $d(e,t,n){var r=t.pendingProps;switch(Qo(t),t.tag){case 2:case 16:case 15:case 0:case 11:case 7:case 8:case 12:case 9:case 14:return ce(t),null;case 1:return ge(t.type)&&Qr(),ce(t),null;case 3:return r=t.stateNode,yn(),B(ye),B(se),li(),r.pendingContext&&(r.context=r.pendingContext,r.pendingContext=null),(e===null||e.child===null)&&(Zr(t)?t.flags|=4:e===null||e.memoizedState.isDehydrated&&!(t.flags&256)||(t.flags|=1024,Ue!==null&&(Oi(Ue),Ue=null))),_i(e,t),ce(t),null;case 5:ni(t);var l=Ut(tr.current);if(n=t.type,e!==null&&t.stateNode!=null)Va(e,t,n,r,l),e.ref!==t.ref&&(t.flags|=512,t.flags|=2097152);else{if(!r){if(t.stateNode===null)throw Error(k(166));return ce(t),null}if(e=Ut(Ye.current),Zr(t)){r=t.stateNode,n=t.type;var o=t.memoizedProps;switch(r[Ke]=t,r[Zn]=o,e=(t.mode&1)!==0,n){case"dialog":V("cancel",r),V("close",r);break;case"iframe":case"object":case"embed":V("load",r);break;case"video":case"audio":for(l=0;l<Yn.length;l++)V(Yn[l],r);break;case"source":V("error",r);break;case"img":case"image":case"link":V("error",r),V("load",r);break;case"details":V("toggle",r);break;case"input":xu(r,o),V("invalid",r);break;case"select":r._wrapperState={wasMultiple:!!o.multiple},V("invalid",r);break;case"textarea":Cu(r,o),V("invalid",r)}no(n,o),l=null;for(var i in o)if(o.hasOwnProperty(i)){var u=o[i];i==="children"?typeof u=="string"?r.textContent!==u&&(o.suppressHydrationWarning!==!0&&Br(r.textContent,u,e),l=["children",u]):typeof u=="number"&&r.textContent!==""+u&&(o.suppressHydrationWarning!==!0&&Br(r.textContent,u,e),l=["children",""+u]):En.hasOwnProperty(i)&&u!=null&&i==="onScroll"&&V("scroll",r)}switch(n){case"input":kr(r),Eu(r,o,!0);break;case"textarea":kr(r),ju(r);break;case"select":case"option":break;default:typeof o.onClick=="function"&&(r.onclick=Wr)}r=l,t.updateQueue=r,r!==null&&(t.flags|=4)}else{i=l.nodeType===9?l:l.ownerDocument,e==="http://www.w3.org/1999/xhtml"&&(e=Pu(n)),e==="http://www.w3.org/1999/xhtml"?n==="script"?(e=i.createElement("div"),e.innerHTML="<script><\/script>",e=e.removeChild(e.firstChild)):typeof r.is=="string"?e=i.createElement(n,{is:r.is}):(e=i.createElement(n),n==="select"&&(i=e,r.multiple?i.multiple=!0:r.size&&(i.size=r.size))):e=i.createElementNS(e,n),e[Ke]=t,e[Zn]=r,$a(e,t,!1,!1),t.stateNode=e;e:{switch(i=ro(n,r),n){case"dialog":V("cancel",e),V("close",e),l=r;break;case"iframe":case"object":case"embed":V("load",e),l=r;break;case"video":case"audio":for(l=0;l<Yn.length;l++)V(Yn[l],e);l=r;break;case"source":V("error",e),l=r;break;case"img":case"image":case"link":V("error",e),V("load",e),l=r;break;case"details":V("toggle",e),l=r;break;case"input":xu(e,r),l=Jl(e,r),V("invalid",e);break;case"option":l=r;break;case"select":e._wrapperState={wasMultiple:!!r.multiple},l=K({},r,{value:void 0}),V("invalid",e);break;case"textarea":Cu(e,r),l=eo(e,r),V("invalid",e);break;default:l=r}no(n,l),u=l;for(o in u)if(u.hasOwnProperty(o)){var s=u[o];o==="style"?Lu(e,s):o==="dangerouslySetInnerHTML"?(s=s?s.__html:void 0,s!=null&&Tu(e,s)):o==="children"?typeof s=="string"?(n!=="textarea"||s!=="")&&Pn(e,s):typeof s=="number"&&Pn(e,""+s):o!=="suppressContentEditableWarning"&&o!=="suppressHydrationWarning"&&o!=="autoFocus"&&(En.hasOwnProperty(o)?s!=null&&o==="onScroll"&&V("scroll",e):s!=null&&$l(e,o,s,i))}switch(n){case"input":kr(e),Eu(e,r,!1);break;case"textarea":kr(e),ju(e);break;case"option":r.value!=null&&e.setAttribute("value",""+ft(r.value));break;case"select":e.multiple=!!r.multiple,o=r.value,o!=null?Jt(e,!!r.multiple,o,!1):r.defaultValue!=null&&Jt(e,!!r.multiple,r.defaultValue,!0);break;default:typeof l.onClick=="function"&&(e.onclick=Wr)}switch(n){case"button":case"input":case"select":case"textarea":r=!!r.autoFocus;break e;case"img":r=!0;break e;default:r=!1}}r&&(t.flags|=4)}t.ref!==null&&(t.flags|=512,t.flags|=2097152)}return ce(t),null;case 6:if(e&&t.stateNode!=null)Ba(e,t,e.memoizedProps,r);else{if(typeof r!="string"&&t.stateNode===null)throw Error(k(166));if(n=Ut(tr.current),Ut(Ye.current),Zr(t)){if(r=t.stateNode,n=t.memoizedProps,r[Ke]=t,(o=r.nodeValue!==n)&&(e=Ce,e!==null))switch(e.tag){case 3:Br(r.nodeValue,n,(e.mode&1)!==0);break;case 5:e.memoizedProps.suppressHydrationWarning!==!0&&Br(r.nodeValue,n,(e.mode&1)!==0)}o&&(t.flags|=4)}else r=(n.nodeType===9?n:n.ownerDocument).createTextNode(r),r[Ke]=t,t.stateNode=r}return ce(t),null;case 13:if(B(Y),r=t.memoizedState,e===null||e.memoizedState!==null&&e.memoizedState.dehydrated!==null){if(H&&Ne!==null&&t.mode&1&&!(t.flags&128))Qs(),pn(),t.flags|=98560,o=!1;else if(o=Zr(t),r!==null&&r.dehydrated!==null){if(e===null){if(!o)throw Error(k(318));if(o=t.memoizedState,o=o!==null?o.dehydrated:null,!o)throw Error(k(317));o[Ke]=t}else pn(),!(t.flags&128)&&(t.memoizedState=null),t.flags|=4;ce(t),o=!1}else Ue!==null&&(Oi(Ue),Ue=null),o=!0;if(!o)return t.flags&65536?t:null}return t.flags&128?(t.lanes=n,t):(r=r!==null,r!==(e!==null&&e.memoizedState!==null)&&r&&(t.child.flags|=8192,t.mode&1&&(e===null||Y.current&1?ne===0&&(ne=3):Ii())),t.updateQueue!==null&&(t.flags|=4),ce(t),null);case 4:return yn(),_i(e,t),e===null&&Xn(t.stateNode.containerInfo),ce(t),null;case 10:return Jo(t.type._context),ce(t),null;case 17:return ge(t.type)&&Qr(),ce(t),null;case 19:if(B(Y),o=t.memoizedState,o===null)return ce(t),null;if(r=(t.flags&128)!==0,i=o.rendering,i===null)if(r)ir(o,!1);else{if(ne!==0||e!==null&&e.flags&128)for(e=t.child;e!==null;){if(i=nl(e),i!==null){for(t.flags|=128,ir(o,!1),r=i.updateQueue,r!==null&&(t.updateQueue=r,t.flags|=4),t.subtreeFlags=0,r=n,n=t.child;n!==null;)o=n,e=r,o.flags&=14680066,i=o.alternate,i===null?(o.childLanes=0,o.lanes=e,o.child=null,o.subtreeFlags=0,o.memoizedProps=null,o.memoizedState=null,o.updateQueue=null,o.dependencies=null,o.stateNode=null):(o.childLanes=i.childLanes,o.lanes=i.lanes,o.child=i.child,o.subtreeFlags=0,o.deletions=null,o.memoizedProps=i.memoizedProps,o.memoizedState=i.memoizedState,o.updateQueue=i.updateQueue,o.type=i.type,e=i.dependencies,o.dependencies=e===null?null:{lanes:e.lanes,firstContext:e.firstContext}),n=n.sibling;return $(Y,Y.current&1|2),t.child}e=e.sibling}o.tail!==null&&J()>Sn&&(t.flags|=128,r=!0,ir(o,!1),t.lanes=4194304)}else{if(!r)if(e=nl(i),e!==null){if(t.flags|=128,r=!0,n=e.updateQueue,n!==null&&(t.updateQueue=n,t.flags|=4),ir(o,!0),o.tail===null&&o.tailMode==="hidden"&&!i.alternate&&!H)return ce(t),null}else 2*J()-o.renderingStartTime>Sn&&n!==1073741824&&(t.flags|=128,r=!0,ir(o,!1),t.lanes=4194304);o.isBackwards?(i.sibling=t.child,t.child=i):(n=o.last,n!==null?n.sibling=i:t.child=i,o.last=i)}return o.tail!==null?(t=o.tail,o.rendering=t,o.tail=t.sibling,o.renderingStartTime=J(),t.sibling=null,n=Y.current,$(Y,r?n&1|2:n&1),t):(ce(t),null);case 22:case 23:return Di(),r=t.memoizedState!==null,e!==null&&e.memoizedState!==null!==r&&(t.flags|=8192),r&&t.mode&1?je&1073741824&&(ce(t),t.subtreeFlags&6&&(t.flags|=8192)):ce(t),null;case 24:return null;case 25:return null}throw Error(k(156,t.tag))}function Vd(e,t){switch(Qo(t),t.tag){case 1:return ge(t.type)&&Qr(),e=t.flags,e&65536?(t.flags=e&-65537|128,t):null;case 3:return yn(),B(ye),B(se),li(),e=t.flags,e&65536&&!(e&128)?(t.flags=e&-65537|128,t):null;case 5:return ni(t),null;case 13:if(B(Y),e=t.memoizedState,e!==null&&e.dehydrated!==null){if(t.alternate===null)throw Error(k(340));pn()}return e=t.flags,e&65536?(t.flags=e&-65537|128,t):null;case 19:return B(Y),null;case 4:return yn(),null;case 10:return Jo(t.type._context),null;case 22:case 23:return Di(),null;case 24:return null;default:return null}}var fl=!1,fe=!1,Bd=typeof WeakSet=="function"?WeakSet:Set,j=null;function kn(e,t){var n=e.ref;if(n!==null)if(typeof n=="function")try{n(null)}catch(r){Z(e,t,r)}else n.current=null}function Ei(e,t,n){try{n()}catch(r){Z(e,t,r)}}var Wa=!1;function Wd(e,t){if(Do=Lr,e=Ss(),Po(e)){if("selectionStart"in e)var n={start:e.selectionStart,end:e.selectionEnd};else e:{n=(n=e.ownerDocument)&&n.defaultView||window;var r=n.getSelection&&n.getSelection();if(r&&r.rangeCount!==0){n=r.anchorNode;var l=r.anchorOffset,o=r.focusNode;r=r.focusOffset;try{n.nodeType,o.nodeType}catch{n=null;break e}var i=0,u=-1,s=-1,c=0,y=0,h=e,p=null;t:for(;;){for(var w;h!==n||l!==0&&h.nodeType!==3||(u=i+l),h!==o||r!==0&&h.nodeType!==3||(s=i+r),h.nodeType===3&&(i+=h.nodeValue.length),(w=h.firstChild)!==null;)p=h,h=w;for(;;){if(h===e)break t;if(p===n&&++c===l&&(u=i),p===o&&++y===r&&(s=i),(w=h.nextSibling)!==null)break;h=p,p=h.parentNode}h=w}n=u===-1||s===-1?null:{start:u,end:s}}else n=null}n=n||{start:0,end:0}}else n=null;for(Io={focusedElem:e,selectionRange:n},Lr=!1,j=t;j!==null;)if(t=j,e=t.child,(t.subtreeFlags&1028)!==0&&e!==null)e.return=t,j=e;else for(;j!==null;){t=j;try{var S=t.alternate;if(t.flags&1024)switch(t.tag){case 0:case 11:case 15:break;case 1:if(S!==null){var E=S.memoizedProps,U=S.memoizedState,f=t.stateNode,a=f.getSnapshotBeforeUpdate(t.elementType===t.type?E:$e(t.type,E),U);f.__reactInternalSnapshotBeforeUpdate=a}break;case 3:var d=t.stateNode.containerInfo;d.nodeType===1?d.textContent="":d.nodeType===9&&d.documentElement&&d.removeChild(d.documentElement);break;case 5:case 6:case 4:case 17:break;default:throw Error(k(163))}}catch(g){Z(t,t.return,g)}if(e=t.sibling,e!==null){e.return=t.return,j=e;break}j=t.return}return S=Wa,Wa=!1,S}function ur(e,t,n){var r=t.updateQueue;if(r=r!==null?r.lastEffect:null,r!==null){var l=r=r.next;do{if((l.tag&e)===e){var o=l.destroy;l.destroy=void 0,o!==void 0&&Ei(t,n,o)}l=l.next}while(l!==r)}}function dl(e,t){if(t=t.updateQueue,t=t!==null?t.lastEffect:null,t!==null){var n=t=t.next;do{if((n.tag&e)===e){var r=n.create;n.destroy=r()}n=n.next}while(n!==t)}}function Ci(e){var t=e.ref;if(t!==null){var n=e.stateNode;switch(e.tag){case 5:e=n;break;default:e=n}typeof t=="function"?t(e):t.current=e}}function Ha(e){var t=e.alternate;t!==null&&(e.alternate=null,Ha(t)),e.child=null,e.deletions=null,e.sibling=null,e.tag===5&&(t=e.stateNode,t!==null&&(delete t[Ke],delete t[Zn],delete t[Vo],delete t[Cd],delete t[Nd])),e.stateNode=null,e.return=null,e.dependencies=null,e.memoizedProps=null,e.memoizedState=null,e.pendingProps=null,e.stateNode=null,e.updateQueue=null}function Qa(e){return e.tag===5||e.tag===3||e.tag===4}function Ka(e){e:for(;;){for(;e.sibling===null;){if(e.return===null||Qa(e.return))return null;e=e.return}for(e.sibling.return=e.return,e=e.sibling;e.tag!==5&&e.tag!==6&&e.tag!==18;){if(e.flags&2||e.child===null||e.tag===4)continue e;e.child.return=e,e=e.child}if(!(e.flags&2))return e.stateNode}}function Ni(e,t,n){var r=e.tag;if(r===5||r===6)e=e.stateNode,t?n.nodeType===8?n.parentNode.insertBefore(e,t):n.insertBefore(e,t):(n.nodeType===8?(t=n.parentNode,t.insertBefore(e,n)):(t=n,t.appendChild(e)),n=n._reactRootContainer,n!=null||t.onclick!==null||(t.onclick=Wr));else if(r!==4&&(e=e.child,e!==null))for(Ni(e,t,n),e=e.sibling;e!==null;)Ni(e,t,n),e=e.sibling}function ji(e,t,n){var r=e.tag;if(r===5||r===6)e=e.stateNode,t?n.insertBefore(e,t):n.appendChild(e);else if(r!==4&&(e=e.child,e!==null))for(ji(e,t,n),e=e.sibling;e!==null;)ji(e,t,n),e=e.sibling}var ie=null,Ve=!1;function Et(e,t,n){for(n=n.child;n!==null;)Ya(e,t,n),n=n.sibling}function Ya(e,t,n){if(Qe&&typeof Qe.onCommitFiberUnmount=="function")try{Qe.onCommitFiberUnmount(Cr,n)}catch{}switch(n.tag){case 5:fe||kn(n,t);case 6:var r=ie,l=Ve;ie=null,Et(e,t,n),ie=r,Ve=l,ie!==null&&(Ve?(e=ie,n=n.stateNode,e.nodeType===8?e.parentNode.removeChild(n):e.removeChild(n)):ie.removeChild(n.stateNode));break;case 18:ie!==null&&(Ve?(e=ie,n=n.stateNode,e.nodeType===8?$o(e.parentNode,n):e.nodeType===1&&$o(e,n),Un(e)):$o(ie,n.stateNode));break;case 4:r=ie,l=Ve,ie=n.stateNode.containerInfo,Ve=!0,Et(e,t,n),ie=r,Ve=l;break;case 0:case 11:case 14:case 15:if(!fe&&(r=n.updateQueue,r!==null&&(r=r.lastEffect,r!==null))){l=r=r.next;do{var o=l,i=o.destroy;o=o.tag,i!==void 0&&(o&2||o&4)&&Ei(n,t,i),l=l.next}while(l!==r)}Et(e,t,n);break;case 1:if(!fe&&(kn(n,t),r=n.stateNode,typeof r.componentWillUnmount=="function"))try{r.props=n.memoizedProps,r.state=n.memoizedState,r.componentWillUnmount()}catch(u){Z(n,t,u)}Et(e,t,n);break;case 21:Et(e,t,n);break;case 22:n.mode&1?(fe=(r=fe)||n.memoizedState!==null,Et(e,t,n),fe=r):Et(e,t,n);break;default:Et(e,t,n)}}function Xa(e){var t=e.updateQueue;if(t!==null){e.updateQueue=null;var n=e.stateNode;n===null&&(n=e.stateNode=new Bd),t.forEach(function(r){var l=qd.bind(null,e,r);n.has(r)||(n.add(r),r.then(l,l))})}}function Be(e,t){var n=t.deletions;if(n!==null)for(var r=0;r<n.length;r++){var l=n[r];try{var o=e,i=t,u=i;e:for(;u!==null;){switch(u.tag){case 5:ie=u.stateNode,Ve=!1;break e;case 3:ie=u.stateNode.containerInfo,Ve=!0;break e;case 4:ie=u.stateNode.containerInfo,Ve=!0;break e}u=u.return}if(ie===null)throw Error(k(160));Ya(o,i,l),ie=null,Ve=!1;var s=l.alternate;s!==null&&(s.return=null),l.return=null}catch(c){Z(l,t,c)}}if(t.subtreeFlags&12854)for(t=t.child;t!==null;)Ga(t,e),t=t.sibling}function Ga(e,t){var n=e.alternate,r=e.flags;switch(e.tag){case 0:case 11:case 14:case 15:if(Be(t,e),Ge(e),r&4){try{ur(3,e,e.return),dl(3,e)}catch(E){Z(e,e.return,E)}try{ur(5,e,e.return)}catch(E){Z(e,e.return,E)}}break;case 1:Be(t,e),Ge(e),r&512&&n!==null&&kn(n,n.return);break;case 5:if(Be(t,e),Ge(e),r&512&&n!==null&&kn(n,n.return),e.flags&32){var l=e.stateNode;try{Pn(l,"")}catch(E){Z(e,e.return,E)}}if(r&4&&(l=e.stateNode,l!=null)){var o=e.memoizedProps,i=n!==null?n.memoizedProps:o,u=e.type,s=e.updateQueue;if(e.updateQueue=null,s!==null)try{u==="input"&&o.type==="radio"&&o.name!=null&&_u(l,o),ro(u,i);var c=ro(u,o);for(i=0;i<s.length;i+=2){var y=s[i],h=s[i+1];y==="style"?Lu(l,h):y==="dangerouslySetInnerHTML"?Tu(l,h):y==="children"?Pn(l,h):$l(l,y,h,c)}switch(u){case"input":ql(l,o);break;case"textarea":Nu(l,o);break;case"select":var p=l._wrapperState.wasMultiple;l._wrapperState.wasMultiple=!!o.multiple;var w=o.value;w!=null?Jt(l,!!o.multiple,w,!1):p!==!!o.multiple&&(o.defaultValue!=null?Jt(l,!!o.multiple,o.defaultValue,!0):Jt(l,!!o.multiple,o.multiple?[]:"",!1))}l[Zn]=o}catch(E){Z(e,e.return,E)}}break;case 6:if(Be(t,e),Ge(e),r&4){if(e.stateNode===null)throw Error(k(162));l=e.stateNode,o=e.memoizedProps;try{l.nodeValue=o}catch(E){Z(e,e.return,E)}}break;case 3:if(Be(t,e),Ge(e),r&4&&n!==null&&n.memoizedState.isDehydrated)try{Un(t.containerInfo)}catch(E){Z(e,e.return,E)}break;case 4:Be(t,e),Ge(e);break;case 13:Be(t,e),Ge(e),l=e.child,l.flags&8192&&(o=l.memoizedState!==null,l.stateNode.isHidden=o,!o||l.alternate!==null&&l.alternate.memoizedState!==null||(zi=J())),r&4&&Xa(e);break;case 22:if(y=n!==null&&n.memoizedState!==null,e.mode&1?(fe=(c=fe)||y,Be(t,e),fe=c):Be(t,e),Ge(e),r&8192){if(c=e.memoizedState!==null,(e.stateNode.isHidden=c)&&!y&&e.mode&1)for(j=e,y=e.child;y!==null;){for(h=j=y;j!==null;){switch(p=j,w=p.child,p.tag){case 0:case 11:case 14:case 15:ur(4,p,p.return);break;case 1:kn(p,p.return);var S=p.stateNode;if(typeof S.componentWillUnmount=="function"){r=p,n=p.return;try{t=r,S.props=t.memoizedProps,S.state=t.memoizedState,S.componentWillUnmount()}catch(E){Z(r,n,E)}}break;case 5:kn(p,p.return);break;case 22:if(p.memoizedState!==null){qa(h);continue}}w!==null?(w.return=p,j=w):qa(h)}y=y.sibling}e:for(y=null,h=e;;){if(h.tag===5){if(y===null){y=h;try{l=h.stateNode,c?(o=l.style,typeof o.setProperty=="function"?o.setProperty("display","none","important"):o.display="none"):(u=h.stateNode,s=h.memoizedProps.style,i=s!=null&&s.hasOwnProperty("display")?s.display:null,u.style.display=zu("display",i))}catch(E){Z(e,e.return,E)}}}else if(h.tag===6){if(y===null)try{h.stateNode.nodeValue=c?"":h.memoizedProps}catch(E){Z(e,e.return,E)}}else if((h.tag!==22&&h.tag!==23||h.memoizedState===null||h===e)&&h.child!==null){h.child.return=h,h=h.child;continue}if(h===e)break e;for(;h.sibling===null;){if(h.return===null||h.return===e)break e;y===h&&(y=null),h=h.return}y===h&&(y=null),h.sibling.return=h.return,h=h.sibling}}break;case 19:Be(t,e),Ge(e),r&4&&Xa(e);break;case 21:break;default:Be(t,e),Ge(e)}}function Ge(e){var t=e.flags;if(t&2){try{e:{for(var n=e.return;n!==null;){if(Qa(n)){var r=n;break e}n=n.return}throw Error(k(160))}switch(r.tag){case 5:var l=r.stateNode;r.flags&32&&(Pn(l,""),r.flags&=-33);var o=Ka(e);ji(e,o,l);break;case 3:case 4:var i=r.stateNode.containerInfo,u=Ka(e);Ni(e,u,i);break;default:throw Error(k(161))}}catch(s){Z(e,e.return,s)}e.flags&=-3}t&4096&&(e.flags&=-4097)}function Hd(e,t,n){j=e,Za(e)}function Za(e,t,n){for(var r=(e.mode&1)!==0;j!==null;){var l=j,o=l.child;if(l.tag===22&&r){var i=l.memoizedState!==null||fl;if(!i){var u=l.alternate,s=u!==null&&u.memoizedState!==null||fe;u=fl;var c=fe;if(fl=i,(fe=s)&&!c)for(j=l;j!==null;)i=j,s=i.child,i.tag===22&&i.memoizedState!==null?ba(l):s!==null?(s.return=i,j=s):ba(l);for(;o!==null;)j=o,Za(o),o=o.sibling;j=l,fl=u,fe=c}Ja(e)}else l.subtreeFlags&8772&&o!==null?(o.return=l,j=o):Ja(e)}}function Ja(e){for(;j!==null;){var t=j;if(t.flags&8772){var n=t.alternate;try{if(t.flags&8772)switch(t.tag){case 0:case 11:case 15:fe||dl(5,t);break;case 1:var r=t.stateNode;if(t.flags&4&&!fe)if(n===null)r.componentDidMount();else{var l=t.elementType===t.type?n.memoizedProps:$e(t.type,n.memoizedProps);r.componentDidUpdate(l,n.memoizedState,r.__reactInternalSnapshotBeforeUpdate)}var o=t.updateQueue;o!==null&&qs(t,o,r);break;case 3:var i=t.updateQueue;if(i!==null){if(n=null,t.child!==null)switch(t.child.tag){case 5:n=t.child.stateNode;break;case 1:n=t.child.stateNode}qs(t,i,n)}break;case 5:var u=t.stateNode;if(n===null&&t.flags&4){n=u;var s=t.memoizedProps;switch(t.type){case"button":case"input":case"select":case"textarea":s.autoFocus&&n.focus();break;case"img":s.src&&(n.src=s.src)}}break;case 6:break;case 4:break;case 12:break;case 13:if(t.memoizedState===null){var c=t.alternate;if(c!==null){var y=c.memoizedState;if(y!==null){var h=y.dehydrated;h!==null&&Un(h)}}}break;case 19:case 17:case 21:case 22:case 23:case 25:break;default:throw Error(k(163))}fe||t.flags&512&&Ci(t)}catch(p){Z(t,t.return,p)}}if(t===e){j=null;break}if(n=t.sibling,n!==null){n.return=t.return,j=n;break}j=t.return}}function qa(e){for(;j!==null;){var t=j;if(t===e){j=null;break}var n=t.sibling;if(n!==null){n.return=t.return,j=n;break}j=t.return}}function ba(e){for(;j!==null;){var t=j;try{switch(t.tag){case 0:case 11:case 15:var n=t.return;try{dl(4,t)}catch(s){Z(t,n,s)}break;case 1:var r=t.stateNode;if(typeof r.componentDidMount=="function"){var l=t.return;try{r.componentDidMount()}catch(s){Z(t,l,s)}}var o=t.return;try{Ci(t)}catch(s){Z(t,o,s)}break;case 5:var i=t.return;try{Ci(t)}catch(s){Z(t,i,s)}}}catch(s){Z(t,t.return,s)}if(t===e){j=null;break}var u=t.sibling;if(u!==null){u.return=t.return,j=u;break}j=t.return}}var Qd=Math.ceil,pl=Je.ReactCurrentDispatcher,Pi=Je.ReactCurrentOwner,Oe=Je.ReactCurrentBatchConfig,I=0,le=null,b=null,ue=0,je=0,wn=kt(0),ne=0,sr=null,Vt=0,ml=0,Ti=0,ar=null,we=null,zi=0,Sn=1/0,ot=null,hl=!1,Li=null,Ct=null,vl=!1,Nt=null,yl=0,cr=0,Ri=null,gl=-1,kl=0;function ve(){return I&6?J():gl!==-1?gl:gl=J()}function jt(e){return e.mode&1?I&2&&ue!==0?ue&-ue:Pd.transition!==null?(kl===0&&(kl=Ku()),kl):(e=A,e!==0||(e=window.event,e=e===void 0?16:ts(e.type)),e):1}function We(e,t,n,r){if(50<cr)throw cr=0,Ri=null,Error(k(185));On(e,n,r),(!(I&2)||e!==le)&&(e===le&&(!(I&2)&&(ml|=n),ne===4&&Pt(e,ue)),Se(e,r),n===1&&I===0&&!(t.mode&1)&&(Sn=J()+500,Yr&&St()))}function Se(e,t){var n=e.callbackNode;Pf(e,t);var r=Pr(e,e===le?ue:0);if(r===0)n!==null&&Wu(n),e.callbackNode=null,e.callbackPriority=0;else if(t=r&-r,e.callbackPriority!==t){if(n!=null&&Wu(n),t===1)e.tag===0?jd(tc.bind(null,e)):$s(tc.bind(null,e)),_d(function(){!(I&6)&&St()}),n=null;else{switch(Yu(r)){case 1:n=co;break;case 4:n=Hu;break;case 16:n=Er;break;case 536870912:n=Qu;break;default:n=Er}n=ac(n,ec.bind(null,e))}e.callbackPriority=t,e.callbackNode=n}}function ec(e,t){if(gl=-1,kl=0,I&6)throw Error(k(327));var n=e.callbackNode;if(xn()&&e.callbackNode!==n)return null;var r=Pr(e,e===le?ue:0);if(r===0)return null;if(r&30||r&e.expiredLanes||t)t=wl(e,r);else{t=r;var l=I;I|=2;var o=rc();(le!==e||ue!==t)&&(ot=null,Sn=J()+500,Wt(e,t));do try{Xd();break}catch(u){nc(e,u)}while(!0);Zo(),pl.current=o,I=l,b!==null?t=0:(le=null,ue=0,t=ne)}if(t!==0){if(t===2&&(l=fo(e),l!==0&&(r=l,t=Mi(e,l))),t===1)throw n=sr,Wt(e,0),Pt(e,r),Se(e,J()),n;if(t===6)Pt(e,r);else{if(l=e.current.alternate,!(r&30)&&!Kd(l)&&(t=wl(e,r),t===2&&(o=fo(e),o!==0&&(r=o,t=Mi(e,o))),t===1))throw n=sr,Wt(e,0),Pt(e,r),Se(e,J()),n;switch(e.finishedWork=l,e.finishedLanes=r,t){case 0:case 1:throw Error(k(345));case 2:Ht(e,we,ot);break;case 3:if(Pt(e,r),(r&130023424)===r&&(t=zi+500-J(),10<t)){if(Pr(e,0)!==0)break;if(l=e.suspendedLanes,(l&r)!==r){ve(),e.pingedLanes|=e.suspendedLanes&l;break}e.timeoutHandle=Uo(Ht.bind(null,e,we,ot),t);break}Ht(e,we,ot);break;case 4:if(Pt(e,r),(r&4194240)===r)break;for(t=e.eventTimes,l=-1;0<r;){var i=31-Ie(r);o=1<<i,i=t[i],i>l&&(l=i),r&=~o}if(r=l,r=J()-r,r=(120>r?120:480>r?480:1080>r?1080:1920>r?1920:3e3>r?3e3:4320>r?4320:1960*Qd(r/1960))-r,10<r){e.timeoutHandle=Uo(Ht.bind(null,e,we,ot),r);break}Ht(e,we,ot);break;case 5:Ht(e,we,ot);break;default:throw Error(k(329))}}}return Se(e,J()),e.callbackNode===n?ec.bind(null,e):null}function Mi(e,t){var n=ar;return e.current.memoizedState.isDehydrated&&(Wt(e,t).flags|=256),e=wl(e,t),e!==2&&(t=we,we=n,t!==null&&Oi(t)),e}function Oi(e){we===null?we=e:we.push.apply(we,e)}function Kd(e){for(var t=e;;){if(t.flags&16384){var n=t.updateQueue;if(n!==null&&(n=n.stores,n!==null))for(var r=0;r<n.length;r++){var l=n[r],o=l.getSnapshot;l=l.value;try{if(!Ae(o(),l))return!1}catch{return!1}}}if(n=t.child,t.subtreeFlags&16384&&n!==null)n.return=t,t=n;else{if(t===e)break;for(;t.sibling===null;){if(t.return===null||t.return===e)return!0;t=t.return}t.sibling.return=t.return,t=t.sibling}}return!0}function Pt(e,t){for(t&=~Ti,t&=~ml,e.suspendedLanes|=t,e.pingedLanes&=~t,e=e.expirationTimes;0<t;){var n=31-Ie(t),r=1<<n;e[n]=-1,t&=~r}}function tc(e){if(I&6)throw Error(k(327));xn();var t=Pr(e,0);if(!(t&1))return Se(e,J()),null;var n=wl(e,t);if(e.tag!==0&&n===2){var r=fo(e);r!==0&&(t=r,n=Mi(e,r))}if(n===1)throw n=sr,Wt(e,0),Pt(e,t),Se(e,J()),n;if(n===6)throw Error(k(345));return e.finishedWork=e.current.alternate,e.finishedLanes=t,Ht(e,we,ot),Se(e,J()),null}function Fi(e,t){var n=I;I|=1;try{return e(t)}finally{I=n,I===0&&(Sn=J()+500,Yr&&St())}}function Bt(e){Nt!==null&&Nt.tag===0&&!(I&6)&&xn();var t=I;I|=1;var n=Oe.transition,r=A;try{if(Oe.transition=null,A=1,e)return e()}finally{A=r,Oe.transition=n,I=t,!(I&6)&&St()}}function Di(){je=wn.current,B(wn)}function Wt(e,t){e.finishedWork=null,e.finishedLanes=0;var n=e.timeoutHandle;if(n!==-1&&(e.timeoutHandle=-1,xd(n)),b!==null)for(n=b.return;n!==null;){var r=n;switch(Qo(r),r.tag){case 1:r=r.type.childContextTypes,r!=null&&Qr();break;case 3:yn(),B(ye),B(se),li();break;case 5:ni(r);break;case 4:yn();break;case 13:B(Y);break;case 19:B(Y);break;case 10:Jo(r.type._context);break;case 22:case 23:Di()}n=n.return}if(le=e,b=e=Tt(e.current,null),ue=je=t,ne=0,sr=null,Ti=ml=Vt=0,we=ar=null,At!==null){for(t=0;t<At.length;t++)if(n=At[t],r=n.interleaved,r!==null){n.interleaved=null;var l=r.next,o=n.pending;if(o!==null){var i=o.next;o.next=l,r.next=i}n.pending=r}At=null}return e}function nc(e,t){do{var n=b;try{if(Zo(),rl.current=ul,ll){for(var r=X.memoizedState;r!==null;){var l=r.queue;l!==null&&(l.pending=null),r=r.next}ll=!1}if($t=0,re=te=X=null,nr=!1,rr=0,Pi.current=null,n===null||n.return===null){ne=1,sr=t,b=null;break}e:{var o=e,i=n.return,u=n,s=t;if(t=ue,u.flags|=32768,s!==null&&typeof s=="object"&&typeof s.then=="function"){var c=s,y=u,h=y.tag;if(!(y.mode&1)&&(h===0||h===11||h===15)){var p=y.alternate;p?(y.updateQueue=p.updateQueue,y.memoizedState=p.memoizedState,y.lanes=p.lanes):(y.updateQueue=null,y.memoizedState=null)}var w=ja(i);if(w!==null){w.flags&=-257,Pa(w,i,u,o,t),w.mode&1&&Na(o,c,t),t=w,s=c;var S=t.updateQueue;if(S===null){var E=new Set;E.add(s),t.updateQueue=E}else S.add(s);break e}else{if(!(t&1)){Na(o,c,t),Ii();break e}s=Error(k(426))}}else if(H&&u.mode&1){var U=ja(i);if(U!==null){!(U.flags&65536)&&(U.flags|=256),Pa(U,i,u,o,t),Xo(gn(s,u));break e}}o=s=gn(s,u),ne!==4&&(ne=2),ar===null?ar=[o]:ar.push(o),o=i;do{switch(o.tag){case 3:o.flags|=65536,t&=-t,o.lanes|=t;var f=Ea(o,s,t);Js(o,f);break e;case 1:u=s;var a=o.type,d=o.stateNode;if(!(o.flags&128)&&(typeof a.getDerivedStateFromError=="function"||d!==null&&typeof d.componentDidCatch=="function"&&(Ct===null||!Ct.has(d)))){o.flags|=65536,t&=-t,o.lanes|=t;var g=Ca(o,u,t);Js(o,g);break e}}o=o.return}while(o!==null)}oc(n)}catch(N){t=N,b===n&&n!==null&&(b=n=n.return);continue}break}while(!0)}function rc(){var e=pl.current;return pl.current=ul,e===null?ul:e}function Ii(){(ne===0||ne===3||ne===2)&&(ne=4),le===null||!(Vt&268435455)&&!(ml&268435455)||Pt(le,ue)}function wl(e,t){var n=I;I|=2;var r=rc();(le!==e||ue!==t)&&(ot=null,Wt(e,t));do try{Yd();break}catch(l){nc(e,l)}while(!0);if(Zo(),I=n,pl.current=r,b!==null)throw Error(k(261));return le=null,ue=0,ne}function Yd(){for(;b!==null;)lc(b)}function Xd(){for(;b!==null&&!kf();)lc(b)}function lc(e){var t=sc(e.alternate,e,je);e.memoizedProps=e.pendingProps,t===null?oc(e):b=t,Pi.current=null}function oc(e){var t=e;do{var n=t.alternate;if(e=t.return,t.flags&32768){if(n=Vd(n,t),n!==null){n.flags&=32767,b=n;return}if(e!==null)e.flags|=32768,e.subtreeFlags=0,e.deletions=null;else{ne=6,b=null;return}}else if(n=$d(n,t,je),n!==null){b=n;return}if(t=t.sibling,t!==null){b=t;return}b=t=e}while(t!==null);ne===0&&(ne=5)}function Ht(e,t,n){var r=A,l=Oe.transition;try{Oe.transition=null,A=1,Gd(e,t,n,r)}finally{Oe.transition=l,A=r}return null}function Gd(e,t,n,r){do xn();while(Nt!==null);if(I&6)throw Error(k(327));n=e.finishedWork;var l=e.finishedLanes;if(n===null)return null;if(e.finishedWork=null,e.finishedLanes=0,n===e.current)throw Error(k(177));e.callbackNode=null,e.callbackPriority=0;var o=n.lanes|n.childLanes;if(Tf(e,o),e===le&&(b=le=null,ue=0),!(n.subtreeFlags&2064)&&!(n.flags&2064)||vl||(vl=!0,ac(Er,function(){return xn(),null})),o=(n.flags&15990)!==0,n.subtreeFlags&15990||o){o=Oe.transition,Oe.transition=null;var i=A;A=1;var u=I;I|=4,Pi.current=null,Wd(e,n),Ga(n,e),hd(Io),Lr=!!Do,Io=Do=null,e.current=n,Hd(n),wf(),I=u,A=i,Oe.transition=o}else e.current=n;if(vl&&(vl=!1,Nt=e,yl=l),o=e.pendingLanes,o===0&&(Ct=null),_f(n.stateNode),Se(e,J()),t!==null)for(r=e.onRecoverableError,n=0;n<t.length;n++)l=t[n],r(l.value,{componentStack:l.stack,digest:l.digest});if(hl)throw hl=!1,e=Li,Li=null,e;return yl&1&&e.tag!==0&&xn(),o=e.pendingLanes,o&1?e===Ri?cr++:(cr=0,Ri=e):cr=0,St(),null}function xn(){if(Nt!==null){var e=Yu(yl),t=Oe.transition,n=A;try{if(Oe.transition=null,A=16>e?16:e,Nt===null)var r=!1;else{if(e=Nt,Nt=null,yl=0,I&6)throw Error(k(331));var l=I;for(I|=4,j=e.current;j!==null;){var o=j,i=o.child;if(j.flags&16){var u=o.deletions;if(u!==null){for(var s=0;s<u.length;s++){var c=u[s];for(j=c;j!==null;){var y=j;switch(y.tag){case 0:case 11:case 15:ur(8,y,o)}var h=y.child;if(h!==null)h.return=y,j=h;else for(;j!==null;){y=j;var p=y.sibling,w=y.return;if(Ha(y),y===c){j=null;break}if(p!==null){p.return=w,j=p;break}j=w}}}var S=o.alternate;if(S!==null){var E=S.child;if(E!==null){S.child=null;do{var U=E.sibling;E.sibling=null,E=U}while(E!==null)}}j=o}}if(o.subtreeFlags&2064&&i!==null)i.return=o,j=i;else e:for(;j!==null;){if(o=j,o.flags&2048)switch(o.tag){case 0:case 11:case 15:ur(9,o,o.return)}var f=o.sibling;if(f!==null){f.return=o.return,j=f;break e}j=o.return}}var a=e.current;for(j=a;j!==null;){i=j;var d=i.child;if(i.subtreeFlags&2064&&d!==null)d.return=i,j=d;else e:for(i=a;j!==null;){if(u=j,u.flags&2048)try{switch(u.tag){case 0:case 11:case 15:dl(9,u)}}catch(N){Z(u,u.return,N)}if(u===i){j=null;break e}var g=u.sibling;if(g!==null){g.return=u.return,j=g;break e}j=u.return}}if(I=l,St(),Qe&&typeof Qe.onPostCommitFiberRoot=="function")try{Qe.onPostCommitFiberRoot(Cr,e)}catch{}r=!0}return r}finally{A=n,Oe.transition=t}}return!1}function ic(e,t,n){t=gn(n,t),t=Ea(e,t,1),e=_t(e,t,1),t=ve(),e!==null&&(On(e,1,t),Se(e,t))}function Z(e,t,n){if(e.tag===3)ic(e,e,n);else for(;t!==null;){if(t.tag===3){ic(t,e,n);break}else if(t.tag===1){var r=t.stateNode;if(typeof t.type.getDerivedStateFromError=="function"||typeof r.componentDidCatch=="function"&&(Ct===null||!Ct.has(r))){e=gn(n,e),e=Ca(t,e,1),t=_t(t,e,1),e=ve(),t!==null&&(On(t,1,e),Se(t,e));break}}t=t.return}}function Zd(e,t,n){var r=e.pingCache;r!==null&&r.delete(t),t=ve(),e.pingedLanes|=e.suspendedLanes&n,le===e&&(ue&n)===n&&(ne===4||ne===3&&(ue&130023424)===ue&&500>J()-zi?Wt(e,0):Ti|=n),Se(e,t)}function uc(e,t){t===0&&(e.mode&1?(t=jr,jr<<=1,!(jr&130023424)&&(jr=4194304)):t=1);var n=ve();e=nt(e,t),e!==null&&(On(e,t,n),Se(e,n))}function Jd(e){var t=e.memoizedState,n=0;t!==null&&(n=t.retryLane),uc(e,n)}function qd(e,t){var n=0;switch(e.tag){case 13:var r=e.stateNode,l=e.memoizedState;l!==null&&(n=l.retryLane);break;case 19:r=e.stateNode;break;default:throw Error(k(314))}r!==null&&r.delete(t),uc(e,n)}var sc;sc=function(e,t,n){if(e!==null)if(e.memoizedProps!==t.pendingProps||ye.current)ke=!0;else{if(!(e.lanes&n)&&!(t.flags&128))return ke=!1,Ud(e,t,n);ke=!!(e.flags&131072)}else ke=!1,H&&t.flags&1048576&&Vs(t,Gr,t.index);switch(t.lanes=0,t.tag){case 2:var r=t.type;cl(e,t),e=t.pendingProps;var l=cn(t,se.current);vn(t,n),l=ui(null,t,r,e,l,n);var o=si();return t.flags|=1,typeof l=="object"&&l!==null&&typeof l.render=="function"&&l.$$typeof===void 0?(t.tag=1,t.memoizedState=null,t.updateQueue=null,ge(r)?(o=!0,Kr(t)):o=!1,t.memoizedState=l.state!==null&&l.state!==void 0?l.state:null,ei(t),l.updater=sl,t.stateNode=l,l._reactInternals=t,mi(t,r,e,n),t=gi(null,t,r,!0,o,n)):(t.tag=0,H&&o&&Ho(t),he(null,t,l,n),t=t.child),t;case 16:r=t.elementType;e:{switch(cl(e,t),e=t.pendingProps,l=r._init,r=l(r._payload),t.type=r,l=t.tag=ep(r),e=$e(r,e),l){case 0:t=yi(null,t,r,e,n);break e;case 1:t=Oa(null,t,r,e,n);break e;case 11:t=Ta(null,t,r,e,n);break e;case 14:t=za(null,t,r,$e(r.type,e),n);break e}throw Error(k(306,r,""))}return t;case 0:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:$e(r,l),yi(e,t,r,l,n);case 1:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:$e(r,l),Oa(e,t,r,l,n);case 3:e:{if(Fa(t),e===null)throw Error(k(387));r=t.pendingProps,o=t.memoizedState,l=o.element,Zs(e,t),tl(t,r,null,n);var i=t.memoizedState;if(r=i.element,o.isDehydrated)if(o={element:r,isDehydrated:!1,cache:i.cache,pendingSuspenseBoundaries:i.pendingSuspenseBoundaries,transitions:i.transitions},t.updateQueue.baseState=o,t.memoizedState=o,t.flags&256){l=gn(Error(k(423)),t),t=Da(e,t,r,n,l);break e}else if(r!==l){l=gn(Error(k(424)),t),t=Da(e,t,r,n,l);break e}else for(Ne=gt(t.stateNode.containerInfo.firstChild),Ce=t,H=!0,Ue=null,n=Xs(t,null,r,n),t.child=n;n;)n.flags=n.flags&-3|4096,n=n.sibling;else{if(pn(),r===l){t=lt(e,t,n);break e}he(e,t,r,n)}t=t.child}return t;case 5:return bs(t),e===null&&Yo(t),r=t.type,l=t.pendingProps,o=e!==null?e.memoizedProps:null,i=l.children,Ao(r,l)?i=null:o!==null&&Ao(r,o)&&(t.flags|=32),Ma(e,t),he(e,t,i,n),t.child;case 6:return e===null&&Yo(t),null;case 13:return Ia(e,t,n);case 4:return ti(t,t.stateNode.containerInfo),r=t.pendingProps,e===null?t.child=mn(t,null,r,n):he(e,t,r,n),t.child;case 11:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:$e(r,l),Ta(e,t,r,l,n);case 7:return he(e,t,t.pendingProps,n),t.child;case 8:return he(e,t,t.pendingProps.children,n),t.child;case 12:return he(e,t,t.pendingProps.children,n),t.child;case 10:e:{if(r=t.type._context,l=t.pendingProps,o=t.memoizedProps,i=l.value,$(qr,r._currentValue),r._currentValue=i,o!==null)if(Ae(o.value,i)){if(o.children===l.children&&!ye.current){t=lt(e,t,n);break e}}else for(o=t.child,o!==null&&(o.return=t);o!==null;){var u=o.dependencies;if(u!==null){i=o.child;for(var s=u.firstContext;s!==null;){if(s.context===r){if(o.tag===1){s=rt(-1,n&-n),s.tag=2;var c=o.updateQueue;if(c!==null){c=c.shared;var y=c.pending;y===null?s.next=s:(s.next=y.next,y.next=s),c.pending=s}}o.lanes|=n,s=o.alternate,s!==null&&(s.lanes|=n),qo(o.return,n,t),u.lanes|=n;break}s=s.next}}else if(o.tag===10)i=o.type===t.type?null:o.child;else if(o.tag===18){if(i=o.return,i===null)throw Error(k(341));i.lanes|=n,u=i.alternate,u!==null&&(u.lanes|=n),qo(i,n,t),i=o.sibling}else i=o.child;if(i!==null)i.return=o;else for(i=o;i!==null;){if(i===t){i=null;break}if(o=i.sibling,o!==null){o.return=i.return,i=o;break}i=i.return}o=i}he(e,t,l.children,n),t=t.child}return t;case 9:return l=t.type,r=t.pendingProps.children,vn(t,n),l=Re(l),r=r(l),t.flags|=1,he(e,t,r,n),t.child;case 14:return r=t.type,l=$e(r,t.pendingProps),l=$e(r.type,l),za(e,t,r,l,n);case 15:return La(e,t,t.type,t.pendingProps,n);case 17:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:$e(r,l),cl(e,t),t.tag=1,ge(r)?(e=!0,Kr(t)):e=!1,vn(t,n),xa(t,r,l),mi(t,r,l,n),gi(null,t,r,!0,e,n);case 19:return Ua(e,t,n);case 22:return Ra(e,t,n)}throw Error(k(156,t.tag))};function ac(e,t){return Bu(e,t)}function bd(e,t,n,r){this.tag=e,this.key=n,this.sibling=this.child=this.return=this.stateNode=this.type=this.elementType=null,this.index=0,this.ref=null,this.pendingProps=t,this.dependencies=this.memoizedState=this.updateQueue=this.memoizedProps=null,this.mode=r,this.subtreeFlags=this.flags=0,this.deletions=null,this.childLanes=this.lanes=0,this.alternate=null}function Fe(e,t,n,r){return new bd(e,t,n,r)}function Ai(e){return e=e.prototype,!(!e||!e.isReactComponent)}function ep(e){if(typeof e=="function")return Ai(e)?1:0;if(e!=null){if(e=e.$$typeof,e===Wl)return 11;if(e===Kl)return 14}return 2}function Tt(e,t){var n=e.alternate;return n===null?(n=Fe(e.tag,t,e.key,e.mode),n.elementType=e.elementType,n.type=e.type,n.stateNode=e.stateNode,n.alternate=e,e.alternate=n):(n.pendingProps=t,n.type=e.type,n.flags=0,n.subtreeFlags=0,n.deletions=null),n.flags=e.flags&14680064,n.childLanes=e.childLanes,n.lanes=e.lanes,n.child=e.child,n.memoizedProps=e.memoizedProps,n.memoizedState=e.memoizedState,n.updateQueue=e.updateQueue,t=e.dependencies,n.dependencies=t===null?null:{lanes:t.lanes,firstContext:t.firstContext},n.sibling=e.sibling,n.index=e.index,n.ref=e.ref,n}function Sl(e,t,n,r,l,o){var i=2;if(r=e,typeof e=="function")Ai(e)&&(i=1);else if(typeof e=="string")i=5;else e:switch(e){case Zt:return Qt(n.children,l,o,t);case Vl:i=8,l|=8;break;case Bl:return e=Fe(12,n,t,l|2),e.elementType=Bl,e.lanes=o,e;case Hl:return e=Fe(13,n,t,l),e.elementType=Hl,e.lanes=o,e;case Ql:return e=Fe(19,n,t,l),e.elementType=Ql,e.lanes=o,e;case gu:return xl(n,l,o,t);default:if(typeof e=="object"&&e!==null)switch(e.$$typeof){case vu:i=10;break e;case yu:i=9;break e;case Wl:i=11;break e;case Kl:i=14;break e;case ct:i=16,r=null;break e}throw Error(k(130,e==null?e:typeof e,""))}return t=Fe(i,n,t,l),t.elementType=e,t.type=r,t.lanes=o,t}function Qt(e,t,n,r){return e=Fe(7,e,r,t),e.lanes=n,e}function xl(e,t,n,r){return e=Fe(22,e,r,t),e.elementType=gu,e.lanes=n,e.stateNode={isHidden:!1},e}function Ui(e,t,n){return e=Fe(6,e,null,t),e.lanes=n,e}function $i(e,t,n){return t=Fe(4,e.children!==null?e.children:[],e.key,t),t.lanes=n,t.stateNode={containerInfo:e.containerInfo,pendingChildren:null,implementation:e.implementation},t}function tp(e,t,n,r,l){this.tag=t,this.containerInfo=e,this.finishedWork=this.pingCache=this.current=this.pendingChildren=null,this.timeoutHandle=-1,this.callbackNode=this.pendingContext=this.context=null,this.callbackPriority=0,this.eventTimes=po(0),this.expirationTimes=po(-1),this.entangledLanes=this.finishedLanes=this.mutableReadLanes=this.expiredLanes=this.pingedLanes=this.suspendedLanes=this.pendingLanes=0,this.entanglements=po(0),this.identifierPrefix=r,this.onRecoverableError=l,this.mutableSourceEagerHydrationData=null}function Vi(e,t,n,r,l,o,i,u,s){return e=new tp(e,t,n,u,s),t===1?(t=1,o===!0&&(t|=8)):t=0,o=Fe(3,null,null,t),e.current=o,o.stateNode=e,o.memoizedState={element:r,isDehydrated:n,cache:null,transitions:null,pendingSuspenseBoundaries:null},ei(o),e}function np(e,t,n){var r=3<arguments.length&&arguments[3]!==void 0?arguments[3]:null;return{$$typeof:Gt,key:r==null?null:""+r,children:e,containerInfo:t,implementation:n}}function cc(e){if(!e)return wt;e=e._reactInternals;e:{if(Mt(e)!==e||e.tag!==1)throw Error(k(170));var t=e;do{switch(t.tag){case 3:t=t.stateNode.context;break e;case 1:if(ge(t.type)){t=t.stateNode.__reactInternalMemoizedMergedChildContext;break e}}t=t.return}while(t!==null);throw Error(k(171))}if(e.tag===1){var n=e.type;if(ge(n))return As(e,n,t)}return t}function fc(e,t,n,r,l,o,i,u,s){return e=Vi(n,r,!0,e,l,o,i,u,s),e.context=cc(null),n=e.current,r=ve(),l=jt(n),o=rt(r,l),o.callback=t??null,_t(n,o,l),e.current.lanes=l,On(e,l,r),Se(e,r),e}function _l(e,t,n,r){var l=t.current,o=ve(),i=jt(l);return n=cc(n),t.context===null?t.context=n:t.pendingContext=n,t=rt(o,i),t.payload={element:e},r=r===void 0?null:r,r!==null&&(t.callback=r),e=_t(l,t,i),e!==null&&(We(e,l,i,o),el(e,l,i)),i}function El(e){if(e=e.current,!e.child)return null;switch(e.child.tag){case 5:return e.child.stateNode;default:return e.child.stateNode}}function dc(e,t){if(e=e.memoizedState,e!==null&&e.dehydrated!==null){var n=e.retryLane;e.retryLane=n!==0&&n<t?n:t}}function Bi(e,t){dc(e,t),(e=e.alternate)&&dc(e,t)}function rp(){return null}var pc=typeof reportError=="function"?reportError:function(e){console.error(e)};function Wi(e){this._internalRoot=e}Cl.prototype.render=Wi.prototype.render=function(e){var t=this._internalRoot;if(t===null)throw Error(k(409));_l(e,t,null,null)},Cl.prototype.unmount=Wi.prototype.unmount=function(){var e=this._internalRoot;if(e!==null){this._internalRoot=null;var t=e.containerInfo;Bt(function(){_l(null,e,null,null)}),t[qe]=null}};function Cl(e){this._internalRoot=e}Cl.prototype.unstable_scheduleHydration=function(e){if(e){var t=Zu();e={blockedOn:null,target:e,priority:t};for(var n=0;n<ht.length&&t!==0&&t<ht[n].priority;n++);ht.splice(n,0,e),n===0&&bu(e)}};function Hi(e){return!(!e||e.nodeType!==1&&e.nodeType!==9&&e.nodeType!==11)}function Nl(e){return!(!e||e.nodeType!==1&&e.nodeType!==9&&e.nodeType!==11&&(e.nodeType!==8||e.nodeValue!==" react-mount-point-unstable "))}function mc(){}function lp(e,t,n,r,l){if(l){if(typeof r=="function"){var o=r;r=function(){var c=El(i);o.call(c)}}var i=fc(t,r,e,0,null,!1,!1,"",mc);return e._reactRootContainer=i,e[qe]=i.current,Xn(e.nodeType===8?e.parentNode:e),Bt(),i}for(;l=e.lastChild;)e.removeChild(l);if(typeof r=="function"){var u=r;r=function(){var c=El(s);u.call(c)}}var s=Vi(e,0,!1,null,null,!1,!1,"",mc);return e._reactRootContainer=s,e[qe]=s.current,Xn(e.nodeType===8?e.parentNode:e),Bt(function(){_l(t,s,n,r)}),s}function jl(e,t,n,r,l){var o=n._reactRootContainer;if(o){var i=o;if(typeof l=="function"){var u=l;l=function(){var s=El(i);u.call(s)}}_l(t,i,e,l)}else i=lp(n,t,e,l,r);return El(i)}Xu=function(e){switch(e.tag){case 3:var t=e.stateNode;if(t.current.memoizedState.isDehydrated){var n=Mn(t.pendingLanes);n!==0&&(mo(t,n|1),Se(t,J()),!(I&6)&&(Sn=J()+500,St()))}break;case 13:Bt(function(){var r=nt(e,1);if(r!==null){var l=ve();We(r,e,1,l)}}),Bi(e,1)}},ho=function(e){if(e.tag===13){var t=nt(e,134217728);if(t!==null){var n=ve();We(t,e,134217728,n)}Bi(e,134217728)}},Gu=function(e){if(e.tag===13){var t=jt(e),n=nt(e,t);if(n!==null){var r=ve();We(n,e,t,r)}Bi(e,t)}},Zu=function(){return A},Ju=function(e,t){var n=A;try{return A=e,t()}finally{A=n}},io=function(e,t,n){switch(t){case"input":if(ql(e,n),t=n.name,n.type==="radio"&&t!=null){for(n=e;n.parentNode;)n=n.parentNode;for(n=n.querySelectorAll("input[name="+JSON.stringify(""+t)+'][type="radio"]'),t=0;t<n.length;t++){var r=n[t];if(r!==e&&r.form===e.form){var l=Hr(r);if(!l)throw Error(k(90));Su(r),ql(r,l)}}}break;case"textarea":Nu(e,n);break;case"select":t=n.value,t!=null&&Jt(e,!!n.multiple,t,!1)}},Fu=Fi,Du=Bt;var op={usingClientEntryPoint:!1,Events:[Jn,sn,Hr,Mu,Ou,Fi]},fr={findFiberByHostInstance:Ot,bundleType:0,version:"18.3.1",rendererPackageName:"react-dom"},ip={bundleType:fr.bundleType,version:fr.version,rendererPackageName:fr.rendererPackageName,rendererConfig:fr.rendererConfig,overrideHookState:null,overrideHookStateDeletePath:null,overrideHookStateRenamePath:null,overrideProps:null,overridePropsDeletePath:null,overridePropsRenamePath:null,setErrorHandler:null,setSuspenseHandler:null,scheduleUpdate:null,currentDispatcherRef:Je.ReactCurrentDispatcher,findHostInstanceByFiber:function(e){return e=$u(e),e===null?null:e.stateNode},findFiberByHostInstance:fr.findFiberByHostInstance||rp,findHostInstancesForRefresh:null,scheduleRefresh:null,scheduleRoot:null,setRefreshHandler:null,getCurrentFiber:null,reconcilerVersion:"18.3.1-next-f1338f8080-20240426"};if(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__<"u"){var Pl=__REACT_DEVTOOLS_GLOBAL_HOOK__;if(!Pl.isDisabled&&Pl.supportsFiber)try{Cr=Pl.inject(ip),Qe=Pl}catch{}}xe.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=op,xe.createPortal=function(e,t){var n=2<arguments.length&&arguments[2]!==void 0?arguments[2]:null;if(!Hi(t))throw Error(k(200));return np(e,t,null,n)},xe.createRoot=function(e,t){if(!Hi(e))throw Error(k(299));var n=!1,r="",l=pc;return t!=null&&(t.unstable_strictMode===!0&&(n=!0),t.identifierPrefix!==void 0&&(r=t.identifierPrefix),t.onRecoverableError!==void 0&&(l=t.onRecoverableError)),t=Vi(e,1,!1,null,null,n,!1,r,l),e[qe]=t.current,Xn(e.nodeType===8?e.parentNode:e),new Wi(t)},xe.findDOMNode=function(e){if(e==null)return null;if(e.nodeType===1)return e;var t=e._reactInternals;if(t===void 0)throw typeof e.render=="function"?Error(k(188)):(e=Object.keys(e).join(","),Error(k(268,e)));return e=$u(t),e=e===null?null:e.stateNode,e},xe.flushSync=function(e){return Bt(e)},xe.hydrate=function(e,t,n){if(!Nl(t))throw Error(k(200));return jl(null,e,t,!0,n)},xe.hydrateRoot=function(e,t,n){if(!Hi(e))throw Error(k(405));var r=n!=null&&n.hydratedSources||null,l=!1,o="",i=pc;if(n!=null&&(n.unstable_strictMode===!0&&(l=!0),n.identifierPrefix!==void 0&&(o=n.identifierPrefix),n.onRecoverableError!==void 0&&(i=n.onRecoverableError)),t=fc(t,null,e,1,n??null,l,!1,o,i),e[qe]=t.current,Xn(e),r)for(e=0;e<r.length;e++)n=r[e],l=n._getVersion,l=l(n._source),t.mutableSourceEagerHydrationData==null?t.mutableSourceEagerHydrationData=[n,l]:t.mutableSourceEagerHydrationData.push(n,l);return new Cl(t)},xe.render=function(e,t,n){if(!Nl(t))throw Error(k(200));return jl(null,e,t,!1,n)},xe.unmountComponentAtNode=function(e){if(!Nl(e))throw Error(k(40));return e._reactRootContainer?(Bt(function(){jl(null,null,e,!1,function(){e._reactRootContainer=null,e[qe]=null})}),!0):!1},xe.unstable_batchedUpdates=Fi,xe.unstable_renderSubtreeIntoContainer=function(e,t,n,r){if(!Nl(n))throw Error(k(200));if(e==null||e._reactInternals===void 0)throw Error(k(38));return jl(e,t,n,!1,r)},xe.version="18.3.1-next-f1338f8080-20240426";function hc(){if(!(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__>"u"||typeof __REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE!="function"))try{__REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE(hc)}catch(e){console.error(e)}}hc(),cu.exports=xe;var Tl=cu.exports,vc,yc=Tl;vc=yc.createRoot,yc.hydrateRoot;function up({activeSubmission:e,ctx:t,workflowBusy:n,workflowActions:r,manualSaving:l,setAddActivityOpen:o,setAddProjectOpen:i,setProjectSearch:u,setActivitySearch:s,runWorkflowAction:c,persistTimesheet:y}){return m.jsxs("div",{style:{display:"flex",gap:8,alignItems:"center"},children:[m.jsx("button",{className:"btn btn-default btn-sm",onClick:()=>e&&frappe.set_route("Form","Timesheet Submission",e),disabled:!e,children:"Open Submission"}),frappe.user_roles.some(h=>["Supervisor","Finance"].includes(h))&&m.jsx("button",{className:"btn btn-default btn-sm",onClick:()=>e&&(t==null?void 0:t.employee)&&frappe.set_route("timesheet_workflow","employee",t.employee,"submission",e),disabled:!e||!(t!=null&&t.employee),children:"Review"}),m.jsxs("div",{className:"btn-group",children:[m.jsx("button",{className:"btn btn-default btn-sm dropdown-toggle","data-toggle":"dropdown","aria-expanded":"false",disabled:n,children:n?__("Working..."):__("Actions")}),m.jsxs("ul",{className:"dropdown-menu dropdown-menu-right ts-actions-menu",children:[m.jsx("li",{children:m.jsxs("a",{className:"ts-actions-link",href:"#",onClick:h=>{h.preventDefault(),o(!1),i(p=>!p),u("")},children:[m.jsx("i",{className:"fa fa-folder-open ts-actions-icon","aria-hidden":"true"}),m.jsx("span",{children:__("Add Project")})]})}),m.jsx("li",{children:m.jsxs("a",{className:"ts-actions-link",href:"#",onClick:h=>{h.preventDefault(),i(!1),o(p=>!p),s("")},children:[m.jsx("i",{className:"fa fa-plus-circle ts-actions-icon","aria-hidden":"true"}),m.jsx("span",{children:__("Add Activity")})]})}),r.length>0&&m.jsx("li",{className:"divider"}),r.map(h=>m.jsx("li",{children:m.jsxs("a",{className:"ts-actions-link",href:"#",onClick:p=>{p.preventDefault(),c(h)},children:[m.jsx("i",{className:"fa fa-random ts-actions-icon","aria-hidden":"true"}),m.jsx("span",{children:h})]})},h))]})]}),m.jsx("button",{className:"btn btn-primary btn-sm",onClick:()=>y(!0),disabled:l,children:l?"Saving...":"Save"})]})}function gc(e){if(!e||typeof e!="string")return"";const[t,n]=e.split("-"),r=parseInt(t,10),l=parseInt(n,10);return!r||r<1||r>12||!l?e:`${new Date(l,r-1,1).toLocaleString("en-US",{month:"long"})} ${l}`}function sp({visibleSubmissions:e,activeSubmission:t,loadContext:n}){return m.jsxs("div",{className:"frappe-card",style:{padding:10,height:"fit-content"},children:[m.jsx("div",{style:{fontWeight:600,marginBottom:8},children:"Available Months"}),e.map(r=>m.jsx("button",{className:`btn btn-xs ${r.name===t?"btn-primary":"btn-default"}`,style:{width:"100%",marginBottom:6,textAlign:"left"},onClick:()=>n(r.name),children:gc(r.month_year)},r.name))]})}let ap=1;function kc(){return"row_"+ap++}function Qi(){return{id:kc(),task:"",hours:{}}}function Ki(e,t){return{type:e,name:t,tasks:[Qi()]}}function cp(e){return(e||[]).map(t=>({type:t.type,name:t.name,tasks:(t.tasks||[]).length?t.tasks.map(n=>({id:kc(),task:n.task||"",hours:n.hours||{}})):[Qi()]}))}function fp({submissionName:e,onContextChange:t}){const[r,l]=R.useState(null),[o,i]=R.useState(e||""),[u,s]=R.useState([]),[c,y]=R.useState(!0),[h,p]=R.useState(!1),[w,S]=R.useState(!1),[E,U]=R.useState(!1),[f,a]=R.useState(!1),[d,g]=R.useState(""),[N,P]=R.useState(""),[T,L]=R.useState(""),[Q,F]=R.useState(""),[Pe,He]=R.useState([]),[it,zt]=R.useState(!1),dr=R.useRef(!0),ut=R.useRef(null),st=R.useRef(!1),x=R.useRef(!1),[M,O]=R.useState(()=>window.innerWidth<768),[W,ee]=R.useState(null);R.useEffect(()=>{const v=()=>O(window.innerWidth<768);return window.addEventListener("resize",v),()=>window.removeEventListener("resize",v)},[]);const q=R.useCallback(()=>{st.current=!0,x.current=!1},[]),Te=R.useCallback(v=>{if(!v){y(!1),F("No submission specified. Please open this page from a Timesheet Submission.");return}y(!0),F(""),frappe.call({method:"corporate_services.icl_corporate_services.page.employee_timesheet_entry.employee_timesheet_entry.get_timesheet_context",args:{submission_name:v},callback(C){if(C.message){const _=C.message;l(_),t&&t(_);const z=cp(_.existing_sections||[]);if(z.length)s(z);else{const G=(_.projects||[]).map(de=>Ki("project",de.project_name));s(G)}i(v)}y(!1),dr.current=!1},error(){F("Failed to load timesheet context."),y(!1),dr.current=!1}})},[t]);R.useEffect(()=>{Te(e)},[e,Te]);const Lt=R.useCallback(v=>{if(!v){He([]);return}frappe.call({method:"frappe.client.get",args:{doctype:"Timesheet Submission",name:v},callback(C){const _=C.message;if(!_){He([]);return}frappe.call({method:"frappe.model.workflow.get_transitions",args:{doc:_},callback(z){const G=(z.message||[]).map(de=>de.action).filter(Boolean);He(G)},error(){He([])}})},error(){He([])}})},[]);R.useEffect(()=>{Lt(o)},[o,Lt]);const De=R.useCallback((v=!1)=>{if(h||!o)return;const C=u.map(_=>({type:_.type,name:_.name,tasks:_.tasks.map(z=>({task:z.task,hours:z.hours}))}));p(!0),v&&S(!0),frappe.call({method:"corporate_services.icl_corporate_services.page.employee_timesheet_entry.employee_timesheet_entry.save_web_timesheet",args:{submission_name:o,sections:JSON.stringify(C)},callback(_){p(!1),S(!1),st.current=!1;const z=_.message||{};z.status==="success"?(x.current=!1,v&&frappe.show_alert({message:`Timesheet saved. Total hours: ${z.total_hours.toFixed(1)}`,indicator:"green"},5)):v?frappe.msgprint({title:z.status==="locked"?"Timesheet Locked":"Save Failed",message:z.message||"Could not save timesheet.",indicator:"red"}):(st.current=!1,x.current||(x.current=!0,frappe.msgprint({title:z.status==="locked"?"Timesheet Locked":"Autosave Failed",message:z.message||"Could not autosave timesheet. Edit the timesheet to try again.",indicator:"red"})))},error(){p(!1),S(!1),v?frappe.msgprint({title:"Save Failed",message:"Could not save timesheet.",indicator:"red"}):(st.current=!1,x.current||(x.current=!0,frappe.msgprint({title:"Autosave Failed",message:"Could not autosave timesheet. Edit the timesheet to try again.",indicator:"red"})))}})},[u,o,h]);R.useEffect(()=>{if(st.current&&!(dr.current||c))return ut.current&&clearTimeout(ut.current),ut.current=setTimeout(()=>De(!1),1200),()=>{ut.current&&clearTimeout(ut.current)}},[u,c,De,!0]);const at=R.useMemo(()=>{if(!r)return{};const v={};for(const C of r.dates)v[C.date]=0;for(const C of u)for(const _ of C.tasks)for(const[z,G]of Object.entries(_.hours)){const de=parseFloat(G)||0;de>0&&(v[z]=(v[z]||0)+de)}return v},[r,u]),Kt=R.useMemo(()=>Object.values(at).reduce((v,C)=>v+C,0),[at]),zl=R.useCallback(v=>Object.values(v.hours).reduce((C,_)=>C+(parseFloat(_)||0),0),[]),wc=R.useCallback(v=>v.tasks.reduce((C,_)=>C+Object.values(_.hours).reduce((z,G)=>z+(parseFloat(G)||0),0),0),[]),Sc=R.useCallback((v,C,_,z)=>{q(),s(G=>G.map((de,pr)=>pr!==v?de:{...de,tasks:de.tasks.map(Ll=>Ll.id!==C?Ll:{...Ll,hours:{...Ll.hours,[_]:z===""?0:parseFloat(z)||0}})}))},[q]),xc=R.useCallback((v,C,_)=>{q(),s(z=>z.map((G,de)=>de!==v?G:{...G,tasks:G.tasks.map(pr=>pr.id!==C?pr:{...pr,task:_})}))},[q]),_c=R.useCallback(v=>{q(),s(C=>C.map((_,z)=>z!==v?_:{..._,tasks:[..._.tasks,Qi()]}))},[q]),Ec=R.useCallback((v,C)=>{q(),s(_=>_.map((z,G)=>G!==v||z.tasks.length<=1?z:{...z,tasks:z.tasks.filter(de=>de.id!==C)}))},[q]),Cc=R.useCallback((v,C)=>{q(),s(_=>_.map((z,G)=>G!==v?z:{...z,name:C}))},[q]),Nc=R.useCallback(v=>{q(),s(C=>C.filter((_,z)=>z!==v))},[q]),dp=R.useCallback(v=>{q(),s(C=>[...C,Ki("project",v)]),U(!1),g("")},[q]),Yi=R.useCallback(v=>{q(),s(C=>[...C,Ki("activity",v)]),a(!1),P("")},[q]),pp=R.useCallback(()=>{const v=(T||"").trim();v&&frappe.call({method:"corporate_services.icl_corporate_services.page.employee_timesheet_entry.employee_timesheet_entry.create_activity_type",args:{name:v},callback:C=>{const _=C.message&&C.message.name||v;l(z=>new Set(((z==null?void 0:z.activity_types)||[]).map(de=>de.toLowerCase())).has(_.toLowerCase())?z:{...z,activity_types:[...(z==null?void 0:z.activity_types)||[],_].sort()}),Yi(_),L("")}})},[T,Yi]),mp=R.useCallback(v=>{!o||!v||it||(zt(!0),frappe.call({method:"frappe.client.get",args:{doctype:"Timesheet Submission",name:o},callback(C){const _=C.message;if(!_){zt(!1),frappe.msgprint(__("Unable to load linked Timesheet Submission."));return}frappe.call({method:"frappe.model.workflow.apply_workflow",args:{doc:_,action:v},callback(){zt(!1),frappe.show_alert({message:__("Workflow action applied: {0}",[v]),indicator:"green"}),Lt(o),Te(o)},error(){zt(!1)}})},error(){zt(!1)}}))},[o,it,Te,Lt]),Xi=R.useMemo(()=>{if(!r)return[];const v=new Set(u.filter(C=>C.type==="project").map(C=>C.name));return(r.all_projects||[]).filter(C=>!v.has(C.project_name))},[r,u]),Gi=R.useMemo(()=>{if(!r)return[];const v=new Set(u.filter(C=>C.type==="activity").map(C=>C.name));return(r.activity_types||[]).filter(C=>!v.has(C))},[r,u]),jc=R.useMemo(()=>{const v=d.toLowerCase();return v?Xi.filter(C=>C.project_name.toLowerCase().includes(v)):Xi},[Xi,d]),Pc=R.useMemo(()=>{const v=N.toLowerCase();return v?Gi.filter(C=>C.toLowerCase().includes(v)):Gi},[Gi,N]),Tc=R.useMemo(()=>{const v=((r==null?void 0:r.workflow_style)||"").toLowerCase();return{success:"green",danger:"red",primary:"blue",warning:"orange",info:"light-blue",inverse:"darkgrey"}[v]||"gray"},[r]);R.useMemo(()=>gc(r==null?void 0:r.month_year),[r]);const hp=R.useMemo(()=>(Array.isArray(r==null?void 0:r.submissions)?r.submissions:[]).slice(0,5),[r]);if(c)return m.jsx("div",{className:"ts-loading",children:"Loading timesheet..."});if(Q)return m.jsx("div",{className:"alert alert-danger",style:{margin:20},children:Q});if(!r)return null;const zc=document.getElementById("timesheet-entry-sidebar-root"),Lc=document.getElementById("timesheet-entry-context-root"),Rc=document.getElementById("timesheet-entry-actions-root"),vp=m.jsx(sp,{visibleSubmissions:hp,activeSubmission:o,loadContext:Te}),yp=m.jsx(up,{activeSubmission:o,ctx:r,workflowBusy:it,workflowActions:Pe,manualSaving:w,setAddActivityOpen:a,setAddProjectOpen:U,setProjectSearch:g,setActivitySearch:P,runWorkflowAction:mp,persistTimesheet:De}),gp=m.jsx("div",{style:{display:"inline-flex",alignItems:"center"},children:m.jsx("span",{className:`indicator-pill ${Tc}`,style:{marginLeft:12,whiteSpace:"nowrap"},children:r.workflow_state||"Draft"})});return m.jsxs(m.Fragment,{children:[zc?Tl.createPortal(vp,zc):null,Lc?Tl.createPortal(gp,Lc):null,Rc?Tl.createPortal(yp,Rc):null,m.jsxs("div",{className:"ts-entry-wrap",children:[m.jsx("div",{className:"ts-entry-header"}),E&&m.jsxs("div",{className:"ts-add-project-panel",children:[m.jsx("input",{className:"form-control form-control-sm",placeholder:"Search projects...",value:d,onChange:v=>g(v.target.value),autoFocus:!0}),m.jsxs("div",{style:{maxHeight:200,overflowY:"auto",marginTop:6},children:[jc.length===0&&m.jsx("div",{style:{padding:"6px 10px",color:"#888",fontSize:12},children:"No projects available"}),jc.map(v=>m.jsx("div",{className:"ts-project-item",onClick:()=>dp(v.project_name),children:v.project_name},v.name))]})]}),f&&m.jsxs("div",{className:"ts-add-project-panel",children:[m.jsx("input",{className:"form-control form-control-sm",placeholder:"Search activity types...",value:N,onChange:v=>P(v.target.value),autoFocus:!0}),m.jsxs("div",{style:{maxHeight:200,overflowY:"auto",marginTop:6},children:[Pc.length===0&&m.jsx("div",{style:{padding:"6px 10px",color:"#888",fontSize:12},children:"No activity types available"}),Pc.map(v=>m.jsx("div",{className:"ts-project-item",onClick:()=>Yi(v),children:v},v))]}),m.jsxs("div",{style:{marginTop:8,borderTop:"1px solid #eee",paddingTop:8},children:[m.jsx("input",{className:"form-control form-control-sm",placeholder:"Create new activity type (e.g. Meetings, Leave, Holiday)",value:T,onChange:v=>L(v.target.value)}),m.jsx("button",{className:"btn btn-default btn-sm",style:{marginTop:6},onClick:pp,children:"Create & Add"})]})]}),M&&m.jsxs("div",{className:"ts-mobile-wrap",children:[u.map((v,C)=>m.jsxs("div",{className:"ts-mobile-section",children:[m.jsxs("div",{className:`ts-mobile-section-header ${v.type==="activity"?"ts-mobile-header-activity":""}`,children:[m.jsxs("div",{className:"ts-mobile-section-title",children:[v.type==="activity"?m.jsx("input",{className:"form-control form-control-sm ts-mobile-activity-input",value:v.name,onChange:_=>Cc(C,_.target.value),placeholder:"Activity name"}):m.jsx("span",{children:v.name}),m.jsx("span",{className:"ts-section-type-badge",children:v.type==="project"?"Project":"Activity"})]}),m.jsxs("div",{className:"ts-mobile-section-meta",children:[m.jsxs("span",{className:"ts-mobile-section-total",children:[wc(v).toFixed(1),"h"]}),v.type==="activity"&&m.jsx("button",{className:"btn btn-default btn-xs ts-mobile-delete-btn",onClick:()=>Nc(C),children:"Delete"})]})]}),v.tasks.map(_=>m.jsxs("div",{className:"ts-mobile-task",children:[m.jsxs("div",{className:"ts-mobile-task-row",children:[m.jsx("input",{className:"ts-task-input ts-mobile-task-name",type:"text",maxLength:500,value:_.task,onChange:z=>xc(C,_.id,z.target.value),placeholder:"Task description"}),m.jsxs("button",{className:`ts-mobile-hours-toggle ${W===_.id?"active":""}`,onClick:()=>ee(W===_.id?null:_.id),children:[zl(_)>0?zl(_).toFixed(1):"0","h ▾"]}),v.tasks.length>1&&m.jsx("button",{className:"ts-remove-btn",onClick:()=>Ec(C,_.id),title:"Remove row",children:"×"})]}),W===_.id&&m.jsx("div",{className:"ts-mobile-dates",children:r.dates.map(z=>m.jsxs("div",{className:`ts-mobile-date-row${z.is_weekend?" ts-weekend":""}`,children:[m.jsxs("span",{className:"ts-mobile-date-label",children:[z.day_short," ",z.date_num]}),m.jsx("input",{className:"ts-mobile-hours-input",type:"number",min:"0",step:"0.5",inputMode:"decimal",value:_.hours[z.date]||"",onChange:G=>Sc(C,_.id,z.date,G.target.value),placeholder:"0"})]},z.date))})]},_.id)),m.jsx("button",{className:"ts-mobile-addrow-btn",onClick:()=>_c(C),children:"+ Add task row"})]},v.type+"_"+v.name+"_"+C)),m.jsxs("div",{className:"ts-mobile-footer-actions",children:[m.jsx("button",{className:"btn btn-default btn-sm ts-mobile-action-btn",onClick:()=>{a(!1),U(v=>!v),g("")},children:"+ Add Project"}),m.jsx("button",{className:"btn btn-default btn-sm ts-mobile-action-btn",onClick:()=>{U(!1),a(v=>!v),P("")},children:"+ Add Activity"})]}),m.jsxs("div",{className:"ts-mobile-save-bar",children:[m.jsxs("div",{style:{fontSize:12,color:"#666"},children:["Total: ",m.jsxs("strong",{children:[Kt>0?Kt.toFixed(1):"0","h"]}),r.workflow_state&&m.jsx("span",{className:`indicator-pill ${Tc}`,style:{marginLeft:10},children:r.workflow_state})]}),m.jsx("button",{className:"btn btn-primary",onClick:()=>De(!0),disabled:w,children:w?"Saving...":"Save Timesheet"})]})]}),!M&&m.jsx("div",{className:"ts-table-wrap",children:m.jsxs("table",{className:"ts-table",children:[m.jsxs("thead",{children:[m.jsxs("tr",{className:"ts-row-dayname",children:[m.jsx("th",{className:"ts-col-sticky",style:{left:0,minWidth:180},children:"Projects / Activities"}),m.jsx("th",{className:"ts-col-sticky",style:{left:180,minWidth:220},children:"Tasks"}),r.dates.map(v=>m.jsx("th",{className:"ts-col-date"+(v.is_weekend?" ts-weekend":""),children:v.day_short},v.date)),m.jsx("th",{className:"ts-col-total",children:"Total Hours"}),m.jsx("th",{className:"ts-col-total",children:"Ratio %"})]}),m.jsxs("tr",{className:"ts-row-datenum",children:[m.jsx("th",{className:"ts-col-sticky",style:{left:0,minWidth:180,top:29}}),m.jsx("th",{className:"ts-col-sticky",style:{left:180,minWidth:220,top:29},children:"Hours"}),r.dates.map(v=>m.jsx("th",{className:"ts-col-date"+(v.is_weekend?" ts-weekend":""),style:{top:29},children:v.date_num},v.date)),m.jsx("th",{className:"ts-col-total",style:{top:29}}),m.jsx("th",{className:"ts-col-total",style:{top:29}})]})]}),m.jsxs("tbody",{children:[u.map((v,C)=>m.jsxs(Gc.Fragment,{children:[m.jsxs("tr",{className:`ts-section-header ${v.type==="activity"?"ts-section-header-activity":""}`,children:[m.jsxs("td",{colSpan:2,className:"ts-col-sticky ts-section-name",style:{left:0},children:[v.type==="activity"?m.jsx("input",{className:"form-control form-control-sm",style:{display:"inline-block",width:220,marginRight:8,height:28},value:v.name,onChange:_=>Cc(C,_.target.value),placeholder:"Activity name"}):v.name,m.jsx("span",{className:"ts-section-type-badge",children:v.type==="project"?"Project":"Activity"}),v.type==="activity"&&m.jsx("button",{className:"btn btn-default btn-xs ts-activity-delete-btn",style:{marginLeft:8},onClick:()=>Nc(C),title:"Delete activity section",children:"Delete"})]}),r.dates.map(_=>m.jsx("td",{className:_.is_weekend?"ts-weekend":""},_.date)),m.jsx("td",{}),m.jsx("td",{className:"ts-col-total",children:Kt>0?(wc(v)/Kt*100).toFixed(1)+"%":""})]}),v.tasks.map(_=>m.jsxs("tr",{children:[m.jsx("td",{className:"ts-col-sticky",style:{left:0,background:"#fff"}}),m.jsx("td",{className:"ts-col-sticky",style:{left:180,background:"#fff"},children:m.jsxs("div",{className:"ts-task-cell",children:[m.jsx("input",{className:"ts-task-input",type:"text",maxLength:500,value:_.task,onChange:z=>xc(C,_.id,z.target.value),placeholder:"Task description"}),v.tasks.length>1&&m.jsx("button",{className:"ts-remove-btn",onClick:()=>Ec(C,_.id),title:"Remove row",children:"×"})]})}),r.dates.map(z=>m.jsx("td",{className:"ts-col-date"+(z.is_weekend?" ts-weekend":""),children:m.jsx("input",{className:"ts-hours-input",type:"number",min:"0",step:"0.5",value:_.hours[z.date]||"",onChange:G=>Sc(C,_.id,z.date,G.target.value)})},z.date)),m.jsx("td",{className:"ts-col-total",children:zl(_)>0?zl(_).toFixed(1):""}),m.jsx("td",{className:"ts-col-total"})]},_.id)),m.jsxs("tr",{className:"ts-row-addtask",children:[m.jsx("td",{className:"ts-col-sticky",style:{left:0,background:"#fff"}}),m.jsx("td",{className:"ts-col-sticky",style:{left:180,background:"#fff"},children:m.jsx("button",{className:"ts-addtask-btn",onClick:()=>_c(C),children:"+ Add row"})}),r.dates.map(_=>m.jsx("td",{className:_.is_weekend?"ts-weekend":""},_.date)),m.jsx("td",{}),m.jsx("td",{})]})]},v.type+"_"+v.name+"_"+C)),m.jsxs("tr",{className:"ts-row-total",children:[m.jsx("td",{colSpan:2,className:"ts-col-sticky",style:{left:0},children:"TOTAL"}),r.dates.map(v=>m.jsx("td",{className:"ts-col-date"+(v.is_weekend?" ts-weekend":""),children:at[v.date]>0?at[v.date].toFixed(1):""},v.date)),m.jsx("td",{className:"ts-col-total",children:Kt>0?Kt.toFixed(1):""}),m.jsx("td",{className:"ts-col-total",children:Kt>0?"100%":""})]})]})]})})]})]})}window.initTimesheetEntry=function(e,t,n){const r=document.createElement("div");r.id="ts-entry-root",e.main[0].appendChild(r),vc(r).render(m.jsx(fp,{submissionName:t,onContextChange:n}))}})();
